% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addons.r
\name{addons}
\alias{addons}
\title{Test if addons directory exists and if an addon is installed}
\usage{
addons(x = NULL, onFail = "fail", verbose = TRUE)
}
\arguments{
\item{x}{Either \code{NULL} or a character specifying the name of a \strong{GRASS} addons module. If \code{NULL}, the existence of the \code{addonsDir} (see \code{\link[=faster]{faster()}}) will be tested. If the module name is provided, the existence of the folder and module will be tested.}

\item{onFail}{Character: What to do if the addons folder is not found:
\itemize{
\item \code{"fail"} (default): Fail with a message.
\item \code{"warning"}: Warn with a message.
}}

\item{verbose}{Logical: If \code{TRUE} (default), display a message on success or warning (the \code{fail} option always displays a message).}
}
\value{
Logical.
}
\description{
This function tests to see if the "addons" directory specified using \code{\link[=faster]{faster()}} actually exists, and if a particular \strong{GRASS} \verb{addons module is available. The }addons\verb{folder and module must exists for methods that rely on particular **GRASS**}addons` to work.
}
\examples{
if (grassStarted()) {

ao <- addons(onFail = "warning")
if (ao) print("Addons is folder is correctly specified.")

addon <- "v.centerpoints"
exten <- addons(addon, onFail = "warning")

if (exten) print("Extension `v.centerpoints` is installed.")

}
}
