% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.lines.r
\name{as.lines,GRaster-method}
\alias{as.lines,GRaster-method}
\alias{as.lines}
\title{Convert a raster to a lines vector}
\usage{
\S4method{as.lines}{GRaster}(x, round = TRUE, smooth = FALSE)
}
\arguments{
\item{x}{A \code{GRaster}. If more than one layer is in the \code{GRaster}, only the first will be used (with a warning).}
}
\value{
A \code{GVector}.
}
\description{
\code{\link[=as.lines]{as.lines()}} converts a \code{GRaster} to a "lines" \code{GVector}. Before you apply this function, you may need to run \code{\link[=thin]{thin()}} on the raster to reduce linear features to a single-cell width. You may also need to use \code{\link[=cleanGeom]{cleanGeom()}} (especially the "duplicated" and "removeDangles" \code{method}s) afterward to remove duplicated vertices and "dangling" lines.
}
\seealso{
\code{\link[=as.points]{as.points()}}, \code{\link[=as.polygons]{as.polygons()}}, \code{\link[terra:as.lines]{terra::as.lines()}}, \code{\link[=thin]{thin()}}, \code{\link[=cleanGeom]{cleanGeom()}}, and \strong{GRASS} module \code{r.to.vect}
}
