% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioclims.r
\name{bioclims,GRaster-method}
\alias{bioclims,GRaster-method}
\alias{bioclims}
\title{Create BIOCLIM rasters}
\usage{
\S4method{bioclims}{GRaster}(ppt, tmin, tmax, tmean = NULL, bios = NULL, sample = TRUE)
}
\arguments{
\item{ppt}{A "stack" of \code{GRaster}s, representing monthly/weekly/daily precipitation.}

\item{tmin, tmax}{A "stack" of \code{GRaster}s, representing monthly/weekly/daily minimum and maximum temperature.}

\item{tmean}{Either \code{NULL} (default), or a "stack" of \code{GRaster}s, representing monthly/weekly/daily average temperature. If \code{NULL}, \code{tmean} will be calculated internally from \code{tmin} and \code{tmax}. Providing these rasters thus saves time if you already have them on hand.}

\item{bios}{Either \code{NULL} (default), \code{"*"}, or a vector of numbers indicating which BIOLCIMs to calculate:
\itemize{
\item \code{NULL}: Calculate BIOCLIMs 1 through 19
\item \code{"*"}: Calculate all BIOCLIMs this function can calculate.
\item \code{"+"}: Calculate BIOCLIMs 41 onward.
\item Numeric values: Calculate these BIOCLIM varoables. For example, \code{bios = c(1, 12)} calculates BIOCLIMs 1 and 12.
}}

\item{sample}{Logical: If \code{TRUE} (default), BIO4 and 15 are calculated with the sample standard deviation. If \code{FALSE}, then the population standard deviation is used.}
}
\value{
A \code{GRaster} with one or more layers.
}
\description{
The BIOCLIM set of variables were created for modeling species' geographic distributions (Booth et al. 2014). The "standard" set includes 19 variables, which can be created with this function:
\itemize{
\item BIO01: Mean annual temperature, calculated using monthly means
\item BIO02: Mean diurnal range across months (average of monthly difference between maximum and minimum temperature)
\item BIO03: Isothermality (100 * BIO02 / BIO07)
\item BIO04 Temperature seasonality (standard deviation across months of average monthly temperature * 100)
\item BIO05: Maximum temperature of the warmest month (based on maximum temperature)
\item BIO06: Minimum temperature of the coldest month (based on minimum temperature)
\item BIO07: Range of annual temperature (BIO05 - BIO06)
\item BIO08: Mean temperature of the wettest quarter (based on mean temperature)
\item BIO09: Mean temperature of the driest quarter (based on mean temperature)
\item BIO10: Mean temperature of the warmest quarter (based on mean temperature)
\item BIO11: Mean temperature of the coldest quarter (based on mean temperature)
\item BIO12: Total annual precipitation
\item BIO13: Precipitation of the wettest month
\item BIO14: Precipitation of the driest month
\item BIO15: Precipitation seasonality (100 * coefficient of variation)
\item BIO16: Precipitation of the wettest quarter
\item BIO17: Precipitation of the driest quarter
\item BIO18: Precipitation of the warmest quarter (based on mean temperature)
\item BIO19: Precipitation of the coldest quarter (based on mean temperature)
\item BIO41: Mean temperature of the quarter following the coldest quarter (based on mean temperature)
\item BIO42: Mean temperature of the quarter following the warmest quarter (based on mean temperature)
\item BIO43: Precipitation of the quarter following the coldest quarter (based on mean temperature)
\item BIO44: Precipitation of the quarter following the warmest quarter (based on mean temperature)
\item BIO45: Temperature of the quarter following the driest quarter (based on mean temperature)
\item BIO46: Temperature of the quarter following the wettest quarter (based on mean temperature)
\item BIO47: Precipitation of the quarter following the driest quarter (based on mean temperature)
\item BIO48: Precipitation of the quarter following the wettest quarter (based on mean temperature)
\item BIO49: Quarter with the highest temperature.
\item BIO50: Quarter with the lowest temperature.
\item BIO51: Quarter with the highest precipitation.
\item BIO52: Quarter with the lowest precipitation.
}

"Quarter" refers to any consecutive run of three months, not a financial quarter. A quarter can thus include November-December-January, or December-January-February, for example.

The variables are defined assuming that the input rasters represent monthly values (12 rasters for min/max temperature and precipitation), but you can also use sets of 52 rasters, representing one per week, in which case "quarter" would be a successive run of 3 weeks. You could also attempt 365 rasters, in which case a "quarter" would be a run of 3 successive days.

BIOCLIMs 41 through 44 are added here to capture the "shoulder" seasons (spring and autumn) important in temperature regions. BIOCLIMs 45 through 48 are also included for consistency.

BIOCLIMs 49 through 52 are not bioclimatic variables per se, but useful for assessing the properties of the variables that are defined based on the "-est" quarter.

The numbering of the new BIOCLIMs was begun at 41 because BIOCLIMs 20 through 40 are taken (Kriticos et al. 2014).
}
\references{
Booth, T.H., Nix, H.A., Busby, J.R., and Hutchinson, M.F.  2014.  BIOCLIM: The first species distribution modeling package, its early applications and relevance to most current MaxEnt studies.  \emph{Diversity and Distributions} 20:1-9 \doi{10.1111/ddi.12144}.

Kriticos, D.J., Jarošik, V., and Otam N.  2014.  Extending the suite of BIOCLIM variables: A proposed registry system and case study using principal components analysis.  \emph{Methods in Ecology and Evolution} 5:956-960 \doi{10.1111/2041-210X.12244}.
}
