% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.r
\name{c,GRaster-method}
\alias{c,GRaster-method}
\title{"Stack" 'GRaster's}
\usage{
\S4method{c}{GRaster}(x, ...)
}
\arguments{
\item{...}{Two or more \code{GRaster}s in the same \strong{GRASS} \link[=location]{location and mapset}.}
}
\value{
A \code{GRaster}.
}
\description{
\code{GRaster}s can be "stacked" using this function, effectively creating a mult-layered raster set.
}
