% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.r
\name{c,GRaster-method}
\alias{c,GRaster-method}
\alias{c}
\alias{c,GVector-method}
\title{"Stack" GRasters and combine GVectors}
\usage{
\S4method{c}{GRaster}(x, ...)

\S4method{c}{GVector}(x, ...)
}
\arguments{
\item{x}{A \code{GRaster} or a \code{GVector}.}

\item{...}{One or more \code{GRaster}s, one or more \code{GVector}s, a list of \code{GRaster}s, or a list of \code{GVector}s. You can use a mix of lists and individual rasters or vectors.}
}
\value{
A \code{GRaster}.
}
\description{
\code{GRaster}s can be "stacked" using this function, effectively creating a multi-layered raster. This is different from creating a 3-dimensional raster, though such an effect can be emulated using stacking. \code{GVector}s can be combined into a single vector.  Stacks can only be created when:
\itemize{
\item All objects are the same class (either all \code{GRaster}s or all \code{GVector}s).
\item All objects have the same coordinate reference system (see crs()).
\item For \code{GRaster}s:
\itemize{
\item Horizontal extents are the same (see \code{\link[=ext]{ext()}}).
\item Horizontal dimensions are the same (see \code{\link[=res]{res()}}).
\item The topology (2- or 3-dimensional) must be the same. If 3D, then all rasters must have the same number of depths and vertical extents (see \code{\link[=topology]{topology()}}).
}
\item For \code{GVector}s:
\itemize{
\item The geometry (points, lines, or polygons) must be the same. See \code{\link[=geomtype]{geomtype()}}.
}
}

If features (boundaries, lines, etc.) of \code{GVector}s that are combined are identical or nearly identical, they can be cleaned using \code{\link[=removeDupNodes]{removeDupNodes()}} and \code{\link[=snap]{snap()}}.

Data tables associated with \code{GVector}s will be combined if each vector has a table and if each table has the same columns and data types. Otherwise, the data table will be dropped.

You can speed operations by putting the largest vector first in \code{c(x, ...)}.
}
\examples{
if (grassStarted()) {

# Setup
library(sf)

# Rivers vector
madRivers <- fastData("madRivers")

# Convert sf to a GVector
rivers <- fast(madRivers)

# Subset rivers vector
rivers1 <- rivers[1:2]
rivers2 <- rivers[10:11]

# Concatenate rivers
riversCombo <- c(rivers1, rivers2)
riversCombo

# Add columns
newCol <- data.frame(x = 1:11)
riversCols <- colbind(rivers, newCol)
riversCol

as.data.frame(riversCol)

}
}
\seealso{
\link{add<-}, \code{\link[terra:c]{terra::c()}}, \code{\link[terra]{add<-}}
}
