% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbind.r
\name{print}
\alias{print}
\alias{cbind}
\title{Add columns to the data table of a GVector}
\usage{
print(x, ...)
}
\arguments{
\item{...}{The first argument must be a \code{GVector}. Subsequent arguments can be \code{data.frame}s, \code{data.table}s, \code{matrices}, or \code{GVector}s. Only the data tables of subsequent \code{GVector}s are added to the table in \code{x}; the geometries are ignored.}
}
\value{
A \code{GVector}.
}
\description{
\code{cbind()} adds columns to the data table of a \code{GVector}. To combine two \code{GVector}s, see \code{\link[=rbind]{rbind()}}. You combine multiple a \code{GVector}'s data table with \code{data.frame}s, \code{data.table}s, \code{matrices}, or the data table(s) from other \code{GVector}(s).
}
\examples{
if (grassStarted()) {

# Setup
library(sf)

# Rivers vector
madRivers <- fastData("madRivers")

# Convert sf to a GVector
rivers <- fast(madRivers)

# Subset rivers vector
rivers1 <- rivers[1:2]
rivers2 <- rivers[10:11]

# Concatenate rivers
riversCombo <- rbind(rivers1, rivers2)
riversCombo

# Add columns
newCol <- data.frame(x = 1:11)
riversCols <- cbind(rivers, newCol)
riversCol

as.data.frame(riversCol)

}
}
\seealso{
\code{\link[=rbind]{rbind()}}, \code{\link[=c]{c()}}
}
