% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeWorkDir.r
\name{changeWorkDir}
\alias{changeWorkDir}
\title{Change the working directory of a 'GRASS' session}
\usage{
changeWorkDir(workDir)
}
\arguments{
\item{workDir}{The name of the new working directory.}
}
\value{
The previous value of \code{workDir} (invisibly).
}
\description{
\strong{GRASS} sessions must have a working directory where the rasters and vectors are stored. Typiclaly, this is set when you use \code{\link[=startFast]{startFast()}} to initiate a \strong{GRASS} session. By default, this is a temporary directory and does not need to be defined by the user. If this is temporary, there is no easy way to restore it if you shut down \strong{R}--all rasters and vectors in \strong{GRASS} are lost. However, you may want to use a custom directory if you want to start \strong{R} later. If you do this, you may also want to copy or move the directory to a new place. When you do that, you \emph{must} call this function so that the proper working directory is used in the \strong{GRASS} session.
}
