% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.r
\name{classify,GRaster-method}
\alias{classify,GRaster-method}
\alias{classify}
\title{Classify GRaster cell values}
\usage{
\S4method{classify}{GRaster}(x, rcl, include.lowest = FALSE, right = TRUE, others = NULL)
}
\arguments{
\item{x}{A \code{GRaster}.}

\item{rcl}{Reclassification system:
\itemize{
\item A single integer: Number of "bins" into which to divide values. Arguments \code{include.lowest} and \code{right} apply.
\item A vector of numeric values: Breakpoints of bins into which to divide values. These will be sorted from lowest to highest before classification. Arguments \code{include.lowest} and \code{right} apply.
\item A 2-column \code{matrix} or \code{data.frame}: The first column provides specific values in \code{x} to be replaced, and the second provides the values they are replaced with. This method is only useful for classifying \link[=tutorial_raster_data_types]{integer} \code{GRaster}s. Arguments \code{include.lowest} and \code{right} are ignored. Cells will be classified in the order inn which values are listed inn the first column (i.e., duplicate values in the first column will be ignored).
\item A 3-column \code{matrix} or \code{data.frame}: The first column provides the lower value of a bin, the second the upper value, and the third the value to assign to the cells in the bin. Arguments \code{include.lowest} and \code{right} apply. Cells will be classified in the order of how intervals are listed (intervals will not be sorted).
}}

\item{include.lowest, right}{Logical: These arguments determine how cells that have values exactly equal to the lower or upper ends of an interval are classified.
\itemize{
\item \code{include.lowest = TRUE} and \code{right = TRUE}: All intervals will be "left-open, right-closed" except for the lowest interval, which will be "left-closed/right-closed".
\item \code{include.lowest = FALSE} and \code{right = FALSE}: Intervals will be "left-closed/right-open". Cells with values equal to the highest higher boundary will not be reclassified.
\item \code{include.lowest = TRUE} and \code{right = FALSE}: All intervals will be "left-closed/right-open", except for the highest interval, which will be "right-closed/left-closed".
\item \code{right = NA}: Only useful for classifying \link[=tutorial_raster_data_types]{integer} \code{GRaster}s. All intervals are "left-closed/right-closed". This is easier than accounting for "open" intervals when dealing with integers.
}}

\item{others}{Numeric, integer, \code{NULL} (default), or \code{NA}: Value to assign to cells that do not fall into the set intervals. Cells with \code{NA} values are not reclassified. Setting \code{others} equal to \code{NULL} or \code{NA} replaces all other values with \code{NA}.}
}
\value{
A \code{GRaster}.
}
\description{
This function classifies a raster so that cells that have values within a given range are assigned a new value. Cells that are not re-assigned a value retain their original value.
}
\seealso{
\code{\link[terra:classify]{terra::classify()}}; \verb{if ()} statements in \code{\link[=app]{app()}}
}
