% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clump.r
\name{clump,GRaster-method}
\alias{clump,GRaster-method}
\alias{clump}
\title{Group adjacent cells with similar values}
\usage{
\S4method{clump}{GRaster}(x, minDiff = 0, minClumpSize = 1, diagonal = TRUE)
}
\arguments{
\item{x}{A \code{GRaster}.}

\item{minDiff}{Numeric in the range [0, 1): Minimum difference between cells in order for them to be assigned to the same clump. This is a proportion of the range across all cells. For example, if \code{minDiff} is set to 0.01, then the maximum difference between cells in a clump can be up to 1\% of the entire range across all cells. Small values can create large clumps. The default is 0, in which case values have to be exactly the same.}

\item{minClumpSize}{Numeric integer >= 1. Minimum number of cells in a clump. The default is 1.}

\item{diagonal}{Logical: If \code{TRUE} (default), then cells "connected" at corners will be included as part of the same clump.}
}
\value{
A \code{GRaster}.
}
\description{
\code{clump()} identifies groups of adjacent cells that have the same value or same approximate value, and assigns them a unique number, creating "clumps" of same- or similar-valued cells.
}
\examples{
\dontrun{
# NB This example is in a "dontrun{}" block because it requires users to have
# GRASS GIS Version 8+ installed on their system.

# IMPORTANT #1: If you already have a GRASS session started, you will need to
# run the line below and the last line in this example to work with it again.
# If you have not started a GRASS session, you can skip this step and go to
# step #2.
opts. <- getFastOptions()

# IMPORTANT #2: Select the appropriate line below and change as necessary to
# where GRASS is installed on your system.
grassDir <- "/Applications/GRASS-8.3.app/Contents/Resources" # Mac
grassDir <- "C:/Program Files/GRASS GIS 8.3" # Windows
grassDir <- "/usr/local/grass" # Linux

# Setup
library(sf)
library(terra)

# Elevation raster
madElev <- fastData("madElev")

# Start GRASS session for examples only:
faster(x = madElev, grassDir = grassDir,
workDir = tempdir(), location = "examples") # line only needed for examples

# Convert a SpatRaster to a GRaster:
elev <- fast(madElev)

# Find clumps based on exact values. This will appear as a gradient because
# most cells are assigned to a group of 1 cell.
exact <- clump(elev)
plot(exact)

# Clump based on approximate values:
approx <- clump(elev, minDiff = 0.0075)
plot(approx)

# Clump based on approximate values with minimum clump size
# (this can take a minute or two):
approx20 <- clump(elev, minDiff = 0.005, minClumpSize = 20)
plot(approx20)

approx
approx20

# IMPORTANT #3: Revert back to original GRASS session if needed.
fastRestore(opts.)
removeSession("examples")

}
}
