% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_comparison.r
\name{Arith,GRaster,GRaster-method}
\alias{Arith,GRaster,GRaster-method}
\alias{Compare-methods}
\alias{Arith,logical,GRaster-method}
\alias{Arith,GRaster,logical-method}
\alias{Arith,numeric,GRaster-method}
\alias{Arith,GRaster,numeric-method}
\alias{Arith,GRaster,integer-method}
\alias{Arith,integer,GRaster-method}
\alias{Arith,GRaster,character-method}
\alias{Arith,character,GRaster-method}
\alias{Arith,GRegion,GRegion-method}
\title{Compare-methods operations on GRasters and GRegions}
\usage{
\S4method{Arith}{GRaster,GRaster}(e1, e2)

\S4method{Arith}{logical,GRaster}(e1, e2)

\S4method{Arith}{GRaster,logical}(e1, e2)

\S4method{Arith}{numeric,GRaster}(e1, e2)

\S4method{Arith}{GRaster,numeric}(e1, e2)

\S4method{Arith}{GRaster,integer}(e1, e2)

\S4method{Arith}{integer,GRaster}(e1, e2)

\S4method{Arith}{GRaster,character}(e1, e2)

\S4method{Arith}{character,GRaster}(e1, e2)

\S4method{Arith}{GRegion,GRegion}(e1, e2)
}
\arguments{
\item{e1, e2}{Values depend on the type of comparison:
\itemize{
\item Comparing a \code{GRegion} to another \code{GRegion}: \code{e1} and \code{e2} must be \code{GRegion}s!
}}
}
\value{
Comparing \code{GRaster}s: An "integer \code{GRaster} with values of 0 (FALSE), 1 (TRUE), or \code{NA} (neither). Comparing \code{GRegion}s: Logical.
}
\description{
You can do comparative operations on \code{GRaster}s using normal operators in \strong{R}: \code{<}, \code{<=}, \code{==}, \code{!=}, \code{>=}, and \code{>}.

You can also compare two \code{GRegion}s using the \code{==} and \code{!=} operators. Most users of \strong{fasterRaster} will not have to work much with \link[=tutorial_regions]{regions}, so can ignore this functionality.
\itemize{
\item \code{GRegion}s are the same if they have the same coordinate reference system, \link[=tutorial_locations_mapsets]{location and mapset}, topology (2D or 3D), extent, and resolution. If both are 3D, then they must also have the same vertical extent and number of depths.
}
}
\examples{
if (grassStarted()) {

# Setup
library(sf)
library(terra)

# Example data
madElev <- fastData("madElev")

# Convert a SpatRaster to a GRaster
elev <- fast(madElev)
elevs <- c(elev, elev, log10(elev) - 1, sqrt(elev))
names(elevs) <- c("elev1", "elev2", "log_elev", "sqrt_elev")

elev
elevs

# do some math
elev + 100
elev - 100
elev * 100
elev / 100
elev ^ 2
elev \%/\% 100 # divide then round down
elev \%\% 100 # modulus

100 + elev
100 \%/\% elev
100 \%\% elev

elevs + 100
100 + elevs

# math with logicals
elev + TRUE
elev - TRUE
elev * TRUE
elev / TRUE
elev ^ TRUE
elev \%/\% TRUE # divide then round down
elev \%\% TRUE # modulus

elevs + TRUE
TRUE + elevs

# raster interacting with raster
elev + elev
elev - elev
elev * elev
elev / elev
elev ^ log(elev)
elev \%/\% sqrt(elev) # divide then round down
elev \%\% sqrt(elev) # modulus

elevs + elev
elev * elevs

# sign
abs(-1 * elev)
abs(elevs)
sqrt(elevs)

# trigonometry
sin(elev)
cos(elev)
tan(elev)

asin(elev)
acos(elev)
atan(elev)

atan(elevs)
atan2(elev, elev^1.2)
atan2(elevs, elev^1.2)
atan2(elev, elevs^1.2)
atan2(elevs, elevs^1.2)

# logarithms
exp(elev)
log(elev)
ln(elev)
log2(elev)
log1p(elev)
log10(elev)
log(elev, 3)

log(elevs)

# rounding
round(elev + 0.5)
floor(elev + 0.5)
ceiling(elev + 0.5)
trunc(elev + 0.5)

# comparison
elev < 100
elev <= 100
elev == 100
elev != 100
elev > 100
elev >= 100

elev + 100 < 2 * elev

elevs > 10
10 > elevs

# mathematical functions on two or more rasters
mean(elevs)
mmode(elevs)
median(elevs)
nunique(elevs)

sum(elevs)
count(elevs)
min(elevs)
max(elevs)
range(elevs)
skewness(elevs)
kurtosis(elevs)

which.min(elevs)
which.max(elevs)

slope(elevs)
intercept(elevs)
r2(elevs)
tvalue(elevs)

sd(elevs)
sdpop(elevs)
var(elevs)
varpop(elevs)
quantile(elevs, 0.1)

}
}
