% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareFloat.r
\name{compareFloat}
\alias{compareFloat}
\title{Compare values accounting for differences due to floating point precision}
\usage{
compareFloat(x, y, op, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x, y}{Numeric}

\item{op}{Operator for comparison (must be in quotes): \code{"<"}, \code{">"}, \code{"<="}, \code{">="}, \code{"=="}, or \code{"!="}}

\item{tol}{Tolerance value: The largest absolute difference between \code{x} and \code{y} that is to be considered equality. Teh default is \code{.Machine$double.eps^0.5}.}
}
\value{
\code{TRUE}, \code{FALSE}, or \code{NA}
}
\description{
This function compares values while accounting for differences in floating point precision. This function is inspired by the \strong{fpCompare} package.
}
\examples{
x <- 0.9 - 0.8
y <- 0.8 - 0.7

x < y
compareFloat(x, y, "<")

x <= y
compareFloat(x, y, "<=")

x == y
compareFloat(x, y, "==")

y > x
compareFloat(y, x, ">")

y >= x
compareFloat(y, x, ">=")

x != y
compareFloat(x, y, "!=")

}
