% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convHull.r
\name{convHull,GVector-method}
\alias{convHull,GVector-method}
\alias{convHull}
\title{Minimum convex hull around a spatial vector}
\usage{
\S4method{convHull}{GVector}(x, by = "")
}
\arguments{
\item{x}{A \code{GVector}.}

\item{by}{Character: If \code{''} (default), then a convex hull is created for all geometries together. Otherwise, this is the name of a field in the vector. Hulls will be created for each set of geometries with the same value in this column.}
}
\value{
A \code{GVector}.
}
\description{
Create a minimum convex hull around a spatial vector.
}
\seealso{
[terra::convHull(), \code{sf::st_convex_hull()}, \strong{GRASS} module \code{v.hull}
}
