% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crds.r
\name{crds,GRaster-method}
\alias{crds,GRaster-method}
\alias{crds}
\alias{crds,GVector-method}
\alias{st_coordinates}
\title{Coordinates of a vector"s features or a raster"s cell centers}
\usage{
\S4method{crds}{GRaster}(x, z = TRUE, na.rm = TRUE)

\S4method{crds}{GVector}(x, z = TRUE)

st_coordinates(x, z = TRUE)
}
\arguments{
\item{x}{A \code{GVector} or a \code{GRaster}.}

\item{z}{If \code{TRUE} (default), return x-, y-, and z-coordinates. If \code{FALSE}, just return x- and y-coordinates. For 2-dimensional objects, z-coordinates will all be 0.}

\item{na.rm}{Logical: If \code{TRUE}, remove cells that are \code{NA} (\code{GRaster}s only).}
}
\value{
A \code{matrix}, \code{data.frame}, or \code{list}.
}
\description{
Returns the coordinates of the center of cells of a \code{GRaster} or coordinates of a \code{GVector}'s vertices. The output will be a \code{matrix}, \code{data.frame}, or \code{list}. If you want the output to be a "points" \code{GVector}, use \code{\link[=as.points]{as.points()}}.
}
\examples{
\dontrun{
# NB This example is in a "dontrun{}" block because it requires users to have
# GRASS GIS Version 8+ installed on their system.

# IMPORTANT #1: If you already have a GRASS session started, you will need to
# run the line below and the last line in this example to work with it again.
# If you have not started a GRASS session, you can skip this step and go to
# step #2.
opts. <- getFastOptions()

# IMPORTANT #2: Select the appropriate line below and change as necessary to
# where GRASS is installed on your system.
grassDir <- "/Applications/GRASS-8.3.app/Contents/Resources" # Mac
grassDir <- "C:/Program Files/GRASS GIS 8.3" # Windows
grassDir <- "/usr/local/grass" # Linux

# Setup
library(sf)
library(terra)

# Plant specimens (points), elevation (raster)
madDypsis <- fastData("madDypsis")
madElev <- fastData("madElev")

# Start GRASS session for examples only:
faster(x = madDypsis, grassDir = grassDir,
workDir = tempdir(), location = "examples") # line only needed for examples

# Convert a SpatRaster to a GRaster, and sf to a GVector
dypsis <- fast(madDypsis)
elev <- fast(madElev)

### Get coordinates:
dypsisPoints <- crds(dypsis)
elevPoints <- crds(elev)

head(dypsisPoints)
head(elevPoints)

# IMPORTANT #3: Revert back to original GRASS session if needed.
restoreSession(opts.)
removeSession("examples")

}
}
\seealso{
\code{\link[terra:crds]{terra::crds()}}
}
