% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossFreq.r
\name{crossFreq,GRaster-method}
\alias{crossFreq,GRaster-method}
\alias{crossFreq}
\title{Frequencies of combinations of cells across two or more rasters}
\usage{
\S4method{crossFreq}{GRaster}(x, na.rm = TRUE, cats = TRUE, verbose = FALSE)
}
\arguments{
\item{x}{A stack of integer/categorical \code{GRaster}s.}

\item{na.rm}{Logical: If \code{TRUE} (default), then only cells that are not \code{NA} across all rasters will be used. If \code{FALSE}, then for each pair of \code{GRaster}s, all cells that are not \code{NA} in both rasters will be used.}

\item{cats}{Logical: If \code{TRUE} (default), then replace the values of categorical rasters with their category names in the output.}

\item{verbose}{Logical: If \code{TRUE}, display progress messages.}
}
\value{
A \code{data.frame} or a named \code{list} of \code{data.frame}s, one per each pair of rasters in \code{x}.
}
\description{
\code{crossFreq()} tabulates the number of cells across two or more rasters for each combination of values in the rasters. Only cells that are not \code{NA} across all rasters will be used.
}
\examples{
if (grassStarted()) {

# Setup
library(terra)

# Example data
madElev <- fastData("madElev") # raster
madCover <- fastData("madCover") # categorical raster

# Convert to GRasters
elev <- fast(madElev) # integer raster
cover <- fast(madCover) # categorical raster

# Frequencies of integer raster values
f1 <- freq(elev)
print(f1) # have to do this sometimes if output is a data table

# Frequencies of categorical raster values
f2 <- freq(cover)
print(f2) # have to do this sometimes if output is a data table

# Frequencies of given values
f3 <- freq(elev, value = 4)
print(f3) # have to do this sometimes if output is a data table

# When a GRaster has non-integer values, they will be binned:
f4 <- freq(elev + 0.1, bins = 10)
print(f4)

# Calculate cross frequencies between rasters... both need to be integer.
elevWgs84 <- project(elev, cover)
elevClasses <- clump(elevWgs84, minDiff = 0.13) # bin elevations
names(elevClasses) <- 'elevClass'
f5 <- crossFreq(c(elevClasses, cover), na.rm = FALSE)
print(f5) # have to do this sometimes if output is a data table

}
}
\seealso{
\code{\link[=freq]{freq()}}, \strong{GRASS} tool \code{r.stats} (see \code{grassHelp("r.stats")})
}
