% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs.r
\name{crs}
\alias{crs}
\alias{st_crs}
\title{Coordinate reference system of a \code{GRaster} or \code{GVector}}
\usage{
crs(x, ...)

st_crs(x, ...)
}
\arguments{
\item{x}{A \code{GRaster} or \code{GVector}.}
}
\value{
Character.
}
\description{
Get the coordinate reference system (CRS) of a \code{GRaster}, \code{GVector}, \code{GSpatial}, or \code{GSession} object.
}
