% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_nrow_ncol_ncell.r
\name{dim,GRaster-method}
\alias{dim,GRaster-method}
\title{Number of rows and columns of a \code{GRaster} raster}
\usage{
\S4method{dim}{GRaster}(x)
}
\arguments{
\item{x}{A \code{GRaster}.}
}
\value{
A numeric vector.
}
\description{
Number of rows, columns, and (for 3D rasters) depths of a \code{GRaster} raster object.
}
