% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_nrow_ncol_ncell.r
\docType{data}
\name{dim3d}
\alias{dim3d}
\alias{dim3d.GRaster}
\title{Number of rows in a 3D \code{GRaster}}
\format{
An object of class \code{character} of length 1.
}
\usage{
dim3d.GRaster
}
\description{
Number of rows in a 3D \code{GRaster}
}
\keyword{datasets}
