% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLocationProject.r
\name{.addLocationProject}
\alias{.addLocationProject}
\title{Add a "project" or "location" argument to an "args" list}
\usage{
.addLocationProject(args, locProj)
}
\arguments{
\item{args}{A \code{list} of arguments to be passed to a \strong{GRASS} module.}

\item{locProj}{Character: The name of the location or project.}
}
\value{
A \code{list}.
}
\description{
Unhelpfully, starting with \strong{GRASS} 8.4, what were previously called "locations" and now called "projects."  These were supposed to be back-compatible, but are not. This function takes a list of arguments that will be passed to \code{\link[rgrass:execGRASS]{rgrass::execGRASS()}} and adds either a "location" or "project" argument, depending on the version of \strong{GRASS} being used.
}
\keyword{internal}
