% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext.r
\name{.ext}
\alias{.ext}
\title{Function to get extent from a "sources" name of a raster or vector}
\usage{
.ext(x, rastOrVect = NULL)
}
\arguments{
\item{x}{A \code{GRaster}, \code{GSpatial}, or a character (\code{\link[=sources]{sources()}} name of a \code{GRaster} or \code{GVector}).}

\item{rastOrVect}{Either \code{NULL} (class taken from \code{x}, but \code{x} cannot be a character), or "\code{raster}" or "\code{vector}" (partial matching is used).}
}
\value{
A numeric vector.
}
\description{
Function to get extent from a "sources" name of a raster or vector
}
\keyword{internal}
