% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_fragDensity.r
\name{.fragDensity}
\alias{.fragDensity}
\title{Calculate density (pf, or proportion of cells occupied) from a binary matrix}
\usage{
.fragDensity(x, na.rm = FALSE)
}
\arguments{
\item{x}{Binary matrix, with values equal to 0, 1, or \code{NA}.}

\item{na.rm}{Logical, if \code{FALSE} (default) count \code{NA} cells towards cells that could be occupied.}
}
\value{
Numeric in the range [0, 1] or \code{NA} if all cells are \code{NA}.
}
\description{
Calculates proportion of cells occupied (values = 1). Riitters, K., J. Wickham, R. O'Neill, B. Jones, and E. Smith. 2000. Global-scale patterns of forest fragmentation. Conservation Ecology 4:3. URL: https://www.jstor.org/stable/26271763. (Also note the erratum to the paper on their classification scheme at https://www.ecologyandsociety.org/vol4/iss2/art3/errata/january26.2001.html).
}
\examples{
x <- runif(100)
x <- x >= 0.7
x <- matrix(x, nrow=10)
fasterRaster:::.fragDensity(x)
diag(x) <- NA
fasterRaster:::.fragDensity(x)
}
\seealso{
\code{\link[fasterRaster]{fragmentation}}, \code{\link[fasterRaster]{fasterFragmentation}}
}
\keyword{internal}
