% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_fragWorker.r
\name{.fragWorker}
\alias{.fragWorker}
\title{Worker function for calculating raster fragmentation statistics in parallel}
\usage{
.fragWorker(rast, w, calcDensity = TRUE, calcConnect = TRUE,
  calcClass = TRUE, na.rm = FALSE, ...)
}
\arguments{
\item{rast}{Raster with binary values (1 or 0 or \code{NA}).}

\item{calcDensity}{Logical, if \code{TRUE} (default) then calculate density raster.}

\item{calcConnect}{Logical, if \code{TRUE} (default) then calculate connectivity raster.}

\item{na.rm}{Logical, if \code{FALSE} (default) then \code{NA} cells count as part of the area potentially occupied in a window (i.e., the count in the denominator when calculating density and they are counted as potential links in the connectance calculations if a neighboring cell has a value of 1). If \code{FALSE} then areas that border \code{NA} cells could still be classified as "interior" or otherwise have less apparent fragmentation if the occupied cells are fully surrounded by other occupied cells (except for the \code{NA} cells).}

\item{...}{Other arguments (not used).}

\item{i}{Integer, block number of \code{rast}.}

\item{clustBlockSize}{List with three elements: \code{row}, \code{nrows}, and \code{n} (can be generated by \code{\link[raster]{blockSize}}).}

\item{size}{Integer, number of cells wide and high of the window used to calculate fragmentation. This must be an odd integer (default is 3).}

\item{calcConnect}{Logical, if \code{TRUE} (default) then calculate classification raster. Note that to calculate the classification raster the density and connectivity rasters must also be calculated (\code{calcDensity} and \code{calcConnect} should both be \code{TRUE}). If they are not then the will be forced to \code{TRUE} with a warning.}
}
\value{
A list object with three elements, each of which is a matrix: density, connectivity, and fragmentation class (if \code{calcDensity}, \code{calcConnect}, and \code{calcClass} are \code{TRUE}, respectively).
}
\description{
This function is meant as a worker function for multi-core calculation of raster fragmentation statistics. It is intended to be used by other functions and thus is not likely to be of use to normal users.
}
