% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.region.r
\name{.g.region}
\alias{.g.region}
\title{Call GRASS \code{g.region} tool}
\usage{
.g.region()
}
\value{
Displays current region information for the active "project/location" in \strong{GRASS}.
}
\description{
This function calls the \strong{GRASS} tool \code{g.region} to display information on the region of the current \strong{GRASS} "project".
}
\keyword{internal}
