% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomtype.r
\name{.geomtype}
\alias{.geomtype}
\title{Get geometry type from the sources() name of a vector}
\usage{
.geomtype(x)
}
\arguments{
\item{x}{A \code{GVector} or the \code{\link[=sources]{sources()}} name of one.}
}
\value{
Character ("point", "line", or "area"; i.e., in \strong{GRASS} nomenclature).
}
\description{
Get geometry type from the sources() name of a vector
}
\keyword{internal}
