% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_getFlags.r
\name{.getFlags}
\alias{.getFlags}
\title{Return a character list of flags}
\usage{
.getFlags(replace = FALSE, flags = "quiet")
}
\arguments{
\item{replace}{Logical. Determines if rasters or vectors imported into or created in \code{GRASS} can overwrite existing rasters or vectors.}

\item{flags}{Either \code{NULL} (default) or a character vector of flags. The flag "overwrite" will be added to this if \code{replace} is \code{TRUE}.}
}
\value{
a Character vector.
}
\description{
This private function adds "overwrite" to a "flags" character list if the argument \code{remove} is \code{TRUE}.
}
\keyword{internal}
