% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_getCommonArgs.r
\name{.getGrassDir}
\alias{.getGrassDir}
\title{Get common arguments}
\usage{
.getGrassDir(grassDir)
}
\arguments{
\item{grassDir, replace, grassToR, outVectClass, autoRegion, grassVer}{Options or NULL.}
}
\value{
A character or logical.
}
\description{
For internal use only.
}
\details{
.getGrassDir / .getReplace / .getGrassToR / .getOutVectClass / .getAutoRegion / .getGrassVer Get common arguments.
}
\keyword{internal}
