% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_getInRastName.r
\name{.getInRastName}
\alias{.getInRastName}
\title{Get/create inrastName argument}
\usage{
.getInRastName(inRastName, rast)
}
\arguments{
\item{inRastName}{Either \code{NULL}, or a character vector, one per layer in \code{rast}.}

\item{rast}{A \code{SpatRaster} object with one or more layers.}
}
\value{
A character vector.
}
\description{
Retrieves or generates \code{inRastName}, which are names for each input raster.
}
\keyword{internal}
