% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_getInRastName_getInVectName.r
\name{.getInVectName}
\alias{.getInVectName}
\title{Get/create inVectName}
\usage{
.getInVectName(inVectName)
}
\arguments{
\item{inVectName}{Any of: NULL (use default: "inputVect") or a character.}
}
\value{
TRUE invisibly
}
\description{
Define inVectName
}
\keyword{internal}
