% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_getSession.r
\name{.getSessionStarted}
\alias{.getSessionStarted}
\title{Get session-wide arguments}
\usage{
.getSessionStarted()
}
\arguments{
\item{x}{TRUE or FALSE.}
}
\value{
A character or logical.
}
\description{
For internal use only.
}
\details{
.getSessionStarted / .setSessionStarted: Sets flag in .fasterRaster indicating if GRASS started or not
.setSession sets session variables
.getDir / .getLocation / .getMapset Get session variables
.getGrassDir / .getReplace / .getGrassToR / .getOutVectClass Get common arguments
.getRegionExtent / .getRegionDim / .getRegionRes Region variables from .faserRaster
.getGrassVer
}
\keyword{internal}
