% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_ls.r
\name{.ls}
\alias{.ls}
\title{Names of all rasters and/or spatial vectors in the active 'GRASS' location}
\usage{
.ls(rastOrVect = c("rasters", "vectors"))
}
\arguments{
\item{rastOrVect}{The type of spatial objects to display. This can include \code{'rasters'} (all rasters) and/or \code{'vectors'} (all spatial vectors). Partial matching is supported.}

\item{...}{Other arguments (not used).}
}
\value{
Nothing (displays names of rasters and/or vectors).
}
\description{
Displays the names of all rasters and/or vectors that have been exported to or created in the active 'GRASS' session's location.
}
\seealso{
\code{\link[=ls]{ls()}}; \strong{GRASS} module \href{`g.list`}{https://grass.osgeo.org/grass82/manuals/g.list.html}
}
\keyword{internal}
