% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_makeGname.r
\name{.makeGname}
\alias{.makeGname}
\title{Make 'GRASS' name for a raster or vector}
\usage{
.makeGname(x = NULL, rastOrVect = NULL, n = 1L)
}
\arguments{
\item{x}{The object for which to make the name. If \code{NULL}, we assume that the object is to be made (it does not yet exist in \code{GRASS}). We should define \code{rastOrVect} in these cases.}

\item{rastOrVect}{'raster' or 'vector'}

\item{n}{Number of names to make.}
}
\description{
Make 'GRASS' name for a raster or vector
}
\keyword{internal}
