% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSourceName.r
\name{.makeSourceName}
\alias{.makeSourceName}
\title{Make unique GRASS name for rasters, vectors, etc.}
\usage{
.makeSourceName(x = NULL, type = NULL, n = 1L, name = NULL)
}
\arguments{
\item{x}{Character or \code{NULL}: Descriptive string. \strong{Developers, please note}: To assist with debugging, \strong{GRASS} objects created by a \strong{GRASS} module have the module named in this argument (with underscores). Example: "v_in_ogr" or "r_resample".}

\item{type}{Character: \code{raster}, \code{raster3D}, \code{vector}, or \code{table}.}

\item{n}{Numeric integer: Number of names to make}

\item{name}{\code{NULL} (default) or \code{character}: Name of the output, attached as an attribute.}
}
\value{
Character vector.
}
\description{
Make unique GRASS name for rasters, vectors, etc.
}
\keyword{internal}
