% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmax.r
\name{.maxVal}
\alias{.maxVal}
\title{Get maximum value from GRaster metadata}
\usage{
.maxVal(x)
}
\arguments{
\item{x}{A \code{GRaster}.}
}
\description{
Get maximum value from GRaster metadata
}
\keyword{internal}
