% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmax.r
\name{.minVal}
\alias{.minVal}
\title{Get minimum value from GRaster metadata}
\usage{
.minVal(x)
}
\arguments{
\item{x}{A \code{GRaster}.}
}
\description{
Get minimum value from GRaster metadata
}
\keyword{internal}
