% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_private_defaults.r
\name{.namesOfHiddenOptions}
\alias{.namesOfHiddenOptions}
\title{Names or table HIDDEN of options}
\usage{
.namesOfHiddenOptions(type = NULL)
}
\arguments{
\item{type}{NULL (show all names), OR 'numeric', 'logical', or 'character', OR 'table' (show table)}
}
\description{
Names or table HIDDEN of options
}
