% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlevels.r
\name{.nlevels}
\alias{.nlevels}
\title{Count number of levels from a data.frame/table, list, or SpatRaster}
\usage{
.nlevels(x)
}
\arguments{
\item{x}{A \code{GRaster}, \code{SpatRaster}, \emph{or} a \code{data.frame}, \code{data.table}, an empty string, or a list thereof.}
}
\description{
Counts number of levels in a character string (specifically, the empty string \code{""}), a \code{data.frame}, \code{data.table}, or list of \code{data.frame}s or \code{data.table}s or empty strings.
}
\keyword{internal}
