% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{.plot}
\alias{.plot}
\title{Plot using a G-object's \code{\link[=sources]{sources()}} name}
\usage{
.plot(x, ...)
}
\arguments{
\item{x}{The \code{\link[=sources]{sources()}} name of a \code{GRaster} or \code{GVector}}

\item{...}{Other arguments to pass to \code{\link[=plot]{plot()}}.}
}
\description{
Plot using a G-object's \code{\link[=sources]{sources()}} name
}
\keyword{internal}
