% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_pmatch.r
\name{.pmatch}
\alias{.pmatch}
\title{Process 'pmatch()' call}
\usage{
.pmatch(x, table, error = TRUE, ...)
}
\arguments{
\item{x}{Character. String to match.}

\item{table}{Character vector. Values to which to match.}

\item{error}{Logical. If no match is found, return an error?}

\item{...}{Arguments to pass to \code{\link[=pmatch]{pmatch()}}.}
}
\value{
One of the values in \code{table}.
}
\description{
Private. Shortcut for several lines of code to process \code{pmatch()}. Ignores case.
}
\keyword{internal}
