% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.r
\name{.projection}
\alias{.projection}
\title{Get "GRASS" projection of raster or vector}
\usage{
.projection(x)
}
\arguments{
\item{x}{A \code{GRaster} or \code{GVector}.}
}
\value{
Character.
}
\description{
"Getter" for the @projection slot.
}
\keyword{internal}
