% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rastInfo_vectInfo.r
\name{.rastInfo}
\alias{.rastInfo}
\title{Information on rasters and vectors in "GRASS"}
\usage{
.rastInfo(x)
}
\arguments{
\item{x}{A \code{GRaster}, \code{GVector}, or \code{sources}.}
}
\value{
Extent, dimensions, resolution, bottom/top, etc.
}
\description{
Information on rasters and vectors in "GRASS"
}
\keyword{internal}
