% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastRestore.r
\name{.restore}
\alias{.restore}
\title{Hidden function to restore location/mapset based on a GSession object}
\usage{
.restore(x)
}
\description{
Hidden function to restore location/mapset based on a GSession object
}
