% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_options.r
\name{.setHiddenOptions}
\alias{.setHiddenOptions}
\title{Set hidden options}
\usage{
.setHiddenOptions(..., restore = FALSE)
}
\arguments{
\item{...}{Name of option and its value. These include:
\itemize{
\item \code{workDir}: Folder in which \strong{GRASS} rasters and vectors are to be saved. The default is to use \code{\link[=tempdir]{tempdir()}}, which works for most users who will export rasters/vectors they want from the \strong{GRASS} session before quitting \strong{R}.
\item \code{location}: A \strong{GRASS} location is a set of rasters and/or vectors that have the same map projection. They can represent the same actual location on the Earth, or they can represent different places. Generally, mosyt users will not need to worry about this setting and should ignore it. The default value is \code{'default'}.
}}
}
\description{
These options are hidden to the typical user
}
\keyword{internal}
