% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_sharedArgs.r
\name{.sharedArgs_grassToR}
\alias{.sharedArgs_grassToR}
\title{Shared argument(s)}
\arguments{
\item{grassToR}{If \code{TRUE} (default) then the output will be returned to \pkg{R}. If \code{FALSE}, then the output is left in the \code{GRASS} session and named the value in \code{outGrassName}. You can then refer to the output in subsequent \pkg{fasterRaster} functions and avoid the need to re-import it. This is called "chaining" functions together, and it can save a lot of computational time.}
}
\description{
Shared argument(s)
}
