% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_sharedArgs.r
\name{.sharedArgs_inits}
\alias{.sharedArgs_inits}
\title{Shared argument(s)}
\arguments{
\item{inits}{Either \code{NULL} (default), or a named list of arguments to pass to \code{\link{initGrass}}. If \code{NULL}, then the current active \code{GRASS} session is used, or a new one in the "default" location is created.}
}
\description{
Shared argument(s)
}
\keyword{internal}
