% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_sharedArgs.r
\name{.sharedArgs_inRastName_plural}
\alias{.sharedArgs_inRastName_plural}
\title{Shared argument(s)}
\arguments{
\item{inRastName}{The name of the input raster(s) in \code{GRASS}. If \code{rast} is a multi-layer raster, then there must be one name per layer. This argument is useful if you are "chaining" \pkg{fasterRaster} functions together and wish to refer to the input raster(s) later. The default value is the name(s) of the raster(s) (i.e., \code{names(rast)}).}
}
\description{
Shared argument(s)
}
