% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_sharedArgs.r
\name{.sharedArgs_inVectName}
\alias{.sharedArgs_inVectName}
\title{Shared argument(s)}
\arguments{
\item{inVectName}{Character. The name of the input vector in \code{GRASS}. If missing or \code{NULL}, \"code{inputVect}" will be used. This is useful if you are \link{chaining} \pkg{fasterRaster} functions together and wish to refer to the vector later in the \code{GRASS} session.}
}
\description{
Shared argument(s)
}
\keyword{internal}
