% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_sharedArgs.r
\name{.sharedArgs_outGrassName}
\alias{.sharedArgs_outGrassName}
\title{Shared argument(s)}
\arguments{
\item{outGrassName}{Character. Name of output in \code{GRASS}. This is useful if you want to refer to the output object in \code{GRASS} later in a session by "chaining" functions.}
}
\description{
Shared argument(s)
}
