% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_sharedArgs.r
\name{.sharedArgs_outVectClass}
\alias{.sharedArgs_outVectClass}
\title{Shared argument(s)}
\arguments{
\item{outVectClass}{Character or \code{NULL}. This sets the class of spatial vectors imported back to \code{R} from \code{GRASS}. If \code{'SpatRaster'} or \code{NULL} 9default), a \code{SpatVector} from the \pkg{terra} package is returned. If \code{'sf'}, an \code{sf} object from the \pkg{sf} package is returned. This argument is ignored if \code{grassToR} is \code{FALSE}. You can set this argument for all functions at once using \code{\link{fasterOptions}}.}
}
\description{
Shared argument(s)
}
\keyword{internal}
