% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_sharedArgs.r
\name{.sharedArgs_rast}
\alias{.sharedArgs_rast}
\title{Shared argument(s)}
\arguments{
\item{rast}{A \code{SpatRaster} or the name of a raster already in an existing \code{GRASS} session.}
}
\description{
Shared argument(s)
}
