% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_sharedArgs.r
\name{.sharedArgs_rast_multiple}
\alias{.sharedArgs_rast_multiple}
\title{Shared argument(s)}
\arguments{
\item{rast}{A \code{SpatRaster} with one or more layers, \emph{or} the name(s) of one or more raster in the active \code{GRASS} session.}
}
\description{
Shared argument(s)
}
