% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_sharedArgs.r
\name{.sharedArgs_vect}
\alias{.sharedArgs_vect}
\title{Shared argument(s)}
\arguments{
\item{vect}{A \code{SpatVector}, an \code{sf} object, or the name of a spatial vector in the active \code{GRASS} session.}
}
\description{
Shared argument(s)
}
\keyword{internal}
