% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rastInfo_vectInfo.r
\name{.vectInfo}
\alias{.vectInfo}
\title{Metadata on a vector in GRASS}
\usage{
.vectInfo(x, integer = TRUE, cats = NULL)
}
\arguments{
\item{x}{A \code{GVector} or the \code{\link[=sources]{sources()}} name of one.}

\item{integer}{Logical. If \code{TRUE} (default), the "cats" of a vector will be returned as type integer. If \code{FALSE}, they will be returned as a character vector.}

\item{cats}{\code{NULL} (default) or an integer vector of category numbers, one per geometry.}
}
\value{
A \code{vectInfo} object (a list).
}
\description{
This function queries \strong{GRASS} to obtain metadata on a vector.
}
\keyword{internal}
