% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vValidCats.r
\name{.vValidCats}
\alias{.vValidCats}
\title{Are the category values of a vector valid?}
\usage{
.vValidCats(x)
}
\arguments{
\item{x}{A \code{GVector} or the \code{\link[=sources]{sources()}} name of one.}
}
\value{
Logical.
}
\description{
Category values of a \strong{GRASS} vector can be invalid if \strong{GRASS} assigns more than one value to a geometry (e.g., "7/12"). This can occur when the vector was created by software that does not use a topological system (e.g., a shapefile).
}
\keyword{internal}
