% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonal.r
\name{.zonal}
\alias{.zonal}
\title{Internal function for zonal()}
\usage{
.zonal(x, z, fun, probs, zones, xnames)
}
\arguments{
\item{x}{\code{\link[=sources]{sources()}} name of \code{GRaster}.}

\item{z}{\code{\link[=sources]{sources()}} name of "zones" \code{GRaster}.}

\item{fun}{Character: Name of function(s).}

\item{probs}{Numeric in the range 0 to 1, inclusive.}

\item{zones}{Vector of zone values (integers).}

\item{xnames}{Character: Names of \code{x}.}
}
\description{
Internal function for zonal()
}
\keyword{internal}
