% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonal.r
\name{.zonalByVector}
\alias{.zonalByVector}
\title{Internal function for zonal() when y is a GVector}
\usage{
.zonalByVector(x, z, fun, probs, gtype)
}
\arguments{
\item{x}{GRaster \code{\link[=sources]{sources()}} name.}

\item{z}{GVector \code{\link[=sources]{sources()}} name.}

\item{fun}{Character}

\item{gtype}{\code{geomtype(z, grass = TRUE)} ("area", "line", or "point")}
}
\description{
Internal function for zonal() when y is a GVector
}
\keyword{internal}
