% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropRows.r
\name{dropRows,data.frame-method}
\alias{dropRows,data.frame-method}
\alias{dropRows}
\alias{dropRows,data.table-method}
\title{Remove rows in a data.table}
\usage{
\S4method{dropRows}{data.frame}(x, drops)

\S4method{dropRows}{data.table}(x, drops)
}
\arguments{
\item{drops}{Numeric, integer, or logical vector: Indices or indicators of rows to remove.

If a logical vector is supplied, rows that correspond to \code{TRUE} will be removed. If the vector is shorter than the number of rows, values of \code{drops} will be recycled.}

\item{X}{A \code{data.table} or \code{data.frame}.}
}
\value{
A \code{data.table} or \code{data.frame}.
}
\description{
As of September of 2023, the \strong{data.table} package does not have a function for removing rows by index. This function does this job and is adapted from an \url{issue}{https://github.com/Rdatatable/data.table/issues/635} to that effect raised on the \strong{data.table} \emph{GitHub} page.
}
