% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportRastToGrass.r
\name{exportRastToGrass}
\alias{exportRastToGrass}
\title{Write raster to an existing GRASS session}
\usage{
exportRastToGrass(rast, grassName = "rast", tempDir = raster::tmpDir())
}
\arguments{
\item{rast}{A \code{SpatRaster} raster.}

\item{grassName}{What to name the raster in GRASS.}

\item{tempDir}{Temporary directory.}
}
\value{
Nothing (exports a raster to a GRASS session so it can be used by other functions).
}
\description{
Write a raster to an existing GRASS session with handling for very big rasters.
}
\examples{

\donttest{

rastFile <- system.file('extdata', 'madForest2000.tif', package='fasterRaster')
madForest2000 <- rast(rastFile)

rastFile <- system.file('extdata', 'madElev.tif', package='fasterRaster')
madElev <- rast(rastFile)

# start a GRASS session
initGrass(madForest2000, grassDir = grassDir)
exportRastToGrass(madElev, grassName = 'madElev', grassDir = grassDir)

}

}
