% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportRastToGrass.r
\name{exportVectToGrass}
\alias{exportVectToGrass}
\title{Export a spatial vector to an existing GRASS session}
\usage{
exportVectToGrass(vect, grassName = "vect")
}
\arguments{
\item{vect}{A \code{SpatVector} (\pkg{terra} package) or \code{sf} (\pkg{sf} package) spatial vector object.}

\item{grassName}{Name of vector in \pkg{GRASS}.}
}
\value{
Nothing (exports a vector to a \code{GRASS} session so it can be used by other functions).
}
\description{
Export a spatial vector to an existing \code{GRASS} session.
}
\examples{
man/examples/ex_exportXToGrass.r

}
