% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faster.r
\name{faster}
\alias{faster}
\title{Generic function to call a \code{GRASS} function (module)}
\usage{
faster(
  mod,
  rast = NULL,
  vect = NULL,
  rastName = "rast",
  vectName = "vect",
  outType = NULL,
  output = "output",
  flags = c("quiet", "overwrite"),
  ...,
  grassDir = options()$grassDir,
  grassToR = TRUE
)
}
\arguments{
\item{mod}{Name of \code{GRASS} module (e.g., \code{'r.latlong'} or \code{'v.buffer'}.}

\item{rast}{A \code{SpatRaster} or the name of a raster already in an existing \code{GRASS} session. If the \code{GRASS} module does not use a raster, this can be \code{NULL}.}

\item{vect}{A \code{Spatvector}, an \code{sf} object, or the name of a spatial vector already in an existing \code{GRASS} session. If the \code{GRASS} module does not use a vector, this can be \code{NULL}.}

\item{rastName}{Character. Name of the input raster (if any) when exported to \pkg{GRASS}. Default is \code{'rast'}.}

\item{vectName}{Character. Name of the input vector (if any) when exported to \pkg{GRASS}. Default is \code{'vect'}.}

\item{outType}{\code{NULL}, \code{'raster'}, or \code{'vector'} (partial string matching is used). Type of output expected from the \code{GRASS} module.}

\item{flags}{List of flags for the \code{GRASS} module. The default (\code{c('quiet', 'overwrite')}) causes the module to report little/no messages and to overwrite existing files of the same name. For more flags, see the help documentation for the respective \code{GRASS} module.}

\item{...}{Arguments to pass to the \code{GRASS} module through \code{\link[rgrass]{execGRASS}}.}

\item{grassDir}{Name of the folder path in which \code{GRASS} is installed. For a Windows system, this might be something like \code{'C:/Program Files/GRASS GIS 8.2'}. For a Mac system, this might be like \code{"/Applications/GRASS-8.2.app/Contents/Resources"}. If you are using a lot of \pkg{fasterRaster} functions, it can be easier to set this once using \code{\link{fasterOptions}}. By default, this is \code{NULL}, in which case the install path for \code{GRASS} will be searched for, but if not found, will result in an error.}

\item{grassToR}{If \code{TRUE} (default) then the output will be returned to \pkg{R}. If \code{FALSE}, then the output is left in the \code{GRASS} session and named the value in \code{outGrassName}. You can then refer to the output in subsequent \pkg{fasterRaster} functions and avoid the need to re-import it. This is called "chaining" functions together, and it can save a lot of computational time.}
}
\value{
A raster, vector, or some other product of the \code{GRASS} module.
}
\description{
This function is wrapper for \code{\link[rgrass]{execGRASS}}, plus code necessary to initiate a \code{GRASS} session. Many of the functions in \pkg{fasterRaster} actually utilize this function.  This function works best for modules that take one raster and/or one vector as input and produce one raster or vector as output.
}
\details{
For a list of vector commands, please see <https://grass.osgeo.org/grass82/manuals/vector.html>. For a list of raster commands, please see <https://grass.osgeo.org/grass82/manuals/raster.html>.
}
\examples{
man/examples/ex_faster.r

}
\seealso{
\code{\link[rgrass]{execGRASS}}
}
