% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterApp.r
\name{fasterApp}
\alias{fasterApp}
\title{User-defined calculations on a single raster}
\usage{
fasterApp(
  rast,
  expression,
  grassDir = options()$grassDir,
  grassToR = TRUE,
  inRastName = ifelse(is.null(names(rast)), "rast", names(rast)),
  outGrassName = "appRast"
)
}
\arguments{
\item{rast}{A single-layer or multi-layer \code{SpatRaster}, or the name(s) of one or more raster(s) already in an existing \code{GRASS} session.}

\item{expression}{Character. Formula to evaluate. Below are examples where the input raster is named "rast" in the \code{GRASS} session and the output raster will be named "out". Note that raster must be referred to in formulae using their raster name (not name in R). In other words, in the examples below we could call a raster in R \code{myRast} but \code{names(myRast)} yields \code{rast}, so we need to call it "\code{rast}" in the formulae. The raster "\code{out}" will be created by the function.
\itemize{
\item Take the input raster "rast", multiplies each cell by 0, then adds 1 to each cell: \code{'out = (rast * 0) + 1'}
\item Take the input raster "rast", multiplies square the value of each cell: \code{'out = rast^2'}
\item Take the input raster "rast" and redefines each cell value to 17: \code{'out = 17'}
\item Take the input rasters "rast1" and "rast2", and multiplies them together. Note that in this case, \code{rast} will actually be a \code{SpatRaster} object with two layers named "rast1" and "rast2": \code{'out = rast1 * rast2'}
\item Take the input raster "rast" and redefines each cell value to equal the sum of it and the four cells in the rooks's neighborhood (north, south, east, west of it): \code{'out = rast[0, 0] + rast[-1, 0] + rast[0, -1] + rast[0, 1] + rast[1, 0]'}
\item Low-pass (averaging) filter across a 3x3 neighborhood: \code{'out = (rast[-1, -1] + rast[-1, 0] + rast[-1, 1] + rast[0, -1] + rast[0, 0] + rast[0, 1] + rast[1, -1] + rast[1, 0] + rast[1, 1]) / 9'}
\item High-pass ("edge-finding") filter across a 3x3 neighborhood: \code{'out = -0.7 * rast[-1, -1] -1 * rast[-1, 0] -0.7 * rast[-1, 1] -1 * rast[0, -1] + 6.8 * rast[0, 0] -1 * rast[0, 1] -0.7 * rast[1, -1] -1 * rast[1, 0] -0.7 * rast[1, 1]'}
}
See the documentation for \code{r.mapcalc}{https://grass.osgeo.org/grass82/manuals/r.mapcalc.html} for more information. Note that some operations that may make sense in \code{GRASS} may not make sense when exported back to R (e.g., changing color).}

\item{grassDir}{Name of the folder path in which \code{GRASS} is installed. For a Windows system, this might be something like \code{'C:/Program Files/GRASS GIS 8.2'}. For a Mac system, this might be like \code{"/Applications/GRASS-8.2.app/Contents/Resources"}. If you are using a lot of \pkg{fasterRaster} functions, it can be easier to set this once using \code{\link{fasterOptions}}. By default, this is \code{NULL}, in which case the install path for \code{GRASS} will be searched for, but if not found, will result in an error.}

\item{grassToR}{If \code{TRUE} (default) then the output will be returned to \pkg{R}. If \code{FALSE}, then the output is left in the \code{GRASS} session and named the value in \code{outGrassName}. You can then refer to the output in subsequent \pkg{fasterRaster} functions and avoid the need to re-import it. This is called "chaining" functions together, and it can save a lot of computational time.}

\item{outGrassName}{Character. Name of output in \code{GRASS}. This is useful if you want to refer to the output object in \code{GRASS} later in a session by "chaining" functions.}
}
\value{
If \code{grassToR} if \code{TRUE}, then a raster object with the same coordinate reference system, extent, and resolution as \code{rast}. Regardless, a raster by a name given in \code{expression} (before the "=" symbol) is written into the \code{GRASS} session.
}
\description{
This function creates a raster from another raster based on a user-defined function.
}
\details{
The function \code{\link[terra]{focal}}, \code{\link[terra]{focalCpp}}, and \code{\link{fasterFocal}} \emph{may} be faster for focal calculations.
}
\examples{
\dontrun{

library(terra)

# change this to where GRASS is installed on your system
grassDir <- 'C:/Program Files/GRASS GIS 8.2' # example for a PC
grassDir <- "/Applications/GRASS-8.2.app/Contents/Resources" # for a Mac

madElev <- fasterData('madElev')
ex <- 'out = madElev^2'
madElev2 <- fasterMapcalc(madElev, expression=ex, grassDir=grassDir)
plot(c(madElev, madElev2))

ex <- 'madElev_17 = 17'
m_17 <- fasterMapcalc(madElev, expression=ex, grassDir=grassDir)
plot(c(madElev, m_17))

# low-pass filter
ex <- 'lowPass = (madElev[-1, -1] + madElev[-1, 0] + madElev[-1, 1] + madElev[0, -1] + madElev[0, 0] + madElev[0, 1] + madElev[1, -1] + madElev[1, 0] + madElev[1, 1]) / 9'
lp <- fasterMapcalc(madElev, expression=ex, grassDir=grassDir)
plot(c(madElev, lp))

# high-pass filter
ex <- 'highPass = -0.7 * madElev[-1, -1] -1 * madElev[-1, 0] -0.7 * madElev[-1, 1] -1 * madElev[0, -1] + 6.8 * madElev[0, 0] -1 * madElev[0, 1] -0.7 * madElev[1, -1] -1 * madElev[1, 0] -0.7 * madElev[1, 1]'
hp <- fasterMapcalc(madElev, expression=ex, grassDir=grassDir)
plot(c(madElev, hp))

# Using some rasters already in a GRASS session:
# Note we are passing raster "hp" as a raster, not as something already in
# GRASS (even though it is). Also, we need to refer to the "hp" raster by
# its raster name in the formula (ie, see "names(hp)") since it is passed in
# as a raster.

input <- initGrass(rast=madElev, rastName='madElev', grassDir=grassDir)
exportRastToGrass(lp, grassName='lp')

ex <- 'out = (lp * highPass) / madElev'
mapcalc <- fasterMapcalc('madElev', 'lp', hp, expression=ex,
grassDir=grassDir)
mapcalc

# note differences between r.mapcalc and terra package math:
mapcalc
rast <- (lp * hp) / madElev
rast

# but they largely look the same
plot(c(mapcalc, rast))
plot(mapcalc - rast)

}
}
\seealso{
\code{\link[terra]{app}} and \code{\link[terra]{focal}} in \pkg{terra}; \code{\link{fasterFocal}} in \pkg{fasterRaster}; \href{https://grass.osgeo.org/grass82/manuals/r.mapcalc.html}{\code{r.mapcalc}} in \code{GRASS}
}
