% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterBufferRast.r
\name{fasterBufferRast}
\alias{fasterBufferRast}
\title{Buffer cells of a raster}
\usage{
fasterBufferRast(
  rast,
  width,
  units = "meters",
  ignore = NA,
  out = "terra",
  lowMemory = FALSE,
  replace = FALSE,
  inRastName = NULL,
  outGrassName = "bufferRast",
  grassToR = TRUE,
  grassDir = options()$grassDir,
  inits = NULL
)
}
\arguments{
\item{rast}{A \code{SpatRaster} or the name of a raster already in an existing \code{GRASS} session.}

\item{width}{Numeric. Maximum distance cells must be from focal cells to be within the buffer. Note that this function can only handle one value of \code{width} (unlike the function \code{r.buffer} in \code{GRASS}).}

\item{units}{Either \code{'meters'} (default), \code{'kilometers'}, \code{'feet'}, \code{'miles'}, or \code{'nautmiles'}. Indicates the units of \code{width}.}

\item{ignore}{Either {NA} (default) or 0. The buffer will be drawn around cells that are not {NA} or not 0, depending on this value.}

\item{out}{Any of:
\itemize{
    \item \code{'terra'} (default): The output raster will be the same as if using the \code{link[terra]{buffer}} function in the \pkg{terra} package. Cells in the buffer plus the cells around which buffers are created have values of 1, and all other cells are \code{NA}.
        \item \code{'grass'}: All cells in the bufer are represented as 2's, all cells around which buffers are created are represented as 1's, and all other cells are \code{NA}.
    \item \code{'buffer'}: All cells in the buffer are represented as 1's and all others as \code{NA}.
}}

\item{lowMemory}{Logical. If \code{FALSE} (default) use faster, memory-intensive procedure. If \code{TRUE} then use the slower, low-memory version. To help decide, consider using the low-memory version on a system with 1 GB of RAM for a raster larger than about 32000x32000 cells, or for a system with  with 8 GB of RAM a raster larger than about 90000x90000 cells.}

\item{replace}{Logical or \code{NULL}. If \code{FALSE} or \code{NULL} (default), then attempting to overwrite rasters or vectors already in a \code{GRASS} session will result in an error. You can allow overwriting by changing this to \code{TRUE}, or by setting \code{inRastName}, \code{inVectName}, and/or \code{outGrassName} to values different from rasters or vectors that already exist in the session (not all functions have these arguments). Use \code{\link{fasterLs}} to see existing rasters and vectors.  You can set this for all functions at once using \code{\link{fasterOptions}}.}

\item{inRastName}{Character, missing, or \code{NULL}. The name of the input raster in \code{GRASS}. If this is missing, the name of the raster or raster file will be used. If \code{NULL}, then "\code{inputRast}" will be used. This argument is useful if you are \link{chaining} \pkg{fasterRaster} functions together and wish to refer to the input raster later in the \code{GRASS} session.}

\item{outGrassName}{Character. Name of output in \code{GRASS}. This is useful if you want to refer to the output object in \code{GRASS} later in a session by \link{chaining} functions.}

\item{grassToR}{Logical or \code{NULL}. If \code{TRUE} or \code{NULL} (default) then the output will be returned to \code{R}. If \code{FALSE}, then the output is left in the \code{GRASS} session (with the name given by \code{outGrassName}). You can set this argument for all functions at once using \code{\link{fasterOptions}}.}

\item{grassDir}{Character or \code{NULL}. Name of the path in which \code{GRASS} is installed. For a Windows system, this might be something like \code{'C:/Program Files/GRASS GIS 8.3'}, for a Mac, like \code{"/Applications/GRASS-8.3.app/Contents/Resources"}, and for Linux like \code{'/usr/local/grass'}. If you are using a lot of \pkg{fasterRaster} functions, you can set this for all functions at once using \code{\link{fasterOptions}}. By default, this is \code{NULL}, in which case the install path for \code{GRASS} will be searched for, but if not found, will result in an error (usually the case).}

\item{inits}{Either \code{NULL} (default), or a named list of arguments to pass to \code{\link{initGrass}}. If \code{NULL}, then the current active \code{GRASS} session is used, or a new one in the "default" location is created.}
}
\value{
If \code{grassToR} if \code{TRUE}, then a raster with the same extent, resolution, and coordinate reference system as \code{rast}. Regardless, a raster with a name given by \code{outGrassName} is written into the \code{GRASS} session.
}
\description{
Create a buffer around non-\code{NA} cells or cells with values of 0 in a raster. The output will be a raster.
This function is a potentially faster version of the \code{\link[terra]{buffer}} function in the \pkg{terra} package for calculating a buffer around non-\code{NA} cells or cells with values of 0 in a raster. The output will be a raster.
}
\examples{
\dontrun{

library(terra)

# change this to where GRASS is installed on your system
grassDir <- 'C:/Program Files/GRASS GIS 8.2' # example for a PC
grassDir <- "/Applications/GRASS-8.2.app/Contents/Resources" # for a Mac

madForest2000 <- fasterData('madForest2000')

# NOTE: If you already have an active GRASS session, running fasterRaster
# functions in examples could modify it. To obviate that, the function below
# 
rastBuff <- fasterBufferRast(rast=madForest2000, width=2,
units='kilometers', grassDir=grassDir, location='examples')

plot(rastBuff, col='forestgreen')
legend('topright', legend=c('buffered + buffer'),
fill='forestgreen')

# GRASS default
rastBuffGrass <- fasterBufferRast(rast=madForest2000, width=2,
units='kilometers', out='grass', grassDir=grassDir)

plot(rastBuffGrass, col=c('gray', 'forestgreen'))
legend('topright', legend=c('buffered', 'buffer'),
fill=c('gray', 'forestgreen'))

# just the buffer
rastBuffBuff <- fasterBufferRast(rast=madForest2000, width=2,
units='kilometers', out='buffer', grassDir=grassDir)

plot(rastBuffBuff, col='forestgreen')
legend('topright', legend=c('buffer only'),
fill='forestgreen')

}
}
\seealso{
\code{\link[terra]{buffer}} in the \pkg{terra} package; \href{https://grass.osgeo.org/grass82/manuals/r.buffer.html}{\code{r.buffer}} in \code{GRASS}
}
