% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterBufferVect.r
\name{fasterBufferVect}
\alias{fasterBufferVect}
\title{Buffer a spatial vector}
\usage{
fasterBufferVect(
  vect,
  width,
  union = TRUE,
  replace = FALSE,
  inVectName = NULL,
  outGrassName = "bufferVect",
  outVectClass = NULL,
  grassToR = TRUE,
  grassDir = options()$grassDir,
  inits = NULL
)
}
\arguments{
\item{vect}{A \code{SpatVector}, an \code{sf} object, or the name of a spatial vector in the active \code{GRASS} session.}

\item{width}{Size of the buffer in map units (usually meters). To create "inward" buffers, you can use negative numbers.}

\item{union}{If \code{TRUE} (default), the union buffers of different features that intersect. If \code{FALSE}, then each input feature will have its own buffer, and attributes from each input feature will be transfered to each buffer feature.}

\item{replace}{Logical or \code{NULL}. If \code{FALSE} or \code{NULL} (default), then attempting to overwrite rasters or vectors already in a \code{GRASS} session will result in an error. You can allow overwriting by changing this to \code{TRUE}, or by setting \code{inRastName}, \code{inVectName}, and/or \code{outGrassName} to values different from rasters or vectors that already exist in the session (not all functions have these arguments). Use \code{\link{fasterLs}} to see existing rasters and vectors.  You can set this for all functions at once using \code{\link{fasterOptions}}.}

\item{inVectName}{Character. The name of the input vector in \code{GRASS}. If missing or \code{NULL}, \"code{inputVect}" will be used. This is useful if you are \link{chaining} \pkg{fasterRaster} functions together and wish to refer to the vector later in the \code{GRASS} session.}

\item{outGrassName}{Character. Name of output in \code{GRASS}. This is useful if you want to refer to the output object in \code{GRASS} later in a session by \link{chaining} functions.}

\item{outVectClass}{Character or \code{NULL}. This sets the class of spatial vectors imported back to \code{R} from \code{GRASS}. If \code{'SpatRaster'} or \code{NULL} 9default), a \code{SpatVector} from the \pkg{terra} package is returned. If \code{'sf'}, an \code{sf} object from the \pkg{sf} package is returned. This argument is ignored if \code{grassToR} is \code{FALSE}. You can set this argument for all functions at once using \code{\link{fasterOptions}}.}

\item{grassToR}{Logical or \code{NULL}. If \code{TRUE} or \code{NULL} (default) then the output will be returned to \code{R}. If \code{FALSE}, then the output is left in the \code{GRASS} session (with the name given by \code{outGrassName}). You can set this argument for all functions at once using \code{\link{fasterOptions}}.}

\item{grassDir}{Character or \code{NULL}. Name of the path in which \code{GRASS} is installed. For a Windows system, this might be something like \code{'C:/Program Files/GRASS GIS 8.3'}, for a Mac, like \code{"/Applications/GRASS-8.3.app/Contents/Resources"}, and for Linux like \code{'/usr/local/grass'}. If you are using a lot of \pkg{fasterRaster} functions, you can set this for all functions at once using \code{\link{fasterOptions}}. By default, this is \code{NULL}, in which case the install path for \code{GRASS} will be searched for, but if not found, will result in an error (usually the case).}

\item{inits}{Either \code{NULL} (default), or a named list of arguments to pass to \code{\link{initGrass}}. If \code{NULL}, then the current active \code{GRASS} session is used, or a new one in the "default" location is created.}
}
\value{
A \code{SpatVector} (\pkg{terra} package) or an object of class \code{sf} (\code{sf} package), if \code{options(grassVectOut = 'sf')}.
}
\description{
Creates a buffer around a spatial vector.
}
\examples{
\dontrun{

library(sf)

# change this to where GRASS is installed on your system
grassDir <- 'C:/Program Files/GRASS GIS 8.2' # example for a PC
grassDir <- "/Applications/GRASS-8.2.app/Contents/Resources" # for a Mac

madRivers <- fasterData('madRivers')

riverBuff <- fasterBufferVect(madRivers, width=1000, grassDir=grassDir)

plot(riverBuff)
plot(st_geometry(madRivers), col='blue', add=TRUE)

}
}
\seealso{
\code{\link{fasterBufferRast}} in \pkg{fasterRaster}; \code{\link[terra]{buffer}} in package \pkg{terra}; \code{\link[sf]{st_buffer}} in package \pkg{sf}; \href{https://grass.osgeo.org/grass82/manuals/v.buffer.html}{\code{v.buffer}} in \code{GRASS}
}
