% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterConvHull.r
\name{fasterConvHull}
\alias{fasterConvHull}
\title{Minimum convex hull around a spatial vector}
\usage{
fasterConvHull(
  vect,
  replace = FALSE,
  inVectName = NULL,
  outGrassName = "convHullVect",
  outVectClass = NULL,
  grassToR = TRUE,
  grassDir = options()$grassDir,
  inits = NULL
)
}
\arguments{
\item{vect}{A \code{SpatVector}, an \code{sf} object, or the name of a spatial vector in the active \code{GRASS} session.}

\item{replace}{Logical or \code{NULL}. If \code{FALSE} or \code{NULL} (default), then attempting to overwrite rasters or vectors already in a \code{GRASS} session will result in an error. You can allow overwriting by changing this to \code{TRUE}, or by setting \code{inRastName}, \code{inVectName}, and/or \code{outGrassName} to values different from rasters or vectors that already exist in the session (not all functions have these arguments). Use \code{\link{fasterLs}} to see existing rasters and vectors.  You can set this for all functions at once using \code{\link{fasterOptions}}.}

\item{inVectName}{Character. The name of the input vector in \code{GRASS}. If missing or \code{NULL}, \"code{inputVect}" will be used. This is useful if you are \link{chaining} \pkg{fasterRaster} functions together and wish to refer to the vector later in the \code{GRASS} session.}

\item{outGrassName}{Character. Name of output in \code{GRASS}. This is useful if you want to refer to the output object in \code{GRASS} later in a session by \link{chaining} functions.}

\item{outVectClass}{Character or \code{NULL}. This sets the class of spatial vectors imported back to \code{R} from \code{GRASS}. If \code{'SpatRaster'} or \code{NULL} 9default), a \code{SpatVector} from the \pkg{terra} package is returned. If \code{'sf'}, an \code{sf} object from the \pkg{sf} package is returned. This argument is ignored if \code{grassToR} is \code{FALSE}. You can set this argument for all functions at once using \code{\link{fasterOptions}}.}

\item{grassToR}{Logical or \code{NULL}. If \code{TRUE} or \code{NULL} (default) then the output will be returned to \code{R}. If \code{FALSE}, then the output is left in the \code{GRASS} session (with the name given by \code{outGrassName}). You can set this argument for all functions at once using \code{\link{fasterOptions}}.}

\item{grassDir}{Character or \code{NULL}. Name of the path in which \code{GRASS} is installed. For a Windows system, this might be something like \code{'C:/Program Files/GRASS GIS 8.3'}, for a Mac, like \code{"/Applications/GRASS-8.3.app/Contents/Resources"}, and for Linux like \code{'/usr/local/grass'}. If you are using a lot of \pkg{fasterRaster} functions, you can set this for all functions at once using \code{\link{fasterOptions}}. By default, this is \code{NULL}, in which case the install path for \code{GRASS} will be searched for, but if not found, will result in an error (usually the case).}

\item{inits}{Either \code{NULL} (default), or a named list of arguments to pass to \code{\link{initGrass}}. If \code{NULL}, then the current active \code{GRASS} session is used, or a new one in the "default" location is created.}
}
\value{
A \code{SpatVector} (\pkg{terra} package) or an object of class \code{sf} (\code{sf} package), if \code{options(grassVectOut = 'sf')}.
}
\description{
This function creates a minimum convex hull around a spatial vector.
}
\examples{
\dontrun{

library(sf)

# change this to where GRASS is installed on your system
grassDir <- 'C:/Program Files/GRASS GIS 8.2' # example for a PC
grassDir <- "/Applications/GRASS-8.2.app/Contents/Resources" # for a Mac

# spatial points object
data('madCoast0')
data('madDypsis')

### minimum convex hull
mcp <- fasterConvHull(madDypsis, grassDir=grassDir)

plot(st_geometry(madCoast0))
plot(st_geometry(madDypsis), pch=16, add=TRUE)
plot(mcp, border='blue', add=TRUE)

### Delaunay triangulation
del <- fasterDelaunay(madDypsis, grassDir=grassDir)

plot(st_geometry(madCoast0))
plot(st_geometry(madDypsis), pch=16, add=TRUE)
plot(del, border='blue', add=TRUE)

### Voronoi diagram for points
vor <- fasterVoronoi(madDypsis, grassDir=grassDir)

plot(st_geometry(madCoast0))
plot(st_geometry(madDypsis), pch=16, add=TRUE)
plot(vor, border='blue', add=TRUE)

### Voronoi diagram for polygons
data(madCoast4)
vor <- fasterVoronoi(madCoast4, type='polygons', grassDir=grassDir)

plot(st_geometry(madCoast0))
plot(st_geometry(madCoast4), add=TRUE)
plot(vor, border='blue', add=TRUE)

}

}
\seealso{
\code{\link[terra]{convHull}} in package \pkg{terra}; \code{\link[sf]{st_convex_hull}} in package \pkg{sf}; \href{https://grass.osgeo.org/grass82/manuals/v.hull.html}{\code{v.hull}} in \code{GRASS}
}
