% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterCopy.r
\name{fasterCopy}
\alias{fasterCopy}
\title{Make a copy of a raster or vector}
\usage{
fasterCopy(from, to, rastOrVect = NULL, replace = FALSE)
}
\arguments{
\item{from}{Name of the raster or vector to copy.}

\item{to}{New name for the raster or vector.}

\item{rastOrVect}{The type of object named in \code{from}. This should be either \code{'raster'} or \code{'vector'} (partial matching is used). Typically, this does not need to be defined and is left \code{NULL}. However, \code{GRASS} allows for rasters and vectors of the same name, so if the object named in \code{from} is ambiguous, then this should be defined.}

\item{replace}{If \code{TRUE}, then replace any object with the same name as \code{to}. The default is \code{FALSE}, in which case a conflict results in an error.}
}
\value{
\code{TRUE} (invisibly). Also copies a raster or vector in the active \code{GRASS} session.
}
\description{
Make a copy of a raster or vector in an active \code{GRASS} session.
}
\examples{
/man/examples/ex_fasterOperations.r

}
\seealso{
\code{\link{fasterRename}} and \code{\link{fasterLs}} in \pkg{fasterRaster}; \code{GRASS} module \href{https://grass.osgeo.org/grass82/manuals/g.copy.html}{g.copy}
}
