% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterCropRast.r
\name{fasterCropRast}
\alias{fasterCropRast}
\title{Crop a raster to the area of overlap with another raster or vector}
\usage{
fasterCropRast(
  rast,
  template,
  rastOrVect = NULL,
  replace = FALSE,
  inRastName = NULL,
  grassToR = TRUE,
  outGrassName = "croppedRast",
  grassDir = options()$grassDir,
  inits = NULL
)
}
\arguments{
\item{rast}{A \code{SpatRaster} or the name of a raster already in an existing \code{GRASS} session.}

\item{template}{Any of:
\itemize{
    \item A \code{SpatExtent}, \code{SpatRaster}, \code{SpatVector}, or object of class \code{sf}, used to define the extent of the crop region. This object is \emph{not} exported to the \code{GRASS} session.
    \item The name of raster or vector already in the \code{GRASS} session.
}}

\item{rastOrVect}{Either \code{'raster'} or \code{'vector'}. If \code{NULL} (default), then the function will attempt to guess whether \code{template} refers to a raster or vector. However, in \code{GRASS}, it is possible to have a raster and a vector of the same name. If this is the case, then you can specify whether \code{template} is a raster or vector (partial matching is supported).}

\item{replace}{Logical or \code{NULL}. If \code{FALSE} or \code{NULL} (default), then attempting to overwrite rasters or vectors already in a \code{GRASS} session will result in an error. You can allow overwriting by changing this to \code{TRUE}, or by setting \code{inRastName}, \code{inVectName}, and/or \code{outGrassName} to values different from rasters or vectors that already exist in the session (not all functions have these arguments). Use \code{\link{fasterLs}} to see existing rasters and vectors.  You can set this for all functions at once using \code{\link{fasterOptions}}.}

\item{inRastName}{Character, missing, or \code{NULL}. The name of the input raster in \code{GRASS}. If this is missing, the name of the raster or raster file will be used. If \code{NULL}, then "\code{inputRast}" will be used. This argument is useful if you are \link{chaining} \pkg{fasterRaster} functions together and wish to refer to the input raster later in the \code{GRASS} session.}

\item{grassToR}{Logical or \code{NULL}. If \code{TRUE} or \code{NULL} (default) then the output will be returned to \code{R}. If \code{FALSE}, then the output is left in the \code{GRASS} session (with the name given by \code{outGrassName}). You can set this argument for all functions at once using \code{\link{fasterOptions}}.}

\item{grassDir}{Character or \code{NULL}. Name of the path in which \code{GRASS} is installed. For a Windows system, this might be something like \code{'C:/Program Files/GRASS GIS 8.3'}, for a Mac, like \code{"/Applications/GRASS-8.3.app/Contents/Resources"}, and for Linux like \code{'/usr/local/grass'}. If you are using a lot of \pkg{fasterRaster} functions, you can set this for all functions at once using \code{\link{fasterOptions}}. By default, this is \code{NULL}, in which case the install path for \code{GRASS} will be searched for, but if not found, will result in an error (usually the case).}

\item{inits}{Either \code{NULL} (default), or a named list of arguments to pass to \code{\link{initGrass}}. If \code{NULL}, then the current active \code{GRASS} session is used, or a new one in the "default" location is created.}
}
\value{
A \code{SpatRaster}. Also creates a raster in a new grass session named \code{outGrassName}.
}
\description{
Crop a raster to the extent of another raster or vector.
}
\examples{
man/examples/ex_fasterCropExtendRast.r

}
\seealso{
\code{\link{fasterRes}}, \code{\link{fasterExtendRast}}, and \code{\link{fasterTrimRast}} in \pkg{fasterRaster}; \code{\link[terra]{crop}} and \code{\link[terra]{extend}} in \pkg{terra}; \code{GRASS} module \href{https://grass.osgeo.org/grass82/manuals/r.clip.html}{\code{r.clip}}
}
