% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterCRS.r
\name{fasterCRS}
\alias{fasterCRS}
\title{Display the coordinate reference system of the current GRASS sesion}
\usage{
fasterCRS()
}
\value{
Nothing (displays information on the coordinate reference system of the current \code{GRASS} session).
}
\description{
Display the coordinate reference system of the current \code{GRASS} sesion.
}
\examples{
\dontrun{

library(sf)
library(terra)

# start a GRASS session with a raster and get information on it
madElev <- fasterData('madElev')
initGrass(madElev, inRastName = 'madElev')
fasterInfoRast('madElev')

# export a vector and get information on it
madCoast0 <- fasterData('madCoast0')
exportVectToGrass(madCoast0, inVectName = 'madCoast0')
fasterInfoVect('madCoast0')

# information on the coordinate rafeternec system
fasterCRS()

# names of all rasters and vectors in the GRASS session
fasterLs()

# remove raster
fasterRm('madElev')
fasterLs()

# remove vector
fasterRm('madCoast0')
fasterLs()

}

}
\seealso{
\href{https://grass.osgeo.org/grass82/manuals/g.proj.html}{\code{g.list}} in \code{GRASS}
}
