% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterData.r
\name{fasterData}
\alias{fasterData}
\title{Get one of the example rasters or spatial vectors}
\usage{
fasterData(x)
}
\arguments{
\item{x}{The name of the raster or spatial vector to get. All of these pertain to a portion of the eastern coast of Madagascar. Options include:
\itemize{
    \item    \code{'madCoast0'}: Outline of the region (\code{sf} spatial vector).
    \item    \code{'madCoast4'}: Outlines of the Fokontanies (Communes) of the region (\code{sf} spatial vector).
    \item    \code{'madElev'}: Elevation (raster).
    \item    \code{'madForest2000'}: Forest cover in year 2000 (raster).
    \item    \code{'madForest2014'}: Forest cover in year 2014 (raster).
    \item    \code{'madRivers'}: Major rivers (\code{sf} spatial vector).
}}
}
\value{
A \code{SpatRaster} or \code{sf} spatial vector.
}
\description{
This function is a simple way to get one the several example rasters or spatial vector datasets that come with \pkg{fasterRaster}. These include.
}
\examples{
man/examples/ex_madData.r

}
\seealso{
\code{\link{madCoast0}}, \code{\link{madCoast4}}, \code{\link{madElev}}, \code{\link{madForest2000}}, \code{\link{madForest2014}}, \code{\link{madRivers}}
}
