% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterDim.r
\name{fasterDim}
\alias{fasterDim}
\alias{fasterNcell}
\alias{fasterNcol}
\alias{fasterNrow}
\title{Number of rows and columns of a raster}
\usage{
fasterDim(rast, names = TRUE)

fasterNcell(rast)

fasterNcol(rast)

fasterNrow(rast)
}
\arguments{
\item{rast}{Name of a raster in the active \code{GRASS} session.}

\item{names}{If \code{TRUE} (default), return a named raster.}
}
\value{
A one- or two-element numeric vector with number of rows and columns, or number of cells in the raster.
}
\description{
Reports the number of cells, rows, and/or columns of a raster in a \code{GRASS} session.
}
\examples{
\dontrun{

library(sf)
library(terra)

# start a GRASS session with a raster
madElev <- fasterData('madElev')
initGrass(madElev, inRastName = 'madElev')

# import a vector
madRivers <- fasterData('madRivers')
fasterVect(madRivers, inVectName = 'madRivers')

# what's in the GRASS session?
fasterLs('rasters') # what rasters?
fasterLs('vectors') # what vectors?
fasterLs() # what rasters and vectors?

# information on the raster and vector
fasterInfo('madElev')
fasterInfo('madCoast0')
fasterInfo() # everything

# information on the coordinate reference system of the GRASS session
fasterCRS() # "plain" format
fasterCRS(TRUE) # nice for the eyes

# spatial extent of the GRASS session
fasterExt('madElev') # extent of madElev raster
fasterExt('madRivers') # extent of madRivers vector
fasterExt() # maximum extent of all spatial objects

}
}
\seealso{
\code{\link{fasterInfo}}, \code{\link{fasterExt}}, and \code{\link{fasterRes}} in \pkg{fasterRaster}; \code{\link[terra]{dim}}, \code{\link[terra]{nrow}}, \code{\link[terra]{ncol}}, and \code{\link[terra]{ncell}} in package \pkg{terra}; \code{GRASS} module \href{https://grass.osgeo.org/grass82/manuals/r.info.html}{\code{r.info}}
}
