% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterExt.r
\name{fasterExt}
\alias{fasterExt}
\title{Spatial extent of the active GRASS region or raster}
\usage{
fasterExt(x, rastOrVect = NULL, terra = FALSE, names = TRUE)
}
\arguments{
\item{x}{Name of a raster or vector in the active \code{GRASS} session.}

\item{rastOrVect}{If \code{NULL} (default), then the function will attempt to guess whether \code{x} refers to a raster or vector. However, in \code{GRASS}, it is possible to have a raster and a vector of the same name. This argument helps disambugiate this case.}

\item{terra}{If \code{TRUE}, the object that is returned is an \code{\link[terra]{SpatExtent}} object. If \code{FALSE} (default), then the bounding coordinates are returned as a vector.}

\item{names}{If \code{TRUE} (default), then the vector will have named values.}
}
\value{
A numeric vector or an object of class \code{\link[terra]{SpatExtent}}.
}
\description{
Report the spatial extent of a raster or vector in an active \code{GRASS} session.
}
\examples{
\dontrun{

library(sf)
library(terra)

# start a GRASS session with a raster
madElev <- fasterData('madElev')
initGrass(madElev, inRastName = 'madElev')

# import a vector
madRivers <- fasterData('madRivers')
fasterVect(madRivers, inVectName = 'madRivers')

# what's in the GRASS session?
fasterLs('rasters') # what rasters?
fasterLs('vectors') # what vectors?
fasterLs() # what rasters and vectors?

# information on the raster and vector
fasterInfo('madElev')
fasterInfo('madCoast0')
fasterInfo() # everything

# information on the coordinate reference system of the GRASS session
fasterCRS() # "plain" format
fasterCRS(TRUE) # nice for the eyes

# spatial extent of the GRASS session
fasterExt('madElev') # extent of madElev raster
fasterExt('madRivers') # extent of madRivers vector
fasterExt() # maximum extent of all spatial objects

}
}
\seealso{
\code{\link{fasterInfo}}, \code{\link{fasterDim}}, and \code{\link{fasterRes}} in \pkg{fasterRaster}; \code{\link[terra]{SpatExtent}} and \code{\link[terra]{ext}} in the \pkg{terra} package; \code{GRASS} modules \href{https://grass.osgeo.org/grass82/manuals/r.info.html}{r.info} and \href{https://grass.osgeo.org/grass82/manuals/v.info.html}{v.info}
}
