% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterFunctions.r
\docType{data}
\name{fasterFunctions}
\alias{fasterFunctions}
\title{Table of functions in fasterRaster and equivalents in \pkg{GRASS} and terra}
\format{
An object of class \code{'data.table'}.
}
\usage{
data(fasterFunctions)
}
\description{
Table of functions in \pkg{fasterRaster} and equivalents in \pkg{GRASS} and \pkg{terra}, plus a link to the \pkg{GRASS} function.
}
\examples{
man/examples/ex_fasterHelp.r

}
\keyword{functions}
