% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterHelp.r
\name{fasterHelp}
\alias{fasterHelp}
\title{Lookup equivalent functions in fasterRaster, \code{GRASS}, and terra}
\usage{
fasterHelp(x = NULL, terra = FALSE, grass = FALSE, approx = FALSE)
}
\arguments{
\item{x}{Name of the function to lookup. This can be a function in \pkg{fasterRaster}, \pkg{terra}, or \pkg{GRASS}. By default, it will be assumed to be the \pkg{fasterRaster} function. If this is \code{NULL} (default), it will return the entire crosswalk table.}

\item{terra, grass, sf}{If \code{TRUE}, the function named in \code{x} is from \pkg{terra}, \pkg{GRASS}, or \pkg{sf}. Only one of \code{terra}, \code{grass}, and \code{sf} can be \code{TRUE}.}
}
\value{
Nothing (displays a equivalent functions on the screen).
}
\description{
Lookup equivalent functions in fasterRaster, \pkg{GRASS}, and terra, plus get the URL of the help file for the \code{GRASS} function, if it exists.
}
\examples{
man/examples/ex_fasterHelp.r

}
