% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterInfo.r
\name{fasterInfo}
\alias{fasterInfo}
\title{Information on a raster or vector in a GRASS session}
\usage{
fasterInfo(x, rastOrVect = NULL)
}
\arguments{
\item{x}{The name of the raster(s) or vector(s) in the active \code{GRASS} session. These must be in quotes. If missing, then information on all rasters and vectors is reported.}

\item{rastOrVect}{Either \code{'rasters'} or \code{'vectors'} (one value per value in \code{x}). If \code{NULL} (default), then the function will attempt to guess whether \code{x} refers to a raster or vector. However, in \code{GRASS}, it is possible to have a raster and a vector of the same name. If this is the case, then you can specify whether \code{x} is a raster or vector (partial matching is supported).}
}
\value{
A \code{data.frame} (or \code{NULL} if no rasters or vectors were found).
}
\description{
Displays information on a raster or vector already in an R session. For further
}
\examples{
\dontrun{

library(sf)
library(terra)

# start a GRASS session with a raster
madElev <- fasterData('madElev')
initGrass(madElev, inRastName = 'madElev')

# import a vector
madRivers <- fasterData('madRivers')
fasterVect(madRivers, inVectName = 'madRivers')

# what's in the GRASS session?
fasterLs('rasters') # what rasters?
fasterLs('vectors') # what vectors?
fasterLs() # what rasters and vectors?

# information on the raster and vector
fasterInfo('madElev')
fasterInfo('madCoast0')
fasterInfo() # everything

# information on the coordinate reference system of the GRASS session
fasterCRS() # "plain" format
fasterCRS(TRUE) # nice for the eyes

# spatial extent of the GRASS session
fasterExt('madElev') # extent of madElev raster
fasterExt('madRivers') # extent of madRivers vector
fasterExt() # maximum extent of all spatial objects

}
}
\seealso{
\code{\link{fasterLs}}, \code{\link{fasterExt}}, \code{\link{fasterDim}}, and \code{\link{fasterRes}} in \pkg{fasterRaster}; \href{https://grass.osgeo.org/grass82/manuals/r.info.html}{\code{r.info}} and \href{https://grass.osgeo.org/grass82/manuals/v.info.html}{\code{v.info}} in \code{GRASS}
}
