% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterLs.r
\name{fasterLs}
\alias{fasterLs}
\title{Names of all rasters and/or spatial vectors in a GRASS session}
\usage{
fasterLs(type = c("raster", "vector"), incTemp = FALSE)
}
\arguments{
\item{type}{The type of spatial dataset for which to display names. This can include \code{'rasters'} (all rasters) and/or \code{'vectors'} (all spatial vectors). Partial matching is supported.}

\item{incTemp}{If \code{TRUE}, then remove any temporary files from the results. By default, these are not returned as part of the results. Some \pkg{fasterRaster} functions create temporary rasters or vectors which begin with the string "\code{TEMPTEMP_}".}
}
\value{
Nothing (displays names of rasters and/or vectors).
}
\description{
Displays the names of all rasters and/or vectors that have been exported to or created in a GRASS session.
}
\examples{
\dontrun{

library(sf)
library(terra)

# start a GRASS session with a raster
madElev <- fasterData('madElev')
initGrass(madElev, inRastName = 'madElev')

# import a vector
madRivers <- fasterData('madRivers')
fasterVect(madRivers, inVectName = 'madRivers')

# what's in the GRASS session?
fasterLs('rasters') # what rasters?
fasterLs('vectors') # what vectors?
fasterLs() # what rasters and vectors?

# information on the raster and vector
fasterInfo('madElev')
fasterInfo('madCoast0')
fasterInfo() # everything

# information on the coordinate reference system of the GRASS session
fasterCRS() # "plain" format
fasterCRS(TRUE) # nice for the eyes

# spatial extent of the GRASS session
fasterExt('madElev') # extent of madElev raster
fasterExt('madRivers') # extent of madRivers vector
fasterExt() # maximum extent of all spatial objects

}
}
\seealso{
\code{\link{fasterInfoRast}}, \code{\link{fasterInfoVect}}, \href{https://grass.osgeo.org/grass82/manuals/g.list.html}{\code{g.list}} in \code{GRASS}
}
