% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterQuantile.r
\name{fasterQuantile}
\alias{fasterQuantile}
\title{Compute quantiles for a raster.}
\usage{
fasterQuantile(rast, probs = c(0.025, 0.25, 0.5, 0.75, 0.975),
  grassLoc = NULL, grassInit = TRUE, ...)
}
\arguments{
\item{rast}{Either a raster or the full path and name of a raster object.}

\item{probs}{A numeric list of quantiles to calculate.}

\item{grassLoc}{Either \code{NULL} or a 3-element character vector. If the latter, the first element is the base path to the installation of GRASS, the second the version number, and the third the install type for GRASS.  For example, \code{c('C:/OSGeo4W64/', 'grass-7.4.1', 'osgeo4W')}. See \code{\link[link2GI]{linkGRASS7}} for further help. If \code{NULL} (default) the an installation of GRASS is searched for; this may take several minutes.}

\item{grassInit}{Logical, if \code{TRUE} (default) then a new GRASS session is initialized. If \code{FALSE} then it is assumed a GRASS session has been initialized using the raster in \code{rast}. The latter is useful if you are chaining \pkg{fasterRaster} functions together and the first function initializes the session.}

\item{...}{Arguments to pass to \code{\link[rgrass7]{execGRASS}}.}

\item{grassToR}{Logical, if \code{TRUE} (default) then the product of the calculations will be returned to R. If \code{FALSE}, then the product is left in the GRASS session and named \code{slope}, \code{aspect}, \code{profileCurve}, \code{tanCurve}, \code{eastWestSlope}, or \code{northSouthSlope}. The latter case is useful (and faster) when chaining several \pkg{fasterRaster} functions together.}
}
\value{
A named vector of the values for each quantile named in \code{probs}.
}
\description{
This function is a potentially faster version of the function \code{quantile(raster, probs)} for calculating the quantiles of a raster. This function will also work on rasters too big to load into memory using the \pkg{raster} package.
}
\details{
See (r.buffer)[https://grass.osgeo.org/grass74/manuals/r.buffer.html] for more details.  Note that if you get an error saying "", then you should add the EPSG code to the beginning of the raster and vector coordinate reference system string (its "proj4string"). For example, \code{proj4string(x) <- CRS('+init=epsg:32738')}. EPSG codes for various projections, datums, and locales can be found at (Spatial Reference)[http://spatialreference.org/].
}
\examples{
\dontrun{
# change this according to where GRASS 7 is installed on your system
grassLoc <- c('C:/OSGeo4W64/', 'grass-7.4.1', 'osgeo4W')

data(madForest2000)

# calculate distance to forest
distToForest <- fasterRastDistance(madForest2000, grassLoc=grassLoc)

# calculate quantiles
probs <- c(0.025, 0.1, 0.25, 0.5, 0.75, 0.9, 0.975)
quants <- fasterQuantile(distToForest, probs=probs, grassLoc=grassLoc)
}
}
\seealso{
\code{\link[stats]{quantile}} in the \pkg{base} package and \code{\link[raster]{quantile}} in the \pkg{raster} package
}
