% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterRast.r
\name{fasterRast}
\alias{fasterRast}
\title{Export raster(s) to an existing GRASS session}
\usage{
fasterRast(rast, inRastName = NULL, replace = FALSE, autoRegion = TRUE, ...)
}
\arguments{
\item{inRastName}{The name of the input raster(s) in \code{GRASS}. If \code{rast} is a multi-layer raster, then there must be one name per layer. If missing or \code{NULL}, the name(s) of the raster(s) will be used (i.e., \code{names(rast)}). This argument is useful if you are \link{chaining} \pkg{fasterRaster} functions together and wish to refer to the raster(s) later in the \code{GRASS} session.}

\item{replace}{Logical or \code{NULL}. If \code{FALSE} or \code{NULL} (default), then attempting to overwrite rasters or vectors already in a \code{GRASS} session will result in an error. You can allow overwriting by changing this to \code{TRUE}, or by setting \code{inRastName}, \code{inVectName}, and/or \code{outGrassName} to values different from rasters or vectors that already exist in the session (not all functions have these arguments). Use \code{\link{fasterLs}} to see existing rasters and vectors.  You can set this for all functions at once using \code{\link{fasterOptions}}.}

\item{autoRegion}{Logical or \code{NULL}. If \code{TRUE} or \code{NULL} (default), then during execution of this function, the \code{GRASS} \link{region} will be resized to encompass the raster/vector/area of interest, and if relevant, resampled to match its spatial resolution. If \code{FALSE}, the the current region's extent and resolution will be used. Most users will want to keep this \code{TRUE} or \code{NULL} (same as \code{TRUE}). If you change this to \code{FALSE}, you will need to keep track of the \code{GRASS} \link{region} properties. This argument can be set for all functions at once using \code{\link{fasterOptions}}.}

\item{...}{Additional arguments (unused).}

\item{A}{\code{SpatRaster} with one or more layers, or a character vector providing the path and names of raster files to import into the active \code{GRASS} session.}
}
\value{
\code{TRUE} (invisibly, if the raster was successfully exported). The function also exports a raster to a \code{GRASS} session so it can be used by other functions.
}
\description{
Export one or more rasters to an existing \code{GRASS} session. To import a raster from \code{GRASS} to \pkg{R}, use \code{\link[rgrass]{read_RAST}}. Note that rasters imported back from \code{GRASS}, when saved using \pkg{terra}'s \code{\link[terra]{writeRaster}} function, often obtain the integer data type, which can truncate values. To obviate this, see \code{\link{writeRaster4}} or save rasters directly from \code{R} using \code{\link{fasterWriteRaster}}.
}
\examples{
man/examples/ex_fasterRast_fasterVect.R

}
\seealso{
\code{\link[terra]{rast}} in package \pkg{terra}; \code{\link[rgrass]{write_RAST}} and \code{\link[rgrass]{read_RAST}} in package \pkg{rgrass}; and \code{GRASS} module \href{https://grass.osgeo.org/grass82/manuals/r.in.gdal.html}{r.in.gdal}
}
