% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterRastDistance.r
\name{fasterRastDistance}
\alias{fasterRastDistance}
\title{Distance between cells with NAs and closest non-NA cells on a raster}
\usage{
fasterRastDistance(
  rast,
  metric = "euclidean",
  meters = TRUE,
  fillNAs = TRUE,
  replace = FALSE,
  inRastName = NULL,
  outGrassName = "distanceRast",
  grassToR = TRUE,
  grassDir = options()$grassDir,
  inits = NULL
)
}
\arguments{
\item{rast}{A \code{SpatRaster} or the name of a raster already in an existing \code{GRASS} session.}

\item{metric}{Character, indicates type of distance to calculate:
\itemize{
\item \code{euclidean} (default): Euclidean distance
\item \code{geodesic}: geographic distance (suggested to use with \code{meters = TRUE}).
\item \code{squared}: Squared Euclidean distance
\item \code{maximum}: Maximum Euclidean distance
\item \code{manhattan}: Manhattan distance (i.e., "taxicab" distance, distance along cells going only north-south and east-west and never along a diagonal).
}}

\item{meters}{Logical, if \code{TRUE} then distance is in meters. If \code{FALSE} then distance is in map units.}

\item{fillNAs}{Logical, if \code{TRUE} (default) then fill code{NA} cells with distance between them and closest non-\code{NA}. If \code{TRUE} then replace value in non-{NA} cells with distance between them and nearest \code{NA} cell.}

\item{replace}{Logical or \code{NULL}. If \code{FALSE} or \code{NULL} (default), then attempting to overwrite rasters or vectors already in a \code{GRASS} session will result in an error. You can allow overwriting by changing this to \code{TRUE}, or by setting \code{inRastName}, \code{inVectName}, and/or \code{outGrassName} to values different from rasters or vectors that already exist in the session (not all functions have these arguments). Use \code{\link{fasterLs}} to see existing rasters and vectors.  You can set this for all functions at once using \code{\link{fasterOptions}}.}

\item{inRastName}{Character, missing, or \code{NULL}. The name of the input raster in \code{GRASS}. If this is missing, the name of the raster or raster file will be used. If \code{NULL}, then "\code{inputRast}" will be used. This argument is useful if you are \link{chaining} \pkg{fasterRaster} functions together and wish to refer to the input raster later in the \code{GRASS} session.}

\item{outGrassName}{Character. Name of output in \code{GRASS}. This is useful if you want to refer to the output object in \code{GRASS} later in a session by \link{chaining} functions.}

\item{grassToR}{Logical or \code{NULL}. If \code{TRUE} or \code{NULL} (default) then the output will be returned to \code{R}. If \code{FALSE}, then the output is left in the \code{GRASS} session (with the name given by \code{outGrassName}). You can set this argument for all functions at once using \code{\link{fasterOptions}}.}

\item{grassDir}{Character or \code{NULL}. Name of the path in which \code{GRASS} is installed. For a Windows system, this might be something like \code{'C:/Program Files/GRASS GIS 8.3'}, for a Mac, like \code{"/Applications/GRASS-8.3.app/Contents/Resources"}, and for Linux like \code{'/usr/local/grass'}. If you are using a lot of \pkg{fasterRaster} functions, you can set this for all functions at once using \code{\link{fasterOptions}}. By default, this is \code{NULL}, in which case the install path for \code{GRASS} will be searched for, but if not found, will result in an error (usually the case).}

\item{inits}{Either \code{NULL} (default), or a named list of arguments to pass to \code{\link{initGrass}}. If \code{NULL}, then the current active \code{GRASS} session is used, or a new one in the "default" location is created.}
}
\value{
If \code{grassToR} if \code{TRUE}, then a raster with the same extent and coordinate reference system as \code{vect}. Regardless, a raster with the name given by \code{outGrassName} is written into the \code{GRASS} session.
}
\description{
This function replaces values in all \code{NA} cells with the distance between them and the closest non-\code{NA} cell. Alternatively, it fills in values of non-\code{NA} cells with the distance between them and the closest \code{NA} cell.
}
\examples{
\dontrun{

# change this according to where GRASS is installed on your system
# grassDir <- 'C:/Program Files/GRASS GIS 8.2' # example for a PC
# grassDir <- "/Applications/GRASS-8.2.app/Contents/Resources" # for a Mac

library(terra)
madForest2000 <- fasterData('madForest2000')

# could also use distance() function from the raster package which is
# slower in this example
distToForest <- fasterRastDistance(rast=madForest2000,
fillNAs=TRUE, grassDir=grassDir)

# terra
# distToForest <- distance(madForest2000)

par(mfrow=c(1, 2))
plot(madForest2000, 'Forest', col='forestgreen')
plot(distToForest, main='Distance to Forest (m)')

}
}
\seealso{
\code{\link[fasterRaster]{fasterVectToRastDistance}} in \pkg{fasterRaster}; \code{\link[terra]{distance}} in \pkg{terra}; \code{GRASS} module \href{https://grass.osgeo.org/grass82/manuals/r.grow.distance.html}{\code{r.grow.distance}}
}
