% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeRaster.r
\name{writeRasterFaster8}
\alias{writeRasterFaster8}
\alias{writeRaster8}
\title{Save a raster exported from GRASS to disk}
\usage{
writeRaster8(x, filename, overwrite = FALSE, datatype = "FLT8S", ...)
}
\description{
Save a raster exported from GRASS to disk
}
