% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterRasterize.r
\name{fasterRasterize}
\alias{fasterRasterize}
\title{Convert a spatial vector to a raster}
\usage{
fasterRasterize(
  vect,
  rast,
  use = "val",
  value = 1,
  field = NULL,
  burn = NULL,
  replace = FALSE,
  inVectName = NULL,
  inRastName = NULL,
  outGrassName = "vectToRast",
  grassToR = TRUE,
  grassDir = options()$grassDir,
  inits = NULL,
  ...
)
}
\arguments{
\item{vect}{A \code{SpatVector}, an \code{sf} object, or the name of a spatial vector in the active \code{GRASS} session.}

\item{rast}{A \code{SpatRaster} or the name of a raster already in an existing \code{GRASS} session.}

\item{use}{Character, indicates the types of values to be "burned" to the raster. Options include
\itemize{
\item \code{value} (default): a user-define value given by \code{value}
\item \code{field}: values directly from a field in \code{vect} named by \code{field}
\item \code{category}: values according to which polygon is covered by a cell named in \code{field}
\item \code{z}: z-coordinate (points or contours only)
\item \code{direction}: flow direction (lines only)
}}

\item{value}{Numeric, value to burn to each cell overlapped by the spatial object in \code{vect}.}

\item{field}{Name of column in \code{vect} with values or category labbels to which to burn to the raster.}

\item{burn}{\code{NULL} or any of \code{'point'}, \code{'line'}, \code{'area'}, \code{'boundary'}, or \code{'centroid'}. This determines the manner in which the vector data is "burned" to the raster. If \code{NULL} (default), then points objects will rasterized as points, lines objects as lines, and polygons objects as areas. See \code{GRASS} module \href{https://grass.osgeo.org/grass82/manuals/v.to.rast.html}{v.to.rast} for more details.}

\item{replace}{Logical or \code{NULL}. If \code{FALSE} or \code{NULL} (default), then attempting to overwrite rasters or vectors already in a \code{GRASS} session will result in an error. You can allow overwriting by changing this to \code{TRUE}, or by setting \code{inRastName}, \code{inVectName}, and/or \code{outGrassName} to values different from rasters or vectors that already exist in the session (not all functions have these arguments). Use \code{\link{fasterLs}} to see existing rasters and vectors.  You can set this for all functions at once using \code{\link{fasterOptions}}.}

\item{inVectName}{Character. The name of the input vector in \code{GRASS}. If missing or \code{NULL}, \"code{inputVect}" will be used. This is useful if you are \link{chaining} \pkg{fasterRaster} functions together and wish to refer to the vector later in the \code{GRASS} session.}

\item{inRastName}{Character, missing, or \code{NULL}. The name of the input raster in \code{GRASS}. If this is missing, the name of the raster or raster file will be used. If \code{NULL}, then "\code{inputRast}" will be used. This argument is useful if you are \link{chaining} \pkg{fasterRaster} functions together and wish to refer to the input raster later in the \code{GRASS} session.}

\item{outGrassName}{Character. Name of output in \code{GRASS}. This is useful if you want to refer to the output object in \code{GRASS} later in a session by \link{chaining} functions.}

\item{grassToR}{Logical or \code{NULL}. If \code{TRUE} or \code{NULL} (default) then the output will be returned to \code{R}. If \code{FALSE}, then the output is left in the \code{GRASS} session (with the name given by \code{outGrassName}). You can set this argument for all functions at once using \code{\link{fasterOptions}}.}

\item{grassDir}{Character or \code{NULL}. Name of the path in which \code{GRASS} is installed. For a Windows system, this might be something like \code{'C:/Program Files/GRASS GIS 8.3'}, for a Mac, like \code{"/Applications/GRASS-8.3.app/Contents/Resources"}, and for Linux like \code{'/usr/local/grass'}. If you are using a lot of \pkg{fasterRaster} functions, you can set this for all functions at once using \code{\link{fasterOptions}}. By default, this is \code{NULL}, in which case the install path for \code{GRASS} will be searched for, but if not found, will result in an error (usually the case).}

\item{inits}{Either \code{NULL} (default), or a named list of arguments to pass to \code{\link{initGrass}}. If \code{NULL}, then the current active \code{GRASS} session is used, or a new one in the "default" location is created.}

\item{...}{Arguments to send to \code{GRASS} module \href{https://grass.osgeo.org/grass82/manuals/v.to.rast.html}{\code{v.to.rast}}.}
}
\value{
If \code{grassToR} if \code{TRUE}, then a raster with the same extent, resolution, and coordinate reference system as \code{rast}. Regardless, a raster with the name given by \code{outRastName} is written into the \code{GRASS} session.
}
\description{
This function converts a spatial points, lines, or polygon vector into a raster based on a "template" raster. All cells covered by the vector can either have values taken from the vector or be set to a user-defined value.
}
\examples{
\dontrun{
# change this according to where GRASS is installed on your system
# grassDir <- 'C:/Program Files/GRASS GIS 8.2' # example for a PC
# grassDir <- "/Applications/GRASS-8.2.app/Contents/Resources" # for a Mac

library(sf)
library(terra)

madCoast <- fasterData('madCoast0')
madForest2000 <- fasterData('madForest2000')

# could also use rasterize() or mask() from the raster package which may
# be faster in this example
madMask <- fasterRasterize(vect=madCoast0, rast=madForest2000, grassDir=grassDir)

# terra
# madMask <- rasterize(madCoast0, madForest2000)
# madMask <- mask(madForest2000, madCoast0)

plot(madMask, main='Portion of Eastern Madagascar')
plot(madCoast0, add=TRUE)

}
}
\seealso{
\code{\link[terra]{rasterize}} in \pkg{terra}; \href{https://grass.osgeo.org/grass82/manuals/v.to.rast.html}{\code{v.to.rast}} in \code{GRASS}
}
