% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterRes.r
\name{fasterRes}
\alias{fasterRes}
\title{Information on a raster or vector in a GRASS session}
\usage{
fasterRes(rast, names = TRUE)
}
\arguments{
\item{names}{If \code{TRUE} (default), return a named vector.}

\item{x}{The name of a raster in the active \code{GRASS} session. This must be in quotes.}
}
\value{
A numeric vector with cell dimensions in the east-west and north-south directions.
}
\description{
Displays information on a raster or vector already in an R session. For further
}
\examples{
\dontrun{

library(sf)
library(terra)

# start a GRASS session with a raster
madElev <- fasterData('madElev')
initGrass(madElev, inRastName = 'madElev')

# import a vector
madRivers <- fasterData('madRivers')
fasterVect(madRivers, inVectName = 'madRivers')

# what's in the GRASS session?
fasterLs('rasters') # what rasters?
fasterLs('vectors') # what vectors?
fasterLs() # what rasters and vectors?

# information on the raster and vector
fasterInfo('madElev')
fasterInfo('madCoast0')
fasterInfo() # everything

# information on the coordinate reference system of the GRASS session
fasterCRS() # "plain" format
fasterCRS(TRUE) # nice for the eyes

# spatial extent of the GRASS session
fasterExt('madElev') # extent of madElev raster
fasterExt('madRivers') # extent of madRivers vector
fasterExt() # maximum extent of all spatial objects

}
}
\seealso{
\code{\link{fasterLs}}, \code{\link{fasterExt}}, \code{\link{fasterDim}}, and \code{\link{fasterRes}} in \pkg{fasterRaster}; \href{https://grass.osgeo.org/grass82/manuals/r.info.html}{\code{r.info}} and \href{https://grass.osgeo.org/grass82/manuals/v.info.html}{\code{v.info}} in \code{GRASS}
}
