% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterResampleRast.r
\name{fasterResampleRast}
\alias{fasterResampleRast}
\title{Resample a raster (change spatial resolution)}
\usage{
fasterResampleRast(
  rast,
  template,
  method = if (catRast) {
     "nearest"
 } else {
     "bilinear"
 },
  catRast = FALSE,
  replace = FALSE,
  inRastName = NULL,
  grassToR = TRUE,
  outGrassName = "croppedRast",
  grassDir = options()$grassDir,
  inits = NULL
)
}
\arguments{
\item{rast}{A \code{SpatRaster} or the name of a raster already in an existing \code{GRASS} session.}

\item{template}{Any of:
\itemize{
    \item A \code{SpatRaster}, used to define the new resolution. This object is \emph{not} exported to the \code{GRASS} session.
    \item The name of raster already in the \code{GRASS} session.
    \item A number indicating the new spatial resolution (cell size in east-west and north-south directions).
    \item Two numbers indicating the new spatial resolution in east-west and north-south directions, respectively.
}}

\item{method}{Method used for interpolation:
\itemize{
\item \code{'bilinear'} (default if \code{catRast} if \code{FALSE}): Uses values from 4 cells
\item \code{'bicubic'}: Uses values from 16 cells
\item \code{'lanczos'}: Uses values from 25 cells
\item \code{'nearest'} (default if \code{catRast} is \code{TRUE}): Uses value from 1 cell
}}

\item{catRast}{If \code{FALSE} (default), then the values in \code{rast} are assumed to be numeric and continuous. If code{FALSE}, then the unique values in \code{rast} are assumed to be categories.}

\item{replace}{Logical or \code{NULL}. If \code{FALSE} or \code{NULL} (default), then attempting to overwrite rasters or vectors already in a \code{GRASS} session will result in an error. You can allow overwriting by changing this to \code{TRUE}, or by setting \code{inRastName}, \code{inVectName}, and/or \code{outGrassName} to values different from rasters or vectors that already exist in the session (not all functions have these arguments). Use \code{\link{fasterLs}} to see existing rasters and vectors.  You can set this for all functions at once using \code{\link{fasterOptions}}.}

\item{inRastName}{Character, missing, or \code{NULL}. The name of the input raster in \code{GRASS}. If this is missing, the name of the raster or raster file will be used. If \code{NULL}, then "\code{inputRast}" will be used. This argument is useful if you are \link{chaining} \pkg{fasterRaster} functions together and wish to refer to the input raster later in the \code{GRASS} session.}

\item{grassToR}{Logical or \code{NULL}. If \code{TRUE} or \code{NULL} (default) then the output will be returned to \code{R}. If \code{FALSE}, then the output is left in the \code{GRASS} session (with the name given by \code{outGrassName}). You can set this argument for all functions at once using \code{\link{fasterOptions}}.}

\item{grassDir}{Character or \code{NULL}. Name of the path in which \code{GRASS} is installed. For a Windows system, this might be something like \code{'C:/Program Files/GRASS GIS 8.3'}, for a Mac, like \code{"/Applications/GRASS-8.3.app/Contents/Resources"}, and for Linux like \code{'/usr/local/grass'}. If you are using a lot of \pkg{fasterRaster} functions, you can set this for all functions at once using \code{\link{fasterOptions}}. By default, this is \code{NULL}, in which case the install path for \code{GRASS} will be searched for, but if not found, will result in an error (usually the case).}

\item{inits}{Either \code{NULL} (default), or a named list of arguments to pass to \code{\link{initGrass}}. If \code{NULL}, then the current active \code{GRASS} session is used, or a new one in the "default" location is created.}
}
\value{
A \code{SpatRaster}. Also creates a raster in a new grass session named \code{outGrassName}.
}
\description{
Resample a raster to a different spatial resolution.
}
\examples{
man/examples/ex_fasterCropExtendRast.r

}
\seealso{
\code{\link{fasterRes}} and \code{\link{fasterAggregate}} in \pkg{fasterRaster}; \code{\link[terra]{resample}} and  \code{\link[terra]{aggregate}} in package \pkg{terra}; \code{GRASS} modules \href{https://grass.osgeo.org/grass82/manuals/r.resample.interp.html}{\code{r.resample.interp}} and \href{https://grass.osgeo.org/grass82/manuals/r.resample.html}{\code{r.resample}}
}
