% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterRm.r
\name{fasterRm}
\alias{fasterRm}
\title{Remove rasters/vectors from a GRASS session}
\usage{
fasterRm(x, pattern = FALSE, rasters = TRUE, vectors = TRUE)
}
\arguments{
\item{x}{Name(s) of the rasters and/or vectors to remove. If this is \code{'*'}, then \emph{everything} will be removed.}

\item{pattern}{If \code{TRUE}, then \code{x} is used as a pattern so that any rasters or vectors matching this pattern will be removed. For example, if an existing \code{GRASS} session has rasters named \code{bio1}, \code{bio2}, and \code{bio3}, plus a vector named \code{bioVector}, then this command will remove them all: \code{fasterRm('bio', pattern = TRUE)}. By default this is \code{FALSE}, so the value(s) in \code{x} must be exact.}

\item{rasters, vectors}{If \code{TRUE}, remove only files of this type.}
}
\value{
Invisibly returns the number of rasters and/or vectors removed. More notably, this function also removes rasters and/or vectors from a \code{GRASS} session.
}
\description{
Remove rasters or vectors from a \code{GRASS} session. This will remove them from \code{GRASS}, but not from R if they exists in R.
}
\examples{
\dontrun{

library(sf)
library(terra)

# start a GRASS session with a raster and get information on it
madElev <- fasterData('madElev')
initGrass(madElev, inRastName = 'madElev')
fasterInfoRast('madElev')

# export a vector and get information on it
madCoast0 <- fasterData('madCoast0')
exportVectToGrass(madCoast0, inVectName = 'madCoast0')
fasterInfoVect('madCoast0')

# information on the coordinate rafeternec system
fasterCRS()

# names of all rasters and vectors in the GRASS session
fasterLs()

# remove raster
fasterRm('madElev')
fasterLs()

# remove vector
fasterRm('madCoast0')
fasterLs()

}

}
\seealso{
\code{\link{fasterLs}}, \href{https://grass.osgeo.org/grass82/manuals/g.remove.html}{\code{g.remove}} in \code{GRASS}
}
