% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterSun.r
\name{fasterSun}
\alias{fasterSun}
\title{Rasters of solar radiance and irradiation}
\usage{
fasterSun(
  elevation,
  coeff_bh,
  coeff_dh,
  horizonHeight = "horizonHeight",
  horizon_step = 90,
  aspect = 180,
  slope = 0,
  albedo = 0.2,
  linke = 3,
  day = 1,
  step = 0.5,
  declination = NULL,
  solar_constant = 1367,
  nprocs = 1,
  distance_step = 1,
  npartitions = 1,
  beam_rad = TRUE,
  diff_rad = TRUE,
  refl_rad = TRUE,
  glob_rad = TRUE,
  insol_time = TRUE,
  grassDir = options()$grassDir,
  grassToR = TRUE,
  ...
)
}
\arguments{
\item{elevation}{Either a raster or the name of a raster in an existing \code{GRASS} session with values representing elevation (typically in meters).}

\item{coeff_bh}{Numeric, character, or a \code{SpatRaster} layer. This is either a raster with values of the real-sky beam radiation coefficient, the name of such a raster in an existing \code{GRASS} session, or a numeric value. Valid values are in the range [0, 1].}

\item{coeff_dh}{Numeric, character, or \code{SpatRaster} layer. This is either a raster with values of the real-sky diffuse radiation coefficient, the name of such a raster in an existing \code{GRASS} session, or a numeric value. Valid values are in the range [0, 1].}

\item{horizonHeight}{Character or stack of \code{SpatRaster} layers. This is either a raster stack/brick, with values in each raster representing the height of the horizon in each direction, or the "basename" of such a raster stack in an existing \code{GRASS} session. If the latter, it is assumed that the raster stack is in an existing \code{GRASS} session and their names are as "horizonHeight_000", "horizonHeight_005", "horizonHeight_010", etc., where the "_xxx" string at the end represents the directional angle in which height is assessed (such that 0 degrees is east, then going counterclockwise so 90 is north, 180 west, and 270 south). It is assume that the first raster is for 0 degrees (east). Rasters with horizon height can be generated using \code{\link{fasterHorizon}}. Values of horizon height are in units of degrees. Note that regardless of the input format, \code{horizon_step} must be specified.}

\item{horizon_step}{Numeric >0. Difference between angular steps in which horizon height is measured in the \code{horizonHeight} rasters. For example, if the \code{horizon_basename} rasters were named "horizonHeight_000", "horizonHeight_005", "horizonHeight_010", etc. then the step size would be 5 degrees between each raster.}

\item{aspect}{Numeric, character, or a \code{SpatRaster} layer. This is either a raster with values of terrain aspect, the name of such a raster in an existing \code{GRASS} session, or a numeric value. Units are in degrees such that a value of 0 is assumed to point east, 90 north, 180 west, and 270 south. This raster can be created using \code{\link{fasterTerrain}}. If aspect is give in degrees such that 0 is north, 90 east, and so on, then it can be converted to the "east = 0" orientation using \code{\link{fasterConvertDegree}}.}

\item{slope}{Numeric, character, or a \code{SpatRaster} layer. This is either a raster with values of terrain slope, the name of such a raster in an existing \code{GRASS} session, or a numeric value. Units are in degrees. It can be created using \code{\link{fasterTerrain}}.}

\item{albedo}{Numeric, character, or a \code{SpatRaster} layer. This is either a raster with values of ground albedo, the name of such a raster in an existing \code{GRASS} session, or a numeric value. Albedo is unit-less. The default value is 0.2.}

\item{linke}{Numeric, character, or a \code{SpatRaster} layer. This is either a raster with values of the Linke atmospheric turbidity coefficient, the name of such a raster in an existing \code{GRASS} session, or a numeric value. The Linke coefficient is unit-less. The default value is 3.}

\item{day}{Positive integer in the range [1, 365]. Day of year for which to calculate ir/radiation. Default is 1 (January 1st).}

\item{step}{Positive integer in the range (0, 24]. Time step in hours for all-day radiation sums. Decimal values are OK.}

\item{declination}{Numeric. Declination value. If \code{NULL} (default), this is calculated automatically. Otherwise, this overrides the internal calculation.}

\item{solar_constant}{Positive numeric. The solar constant (solar energy hitting the top of the atmosphere). Default is 1367. Units are W / m^2.}

\item{nprocs}{Positive numeric in the range [1, 1000]. Number of processor cores to use. Default is 1.}

\item{distance_step}{Positive numeric in the range [0.5, 1.5]. Sampling distance coefficient. Default is 1.}

\item{npartitions}{Positive numeric. Number of chunks in which to read input files. Default is 1.}

\item{beam_rad}{Logical. If \code{TRUE} (default), generate a raster with beam irradiation with units of Wh * m^2 / day ("mode 2" of the \code{r.sun} \code{GRASS} module).}

\item{refl_rad}{Logical. If \code{TRUE} (default), generate a raster with ground-reflected irradiation with units of Wh * m^2 / day ("mode 2" of the \code{r.sun} \code{GRASS} module).}

\item{glob_rad}{Logical. If \code{TRUE} (default), generate a raster with total irradiance/irradiation with units of Wh * m^2 / day ("mode 2" of the \code{r.sun} \code{GRASS} module).}

\item{insol_time}{Logical. If \code{TRUE} (default), generate a raster with total insolation time in hours ("mode 2" of the \code{r.sun} \code{GRASS} module).}

\item{grassDir}{Name of the folder path in which \code{GRASS} is installed. For a Windows system, this might be something like \code{'C:/Program Files/GRASS GIS 8.2'}. For a Mac system, this might be like \code{"/Applications/GRASS-8.2.app/Contents/Resources"}. If you are using a lot of \pkg{fasterRaster} functions, it can be easier to set this once using \code{\link{fasterOptions}}. By default, this is \code{NULL}, in which case the install path for \code{GRASS} will be searched for, but if not found, will result in an error.}

\item{grassToR}{If \code{TRUE} (default) then the output will be returned to \pkg{R}. If \code{FALSE}, then the output is left in the \code{GRASS} session and named the value in \code{outGrassName}. You can then refer to the output in subsequent \pkg{fasterRaster} functions and avoid the need to re-import it. This is called "chaining" functions together, and it can save a lot of computational time.}

\item{time}{Positive numeric. Local solar time. Used for "mode 1" in the \code{r.sun} \code{GRASS} module. Default is \code{NULL} (i.e., use "mode 2").}
}
\value{
If \code{grassToR} if \code{TRUE}, then a raster or raster stack stack with the same extent, resolution, and coordinate reference system as \code{elevation}. Otherwise, a raster or a set of rasters is written into an existing \code{GRASS} session. The names of these rasters are (assuming they are generated): \code{beam_rad}, \code{diff_rad}, \code{refl_rad}, \code{glob_rad}, and/or \code{insol_time}.
}
\description{
This function calculates beam (direct), diffuse and ground reflected solar irradiation for a given day and set of topographic and atmospheric conditions. It utilizes the \code{GRASS} function \code{r.sun}. This function only works in "mode 2" of \code{r.run}.
}
\details{
See the documentation for the \code{GRASS} module \code{r.sun}{https://grass.osgeo.org/grass82/manuals/r.sun.html}.
}
\examples{
man/examples/ex_fasterSun.r

}
\seealso{
\code{\link{fasterHorizon}}, \code{\link{fasterTerrain}}
}
