% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterTerrain.r
\name{fasterTerrain}
\alias{fasterTerrain}
\title{Rasters of slope, aspect, curvature, and partial slopes}
\usage{
fasterTerrain(
  rast,
  slope = TRUE,
  slopeUnits = "degrees",
  aspect = FALSE,
  northIs0 = TRUE,
  profileCurve = FALSE,
  tanCurve = FALSE,
  eastWestSlope = FALSE,
  northSouthSlope = FALSE,
  grassDir = options()$grassDir,
  grassToR = TRUE,
  outGrassNameSlope = "slope",
  outGrassNameAspect = ifelse(northIs0, "aspectNorthIs0", "aspectEastIs0"),
  outGrassNameProfileCurve = "profileCurve",
  outGrassNameTanCurve = "tanCurve",
  outGrassNameEastWestSlope = "eastWestSlope",
  outGrassNameNorthSouthSlope = "northSouthSlope",
  ...
)
}
\arguments{
\item{rast}{A \code{SpatRaster} or the name of a raster already in an existing \code{GRASS} session.}

\item{slope}{Logical, if \code{TRUE} (default) then calculate slope.}

\item{slopeUnits}{Character, "units" in which to calculate slope: either \code{degrees} for degrees or \code{percent}.}

\item{aspect}{Logical, if \code{TRUE} then calculate aspect. Aspect is given in degrees from North going clockwise (0 = north, 90 = east, 180 = south, 270 = west).}

\item{northIs0}{Logical. If \code{TRUE} (default), aspect will be reported such that 0 is north, and degrees run clockwise (90 is east, 180 south, 270 west). If \code{FALSE}, then aspect will be reported such that 0 is east, and degrees run counterclockwise (90 is north, 180 west, 270 south). The latter is the default in \code{GRASS}7.}

\item{profileCurve}{Logical, if \code{TRUE}, calculate profile curvature. Default is \code{FALSE}.}

\item{tanCurve}{Logical, if \code{TRUE}, calculate tangential curvature. Default is \code{FALSE}.}

\item{eastWestSlope}{Logical, if \code{TRUE}, calculate slope in east-west direction. Default is \code{FALSE}.}

\item{northSouthSlope}{Logical, if \code{TRUE}, calculate slope in north-south direction. Default is \code{FALSE}.}

\item{grassDir}{Name of the folder path in which \code{GRASS} is installed. For a Windows system, this might be something like \code{'C:/Program Files/GRASS GIS 8.2'}. For a Mac system, this might be like \code{"/Applications/GRASS-8.2.app/Contents/Resources"}. If you are using a lot of \pkg{fasterRaster} functions, it can be easier to set this once using \code{\link{fasterOptions}}. By default, this is \code{NULL}, in which case the install path for \code{GRASS} will be searched for, but if not found, will result in an error.}

\item{grassToR}{If \code{TRUE} (default) then the output will be returned to \pkg{R}. If \code{FALSE}, then the output is left in the \code{GRASS} session and named the value in \code{outGrassName}. You can then refer to the output in subsequent \pkg{fasterRaster} functions and avoid the need to re-import it. This is called "chaining" functions together, and it can save a lot of computational time.}

\item{outGrassNameSlope}{Character. Name of slope raster in \pkg{GRASS}. Useful for referring to later in the same \code{GRASS} session.}

\item{outGrassNameAspect}{Character. Name of aspect raster in \pkg{GRASS}. Useful for referring to later in the same \code{GRASS} session.}

\item{outGrassNameProfileCurve}{Character. Name of profile curve raster in \pkg{GRASS}. Useful for referring to later in the same \code{GRASS} session.}

\item{outGrassNameTanCurve}{Character. Name of tangent curve raster in \pkg{GRASS}. Useful for referring to later in the same \code{GRASS} session.}

\item{outGrassNameEastWestSlope}{Character. Name of east-west slope raster in \pkg{GRASS}. Useful for referring to later in the same \code{GRASS} session.}

\item{outGrassNameNorthSouthSlope}{Character. Name of north-south slope raster in \pkg{GRASS}. Useful for referring to later in the same \code{GRASS} session.}

\item{...}{Arguments to pass to \code{\link[rgrass]{execGRASS}} when used for rasterizing (i.e., function \code{r.slope.aspect} in \code{GRASS}).}
}
\value{
If \code{grassToR} if \code{TRUE}, then a raster or raster stack with the same extent, resolution, and coordinate reference system as \code{rast}. Regardless, raster(s) given by the names in the \code{outGrassName*} arguments are used are written into the \code{GRASS} session.
}
\description{
This function is a potentially faster version of the \code{\link[terra]{terrain}} function in the \pkg{terra} package for calculating slope and aspect of a raster. It can also calculate profile curvature, tangential curvature, and slope in the east-west or north-south directions.
}
\details{
See the documentation for the \code{GRASS} module \code{r.slope.aspect}{https://grass.osgeo.org/grass82/manuals/r.slope.aspect.html}.
}
\examples{
man/examples/ex_fasterTerrain.r

}
\seealso{
\code{\link[terra]{terrain}}
}
