% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterTrimRast.r
\name{fasterTrimRast}
\alias{fasterTrimRast}
\title{Remove all rows and columns of a raster that are entirely NA}
\usage{
fasterTrimRast(
  rast,
  replace = FALSE,
  inRastName = NULL,
  grassToR = TRUE,
  grassDir = options()$grassDir,
  inits = NULL
)
}
\arguments{
\item{rast}{A \code{SpatRaster} or the name of a raster already in an existing \code{GRASS} session.}

\item{replace}{Logical or \code{NULL}. If \code{FALSE} or \code{NULL} (default), then attempting to overwrite rasters or vectors already in a \code{GRASS} session will result in an error. You can allow overwriting by changing this to \code{TRUE}, or by setting \code{inRastName}, \code{inVectName}, and/or \code{outGrassName} to values different from rasters or vectors that already exist in the session (not all functions have these arguments). Use \code{\link{fasterLs}} to see existing rasters and vectors.  You can set this for all functions at once using \code{\link{fasterOptions}}.}

\item{inRastName}{Character, missing, or \code{NULL}. The name of the input raster in \code{GRASS}. If this is missing, the name of the raster or raster file will be used. If \code{NULL}, then "\code{inputRast}" will be used. This argument is useful if you are \link{chaining} \pkg{fasterRaster} functions together and wish to refer to the input raster later in the \code{GRASS} session.}

\item{grassToR}{Logical or \code{NULL}. If \code{TRUE} or \code{NULL} (default) then the output will be returned to \code{R}. If \code{FALSE}, then the output is left in the \code{GRASS} session (with the name given by \code{outGrassName}). You can set this argument for all functions at once using \code{\link{fasterOptions}}.}

\item{grassDir}{Character or \code{NULL}. Name of the path in which \code{GRASS} is installed. For a Windows system, this might be something like \code{'C:/Program Files/GRASS GIS 8.3'}, for a Mac, like \code{"/Applications/GRASS-8.3.app/Contents/Resources"}, and for Linux like \code{'/usr/local/grass'}. If you are using a lot of \pkg{fasterRaster} functions, you can set this for all functions at once using \code{\link{fasterOptions}}. By default, this is \code{NULL}, in which case the install path for \code{GRASS} will be searched for, but if not found, will result in an error (usually the case).}

\item{inits}{Either \code{NULL} (default), or a named list of arguments to pass to \code{\link{initGrass}}. If \code{NULL}, then the current active \code{GRASS} session is used, or a new one in the "default" location is created.}

\item{warn}{If \code{TRUE} (default), display a warning when using this function about cropping any other rasters in the \code{GRASS} session.}
}
\value{
A \code{SpatRaster}. Also exports a raster to a new or existing \code{GRASS} session.
}
\description{
This function removes all rows and columns from a raster that are entirely \code{NA} values.\cr
IMPORTANT: This function will also crop all rasters in an existing, active \code{GRASS} session to the extent of the trimmed raster.
}
\examples{
man/examples/ex_fasterResizeRast.r

}
\seealso{
\code{\link{fasterExtendRast}} and \code{\link{fasterCropExtendRast}} in \pkg{fasterRaster}; \code{\link[terra]{crop}} and \code{\link[terra]{extend}} in \pkg{terra}; \href{https://grass.osgeo.org/grass82/manuals/r.clip.html}{\code{r.clip}} in \code{GRASS}
}
