% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterVect.r
\name{fasterVect}
\alias{fasterVect}
\title{Export a spatial vector to an active GRASS session}
\usage{
fasterVect(vect, inVectName = NULL, replace = FALSE, autoRegion = TRUE, ...)
}
\arguments{
\item{vect}{A \code{SpatVector} (\pkg{terra} package) or \code{sf} (\pkg{sf} package) spatial vector object.}

\item{inVectName}{Name of vector in \code{GRASS}. If \code{NULL} (default), the name will be set to \code{'inputVect'}.}

\item{replace}{Logical or \code{NULL}. If \code{FALSE} or \code{NULL} (default), then attempting to overwrite rasters or vectors already in a \code{GRASS} session will result in an error. You can allow overwriting by changing this to \code{TRUE}, or by setting \code{inRastName}, \code{inVectName}, and/or \code{outGrassName} to values different from rasters or vectors that already exist in the session (not all functions have these arguments). Use \code{\link{fasterLs}} to see existing rasters and vectors.  You can set this for all functions at once using \code{\link{fasterOptions}}.}

\item{autoRegion}{Logical or \code{NULL}. If \code{TRUE} or \code{NULL} (default), then during execution of this function, the \code{GRASS} \link{region} will be resized to encompass the raster/vector/area of interest, and if relevant, resampled to match its spatial resolution. If \code{FALSE}, the the current region's extent and resolution will be used. Most users will want to keep this \code{TRUE} or \code{NULL} (same as \code{TRUE}). If you change this to \code{FALSE}, you will need to keep track of the \code{GRASS} \link{region} properties. This argument can be set for all functions at once using \code{\link{fasterOptions}}.}

\item{...}{Additional arguments (unused).}
}
\value{
\code{TRUE} (invisibly, if the vector was successfully exported). The function also exports a vector to a \code{GRASS} session so it can be used by other functions.
}
\description{
Export a spatial vector to an active \code{GRASS} session. To import a vector from \code{GRASS} to \pkg{R}, use \code{\link[rgrass]{read_VECT}}. Vectors that are larger than the active \code{GRASS} region are not cropped to the region (unlike when exporting rasters).
}
\examples{
man/examples/ex_fasterRast_fasterVect.R

}
\seealso{
\code{\link[rgrass]{write_VECT}} and \code{\link[rgrass]{read_VECT}} in package \pkg{rgrass}
}
