% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterVectorize.r
\name{fasterVectorize}
\alias{fasterVectorize}
\title{Convert a raster to a vector (points, lines, or polygons)}
\usage{
fasterVectorize(
  rast,
  vectType,
  agg = TRUE,
  smooth = FALSE,
  grassDir = options()$grassDir,
  grassToR = TRUE,
  inRastName = "rast",
  outGrassName = "rastToVect",
  ...
)
}
\arguments{
\item{rast}{A \code{SpatRaster} or the name of a raster already in an existing \code{GRASS} session.}

\item{vectType}{Character. Indicates type of output: \code{point}, \code{line} (not supported yet), or \code{area}.}

\item{agg}{Logical. If \code{TRUE} (default) then union all points/lines/polygons with the same value into the same "multipart" polygon. This may or may not be desirable. For example, if the raster is vectorized into a polygons object each cell will become a separate polygon. Using this option will merge cells with the same value (even if they are not spatially adjacent one another).}

\item{smooth}{Logical. If \code{TRUE} then "round" cell corners by connecting the midpoints of corner cells (which leaves out the corner-most triangle of that cell). This option only applies if \code{vectType} is \code{area}. Default is \code{FALSE}.}

\item{grassDir}{Name of the folder path in which \code{GRASS} is installed. For a Windows system, this might be something like \code{'C:/Program Files/GRASS GIS 8.2'}. For a Mac system, this might be like \code{"/Applications/GRASS-8.2.app/Contents/Resources"}. If you are using a lot of \pkg{fasterRaster} functions, it can be easier to set this once using \code{\link{fasterOptions}}. By default, this is \code{NULL}, in which case the install path for \code{GRASS} will be searched for, but if not found, will result in an error.}

\item{grassToR}{If \code{TRUE} (default) then the output will be returned to \pkg{R}. If \code{FALSE}, then the output is left in the \code{GRASS} session and named the value in \code{outGrassName}. You can then refer to the output in subsequent \pkg{fasterRaster} functions and avoid the need to re-import it. This is called "chaining" functions together, and it can save a lot of computational time.}

\item{inRastName}{The name of the input raster in \code{GRASS}. This argument is useful if you are "chaining" \pkg{fasterRaster} functions together and wish to refer to the input raster later. The default value is the name of the raster (i.e., \code{names(rast)}).}

\item{outGrassName}{Character. Name of output in \code{GRASS}. This is useful if you want to refer to the output object in \code{GRASS} later in a session by "chaining" functions.}

\item{...}{Arguments to pass to \code{\link[rgrass]{execGRASS}} when used for converting a raster to a vector (i.e., function \code{r.to.vect} in \pkg{GRASS}).}

\item{calcDensity}{Logical, if \code{TRUE} then calculate density in the moving window. This will create a raster named \code{density} in the \code{GRASS} environment if \code{grassToR} is \code{FALSE} or return a raster named \code{density} if \code{grassToR} is \code{TRUE}. Default is \code{FALSE}.}

\item{calcConnect}{Logical. If \code{TRUE} then calculate a connectivity raster (conditional probability a cell with a value of 1 has a value that is also 1) in the moving window. This will create a raster named \code{connect} in the \code{GRASS} environment if \code{grassToR} is \code{FALSE} or return a raster named \code{connect} if \code{grassToR} is \code{TRUE}. Default is \code{FALSE}.}
}
\value{
If \code{grassToR} if \code{TRUE}, then a SpatialPointsDataFrame, SpatialLinesDataFrame, or a SpatialPolygonsDataFrame with the same coordinate reference system as \code{rast}. The field named \code{value} will have the raster values. Regardless, vector object with the name given by \code{outGrassName} will be written into the \code{GRASS} session.
}
\description{
Convert a raster to a spatial polygons object (points, lines, or polygons).
}
\examples{
\dontrun{

# change this according to where GRASS is installed on your system
# grassDir <- 'C:/Program Files/GRASS GIS 8.2' # example for a PC
# grassDir <- "/Applications/GRASS-8.2.app/Contents/Resources" # for a Mac

library(terra)

madForest2000 <- fasterData('madForest2000')

# could also use rasterToPolygons() which is
# probably faster in this example
forestPolyGrass <- fasterVectorize(rast=madForest2000,
	vectType='area', grassDir=grassDir)

forestPolyTerra <- rasterToPolygons(madForest2000, dissolve=TRUE)

par(mfrow=c(1, 3))
plot(madForest2000, main='Forest as Raster', col='forestgreen')
plot(forestPolyGrass, main='Forest as Polygon', col='forestgreen', border=NA)
plot(forestPolyTerra, main='Forest as Polygon', col='forestgreen', border=NA)

}
}
\seealso{
\code{\link[terra]{as.points}}, \code{\link[terra]{as.polygons}}, and \code{\link[terra]{as.lines}} in \pkg{terra}; \code{\link[fasterRaster]{fasterRasterize}} in \code{fasterRaster}; \href{https://grass.osgeo.org/grass82/manuals/r.to.vect.html}{\code{r.to.vect}} in \code{GRASS}
}
