% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterWriteVector.r
\name{fasterWriteVector}
\alias{fasterWriteVector}
\title{Save one or more vectors to disk directly from a GRASS session}
\usage{
fasterWriteVector(vect, filename, overwrite = FALSE, ...)
}
\arguments{
\item{vect}{The name(s) or one or more spatial vectors in the active \code{GRASS} session.}

\item{filename}{Path(s) and file name(s). If \code{vect} has more than one name, then you should state multiple file names, one per vector. The function will attempt to ascertain the type of the vector file from the filename extension, but you can also set it using \code{format} (see the \code{...} argument).}

\item{overwrite}{If \code{FALSE} (default), do not save over existing file(s).}

\item{...}{Arguments to send to \href{https://grass.osgeo.org/grass82/manuals/v.out.ogr.html}{v.out.ogr}. This \code{GRASS} modole includes a \code{format} argument, which explicitly states the file format. Some common formats include:
\itemize{
}
For the full list, please see \href{https://grass.osgeo.org/grass82/manuals/v.out.ogr.html}{v.out.ogr}. When using other formats, you may have to specify the \code{createopts} argument, too. \cr}
}
\value{
Nothing. Writes one or more files to disk.
}
\description{
This function saves a spatial vector to disk directly from a \code{GRASS} session. If you want to save a vector from \code{R}, use \code{\link{save}} or \code{\link{saveRDS}}.
}
\seealso{
\code{\link{fasterWriteRaster}} in \pkg{fasterRaster}; \code{\link[terra]{writeRaster}} in package \pkg{terra}; \code{GRASS} module \href{https://grass.osgeo.org/grass82/manuals/v.out.ogr.html}{\code{v.out.ogr}}
}
