% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastRestore.r
\name{fastRestore}
\alias{fastRestore}
\title{Revert to a previous 'GRASS' session (working folder, location and mapset)}
\usage{
fastRestore(...)
}
\arguments{
\item{...}{Either a sequence of arguments with the pattern \verb{<argument = value>}, or a list of arguments. These arguments can be any option available in \code{\link[=setFastOptions]{setFastOptions()}}. Of particular note, the \code{workDir}, \code{location}, and \code{mapset} options allow one to revert to a previous \strong{GRASS} session. The current session's \code{workDir}, \code{location}, and \code{mapset} can be seen using \code{\link[=getFastOptions]{getFastOptions()}}.}
}
\value{
An object of class \code{gmeta} (invisibly) if successful. An error will likely result if not.
}
\description{
This function is useful for reverting to a previous \strong{GRASS} session (\code{workDir}, \code{location}, and/or \code{mapset}). The session must have been already initiated using \code{\link[=fastStart]{fastStart()}} in the current \strong{R} session or a previous one.
}
\seealso{
\strong{GRASS} \link[=https://grass.osgeo.org/grass82/manuals/grass_database.html]{locations and mapsets}
}
