% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastStart.r
\name{fastStart}
\alias{fastStart}
\title{Initialize a 'GRASS' session}
\usage{
fastStart(
  crs,
  grassDir = NULL,
  workDir = NULL,
  overwrite = FALSE,
  warn = TRUE,
  ...
)
}
\arguments{
\item{crs}{Any object from which a coordinate reference system (CRS) can be acquired. Ergo, any of:
\itemize{
\item A \code{SpatRaster}, \code{SpatVector}, \code{SpatExtent}, \code{stars} or \code{sf} object
\item A CRS (coordinate reference system) string
}}

\item{grassDir}{Character or \code{NULL} (default): Folder in which \strong{GRASS} is installed on your computer. This will look different depending on the operating system and verison of \strong{GRASS} you have installed. Here are some examples:
\itemize{
\item Windows: \code{'C:/Program Files/GRASS GIS 8.3'}
\item Mac OS: \code{"/Applications/GRASS-8.3.app/Contents/Resources"}
\item Linux: \code{'/usr/local/grass'}
If \code{NULL}, then the function will use \code{\link[=getFastOption]{getFastOption()}} to attempt to get it. If it fails, \code{grassDir} will stay as \code{NULL} and likely reseult in an error.
}}

\item{workDir}{\code{NULL} or character: The name of the folder in which \strong{GRASS} will store rasters and vectors. If this is \code{NULL} (default), then the \code{\link[=tempdir]{tempdir()}} on the user's system will be used. If users wish to create persistent \strong{GRASS} sesssions that can be used in a different instance of \strong{R} (i.e., if \strong{R} is stopped then restarted), then this needs to be specified.}

\item{overwrite}{Logical: If \code{FALSE} (default), and a \strong{GRASS} session in the stated (or default) location and mapset has already been started, then the function will fail. If \code{TRUE}, then any existing \strong{GRASS} session will be overwritten. \emph{NOTE}: This will \strong{not} remove any \strong{R} objects associated with rasters or vectors in the session, but they will no longer work because the objects they point to will be overwritten.}

\item{warn}{Logical: If \code{TRUE} (default) and \code{overwrite} is \code{TRUE}, then display a warning.}

\item{...}{Options to send to \code{\link[=setFastOptions]{setFastOptions()}}. These should be in \code{option = value} format.}
}
\value{
An \code{\link[rgrass:gmeta]{rgrass::gmeta()}} object (a list) if successful (invisibly). If not successful, the function will usually either fail or return \code{FALSE} with a warning.
}
\description{
This function initializes a \strong{GRASS} session in a particulatr folder. You need to run this function before you use most functions in \strong{fasterRaster} (just once). You can also use \code{\link[=sessionRestore]{sessionRestore()}} to switch between working folders, \link[=sessions]{locations, and mapsets}.
}
\examples{
man/examples/ex_fastStart.r

}
\seealso{
Guide to getting \href{tutorial_starting}{started} with \strong{fasterRaster}.
}
