% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forwardSlash.r
\name{forwardSlash}
\alias{forwardSlash}
\title{Replace backslash with forward slash}
\usage{
forwardSlash(x)
}
\arguments{
\item{x}{A string.}
}
\value{
Character.
}
\description{
This function is helpful for Windows systems, where paths are usually expressed with left slashes, whereas \strong{R} requires right slashes.
}
\examples{

forwardSlash("C:\\\\ecology\\\\main project")

}
