% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grassInfo.r
\name{grassInfo}
\alias{grassInfo}
\title{GRASS citation, version, and copyright information}
\usage{
grassInfo(x = "citation")
}
\arguments{
\item{x}{Character: What to return. Any of:
\itemize{
\item \code{"citation"} (default)
\item \code{"copyright"}: Copyright information
\item \code{"version"}: Version number and release year
\item \code{"versionNumber"}: Version number as numeric, major and minor only (e.g., 8.4)
}

Partial matching is used and case is ignored.}
}
\value{
Character.
}
\description{
Report the \strong{GRASS} citation, version/release year, version number, or copyright information.
}
\examples{
if (grassStarted()) {

# Citation
grassInfo()

# Version number
grassInfo("version")

# Version number
grassInfo("versionNumber")

# Version number
grassInfo("versionNumber")

# Copyright
grassInfo("copyright")

}
}
