% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grassStarted.r
\name{grassStarted}
\alias{grassStarted}
\title{Has "GRASS" been started or not?}
\usage{
grassStarted()
}
\value{
Logical.
}
\description{
Returns \code{TRUE} or \code{FALSE}, depending on whether a \strong{GRASS} connection has been made or not within the current \strong{R} session. Usually used only by developers. \strong{GRASS} is started the first time \code{\link[=fast]{fast()}} is used.
}
\examples{

grassStarted()

}
