% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00d_GRaster_class.r
\docType{class}
\name{GRaster-class}
\alias{GRaster-class}
\alias{GRaster}
\title{Classes for "fasterRaster" locations, rasters, and vectors}
\description{
Classes for "fasterRaster" locations, rasters, and vectors
}
