% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_classes_and_show_methods_a_GSession.r,
%   R/00_classes_and_show_methods_b_GSpatial.r,
%   R/00_classes_and_show_methods_c_GRaster.r,
%   R/00_classes_and_show_methods_d_GVector.r
\docType{class}
\name{GSession-class}
\alias{GSession-class}
\alias{GSession}
\alias{GSpatial-class}
\alias{GSpatial}
\alias{GRaster-class}
\alias{GRaster}
\alias{GVector-class}
\alias{GVector}
\title{Classes for fasterRaster locations, rasters, and vectors}
\value{
An object of class \code{GSession}, \code{GSpatial}, \code{GRaster}, or \code{GVector}.
}
\description{
S4 classes that contain pointers to \strong{GRASS} objects. Most users will manipulate objects using these classes, but do not need to know the details.
}
\details{
\code{GRaster} and \code{GVector} are the main "working" classes for \strong{fasterRaster} functions--most user will not need to engage directly with the others. \code{GSpatial} is contained by \code{GRaster} and \code{GVector} and is the general spatial class. It may be extended to include \strong{GRASS} regions later. \code{GSession} is the location/mapset class and contained by the \code{GSpatial} class.
}
\section{Functions}{
\itemize{
\item \code{GSpatial}: 

\item \code{GRaster}: 

\item \code{GVector}: 

}}
\section{Slots}{

\describe{
\item{\code{location}}{Character: The \strong{GRASS} \link[=tutorial_sessions]{"location"} of the object. The default value is \code{default}. Can be obtained using \code{\link[=location]{location()}}.}

\item{\code{mapset}}{Character: The \strong{GRASS} \link[=tutorial_sessions]{"mapset"}. Default value is \code{PERMANENT}. Typically hidden to users. Can be obtained using \code{\link[=mapset]{mapset()}}.}

\item{\code{gnames}}{Character: Name of the object in \strong{GRASS}. These names are made on-the-fly and provide the pointer to the object from \strong{R} to \strong{GRASS}. Changing them will break the connection. Can be obtained using \code{\link[=gnames]{gnames()}}.}

\item{\code{names}}{Character  (\code{GRaster}s only): Name of a raster or each raster layer in. Can be obtained using \code{\link[=names]{names()}}.}

\item{\code{crs}}{Character: Coordinate reference systems string (preferably in WKT2 format). Can be obtained using \code{\link[=crs]{crs()}} or \code{\link[=st_crs]{st_crs()}}.}

\item{\code{dimensions}}{Dimensions:
\itemize{
\item \code{GRaster}s: Vector of four integers indicating number of rows, columns, depths (for 3D rasters), and layers of a "stack" of rasters. Can be obtained using \code{\link[=dim]{dim()}}, plus \code{\link[=nrow]{nrow()}}, \code{\link[=ncol]{ncol()}}, and \code{\link[=ndepth]{ndepth()}}.
\item \code{GVectors}s: Vector of two integers indicating number of geometries and number of fields. Can be obtained using \code{\link[=dim]{dim()}}, plus \code{\link[=nrow]{nrow()}} and \code{\link[=ncol]{ncol()}}.
}}

\item{\code{extent}}{Numeric vector with four values: Extent of the object listed in order from westernmost longitude, easternmost longitude, southernmost latitude, northernmost latitude. Can be obtained using \code{\link[=ext]{ext()}}.}

\item{\code{ztop,zbottom}}{Numeric: Top- and bottom-most extents of 3D \code{GRaster}s and \code{GVector}s. Can be obtained using \code{\link[=zExt]{zExt()}}.}

\item{\code{topology}}{Character: Valid values are \verb{2D} (2-dimensional--most rasters and vectors) or \verb{3D} (3-dimensional--e.g., as in LIDAR data). Can be obtained using \code{\link[=topology]{topology()}}.}

\item{\code{geometry}}{Character (\code{GVectors}s only): Either \code{points}, \code{lines}, or \code{polygons}. Can be obtained using \code{\link[=geomtype]{geomtype()}}.}

\item{\code{datatypeGRASS}}{Character (\code{GRaster}s only): Type of data stored in a raster, as interpreted by \code{GRASS}. This is either \code{CELL} (integers), \code{FCELL} (floating-point values), or \code{DCELL} (double-values). Can be obtained using \code{\link[=datatype]{datatype()}}.}

\item{\code{resolution}}{Vector of two numeric values (\code{GRaster}s only): Size of a raster cell in the east-west direction and in the north-south direction. Can be obtained using \code{\link[=res]{res()}} and \code{\link[=res3d]{res3d()}}.}

\item{\code{nCats}}{Integer (\code{GRaster}s only): Number of categories. Must be >0. Can be obtained using \code{\link[=ncat]{ncat()}}.}

\item{\code{minVal,maxVal}}{Numeric (\code{GRaster}s only): Minimum and maximum value across all cells. Can be obtained using \code{\link[=minmax]{minmax()}}.}

\item{\code{fields}}{Names of fields (\code{GVector}s only). Can be obtained using \code{\link[=names]{names()}}.}

\item{\code{numFields}}{Number of fields (\code{GVector}s only). Can be obtained using \code{\link[=ncol]{ncol()}}.}

\item{\code{fieldClasses}}{Classes of fields (\code{GVector}s only). Can be obtained using \code{\link[=datatype]{datatype()}}.}
}}

