% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_classes_and_show_methods_e_GVector.r
\docType{class}
\name{GMetaTable-class}
\alias{GMetaTable-class}
\alias{GMetaTable}
\alias{GEmptyMetaTable-class}
\alias{GEmptyMetaTable}
\alias{GFullMetaTable-class}
\alias{GFullMetaTable}
\alias{GVector-class}
\alias{GVector}
\title{Classes for "fasterRaster" locations, rasters, and vectors}
\description{
Classes for "fasterRaster" locations, rasters, and vectors
}
