% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00b_GSpatial_class.r
\docType{class}
\name{GSpatial-class}
\alias{GSpatial-class}
\alias{GSpatial}
\title{Classes for "fasterRaster" locations, rasters, and vectors}
\description{
Classes for "fasterRaster" locations, rasters, and vectors
}
