% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillshade.r
\name{hillshade,GRaster-method}
\alias{hillshade,GRaster-method}
\alias{shade}
\title{Hillshading}
\usage{
\S4method{hillshade}{GRaster}(x, angle = 45, direction = 0, zscale = 1)
}
\arguments{
\item{x}{A \code{GRaster} (typically representing elevation).}

\item{angle}{Numeric: The altitude of the sun above the horizon in degrees. Valid values are in the range \link{0, 90}, and the default value is 45 (half way from the horizon to overhead).}

\item{direction}{The direction (azimuth) in which the sun is shining in degrees. Valid values are in the range \link{0, 360}. The default is 0, meaning the sus is at due south (180 degrees) and shining due north (0 degrees). Note that in this function, 0 corresponds to north and 180 to south, but in the \strong{GRASS} module \code{r.relief}, "east orientation" is used (0 is east, 90 is north, etc.).}

\item{zscale}{Numeric: Value by which to exaggerate terrain. The default is 1.  Numbers greater than this will increase apparant relief, and lessthan this (even negative) will diminish it.}
}
\value{
A \code{GRaster}.
}
\description{
Hillshade rasters are often used for display purposes becayse they make topographical relief look "real" to teh eye.
}
\seealso{
\code{\link[terra:shade]{terra::shade()}}, module \code{r.relief} in \strong{GRASS}
}
