% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layerChiSq.r
\name{layerChiSq,GRaster-method}
\alias{layerChiSq,GRaster-method}
\alias{layerChiSq}
\title{Pearson's chi-squared for two or more integer of categorical rasters}
\usage{
\S4method{layerChiSq}{GRaster}(
  x,
  ...,
  correct = TRUE,
  simulate.p.value = FALSE,
  B = 2000,
  verbose = FALSE
)
}
\arguments{
\item{x}{A \code{GRaster} with one or more layers, each of which must be have cells that represent integers or categories (factors). If at least one cell in a pair of rasters has \code{NA}, then the cell will be ignored.}

\item{...}{Either missing or integer/categorical (factor) \code{GRaster}s.}

\item{correct}{Logical: If \code{TRUE} (default), then apply continuity correction when computing the test statistic for 2 by 2 tables: one half is subtracted from all |\emph{O - E}| differences; however, the correction will not be bigger than the differences themselves. No correction is performed if \code{simulate.p.value = TRUE}.}

\item{simulate.p.value}{Logical: If \code{TRUE}, then the \emph{p}-value will be estimated by Monte Carlo simulation, using \code{\link[stats:chisq.test]{stats::chisq.test()}}]. This is recommended when there are many categories, because the chi-squared distribution is not a good approximation of the distribution of the test statistic in this case. The default is \code{FALSE} because simulating \emph{p}-values can be time-consuming.}

\item{B}{Numeric or integer: Number of replicates for Monte Carlo simulation when \code{simulate.p.value} is \code{TRUE}. The default is 2000.}

\item{verbose}{Logical: If \code{TRUE}, then display progress. Default is \code{FALSE}.}
}
\value{
A list with two or three elements: \code{chisq} is a matrix of chi-squared values for each pairwise comparison; \code{df} is a matrix of degrees of freedom for each pairwise comparison (only included if \code{simulate.p.values = FALSE}); and \code{p.value} is a matrix of \emph{p}-values for each pairwise comparison.
}
\description{
This function takes two or more integer or categorical (factor) \code{GRaster}s and compares them using a Chi-squared test using \code{\link[stats:chisq.test]{stats::chisq.test()}}.
}
\examples{
if (grassStarted()) {

# Setup
library(terra)

# Example data
madChelsa <- fastData("madChelsa")

# Convert SpatRaster to GRaster:
chelsa <- fast(madChelsa)

# Force the values to be integers (required by layerChiSq()):
chelsaInteger <- as.int(chelsa)

# Chi-squared test
layerChiSq(chelsaInteger, verbose = TRUE)

}
}
\seealso{
\code{\link[=layerCor]{layerCor()}}
}
