% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls.r
\name{.ls}
\alias{.ls}
\title{List objects in the active GRASS session}
\usage{
.ls(type = c("rasters", "vectors", "rasters3d", "groups"))
}
\arguments{
\item{type}{The type of spatial objects to display. This can include \code{"rasters"} (all rasters), \code{"vectors"} (all spatial vectors), \code{"rasters3d"} (3D-rasters), and/or \code{"groups"} (groups). Partial matching is supported. If missing, all objects are displayed.}
}
\value{
Character vector of names of \strong{GRASS} objects.
}
\description{
Display the \code{sources} (see \code{sources()}) of all rasters and/or vectors that have been exported to or created in the active \strong{GRASS} session"s location and mapset.
}
\seealso{
\code{\link[=ls]{ls()}}
}
\keyword{internal}
