% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mad0.r
\docType{data}
\name{mad0}
\alias{mad0}
\title{Spatial data: national borders of Madagascar}
\format{An object of class \code{'SpatialPolygonsDataFrame'}. See \code{\link[sp]{SpatialPolygons}}.}
\source{
Database of Global Administrative Areas Version 2.8 or (\href{https://gadm.org/}{GADM})
}
\usage{
data(mad0)
}
\description{
The national borders of Madagascar.
}
\examples{
data(mad0)
plot(mad0)
}
\keyword{datasets}
