% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mad4.r
\docType{data}
\name{mad4}
\alias{mad4}
\title{Spatial data: borders of selected districts of Madagascar}
\format{
An object of class \code{'SpatialPolygonsDataFrame'}. See \code{\link[sp]{SpatialPolygons}}.
}
\source{
Database of Global Administrative Areas Version 2.8 or (\href{https://gadm.org/}{GADM})
}
\usage{
data(mad4)
}
\description{
Borders of selected districts of Madagascar
}
\examples{
data(mad4)
plot(mad4)
}
\keyword{datasets}
