% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madChelsa.r
\docType{data}
\name{madChelsa}
\alias{madChelsa}
\title{Rasters of bioclimatic variables for an eastern portion of Madagascar}
\format{
An object of class \code{'SpatRaster'} in unprojected (WGS84) coordinates.
}
\source{
\doi{https://doi.org/10.1038/sdata.2017.122}
}
\description{
Rasters of bioclimatic variables for an eastern portion of Madagascar from CHELSA version 2.1 in unprojected (WGS84) coordinates. Values represent averages across 1980-2010. Only these BIOCLIM variables are included:
\itemize{
	\item \code{bio1}: Mean annual temperature (deg C)
	\item \code{bio7}: Temperature annual range (hottest - coldest month temperature; deg C)
	\item \code{bio12}: Total annual precipitation (mm)
	\item \code{bio15}: Precipitation seasonality (unitless)
}
}
\examples{

### vector data
library(sf)

# for vector data, we can use data(*) or fasterData(*):
data(madCoast0) # same as next line
madCoast <- fasterData('madCoast0') # same as previous
plot(madCoast0)

madCoast <- fasterData('madCoast4')
plot(madCoast4, add = TRUE)

madRivers <- fasterData('madRivers')
plot(madRivers, add = TRUE)

### raster data
library(terra)

# for raster data, we can get the file directly or using fasterData(*):
rastFile <- system.file('extdata/madElev.tif'), package='fasterRaster')
madElev <- terra::rast(rastFile)

madElev <- fasterData('madElev') # same as previus two lines
plot(madElev)

madForest2000 <- fasterData('madForest2000')
plot(madForest2000, col = 'forestgreen', add = TRUE)

madForest2014 <- fasterData('madForest2014')
plot(madForest2014, col = 'green', add = TRUE)
}
\references{
Karger, D.N., Conrad, O., Böhner, J., Kawohl, T., Kreft, H., Soria-Auza, R.W., Zimmermann, N.E., Linder, H.P., and Kessler, M.  2017.  Climatologies at high resolution for the earth’s land surface areas. \emph{Scientific Data} 4:170122. \doi{https://doi.org/10.1038/sdata.2017.122}
}
\keyword{Madagascar}
\keyword{climate}
