% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madCoast4.r
\docType{data}
\name{madCoast4}
\alias{madCoast4}
\title{Spatial vector of a portion of the coastline of Madagascar}
\format{
An object of class \code{'sf'}.
}
\source{
Database of Global Administrative Areas Version 2.8 (\href{https://gadm.org}{GADM})
}
\usage{
data(madCoast4)
}
\description{
Borders of a selected portion of Madagascar
}
\examples{

library(sf)
data(madCoast4)
plot(madCoast4)

}
\keyword{Madagascar}
