% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madElev.r
\docType{data}
\name{madElev}
\alias{madElev}
\title{Elevation raster for an eastern portion of Madagascar}
\format{
An object of class \code{SpatRaster}. Values are mean meters above sea level.
}
\source{
\link[=https://worldclim.org]{WorldClim}
}
\description{
Elevation raster for an eastern portion of Madagascar.
}
\keyword{Madagascar}
\keyword{elevation}
