% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madElevAnt.r
\docType{data}
\name{madElevAnt}
\alias{madElevAnt}
\title{Elevation raster for the Antanambe Commune of Madagascar}
\format{
An object of class \code{SpatRaster}. Values are mean meters above sea level.
}
\source{
\link[=https://worldclim.org]{WorldClim}
}
\description{
Elevation raster for the Antanambe Commune of Madagascar.
}
\keyword{Madagascar}
\keyword{elevation}
