% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madElevMan.r
\docType{data}
\name{madElevMan}
\alias{madElevMan}
\title{Elevation raster for the Manompana Commune of Madagascar}
\format{
An object of class \code{SpatRaster}. Values are mean meters above sea level.
}
\source{
\link[=https://worldclim.org]{WorldClim}
}
\description{
Elevation raster for the Manompana Commune of Madagascar.
}
\keyword{Madagascar}
\keyword{elevation}
