% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.r
\name{mask,GRaster,GRaster-method}
\alias{mask,GRaster,GRaster-method}
\alias{mask}
\alias{mask,GRaster,GVector-method}
\title{Mask values in a raster}
\usage{
\S4method{mask}{GRaster,GRaster}(x, mask, inverse = FALSE, maskvalues = NA, updatevalue = NA)

\S4method{mask}{GRaster,GVector}(x, mask, inverse = FALSE, updatevalue = NA)
}
\arguments{
\item{x}{A \code{GRaster}.}

\item{mask}{A \code{GRaster} or \code{GVector}.}

\item{inverse}{Logical: If \code{TRUE}, the effect of the mask is inverted. That is, a copy of the input raster is made, but cells that overlap with an \code{NA} in the mask raster or are not covered by the mask vector retain their values. Cells that overlap with an \code{NA} in the mask raster or overlap with the mask vector are forced to \code{NA}.}

\item{maskvalues}{Numeric vector, including \code{NA} (only for when \code{mask} is a \code{GRaster}): The value(s) in the mask raster cells that serve as the mask. The default is \code{NA}, in which case cells in the input raster that overlap with \code{NA} cells in the mask are forced to \code{NA}.}

\item{updatevalue}{Numeric, including \code{NA} (default): The values assigned to masked cells.}
}
\value{
A \code{GRaster}.
}
\description{
The output of \code{mask()} is a \code{GRaster} that has the same as values as the input raster. However, if the \code{mask} argument is a \code{GRaster}, the output will have \code{NA} values in the sample cells that the \code{mask} raster has \code{NA} cells. If the \code{mask} argument is a \code{GVector}, then the output raster will have \code{NA} values in cells the \code{GVector} does not cover.
}
\examples{
\dontrun{
# NB This example is in a "dontrun{}" block because it requires users to have
# GRASS GIS Version 8+ installed on their system.

# IMPORTANT #1: If you already have a GRASS session started, you will need to
# run the line below and the last line in this example to work with it again.
# If you have not started a GRASS session, you can skip this step and go to
# step #2.
opts. <- getFastOptions()

# IMPORTANT #2: Select the appropriate line below and change as necessary to
# where GRASS is installed on your system.

grassDir <- "/Applications/GRASS-8.3.app/Contents/Resources" # Mac
grassDir <- "C:/Program Files/GRASS GIS 8.3" # Windows
grassDir <- "/usr/local/grass" # Linux

# setup
library(sf)
library(terra)

# example data
madElev <- fastData("madElev") # raster
madForest <- fastData("madForest2000") # raster
madCoast <- fastData("madCoast4") # vector
madAnt <- madCoast[madCoast$NAME_4 == "Antanambe", ]

# start GRASS session for examples only
faster(x = madElev, grassDir = grassDir,
workDir = tempdir(), location = "examples") # line only needed for examples

# convert to GRasters and GVectors
elev <- fast(madElev)
forest <- fast(madForest)
ant <- fast(madAnt)

# Mask by a raster or  vector:
byRast <- mask(elev, forest)
plot(byRast)

byVect <- mask(elev, ant)
plot(byVect)

# Mask by a raster or vector, but invert mask:
byRastInvert <- mask(elev, forest, inverse = TRUE)
plot(byRastInvert)

byVectInvert <- mask(elev, ant, inverse = TRUE)
plot(byVectInvert)

# Mask by a raster, but use custom values for the mask:
byRastCustomMask <- mask(elev, elev, maskvalues = 1:20)
plot(byRastCustomMask)

# Mask by a raster or vector, but force masked values to a custom value:
byRastCustomUpdate <- mask(elev, forest, updatevalue = 7)
plot(byRastCustomUpdate)

byVectCustomUpdate <- mask(elev, ant, updatevalue = 7)
plot(byVectCustomUpdate)

# Mask by a raster, inverse, custom values, and custom update:
byVectAll <-
   mask(elev, forest, inverse = TRUE, maskvalues = 1:20, updatevalue = 7)
plot(byVectAll)

# IMPORTANT #3: Revert back to original GRASS session if needed.
restoreSession(opts.)
removeSession("examples")

}
}
\seealso{
\code{\link[terra:mask]{terra::mask()}}, \strong{GRASS} module \code{r.mask}
}
