% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maskRand.r
\name{maskRand,GRaster-method}
\alias{maskRand,GRaster-method}
\alias{maskRand}
\title{Randomly mask cells of a raster}
\usage{
\S4method{maskRand}{GRaster}(
  x,
  n,
  fail = TRUE,
  selectNA = FALSE,
  maskvalue = NA,
  updatevalue = NULL
)
}
\arguments{
\item{x}{A \code{GRaster}.}

\item{n}{Integer > 0: Number of cells to retain.}

\item{fail}{Logical: If \code{TRUE} (default), and \code{n} is greater than the number of non-\code{NA} cells in \code{x}, then fail.}

\item{selectNA}{Logical: By default, only non-\code{NA} cells are retained. If this is \code{TRUE}, then \code{NA} cells can also be selected. If \code{FALSE} (default), then only non-\code{NA} cells are selected for retention.}

\item{maskvalue}{Numeric vector, including \code{NA}, or \code{NULL} (default): Values in the raster to select from. All others will be ignored. If this is \code{NULL}, then only non-\code{NA} cells will be selected for retention.}

\item{updatevalue}{Numeric or \code{NULL} (default): Value to assign to masked cells. If \code{NULL}, then the values in the input raster are retained.}
}
\value{
A \code{GRaster}.
}
\description{
\code{rmask()} randomly masks cells of a raster, retaining a user-specified number of cells.
}
\seealso{
\code{\link[=mask]{mask()}}, \strong{GRASS} module \code{r.random}
}
