% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.r
\name{.message}
\alias{.message}
\title{Display warning or message}
\usage{
.message(msg, message)
}
\arguments{
\item{msg}{Character: Name for the message (used internally). Should be able to be assigned to a list (i.e., no spaces, punctuation, etc.).}

\item{message}{Text for the message.}
}
\value{
\code{TRUE} (invisibly).
}
\description{
Display a warning or message if the given warning has not been displayed since \strong{fasterRaster} was attached or if a given number or hours has passed since then.
}
\keyword{internal}
