% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmax.r
\docType{data}
\name{minmax.GRaster}
\alias{minmax.GRaster}
\title{Minimum and maximum values across all non-NA cells of a 'GRaster'}
\format{
An object of class \code{character} of length 1.
}
\usage{
minmax.GRaster
}
\arguments{
\item{x}{A \code{GRaster}.}
}
\value{
A numeric matrix.
}
\description{
Minimum and maximum values across all non-NA cells of a 'GRaster'.
}
\keyword{datasets}
