% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.omit.r
\name{na.omit,GRaster-method}
\alias{na.omit,GRaster-method}
\alias{na.omit}
\title{Force cells with NA in any layer of a GRaster stack to NA}
\usage{
\S4method{na.omit}{GRaster}(object, verbose = FALSE)
}
\arguments{
\item{object}{A "stack" of \code{GRaster}s.}

\item{verbose}{Logical: If \code{TRUE}, display progress.}
}
\value{
A \code{GRaster}.
}
\description{
'na.omit()' returns a \code{GRaster} stack with the same number of layers as the input, but with all cells that have \code{NA} in any layer set to \code{NA} in all layers. This is useful for masking out areas where data are missing in any of the rasters.
}
\examples{
if (grassStarted()) {

# Setup
library(terra)

### Mask layer-by-layer

# Elevation raster
madElev <- fastData("madElev")
madForest2000 <- fastData("madForest2000")

# Convert SpatRasters to GRasters:
elev <- fast(madElev)
forest2000 <- fast(madForest2000)

# Make a random layer and stack it with elevation:
x <- c(elev, forest2000)

# Put NAs in cells where any layer has an NA:
masked <- na.omit(x, verbose = TRUE)
plot(masked)

}
}
\seealso{
\code{\link[terra:na.omit]{terra::na.omit()}}, \code{\link[stats:na.omit]{stats::na.omit()}}, \strong{GRASS} tool \code{r.mapcalc} (see \code{grassHelp("r.mapcalc")})
}
