% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncat.r
\name{ncat,GRaster-method}
\alias{ncat,GRaster-method}
\alias{ncat}
\alias{ncat,SpatRaster-method}
\title{Number of levels in a categorical raster}
\usage{
\S4method{ncat}{GRaster}(x)

\S4method{ncat}{SpatRaster}(x, dropLevels = TRUE)
}
\arguments{
\item{x}{A \code{GRaster} or \code{SpatRaster}.}

\item{droplevels}{Logical: If \code{TRUE} (default), drop levels that do not appear in the raster.}
}
\value{
A numeric vector of integers. These represent the number of levels that appear in the raster (always for \code{GRaster}s, and for \code{SpatRaster}s when \code{dropLevels = TRUE}), or the number of levels that appear or could appear in the raster (for \code{SpatRaster}s only, when \code{dropLevels = FALSE}).
}
\description{
This function reports the number of categories (levels) in a \link[=tutorial_categorical_rasters]{categorical raster}.
}
\examples{
\dontrun{
# NB This example is in a "dontrun{}" block because it requires users to have
# GRASS GIS Version 8+ installed on their system.

# IMPORTANT #1: If you already have a GRASS session started, you will need to
# run the line below and the last line in this example to work with it again.
# If you have not started a GRASS session, you can skip this step and go to
# step #2.
opts. <- getFastOptions()

# IMPORTANT #2: Select the appropriate line below and change as necessary to
# where GRASS is installed on your system.
grassDir <- "/Applications/GRASS-8.3.app/Contents/Resources" # Mac
grassDir <- "C:/Program Files/GRASS GIS 8.3" # Windows
grassDir <- "/usr/local/grass" # Linux

# Setup
library(terra)

# Example data: Land cover raster
madCover <- fastData("madCover")

# Start GRASS session for examples only
faster(x = madCover, grassDir = grassDir,
workDir = tempdir(), location = "examples") # line only needed for examples

# Convert categorical SpatRaster to categorical GRaster:
cover <- fast(madCover)

# set categories
cover # note... no categories
levs <- levels(madCover)
levs <- levs[[1]]
cover <- categories(cover, layer = 1, value = levs)

cover # Note display of "min"/"max" categories and number of categories
is.factor(cover) # Is the raster categorical?
ncat(cover) # Number of categories
levels(cover) # Categories

# Re-assign levels
value <- c(20, 30, 40, 50, 120, 130, 140, 170, 210)
label <- c("Cropland", "Cropland", "Forest", "Forest", "Forest", "Shrubland", "Herbaceous", "Flooded", "Water")

cats <- data.frame(value = value, label = label)
cover <- categories(cover, layer = 1, value = cats)
cover

# We can implement logical operations on categorical rasters
cover < "Forest" # 1 for cells with a value < 40, 0 otherwise
cover <= "Forest" # 1 for cells with a value < 120, 0 otherwise
cover == "Forest" # 1 for cells with value of 40-120, 0 otherwise
cover != "Forest" # 1 for cells with value that is not 40-120, 0 otherwise
cover > "Forest" # 1 for cells with a value > 120, 0 otherwise
cover >= "Forest" # 1 for cells with a value >= 120, 0 otherwise

cover \%in\% c("Cropland", "Forest") # 1 for cropland/forest cells, 0 otherwise

# IMPORTANT #3: Revert back to original GRASS session if needed.
fastRestore(opts.)
removeSession("examples")

}
}
\seealso{
\code{\link[=levels]{levels()}}, \code{\link[terra:factors]{terra::cats()}}, \code{\link[terra:factors]{terra::levels()}}, \code{\link[terra:factors]{terra::addCats()}}, \code{\link[terra:factors]{terra::droplevels()}}, \link[=tutorial_categorical_rasters]{categorical rasters}
}
