% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlyr.r
\name{nlyr}
\alias{nlyr}
\title{Number of layers of a \code{GRaster} raster}
\usage{
nlyr(x)
}
\arguments{
\item{x}{A \code{GRaster}.}
}
\value{
A numeric value.
}
\description{
Number of layers of a \code{GRaster} raster object.
}
