% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiet.r
\name{.quiet}
\alias{.quiet}
\title{Returns .quiet() or NULL for "flags" argument to GRASS modules}
\usage{
.quiet()
}
\value{
A string (.quiet()) or \code{NULL}.
}
\description{
A function for developers used for setting the "quiet' argument in \code{flags} arguments passed to \code{\link[rgrass:execGRASS]{rgrass::execGRASS()}}. If \code{faster("verbose")} is \code{TRUE}, the string "quiet" is returned. If \code{FALSE}, then \code{NULL} is returned.
}
\examples{

if (grassStarted()) {

faster("verbose")
.quiet()

}

}
\keyword{internal}
