% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activeCat.r, R/buffer.r, R/crs.r, R/distance.r,
%   R/droplevels.r, R/ext.r, R/vect.r
\docType{import}
\name{reexports}
\alias{reexports}
\alias{activeCat}
\alias{st_buffer}
\alias{st_crs}
\alias{st_distance}
\alias{droplevels}
\alias{st_bbox}
\alias{st_as_sf}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{base}{\code{\link[base]{droplevels}}}

  \item{sf}{\code{\link[sf]{st_as_sf}}, \code{\link[sf]{st_bbox}}, \code{\link[sf:geos_unary]{st_buffer}}, \code{\link[sf]{st_crs}}, \code{\link[sf:geos_measures]{st_distance}}}

  \item{terra}{\code{\link[terra]{activeCat}}}
}}

