% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.r
\name{refresh,GRaster-method}
\alias{refresh,GRaster-method}
\alias{refresh}
\alias{refresh,GVector-method}
\title{Refresh metadata in a GRaster or GVector}
\usage{
\S4method{refresh}{GRaster}(x)

\S4method{refresh}{GVector}(x)
}
\arguments{
\item{x}{A \code{GRaster} or \code{GVector}.}
}
\value{
A \code{GRaster} or \code{GVector}.
}
\description{
\code{GRaster}s and \code{GVector}s are really pointers to objects in \strong{GRASS}. The values displayed when you use \code{show()} or \code{print()} for a \code{GRaster} or \code{GVector} are stored in \strong{R}. If, on the odd chance that you make a change to a \code{GRaster} or \code{GVector} in \strong{GRASS} (e.g., using the \strong{GRASS} GUI using \code{\link[=grass]{grass()}}, the changes will not be automatically reflected in the \code{GRaster} or \code{GVector}. This function can be used to refresh the objects in \strong{R} to reflect their proper values.
}
\examples{
man/examples/ex_GRaster.r 

}
