% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionReshape.r
\name{regionReshape}
\alias{regionReshape}
\title{Change or report the extent and dimensions/resolution of a region}
\usage{
regionReshape(x, trimTo = NULL)
}
\arguments{
\item{x}{Any of:
\itemize{
\item Missing (default): Reports the extent, resolution, and dimensions of the current region. All other arguments will be ignored.
\item A \code{GRaster}, \code{SpatRaster}, or \code{stars} raster: Sets the region's extent and dimensions to those of the raster.
}}

\item{trimTo:}{A \code{GRaster} or \code{NULL} (default). If a \code{GRaster}, then the region will be trimmed to the non-\code{NA} cells in this raster. \code{trimTo} can only be non-\code{NULL} if \code{x} is a \code{GRaster}. Ignored if \code{NULL}.}
}
\value{
Either a list, or \code{TRUE} (invisibly) if resizing and resampling was successful. Also resizes and resamples the region in the active \strong{GRASS} session.
}
\description{
This function simultaneously changes the extent and dimensions of a \strong{GRASS} "region". This function is mostly used internally and rarely of direct use to most users.
}
\seealso{
\code{\link[=regionExt]{regionExt()}}, \code{\link[=regionDim]{regionDim()}}, and \code{\link[=regionRes]{regionRes()}} in \strong{fasterRaster}; \code{\link[terra:ext]{terra::ext()}}, \code{\link[terra:dimensions]{terra::dim()}}, and \code{\link[terra:dimensions]{terra::res()}}; \code{\link[sf:st_bbox]{sf::st_bbox()}}; \strong{GRASS} module \link[=https://grass.osgeo.org/grass82/manuals/g.region.html]{g.region}
}
