% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.r
\name{region,missing-method}
\alias{region,missing-method}
\alias{regionsReshape}
\alias{region,SpatRaster-method}
\alias{region}
\alias{region,stars-method}
\alias{region,GRaster-method}
\alias{region,GVector-method}
\alias{regionExt,missing-method}
\alias{regionExt}
\alias{regionExt,numeric-method}
\alias{regionExt,GRaster-method}
\alias{regionExt,GVector-method}
\alias{regionExt,SpatRaster-method}
\alias{regionExt,SpatVector-method}
\alias{regionExt,sf-method}
\alias{regionExt,stars-method}
\alias{regionZExt,missing-method}
\alias{regionZExt}
\alias{regionZExt,numeric-method}
\alias{regionZExt,GRaster-method}
\alias{regionZExt,GVector-method}
\alias{regionDim,missing-method}
\alias{regionDim}
\alias{regionDim,numeric-method}
\alias{regionDim,GRaster-method}
\alias{regionDim,SpatRaster-method}
\alias{regionDim,stars-method}
\alias{regionRes,missing-method}
\alias{regionRes}
\alias{regionRes,numeric-method}
\alias{regionRes,GRaster-method}
\alias{regionRes,SpatRaster-method}
\alias{regionRes,stars-method}
\title{Change or report the extent and dimensions/resolution of a region}
\usage{
\S4method{region}{missing}(x)

\S4method{region}{SpatRaster}(x)

\S4method{region}{stars}(x)

\S4method{region}{GRaster}(x, trimTo = NULL)

\S4method{region}{GVector}(x)

\S4method{regionExt}{missing}(x)

\S4method{regionExt}{numeric}(x)

\S4method{regionExt}{GRaster}(x)

\S4method{regionExt}{GVector}(x)

\S4method{regionExt}{SpatRaster}(x)

\S4method{regionExt}{SpatVector}(x)

\S4method{regionExt}{sf}(x)

\S4method{regionExt}{stars}(x)

\S4method{regionZExt}{missing}(x)

\S4method{regionZExt}{numeric}(x)

\S4method{regionZExt}{GRaster}(x)

\S4method{regionZExt}{GVector}(x)

\S4method{regionDim}{missing}(x)

\S4method{regionDim}{numeric}(x)

\S4method{regionDim}{GRaster}(x)

\S4method{regionDim}{SpatRaster}(x)

\S4method{regionDim}{stars}(x)

\S4method{regionRes}{missing}(x)

\S4method{regionRes}{numeric}(x)

\S4method{regionRes}{GRaster}(x)

\S4method{regionRes}{SpatRaster}(x)

\S4method{regionRes}{stars}(x)
}
\arguments{
\item{x}{Any of:
\itemize{
\item Missing (default): Reports the extent, resolution, and dimensions of the current region. All other arguments will be ignored.
\item A \code{GRaster}, \code{GVector}, \code{SpatRaster}, or \code{stars} raster: Sets the region's extent and dimensions to those of the raster or vector. If a raster is supplied, the region will be resized and resampled to match its extent, resolution, and dimensions. If a vector is supplied, the region will be resized and the number of cells will be increased/decreased to fully encompass this new extent, but resolution will not be changed.
\item A \code{numeric} vector. This will resize the region's extent, resample the region's resolution/dimensions, or both to ensure the desired dimensions or resolution are retained:
* 4 values for \code{regionExt()}: Westernmost and easternmost easting (longitude), and southernmost and northernmost northing (latitude)
* 2 values for \code{regionDim()}: Number of rows and columns
* 2 values for \code{regionRes()}: Size of cells in the x- and y-dimensions
* 2 values for \code{regionZExt()}: Topmost and bottom-most elevations
* 3 values for \code{regionDim3d()}: Number of rows, columns, and depths
* 3 values for \code{regionRes3d()}: Size of cells in the x-, y-, and z-dimensions
}}

\item{trimTo}{A \code{GRaster} or \code{NULL} (default). If a \code{GRaster}, then the region will be trimmed to the non-\code{NA} cells in this raster. \code{trimTo} can only be non-\code{NULL} if \code{x} is a \code{GRaster}. Ignored if \code{NULL}.}
}
\value{
A list with values reflecting the region's horizontal extent, vertical extent, dimensions, and resolution. If the function is used to resize/resample the region, the values \emph{before} resizing/resampling are returned invisibly. The function also resizes/resamples the \link[=tutorial_regions]{region}.
}
\description{
This function either changes the extent and dimensions of a \strong{GRASS} \link[=tutorial_regions]{"region"} or reports the current region's extent, resolution, and dimensions. This function is mostly used internally and rarely of direct use to most users.
}
\examples{
\dontrun{
# NB This example is in a "dontrun{}" block because it requires users to have
# GRASS GIS Version 8+ installed on their system.

# IMPORTANT #1: If you already have a GRASS session started, you will need to
# run the line below and the last line in this example to work with it again.
# If you have not started a GRASS session, you can skip this step and go to
# step #2.
opts. <- getFastOptions()

# IMPORTANT #2: Select the appropriate line below and change as necessary to
# where GRASS is installed on your system.
grassDir <- "/Applications/GRASS-8.2.app/Contents/Resources" # Mac
grassDir <- 'C:/Program Files/GRASS GIS 8.2' # Windows
grassDir <- '/usr/local/grass' # Linux

# setup
library(terra)

# example data
madElev <- fastData('madElev')

# set up one location
faster(crs = madElev, grassDir = grassDir,
workDir = tempdir(), location = 'examples1') # line only needed for examples


TBD

# IMPORTANT #3: Revert back to original GRASS session if needed.
fastRestore(opts.)
removeSession('examples1')
removeSession('examples2')

}
}
