% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rightSlash.r
\name{rightSlash}
\alias{rightSlash}
\title{Replace left slash with right slash in a string}
\usage{
rightSlash(x)
}
\arguments{
\item{x}{A string.}
}
\value{
Character.
}
\description{
This function is helpful for Windows systems, where paths are usually expressed with left slashes, whereas \strong{R} requires right slashes.
}
\examples{

rightSlash('C:\\\\ecology\\\\main project')

}
