% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rRandWalk.r
\name{rRandWalk,GRaster-method}
\alias{rRandWalk,GRaster-method}
\alias{rRandWalk}
\title{Create raster representing one or more random walks}
\usage{
\S4method{rRandWalk}{GRaster}(
  x,
  n = 1,
  steps = 1e+05,
  directions = 8,
  avoid = FALSE,
  sameStart = FALSE,
  seed = NULL,
  check = TRUE
)
}
\arguments{
\item{x}{A \code{GRaster} to serve as a template.}

\item{n}{Numeric: Number of walkers. Default  is 1.}

\item{steps}{Numeric: Number of steps taken by each walker. Default is 100000.}

\item{directions}{Either 4 or 8: Directions in which a walker can turn at any point. If 4, then walks are confined to north/south/east/west directions (Rook's case). If 8, then the cardinal and subcardinal directions are allowed (Queen's case).}

\item{avoid}{Logical: If \code{FALSE} (default), then walkers can traverse their own walks. If \code{TRUE}, walkers avoid their own trails. A self-avoiding random walk can take much longer to compute.}

\item{sameStart}{Logical: If \code{FALSE} (default), walkers can begin anywhere. If \code{TRUE}, then walkers start from the same place.}

\item{seed}{Integer or \code{NULL} (default): If \code{NULL}, then a random seed is generated by the function. If numeric, results will be deterministic. In either case, the value will be rounded to the nearest integer.}

\item{check}{Logical: If \code{TRUE} (default), function will check to see if the addon \code{r.random.walk} has been installed. If it has not, it will attempt to install it.}
}
\value{
A \code{GRaster} with cell values representing the number of times one or more walkers traversed the cell.
}
\description{
This function creates a raster where the cell values represent the number of times one or more random "walkers" traverse the cell.
}
\details{
\strong{Note}: To use this function, you must have installed the \code{r.random.walk} addon using \code{installAddons("r.random.walk")}. If you are not sure if this is already installed, the function will try to do this. You can also do it manually using \code{addons("r.random.walk")}.
}
\examples{
man/examples/ex_rRandWalk/r

}
