% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstring.r
\name{rstring}
\alias{rstring}
\title{Make a nearly-guaranteed unique string}
\usage{
rstring(n, x = 12, filesafe = TRUE)
}
\arguments{
\item{n}{Numeric integer: How many strings to make (default is 1).}

\item{x}{Numeric integer: Number of letters and digits to use to make the string. Default is 12, leading to a probability of two matching random strings of <3.7E-18 if \code{filesafe = TRUE} and <3.1E-22 if \code{FALSE}.}

\item{filesafe}{Logical: If \code{TRUE} (default), make file-safe names (leading character is a letter, only use letters and digits).}
}
\value{
Character.
}
\description{
\code{rstring()} makes a string that is statically extremely likely to be unique.
}
\examples{

rstring(1)
rstring(5)
rstring(5, 3)

}
