% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqToSQL.r
\name{seqToSQL}
\alias{seqToSQL}
\title{Format a numeric series into an SQL value call}
\usage{
seqToSQL(x, maxChar = 64000)
}
\arguments{
\item{x}{A vector of numerical values. The vector should be sorted from  lowers to highest for the most efficient "compression" of sequential ranges. Values will be coerced to class \code{integer}.}

\item{maxChar}{Integer or numeric: Maximum number of characters to include in the output. If the output has more than this number of characters, the remainder is dropped, and the \code{trim} attribute of the output is set to \code{TRUE}. The default is 64000, which is the maximum length of an SQL statement (minus 1000 characters for other functions). Note that this limit is "soft"--the output can exceed this number of characters by a bit.}
}
\value{
A character string. The string has two attributes. The \code{trim} attribute is \code{TRUE} or \code{FALSE}, depending on whether \code{maxChar} was reached or not (and subsequent numbers dropped from the string). The \code{lastIndex} attribute is the last index of \code{x} that was processed (i.e., the index of the last value in the output).
}
\description{
This function takes as its argument a vector of integers or numeric values, and converts sequential runs to a range while keeping non-sequential values as-is. For example, \code{c(1, 5, 6, 7, 8, 9, 15, 16, 20)} becomes \code{"1,5-9,15-16,20"}. This reduces the number of characters necessary to supply to a SQL condition. This function is mainly of use to developers.
}
\examples{

x <- 1:5
seqToSQL(x)

x <- c(1:5, 7)
seqToSQL(x)

x <- c(1:5, 7, 15:16)
y <- c(1:5, 7, 15:16, 20)
seqToSQL(x)
seqToSQL(y)

seqToSQL(x, maxChar = 5)
seqToSQL(y, maxChar = 8)

}
