% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessionLocation_sessionMapset.r
\name{sessionLocation}
\alias{sessionLocation}
\alias{sessionMapset}
\title{"Location" of a 'G-' object}
\usage{
sessionLocation(x)

sessionMapset(x)
}
\arguments{
\item{x}{A \code{GSession} object or and object that contains a \code{GSession} (i.e., a \code{GSpatial} object: a \code{GRaster} or \code{GVector}).}
}
\value{
A character string.
}
\description{
\strong{GRASS} "locations" are sets of one or more rasters and/or vectors with the same coordinate reference systems, and may or may not represent the same actual location on Earth. \strong{GRASS} "mapsets" are like subfolders of locations, and are collections of rasters and/or vectors typically related to the same general project.  Typical users will not need to make changes to the default location (called "location") or mapset (called "PERMANENT").
}
\seealso{
\strong{GRASS} \link[=https://grass.osgeo.org/grass82/manuals/grass_database.html]{locations and mapsets}
}
