% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.r
\name{sources,GRaster-method}
\alias{sources,GRaster-method}
\alias{sources}
\alias{sources,GVector-method}
\alias{sources,character-method}
\title{Name of a raster or vector in a GRASS session}
\usage{
\S4method{sources}{GRaster}(x)

\S4method{sources}{GVector}(x)

\S4method{sources}{character}(x)
}
\arguments{
\item{x}{Either a \code{GSpatial} object or one that inherits from it (i.e., a \code{GRaster} or \code{GVector}), \emph{or} a character. If a character, then the character itself is returned.}
}
\value{
Character.
}
\description{
\code{sources()} retrieves the name of a raster or vector in \strong{GRASS}. \code{GRaster}s and \code{GVector}s are actually pointers to objects stored in a *\emph{GRASS} database. When using \strong{fasterRaster} functions on rasters and vectors, the commands are translated into \strong{GRASS} commands and executed on the objects named in the pointers. These objects use a "source" (which is really a filename) to refer to the \strong{GRASS} objects. In select cases, it can help to get the "sources" of a \code{GRaster} or \code{GVector}. This function is not of use to most users.
}
\examples{
\dontrun{
# NB This example is in a "dontrun{}" block because it requires users to have
# GRASS GIS Version 8+ installed on their system.

# IMPORTANT #1: If you already have a GRASS session started, you will need to
# run the line below and the last line in this example to work with it again.
# If you have not started a GRASS session, you can skip this step and go to
# step #2.
opts. <- getFastOptions()

# IMPORTANT #2: Select the appropriate line below and change as necessary to
# where GRASS is installed on your system.
grassDir <- "/Applications/GRASS-8.3.app/Contents/Resources" # Mac
grassDir <- "C:/Program Files/GRASS GIS 8.3" # Windows
grassDir <- "/usr/local/grass" # Linux

# setup
library(terra)

# example data
madElev <- fastData("madElev")
madForest2000 <- fastData("madForest2000")

# start GRASS session for examples only
faster(x = madElev, grassDir = grassDir,
workDir = tempdir(), location = "examples") # line only needed for examples

# convert SpatRasters to GRasters
elev <- fast(madElev)
forest <- fast(madForest2000)

### GRaster properties
######################

# plotting
plot(elev)

# dimensions
dim(elev) # rows, columns, depths, layers
nrow(elev) # rows
ncol(elev) # columns
ndepth(elev) # depths
nlyr(elev) # layers

res(elev) # resolution (2D)
res3d(elev) # resolution (3D)
zres(elev) # vertical resolution
xres(elev) # vertical resolution
yres(elev) # vertical resolution
zres(elev) # vertical resolution

# cell counts
ncell(elev) # cells
ncell3d(elev) # cells (3D rasters only)

# number of NA and non-NA cells
nacell(elev)
nonnacell(elev)

# topology
topology(elev) # number of dimensions
is.2d(elev) # is it 2-dimensional?
is.3d(elev) # is it 3-dimensional?

minmax(elev) # min/max values

# information on the GRASS session in which the GRaster is located
location(elev) # location
mapset(elev) # mapset

# "names" of the object
names(elev)

# coordinate reference system
crs(elev)
st_crs(elev)

# extent (bounding box)
ext(elev)
st_bbox(elev)

# vertical extent (not defined for this raster)
zext(elev)

# data type
datatype(elev) # fasterRaster type
datatype(elev, "GRASS") # GRASS type
datatype(elev, "terra") # terra type
datatype(elev, "GDAL") # GDAL type

is.integer(elev)
is.float(elev)
is.double(elev)
is.factor(elev)

# convert data type
as.integer(elev) # integer; note that "elev" is already of type "integer"
as.float(elev) # floating-precision
as.double(elev) # double-precision

# assigning
pie <- elev
pie[] <- pi # assign all cells to the value of pi
pie

# concatenating multiple GRasters
rasts <- c(elev, forest)
rasts

# subsetting
rasts[[1]]
rasts[["madForest2000"]]

plot(madElev == 10)
plot(madElev <= 10)
plot(madElev \%in\% c(10, 20, 30, 40, 50))

# replacing
rasts[[2]] <- 2 * forest

# adding layers
rasts[[4]] <- elev > 500 # add a layer
rasts <- c(rasts, sqrt(elev)) # add another
add(rasts) <- ln(elev)

# names
names(rasts)
names(rasts) <- c("elev_meters", "2_x_forest", "high_elevation", "sqrt_elev", "ln_elev")
rasts

# remove a layer
rasts[["2_x_forest"]] <- NULL
rasts

# number of layers
nlyr(rasts)

# cell frequencies
freq(elev)
freq(2 * elev)
freq(rasts)
freq(rasts, value = 1)

# projecting a raster from disk to location's CRS
madLANDSAT <- fastData("madLANDSAT")
landsat <- fast(madLANDSAT) # projects matrix

# correlation and covariance matrices
cor(landsat) # correlation
cov(landsat) # covariance



# IMPORTANT #3: Revert back to original GRASS session if needed.
restoreSession(opts.)
removeSession("examples")

}
}
