% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startFast.r
\name{startFast}
\alias{startFast}
\title{Initialize a 'GRASS' session}
\usage{
startFast(grassDir, crs, workDir = NULL, ...)
}
\arguments{
\item{grassDir}{Character: Folder in which \strong{GRASS} is installed on your computer. This will look different depending on the operating system and verison of \strong{GRASS} you have installed. Here are some examples:
\itemize{
\item Windows: \code{'C:/Program Files/GRASS GIS 8.3'}
\item Mac OS: \code{"/Applications/GRASS-8.3.app/Contents/Resources"}
\item Linux: \code{'/usr/local/grass'}
}}

\item{crs}{Any object from which a coordinate reference system (CRS) can be acquired. Ergo, any of:
\itemize{
\item A \code{SpatRaster}, \code{SpatVector}, \code{SpatExtent}, \code{stars} or \code{sf} object
\item A CRS (coordinate reference system) string
}}

\item{workDir}{\code{NULL} or character: The name of the folder in which \strong{GRASS} will store rasters and vectors. If this is \code{NULL} (default), then the \code{\link[=tempdir]{tempdir()}} on the user's system will be used. If users wish to create persistent \strong{GRASS} sesssions that can be used in a different instance of \strong{R} (i.e., if \strong{R} is stopped then restarted), then this needs to be specified.}

\item{...}{Options to send to \code{\link[=setFastOptions]{setFastOptions()}}. These should be in \code{option = value} format.}
}
\value{
\code{TRUE} if successful (invisibly). If not successful, the function will usually either fail or return \code{FALSE} with a warning.
}
\description{
This function initializes a \strong{GRASS} session in a particulatr folder. You need to run this function before you use most functions in \strong{fasterRaster} (just once). You can also use the function to switch your \strong{GRASS} session to a different folder, even if one is already started.
}
\seealso{
Guide to getting \href{tutorial_starting}{started} with \strong{fasterRaster}.
}
