% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.r
\docType{data}
\name{topology.GSpatial}
\alias{topology.GSpatial}
\title{Topology (2- or 3-dimensions) of a 'GRaster' or 'GVector'}
\format{
An object of class \code{character} of length 1.
}
\usage{
topology.GSpatial
}
\arguments{
\item{x}{A \code{GSpatial} object (i.e., a \code{GRaster} or \code{GVector}).}
}
\description{
Rasters and vectors in \strong{GRASS} can have 2-dimensional or 3-dimensional coordinates. This function returns the dimension of the coordinates of a \code{GSpatial} object or any that inherits from it (i.e., a \code{GRaster} or \code{GVector}).
}
\keyword{datasets}
