% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topoWetnessIndex.r
\name{topoWetnessIndex,GRaster-method}
\alias{topoWetnessIndex,GRaster-method}
\alias{topoWetnessIndex}
\title{Topographic wetness index}
\usage{
\S4method{topoWetnessIndex}{GRaster}(x)
}
\arguments{
\item{x}{A \code{GRaster} (typically representing elevation).}
}
\value{
A \code{GRaster}.
}
\description{
This function creates a raster map with values equal to the topographic wetness index (TWI), which is a measure of how much overland water flow tends to accumulate in or flow away from a location.
}
\examples{
if (grassStarted()) {

# Setup
library(terra)

# Elevation raster
madElev <- fastData("madElev")

# Convert to GRaster:
elev <- fast(madElev)

# Topographic wetness index:
tpi <- topoWetnessIndex(elev)

plot(c(elev, tpi))

}
}
\seealso{
\code{\link[=terrain]{terrain()}}, module \href{https://grass.osgeo.org/grass84/manuals/r.topidx.html}{\code{r.topidx}} in \strong{GRASS}
}
