% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorial_hidden_functions.r
\name{tutorial_hidden_functions}
\alias{tutorial_hidden_functions}
\title{Hidden functions in "fasterRester"}
\description{
\strong{fasterRaster} contains a set of "private" functions that users can access using \code{fasterRaster:::functionName}. These functions are useful for power users and developers. Not all hidden functions are listed here. Often, a method will have a hidden function of the same name that starts with a period. This "period" function is intended to be supplied the \code{\link[=sources]{sources()}} name of a \code{GRaster} or \code{GVector} from other functions so that the calling function does not need to spend the time creating the \code{GRaster} or \code{GVector} pointer before calling the function. "Period" functions will, though, often work on \code{GRaster}s or \code{GVector}s, though some error-checking and region definitions are not conducted.

\code{.copyGSpatial()}: Make a copy of the \strong{GRASS} file pointed to by a \code{GRaster} or \code{GVector}\cr
\code{.fileExt()}: Get file extension\cr
\code{.exists()}: Does the \strong{GRASS} file of a \code{GRaster} or \code{GVector} exist?\cr
\code{.ext()}: Extent from the \code{\link[=sources]{sources()}} name of a \code{GRaster} or \code{GVector}\cr
\code{.geomtype()}: Geometry type ("point", "line", or "area") from the \code{\link[=sources]{sources()}} name of a \code{GVector}\cr
\code{.layerIndex()}: Gets the index of \code{GRaster} layers from a numeric, integer, character, or logical vector\cr
\code{.locationCreate()} Make a connection to \strong{GRASS} (i.e., start \strong{GRASS} from within \strong{R}) and create a location\cr
\code{.locationDelete()} Deletes all files associated with a \strong{GRASS} "location" and mapset\cr
\code{.locationFind()}: Find a specific \strong{GRASS} "location" that already exists\cr
\code{.locationRestore()} Reconnect \strong{GRASS} to a previously-created \strong{GRASS} "location"\cr
\code{.locations()}: List of all available "locations"\cr
\code{.ls()}: Lists the \code{sources} of all objects in the active \strong{GRASS} "location"\cr
\code{.makeGRaster()} and \code{.makeGVector()}: Make \code{GRaster}s or \code{GVector}s from a vector of \code{sources}, which are pointers to files in \strong{GRASS}\cr
\code{.makeSourceNames()}: Makes one or more statistically unique strings that can be used as file names to represent rasters or vectors in \strong{GRASS}\cr
\code{.mapset()}: \strong{GRASS} "mapset" of an object or the active session\cr
\code{.message()}: Display a warning or message if the given warning has not been displayed since \strong{fasterRaster} was attached or if a given number or hours has passed since then\cr
\code{.minVal()} and \code{.maxVal()}: Values in the \verb{@minVal} and \verb{@maxVal} slots in a \code{GRaster}\cr
\code{.nlevels()}: Number of levels in a \code{SpatVector}, \code{data.frame}, \code{data.table}, empty string, or a list of \code{data.frame}s, \code{data.table}s, and/or empty strings.\cr
\code{.plot()}: Plot using the \code{\link[=sources]{sources()}} name of a \code{GRaster} or \code{GVector}\cr
\code{.projection()}: Value of the \verb{@projection} slot in a \code{GRaster} or \code{GVector}\cr
\code{.quiet()}: Returns "quiet" if \code{faster("verbose")} is \code{TRUE}\cr
\code{.rastInfo()} and \code{.vectInfo()}: Metadata for a \strong{GRASS} raster or vector\cr
\code{.region()}: Change or report the active region's extent and resolution\cr
\verb{.regionDim()]}: Change or report the active region's resolution (also \code{\link[=dim]{dim()}} and related functions, with no arguments)\cr
\code{.regionExt()}: Change or report the active region's extent (also \code{\link[=ext]{ext()}} and related functions, with no arguments)\cr
\code{.regionRes()}: Change or report the active region's dimensions (also \code{\link[=res]{res()}} and related functions, with no arguments)\cr
\code{.rename()}: Rename a \strong{GRASS} raster or vector\cr
\code{.rm()}: Delete rasters or vectors in \strong{GRASS}\cr
\code{.vAsDataTable()}: Convert the attribute table linked to a vector in \strong{GRASS} to a \code{data.table}. This table is distinct from the attribute table attached to a \code{GVector}\cr
\code{.vAttachDatabase()}: Add a database table to the \strong{GRASS} representation of a \code{GVector}\cr
\code{.vCats()}: Get a \code{data.table} with a single column named \code{cat}, which corresponds to the \strong{GRASS} attribute table's \code{cat} column\cr
\verb{.vDetachDatabase(): Detach the **GRASS** database from a **GRASS** vector\\cr }.vHasDatabase()\verb{: Tests if **GRASS** vector has a database\\cr }.vIncrementCats()\verb{: Increment category values of a }GVector\verb{\\cr }.vNames()\verb{: "**GRASS**" vector attribute table column names\\cr }.vRecat()\verb{: Change **GRASS** category indices of a **GRASS** vector\\cr }.vValidCats()`: Are category values of a vector valid?\cr
}
\keyword{tutorial}
