% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vDetachDatabase.r
\name{.vDetachDatabase}
\alias{.vDetachDatabase}
\title{Add a database table to a GRASS attribute table}
\usage{
.vDetachDatabase(x)
}
\arguments{
\item{x}{A \code{GVector} or the name of a vector in \strong{GRASS}.}
}
\value{
Invisibly returns the \code{\link[=sources]{sources()}} name of a vector in \strong{GRASS}.
}
\description{
\code{.vDetachDatabase()} detaches the database from a \strong{GRASS} vector and deletes it. This table is meant to be "invisible" to most users--they should use interact with attribute tables using the \code{GVector} slot \verb{@table}. Some functions do require tables (e.g., \code{\link[=extract]{extract()}} and \code{\link[=spatSample]{spatSample()}}). \strong{This function is mostly of use to developers.}
}
\keyword{internal}
