% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vDropColumn.r
\name{.vDropColumn}
\alias{.vDropColumn}
\title{Delete column(s) from a GRASS vector's attribute table}
\usage{
.vDropColumn(x, col)
}
\arguments{
\item{x}{A \code{GVector} or the name of a vector in \strong{GRASS}.}

\item{col}{Name or indices of columns to drop.}
}
\value{
Invisibly returns a \code{GVector} or the \strong{GRASS} name of a vector.
}
\description{
Removes one or more columns from a \strong{GRASS} vector's attribute table.
}
