% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegIndex.r
\name{vegIndex,GRaster-method}
\alias{vegIndex,GRaster-method}
\alias{vegIndex}
\title{Vegetation indices from surface reflectance}
\usage{
\S4method{vegIndex}{GRaster}(
  x,
  metric = "NDVI",
  r = NULL,
  g = NULL,
  b = NULL,
  nir = NULL,
  b5 = NULL,
  b7 = NULL,
  soilSlope = NULL,
  soilIntercept = NULL,
  soilNR = 0.08
)
}
\arguments{
\item{x}{A \code{GRaster} with one layer per required band. Values should be between 0 and 1.}

\item{metric}{Character or character vector: The vegetation index or indices to calculate. Partial matching is used, and case is ignored. More infomration on each index can be found on the \strong{GRASS} help page for the \code{i.vi} module.
\itemize{
\item \code{ARVI}: Atmospherically Resistant Vegetation Index - Less sensitive to atmospheric interference than NDVI. Requires RGB channels.
\item \code{CI}: Crust Index - Suitable for detecting biogenic crust on soil. Requires RB channels.
\item \code{DVI}: Difference Vegetation Index - Requires R and NIR channels.
\item \code{EVI}: Enhanced Vegetation Index - Saturates more slowly than NDVI in high-biomass areas.  Requires RB an NIR channels.
\item \code{EVI2}: Enhanced Vegetation Index 2 - As EVI.  Requires R and NIR channels.
\item \code{GARI}: Green atmospherically resistant vegetation index.  Requires RGB and NIR channels.
\item \code{GEMI}: Global Environmental Monitoring Index.  Requires R and NIR channels.
\item \code{GVI}: Green Vegetation Index - Requires RGB, NIR, plus channels 5 and 7.
\item \code{IPVI}: Infrared Percentage Vegetation Index.  Requires R and NIR channels.
\item \code{MSAVI}: Modified Soil Adjusted Vegetation Index. Requires R and NIR channels, plus soil line slope, intercept, and a noise reduction factor.
\item \code{MSAVI2}: Second Modified Soil Adjusted Vegetation Index. Requires R and NIR channels.
\item \code{NDVI}: Normalized Difference Vegetation Index.  Requires R and NIR channels.
\item \code{NDWI}: Normalized Difference Water Index. Requires G and NIR channels.
\item \code{PVI}: Perpendicular Vegetation Index. Requires R and NIR channels.
\item \code{SAVI}: Soil Adjusted Vegetation Index. Requires R and NIR channels.
\item \code{SR}: Simple Vegetation ratio. Requires R and NIR channels.
\item \code{WDVI}: Weighted Difference Vegetation Index. Requires R, NIR, and soil line weight channels.
}}

\item{r, g, b, nir}{Numeric or character: Index or \code{\link[=names]{names()}} of the layers in \code{x} that represent the red, green, blue, and near infrared channels.}

\item{b5, b7}{Numeric or character: Index of names of the layers representing bands 5 and 7. These are used only for GVI.}

\item{soilSlope, soilIntercept, soilNR}{Numeric: Values of the soil slope, intercept, and soil noise reduction factor (0.08, by default). Used only for calculation of MSAVI.}
}
\value{
A \code{GRaster}.
}
\description{
This function calculates one of many types of vegetation indices from a raster with four bands representing blue (B), green (G), red (R), and near infrared (NIR), plus possibly channels 5 and 7. The function requires rasters that represent reflectance, so should have values that fall in the range 0 to 1.

The \code{\link[=toarLandsat]{toarLandsat()}} or \code{\link[=toarAster]{toarAster()}} functions can convert LANDSAT and ASTER digital-number rasters to radiance rasters, and then to reflectance rasters.
}
