% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewshed.r
\name{viewshed,GRaster-method}
\alias{viewshed,GRaster-method}
\alias{viewshed}
\title{Areas visible from points on a raster}
\usage{
\S4method{viewshed}{GRaster}(
  x,
  loc,
  observer = 1.8,
  target = 0,
  curvature = TRUE,
  refraction = FALSE,
  refractCoeff = 0.14286,
  maxDist = Inf
)
}
\arguments{
\item{x}{A \code{GRaster} typically representing elevation.}

\item{loc}{A "points" \code{GVector}, a two-column matrix with coordinate pairs (longitude then latitude), or a pair of numeric values indicating longitude and latitude of the point from which the viewshed should be calculated. If more than one point is supplied, the final map will reflect the combined viewshed from all points.}

\item{observer}{Numeric: The altitude of the observer above the surface (in meters). The default is 1.8 (approximately eye-level).}

\item{target}{Numeric: The altitude of the target above the surface (in meters). The default is 0.}

\item{curvature}{Logical: If \code{TRUE} (default), account for curvature of the Earth.}

\item{refraction}{Logical: If \code{TRUE}, account for atmospheric refraction. Default is \code{FALSE}.}

\item{refractCoeff}{Numeric between 0 and 1: Refraction coefficient. The default is 0.14286.}

\item{maxDist}{Numeric >= 0: Maximum distance for which to calculate viewshed. Default is \code{Inf} (infinite). This is in map units (usually meters).}
}
\value{
A \code{GRaster}.
}
\description{
\code{viewshed()} calculates the area that can be seen from a given point on a raster. It accounts for elevation, and optionally) curvature of the Earth and atmospheric refraction.
}
\examples{
\dontrun{
# NB This example is in a "dontrun{}" block because it requires users to have
# GRASS GIS Version 8+ installed on their system.

# IMPORTANT #1: If you already have a GRASS session started, you will need to
# run the line below and the last line in this example to work with it again.
# If you have not started a GRASS session, you can skip this step and go to
# step #2.
opts. <- getFastOptions()

# IMPORTANT #2: Select the appropriate line below and change as necessary to
# where GRASS is installed on your system.
grassDir <- "/Applications/GRASS-8.3.app/Contents/Resources" # Mac
grassDir <- 'C:/Program Files/GRASS GIS 8.3' # Windows
grassDir <- '/usr/local/grass' # Linux

# setup
library(sf)
library(terra)

# plant specimens (points), elevation (raster)
madDypsis <- fastData('madDypsis')
madElev <- fastData('madElev')

# start GRASS session for examples only
faster(x = madElev, grassDir = grassDir,
workDir = tempdir(), location = 'examples') # line only needed for examples

# convert a SpatRaster to a GRaster, and sf to a GVector
dypsis <- fast(madDypsis)
elev <- fast(madElev)

# Calculate viewshed from two points:
loc <- dypsis[4:5]
vs <- viewshed(elev, loc)

plot(vs)
plot(loc, add = TRUE)

# IMPORTANT #3: Revert back to original GRASS session if needed.
fastRestore(opts.)
removeSession('examples')

}
}
