% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vIncrementCats.r
\name{.vIncrementCats}
\alias{.vIncrementCats}
\title{Increment category values of a "GRASS" vector}
\usage{
.vIncrementCats(x, add)
}
\arguments{
\item{x}{A \code{GVector} or the \code{\link[=sources]{sources()}} name of one.}

\item{add}{Integer: Value to add to each category value.}
}
\value{
The \code{\link[=sources]{sources()}} name of a \strong{GRASS} vector with category values incremented.
}
\description{
Adds a constant to all category values of a \strong{GRASS} vector. \strong{This function is mostly of use to developers.}
}
\keyword{internal}
