% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vNames.r
\name{.vNames}
\alias{.vNames}
\title{Names of columns of a GRASS vector's attribute table}
\usage{
.vNames(x)
}
\arguments{
\item{x}{A \code{GVector} or the name of a vector in \strong{GRASS}.}
}
\value{
Character vector.
}
\description{
This function returns the column names of a \strong{GRASS} vector's attribute table.
}
\examples{
if (grassStarted()) {

madRivers <- fastData("madRivers")
rivers <- fast(madRivers)

.vCats(rivers)
.vHasDatabase(rivers)
.vAsDataTable(rivers)
.vRecat(rivers) # dangerous--re-categorizes geometries!
.vValidCats(rivers)

.vDetachDatabase(rivers) # dangerous--detaches database!
.vAttachDatabase(rivers)

# Ensure we don't use this for other examples!
.rm(rivers) # delete in GRASS
rm(rivers) # delete in R

}
}
\keyword{internal}
