% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workDir.r
\name{workDir,GSession-method}
\alias{workDir,GSession-method}
\alias{workDir}
\title{Get a GSession's working directory}
\usage{
\S4method{workDir}{GSession}(x)
}
\arguments{
\item{x}{A \code{GSession} object.}
}
\value{
Character.
}
\description{
This function returns the working directory of a \code{GSession} object.
}
