% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeRaster.r
\docType{data}
\name{writeRaster.GRaster}
\alias{writeRaster.GRaster}
\title{Save one or more 'GRaster's to disk}
\format{
An object of class \code{character} of length 1.
}
\usage{
writeRaster.GRaster
}
\arguments{
\item{x}{A \code{GRaster} or missing.}

\item{filename}{Path and file name.}

\item{overwrite}{If \code{FALSE} (default), do not save over existing file(s).}

\item{...}{Additional arguments. These can include:
\itemize{
\item \code{datatype} The datatype of the values stored in non-ASCII rasters. If \code{NULL}, this will be ascertained from the raster . This can any of:
| \strong{\code{terra}} | \strong{\code{GRASS}} | \strong{\code{GDAL}} | Values |
| \code{INT1U}     | \code{CELL}      | \code{Byte}     | Integer values from 0 to 255 |
| \code{INT2U}     | \code{CELL}      | \code{UInt16}   | Integer values from 0 to 65,534 |
| \code{INT2S}     | \code{CELL}     | \code{Int16}    | Integer values from -32,767 to -32,767 |
| \code{INT4S}     | \code{CELL}     | \code{Int32}    | Integer values from -2,147,483,647 to 2,147,483,647 |
| \code{FLT4S}     | \code{FCELL}   | \code{Float32}    | Values from -3.4e+38 to 3.4e+38, including decimal values |
| \code{FLT8S}     | \code{DCELL}   | \code{Float64}    | Values from -1.7e+308 to 1.7e+308, including decimal values |
\item \code{compressTiff}: Character or \code{NULL}: Type of compression for GeoTIFF files. The default is \code{'DEFLATE'}, but can be any of \code{'LZW'}, \code{'PACKBITS'}, \code{'LZMA'}, or \code{NULL}. If \code{NULL}, then no compression is used, but the file can still be reduced in size by using zip, gzip, or other compressions.
\item \code{bigTiff}: Logical: If \code{TRUE}, and the file format is a GeoTIFF and would be larger than 4 GB (regardless of compression), then the file will be saved in BIGTIFF format.
\item \code{format}: Character, indicating file format. This is usually ascertained from the file extension, but in case this fails, it can be stated explicitly. When using other formats, you may have to specify the \code{createopts} argument, too (see \href{r.out.gdal}{https://grass.osgeo.org/grass82/manuals/r.out.gdal.html}). Some common formats include:
\itemize{
\item \code{'GTiff'} (default): GeoTIFF \code{filename} ends in \code{.tif}
\item \code{'ASC'}: ASCII \code{filename} ends in \code{.asc}
}
\item Additional arguments to send to \href{r.out.gdal}{https://grass.osgeo.org/grass82/manuals/r.out.gdal.html} and \href{r.out.ascii}{https://grass.osgeo.org/grass82/manuals/r.out.ascii.html}.
\item \code{precision}: Numeric: For ASCII files, you may need to state the number of significant digits. 32-bit values have 7 digits and 64-bit values have 16. So in these cases the argument would be \code{precision=7} or \code{precision=16}.
}}
}
\value{
A \code{GRaster} or a \code{stars} raster. A raster is also saved to disk.
}
\description{
This function saves a raster to disk directly from a \strong{GRASS} session. If the \code{GRaster} was produced in \code{GRASS} (i.e., not created using \code{terra} or \code{stars} then converted to a \code{GRaster}), and you convert it to a \code{SpatRaster} or \code{stars} raster, please consult \code{\link[=writeRaster4]{writeRaster4()}} for important details!
}
\details{
The file type will be attempted to be ascertained from the file extension, but you can specify the format using the \code{format} argument (see entry for \code{...}). You can see a list of supported formats by simply using this function with no arguments, as in \code{writeRaster()}, or also at \href{r.out.gdal}{https://grass.osgeo.org/grass82/manuals/r.out.gdal.html}. Only the \code{GeoTIFF} file format is guaranteed to work for multi-layered rasters.
}
\seealso{
\code{\link[=writeRaster4]{writeRaster4()}} and \code{\link[=writeRaster8]{writeRaster8()}} for saving a raster created in \strong{GRASS} then imported to \strong{R} using \code{\link[=rast]{rast()}}; \code{\link[terra:writeRaster]{terra::writeRaster()}}; \strong{GRASS} modules \href{r.out.gdal}{https://grass.osgeo.org/grass82/manuals/r.out.gdal.html} and \href{r.out.ascii}{https://grass.osgeo.org/grass82/manuals/r.out.ascii.html}.
}
\keyword{datasets}
