% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeVector.r
\name{writeVector,GVector,character-method}
\alias{writeVector,GVector,character-method}
\alias{writeVector}
\alias{writeVector,missing,missing-method}
\title{Save a GVector to disk}
\usage{
\S4method{writeVector}{GVector,character}(x, filename, overwrite = FALSE, format = "GPKG", ...)

\S4method{writeVector}{missing,missing}(x, filename)
}
\arguments{
\item{x}{A \code{GVector}.}

\item{filename}{Character: Path and file name.}

\item{overwrite}{Logical: If \code{FALSE} (default), do not save over existing files.}

\item{format}{File format. Some common formats include:
\itemize{
\item \code{'GPKG'}: OGC GeoPackage (default).
\item \code{'CSV'}: Comma-separated value... saves the data table only, not the geometries.
\item \code{'ESRI_Shapefile'}: ESRI shapefile... \code{filename} should not end in an extension.
\item \code{'GeoJSON'}: GeoJSON
\item \code{'KML'}: Keyhole Markup Language (KML)
\item \code{'netCDF'}: NetCDF (argument \code{filename} should not end in an extension).
\item \code{'XLSX'}: MS Office Open XML spreadsheet
}}

\item{...}{Additional arguments to send to \strong{GRASS} module \code{v.out.ogr}.}
}
\value{
Invisibly returns a \code{SpatVector}. Importantly, the function also writes one or more files to disk.
}
\description{
This function saves a spatial vector to disk directly from a \strong{GRASS} session.

By default, files will be of OGC GeoPackage format (extension "\code{.gpkg}"), but this can be changed with the \code{format} argument. You can see a list of supported formats by simply using this function with no arguments, as in \code{writeVector()}, or by consulting the online help page for \strong{GRASS} module \code{v.out.ogr}.
}
\examples{
\dontrun{
# NB This example is in a "dontrun{}" block because it requires users to have
# GRASS GIS Version 8+ installed on their system.

# IMPORTANT #1: If you already have a GRASS session started, you will need to
# run the line below and the last line in this example to work with it again.
# If you have not started a GRASS session, you can skip this step and go to
# step #2.

# IMPORTANT #2: Select the appropriate line below and change as necessary to
# where GRASS is installed on your system.
grassDir <- "/Applications/GRASS-8.2.app/Contents/Resources" # Mac
grassDir <- 'C:/Program Files/GRASS GIS 8.2' # Windows
grassDir <- '/usr/local/grass' # Linux

# setup
library(terra)

# example data
madRivers <- fastData('madRivers')

# start GRASS session for examples only
wd <- forwardSlash(tempdir()) # only for examples

fastStart(crs = madRivers, grassDir = grassDir,
workDir = wd, location = 'examples') # line only needed for examples

# convert SpatRasters to GRasters
rivers <- fast(madRivers)
rivers

### save GVector to disk (using temporary file)
filename <- tempfile()
filename <- paste0(filename, '.gpkg')
writeRaster(rivers, filename)

### load raster from disk
rivers2 <- fast(filename)
rivers2

# IMPORTANT #3: Revert back to original GRASS session if needed.
fastRestore(opts.)
removeSession('examples')

}
}
\seealso{
\code{\link[terra:writeVector]{terra::writeVector()}}, \code{\link[=sf:st_write]{sf:st_write()}}
}
