% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xor.r
\name{xor,GVector,GVector-method}
\alias{xor,GVector,GVector-method}
\alias{xor}
\title{Select parts of polygons not shared by two GVectors}
\usage{
\S4method{xor}{GVector,GVector}(x, y)
}
\arguments{
\item{x, y}{\code{GVector}s.}
}
\value{
A \code{GVector}.
}
\description{
The \code{xor()} function selects the area that does not overlap between two "polygon" \code{GVector}s. You can also use the \code{/} operator, as in \verb{vect 1 / vect2}.
}
\examples{
\dontrun{
# NB This example is in a "dontrun{}" block because it requires users to have
# GRASS GIS Version 8+ installed on their system.

# IMPORTANT #1: If you already have a GRASS session started, you will need to
# run the line below and the last line in this example to work with it again.
# If you have not started a GRASS session, you can skip this step and go to
# step #2.
opts. <- getFastOptions()

# IMPORTANT #2: Select the appropriate line below and change as necessary to
# where GRASS is installed on your system.
grassDir <- "/Applications/GRASS-8.3.app/Contents/Resources" # Mac
grassDir <- "C:/Program Files/GRASS GIS 8.3" # Windows
grassDir <- "/usr/local/grass" # Linux

# Setup
library(sf)

# Polygon of coastal Madagascar and Dypsis specimens
madCoast4 <- fastData("madCoast4") # polygons
madDypsis <- fastData("madDypsis") # points

# Start GRASS session for examples only:
faster(x = madCoast4, grassDir = grassDir,
workDir = tempdir(), location = "examples") # line only needed for examples

# Convert vectors:
coast4 <- fast(madCoast4)
dypsis <- fast(madDypsis)

# Create another polygons vector from a convex hull around Dypsis points
hull <- convHull(dypsis)

### union()
###########

unioned <- union(coast4, hull)
plot(unioned)

plus <- coast4 + hull # same as union()

### intersect
#############

inter <- intersect(coast4, hull)
plot(coast4)
plot(hull, border = "red", add = TRUE)
plot(inter, border = "blue", add = TRUE)

### xor
#######

xr <- xor(coast4, hull)
plot(coast4)
plot(xr, border = "blue", add = TRUE)

### not
#######

no <- not(coast4, hull)
plot(coast4)
plot(no, border = "blue", add = TRUE)

minus <- coast4 - hull # same as not()

# IMPORTANT #3: Revert back to original GRASS session if needed.
restoreSession(opts.)
removeSession("examples")

}
}
\seealso{
\code{\link[=c]{c()}}, \code{\link[=aggregate]{aggregate()}}, \code{\link[=crop]{crop()}}, \code{\link[=intersect]{intersect()}}, \code{\link[=union]{union()}}, \code{\link[=erase]{erase()}}
}
