% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zExt.r
\name{zExt}
\alias{zExt}
\title{Vertical extent of a \code{GVector} or \code{GRaster}}
\usage{
zExt(x)
}
\arguments{
\item{x}{A \code{GRaster} or \code{GVector}.}
}
\value{
A numeric matrix.
}
\description{
\code{GRasters} and \code{GVectors} can have 2-dimensional or 3-dimensional coordinate systems. If 2-dimensional, the vertical extent is assumed to be 0. If 3-dimensional, this function reports the lowest and highest coordinates.
}
