% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTable_dropTable.r
\name{addTable<-,GVector,data.frame-method}
\alias{addTable<-,GVector,data.frame-method}
\alias{addTable<-}
\alias{addTable<-,GVector,data.table-method}
\alias{addTable<-,GVector,matrix-method}
\alias{dropTable,GVector-method}
\alias{dropTable}
\title{Attach or detach GVector's data table}
\usage{
\S4method{addTable}{GVector,data.frame}(x) <- value

\S4method{addTable}{GVector,data.table}(x) <- value

\S4method{addTable}{GVector,matrix}(x) <- value

\S4method{dropTable}{GVector}(x)
}
\arguments{
\item{x}{A `GVector.}

\item{value}{A \code{data.frame}, \code{data.table}, or \code{matrix}.}
}
\value{
A \code{GVector}.
}
\description{
\code{addTable()} adds an entire table to a \code{GVector}. It will replace any existing table. There must be one row in the table for each geometry (see \code{\link[=ngeom]{ngeom()}}). You can also add a table column-by-column using the \code{\link[fasterRaster]{$<-}} operator.

\code{dropTable()} removes a data table associated with a \code{GVector}.
}
\examples{
if (grassStarted()) {

# Setup
library(sf)

# Points vector of specimens of species in the plant genus Dypsis
madDypsis <- fastData("madDypsis")

# Convert sf to a GVector:
dypsis <- fast(madDypsis)

# Get just the data table:
df <- as.data.frame(dypsis)
head(df)

# Remove the table
dypsisTable <- dropTable(dypsis)
dypsis
dypsisTable

# Add table:
addTable(dypsisTable) <- df
dypsisTable


}
}
\seealso{
\code{\link[fasterRaster]{$<-}}, \code{\link[=as.data.frame]{as.data.frame()}}, \code{\link[=as.data.table]{as.data.table()}}
}
