% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_arithmetic.r
\name{Arith,GRaster,logical-method}
\alias{Arith,GRaster,logical-method}
\alias{Arith}
\alias{Arith,logical,GRaster-method}
\alias{Arith,GRaster,numeric-method}
\alias{Arith,GRaster,integer-method}
\alias{Arith,numeric,GRaster-method}
\alias{Arith,integer,GRaster-method}
\alias{Arith,GRaster,GRaster-method}
\alias{Arith,GVector,GVector-method}
\title{Arithmetic operations on GRasters}
\usage{
\S4method{Arith}{GRaster,logical}(e1, e2)

\S4method{Arith}{logical,GRaster}(e1, e2)

\S4method{Arith}{GRaster,numeric}(e1, e2)

\S4method{Arith}{GRaster,integer}(e1, e2)

\S4method{Arith}{numeric,GRaster}(e1, e2)

\S4method{Arith}{integer,GRaster}(e1, e2)

\S4method{Arith}{GRaster,GRaster}(e1, e2)

\S4method{Arith}{GVector,GVector}(e1, e2)
}
\arguments{
\item{name}{Character: Name of the new \code{GRaster}.}

\item{src}{\code{sources} of the \code{GRaster} being operated on}

\item{ex}{expression for \code{r.mapcalc}}
}
\value{
A \code{GRaster}.
}
\description{
\strong{\code{GRaster}s}: You can do arithmetic operations on \code{GRaster}s and using normal operators in \strong{R}: \code{+}, \code{-}, \code{*}, \code{/}, \code{^}, \code{\%\%} (modulus), and \code{\%/\%} (integer division).

Note that for \code{GRaster}s, the precision of the result will be determined by the \code{rasterPrecision} option, which can be set using \code{\link[=faster]{faster()}}. The default is \code{"float"}, which is precise to about the 7th decimal place. This be increased to about the 15th decimal place by setting this to \code{"double"}, though it can substantially increase the size of the raster output in memory and when saved to disk.

\strong{\code{GVector}s}: You can also do arithmetic operations on \code{GVector}s:
\code{+} operator: Same as \code{\link[=union]{union()}}
\code{-} operator: Same as \code{\link[=erase]{erase()}}.
\code{*} operator: Same as \code{\link[=intersect]{intersect()}}.
\code{/} operator: Same as \code{\link[=xor]{xor()}}.
}
\examples{
if (grassStarted()) {

# Setup
library(sf)
library(terra)

# Example data
madElev <- fastData("madElev")

# Convert a SpatRaster to a GRaster
elev <- fast(madElev)
elevs <- c(elev, elev, log10(elev) - 1, sqrt(elev))
names(elevs) <- c("elev1", "elev2", "log_elev", "sqrt_elev")

elev
elevs

# do some math
elev + 100
elev - 100
elev * 100
elev / 100
elev ^ 2
elev \%/\% 100 # divide then round down
elev \%\% 100 # modulus

100 + elev
100 \%/\% elev
100 \%\% elev

elevs + 100
100 + elevs

# math with logicals
elev + TRUE
elev - TRUE
elev * TRUE
elev / TRUE
elev ^ TRUE
elev \%/\% TRUE # divide then round down
elev \%\% TRUE # modulus

elevs + TRUE
TRUE + elevs

# raster interacting with raster
elev + elev
elev - elev
elev * elev
elev / elev
elev ^ log(elev)
elev \%/\% sqrt(elev) # divide then round down
elev \%\% sqrt(elev) # modulus

elevs + elev
elev * elevs

# sign
abs(-1 * elev)
abs(elevs)
sqrt(elevs)

# trigonometry
sin(elev)
cos(elev)
tan(elev)

asin(elev)
acos(elev)
atan(elev)

atan(elevs)
atan2(elev, elev^1.2)
atan2(elevs, elev^1.2)
atan2(elev, elevs^1.2)
atan2(elevs, elevs^1.2)

# logarithms
exp(elev)
log(elev)
ln(elev)
log2(elev)
log1p(elev)
log10(elev)
log(elev, 3)

log(elevs)

# rounding
round(elev + 0.5)
floor(elev + 0.5)
ceiling(elev + 0.5)
trunc(elev + 0.5)

# comparison
elev < 100
elev <= 100
elev == 100
elev != 100
elev > 100
elev >= 100

elev + 100 < 2 * elev

elevs > 10
10 > elevs

# mathematical functions on two or more rasters
mean(elevs)
mmode(elevs)
median(elevs)
nunique(elevs)

sum(elevs)
count(elevs)
min(elevs)
max(elevs)
range(elevs)
skewness(elevs)
kurtosis(elevs)

which.min(elevs)
which.max(elevs)

slope(elevs)
intercept(elevs)
r2(elevs)
tvalue(elevs)

sd(elevs)
sdpop(elevs)
var(elevs)
varpop(elevs)
quantile(elevs, 0.1)

}
}
