% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_arithmetic.r
\name{Arith,GRaster,logical-method}
\alias{Arith,GRaster,logical-method}
\alias{Arith}
\alias{Arith,logical,GRaster-method}
\alias{Arith,GRaster,numeric-method}
\alias{Arith,GRaster,integer-method}
\alias{Arith,numeric,GRaster-method}
\alias{Arith,integer,GRaster-method}
\alias{Arith,GRaster,GRaster-method}
\alias{Arith,GVector,GVector-method}
\title{Arithmetic operations on GRasters}
\usage{
\S4method{Arith}{GRaster,logical}(e1, e2)

\S4method{Arith}{logical,GRaster}(e1, e2)

\S4method{Arith}{GRaster,numeric}(e1, e2)

\S4method{Arith}{GRaster,integer}(e1, e2)

\S4method{Arith}{numeric,GRaster}(e1, e2)

\S4method{Arith}{integer,GRaster}(e1, e2)

\S4method{Arith}{GRaster,GRaster}(e1, e2)

\S4method{Arith}{GVector,GVector}(e1, e2)
}
\arguments{
\item{e1, e2}{\code{GRaster}s, \code{numeric}s, \code{integer}s, or \code{logical}s.}
}
\value{
A \code{GRaster}.
}
\description{
\strong{\code{GRaster}s}: You can do arithmetic operations on \code{GRaster}s and using normal operators in \strong{R}: \code{+}, \code{-}, \code{*}, \code{/}, \code{^}, \code{\%\%} (modulus), and \code{\%/\%} (integer division).

\strong{\code{GVector}s}: You can also do arithmetic operations on \code{GVector}s:\cr\cr
\code{+} operator: Same as \code{\link[=union]{union()}}\cr
\code{-} operator: Same as \code{\link[=erase]{erase()}}\cr
\code{*} operator: Same as \code{\link[=intersect]{intersect()}}\cr
\code{/} operator: Same as \code{\link[=xor]{xor()}}\cr
}
