% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.contour.r
\name{as.contour}
\alias{as.contour}
\title{Contour lines from a 'GRaster'}
\usage{
as.contour(x, nlevels, levels)
}
\description{
Create a spatial vector of contour lines from a raster.
}
\seealso{
\code{\link[terra:contour]{terra::as.contour()}}; \strong{GRASS} module \link[=https://grass.osgeo.org/grass82/manuals/r.contour.html]{r.contour}
}
