% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.contour.r
\name{as.contour,GRaster-method}
\alias{as.contour,GRaster-method}
\alias{as.contour}
\title{Contour lines from a "GRaster"}
\usage{
\S4method{as.contour}{GRaster}(x, nlevels, levels)
}
\arguments{
\item{x}{A \code{GRaster}.}

\item{nlevels}{Numeric: A positive integer or missing (default). Number of levels at which to calculate contours. Levels will be calculated in equal-sized steps from the smallest to the largest value of \code{x}. Either \code{nlevels} or \code{levels} must be specified.}

\item{levels}{Numeric vector: A numeric vector of values at which to calculate contour lines. Either \code{nlevels} or \code{levels} must be specified.}
}
\description{
Create a \code{GVector} of contour lines from a \code{GRaster}.
}
\examples{
\dontrun{
# NB This example is in a "dontrun{}" block because it requires users to have
# GRASS GIS Version 8+ installed on their system.

# IMPORTANT #1: If you already have a GRASS session started, you will need to
# run the line below and the last line in this example to work with it again.
# If you have not started a GRASS session, you can skip this step and go to
# step #2.
opts. <- getFastOptions()

# IMPORTANT #2: Select the appropriate line below and change as necessary to
# where GRASS is installed on your system.
grassDir <- "/Applications/GRASS-8.3.app/Contents/Resources" # Mac
grassDir <- "C:/Program Files/GRASS GIS 8.3" # Windows
grassDir <- "/usr/local/grass" # Linux

# Setup
library(terra)

# Elevation raster
madElev <- fastData("madElev")

# Start GRASS session for examples only:
faster(x = madElev, grassDir = grassDir,
workDir = tempdir(), location = "examples") # line only needed for examples

# Convert a SpatRaster to a GRaster:
elev <- fast(madElev)

# Calculate contour lines:
conts <- as.contour(elev, nlevels = 10)

plot(madElev)
plot(conts, add = TRUE)

# IMPORTANT #3: Revert back to original GRASS session if needed.
restoreSession(opts.)
removeSession("examples")

}
}
\seealso{
\code{\link[terra:contour]{terra::as.contour()}}, module \code{r.contour} in \strong{GRASS}
}
