% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.r
\name{as.data.table,GRaster-method}
\alias{as.data.table,GRaster-method}
\alias{as.data.table}
\alias{as.data.frame,GRaster-method}
\alias{as.data.frame}
\alias{as.data.frame,GVector-method}
\alias{as.data.table,GVector-method}
\title{Convert GRaster or GVector to a data frame}
\usage{
\S4method{as.data.table}{GRaster}(x, na.rm = TRUE, labels = TRUE, xy = FALSE, cells = FALSE)

\S4method{as.data.frame}{GRaster}(x, na.rm = TRUE, labels = TRUE, xy = FALSE, cells = FALSE)

\S4method{as.data.frame}{GVector}(x)

\S4method{as.data.table}{GVector}(x)
}
\arguments{
\item{x}{A \code{GRaster} or \code{GVector}.}

\item{na.rm}{Logical: Indicates whether to remove \code{NA} values (default is \code{TRUE}; \code{GRaster}s only)`.}

\item{labels}{Logical: If \code{TRUE}, a "factor" (categorical) \code{GRaster} will have its factor level labels returned. If \code{FALSE}, the integer codes are returned. Default is \code{TRUE} (for \code{GRaster}s only, and only has an effect if the \code{GRaster} is of type "factor").}

\item{xy}{Logical: If \code{TRUE}, return coordinates of cell centers (default is \code{FALSE}, for \code{GRaster}s only).}

\item{cells}{Logical: If \code{TRUE}, return cell columns and rows (default is \code{FALSE}, for \code{GRaster}s only).}
}
\value{
A \code{data.frame} or \code{NULL} (if the \code{GVector} has no data table, or if \code{GRaster}'s values are all \code{NA}).
}
\description{
Convert a \code{GRaster} to a table, or a \code{GVector}'s data table to a \code{data.frame} or \code{data.table}.
}
\examples{
if (grassStarted()) {

# Setup
library(sf)
library(terra)

# Example data
madElev <- fastData("madElev")
madForest2000 <- fastData("madForest2000")
madCoast0 <- fastData("madCoast0")
madRivers <- fastData("madRivers")
madDypsis <- fastData("madDypsis")

### GRaster properties

# convert SpatRasters to GRasters
elev <- fast(madElev)
forest <- fast(madForest2000)

# plot
plot(elev)

dim(elev) # rows, columns, depths, layers
nrow(elev) # rows
ncol(elev) # columns
ndepth(elev) # depths
nlyr(elev) # layers

res(elev) # resolution

ncell(elev) # cells
ncell3d(elev) # cells (3D rasters only)

topology(elev) # number of dimensions
is.2d(elev) # is it 2D?
is.3d(elev) # is it 3D?

minmax(elev) # min/max values

# name of object in GRASS
sources(elev)

# "names" of the object
names(elev)

# coordinate reference system
crs(elev)

# extent (bounding box)
ext(elev)

# data type
datatype(elev)

### operations on GRasters

# assigning
pie <- elev
pie[] <- pi # assign all cells to the value of pi
pie

# concatenating multiple GRasters
rasts <- c(elev, forest)
rasts

# number of layers
nlyr(elev)

# adding a raster "in place"
add(rasts) <- ln(elev)
rasts

# subsetting
rasts[[1]]
rasts[["madForest2000"]]

# assigning
rasts[[4]] <- elev > 500

# names
names(rasts)
names(rasts) <- c("elev_meters", "forest", "ln_elev", "high_elevation")
rasts

# converting to data.tables/data.frames
as.data.table(rasts)
head(as.data.frame(rasts))

### GVector properties

# convert sf vectors to GVectors
coast <- fast(madCoast4)
rivers <- fast(madRivers)
dypsis <- fast(madDypsis)

# extent
ext(rivers)

W(rivers) # western extent
E(rivers) # eastern extent
S(rivers) # southern extent
N(rivers) # northern extent
top(rivers) # top extent (NA for 2D rasters like this one)
bottom(rivers) # bottom extent (NA for 2D rasters like this one)

# coordinate reference system
crs(rivers)
st_crs(rivers)

# column names and data types
names(coast)
datatype(coast)

# name of object in GRASS
sources(rivers)

# points, lines, or polygons?
geomtype(dypsis)
geomtype(rivers)
geomtype(coast)

is.points(dypsis)
is.points(coast)

is.lines(rivers)
is.lines(dypsis)

is.polygons(coast)
is.polygons(dypsis)

# dimensions
nrow(rivers) # how many spatial features
ncol(rivers) # hay many columns in the data frame

# number of geometries and sub-geometries
ngeom(coast)
nsubgeom(coast)

# 2- or 3D
topology(rivers) # dimensionality
is.2d(elev) # is it 2D?
is.3d(elev) # is it 3D?

# Update values from GRASS
# (Reads values from GRASS... will not appear to do anything in this case)
coast <- update(coast)

### operations on GVectors

# convert to data frame
as.data.frame(rivers)
as.data.table(rivers)

# subsetting
rivers[c(1, 3)] # select 2 rows/geometries
rivers[-3] # remove row/geometry 3
rivers[ , 1] # column 1
rivers[ , "TopElev"] # select column
rivers[["TopElev"]] # select column
rivers[1, 2:3] # row/geometry 1 and column 2 and 3
rivers[c(TRUE, FALSE)] # select every other geometry (T/F vector is recycled)
rivers[ , c(TRUE, FALSE)] # select every other column (T/F vector is recycled)

# removing data table
noTable <- dropTable(rivers)
noTable
nrow(rivers)
nrow(noTable)

# Refresh values from GRASS
# (Reads values from GRASS... will not appear to do anything in this case
# since the rivers object is up-to-date):
rivers <- update(rivers)

# Concatenating multiple vectors
rivers2 <- rbind(rivers, rivers)
dim(rivers)
dim(rivers2)

}
}
\seealso{
\code{\link[terra:as.data.frame]{terra::as.data.frame()}}, \code{\link[data.table:as.data.table]{data.table::as.data.table()}}
}
