% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.points.r
\name{as.points,GVector-method}
\alias{as.points,GVector-method}
\alias{as.points}
\alias{as.points,GRaster-method}
\title{Convert a raster or lines or polygons vector to a points vector}
\usage{
\S4method{as.points}{GVector}(x)

\S4method{as.points}{GRaster}(x, values = TRUE)
}
\arguments{
\item{x}{A \code{GRaster}, \code{GVector}.}

\item{values}{Logical: If \code{TRUE} (default), create an attribute table with raster cell values, with one row per point.}
}
\value{
A \code{points} \code{GVector}.
}
\description{
'as.points()' converts a \code{GRaster} or \code{GVector} to a points \code{GVector}. For rasters, the points have the coordinates of cell centers and are assigned the cells' values. Only non-\code{NA} cells will be converted to a point. For vectors, each point will have the attributes of the line or polygon to which it belonged. Points are extracted from each vertex.
}
\examples{
\dontrun{
# NB This example is in a "dontrun{}" block because it requires users to have
# GRASS GIS Version 8+ installed on their system.

# IMPORTANT #1: If you already have a GRASS session started, you will need to
# run the line below and the last line in this example to work with it again.
# If you have not started a GRASS session, you can skip this step and go to
# step #2.
opts. <- getFastOptions()

# IMPORTANT #2: Select the appropriate line below and change as necessary to
# where GRASS is installed on your system.
grassDir <- "/Applications/GRASS-8.3.app/Contents/Resources" # Mac
grassDir <- "C:/Program Files/GRASS GIS 8.3" # Windows
grassDir <- "/usr/local/grass" # Linux

# Setup
library(sf)
library(terra)

# Elevation raster, outline of a part of Madagascar, and rivers vector:
madElev <- fastData("madElev")
madCoast0 <- fastData("madCoast0")
madRivers <- fastData("madRivers")

# Start GRASS session for examples only:
faster(x = madElev, grassDir = grassDir,
workDir = tempdir(), location = "examples") # line only needed for examples

# Convert to GRaster and GVectors:
elev <- fast(madElev)
coast <- fast(madCoast0)
rivers <- fast(madRivers)

# Extract points from raster. To make this fast, we will first crop the raster
# to a small extent defined by a single river.
river <- rivers[1]
elevCrop <- crop(elev, river)
elevPoints <- as.points(elevCrop)
elevPoints

plot(elevCrop)
plot(elevPoints, pch = '.', add = TRUE)

# Extract points from vectors:
coastPoints <- as.points(coast)
riversPoints <- as.points(rivers)

plot(coast)
plot(coastPoints, add = TRUE)

plot(rivers, col = "blue", add = TRUE)
plot(riversPoints, col = "blue", add = TRUE)

# IMPORTANT #3: Revert back to original GRASS session if needed.
restoreSession(opts.)
removeSession("examples")

}
}
\seealso{
\code{\link[=as.lines]{as.lines()}}, \code{\link[=as.polygons]{as.polygons()}}, \code{\link[terra:as.points]{terra::as.points()}}, and ** GRASS** modules \code{v.to.points} and \code{r.to.vect}
}
