% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.polygons.r
\name{as.polygons,GRaster-method}
\alias{as.polygons,GRaster-method}
\alias{as.polygons}
\title{Convert a raster to a polygons vector}
\usage{
\S4method{as.polygons}{GRaster}(x, round = TRUE, smooth = FALSE)
}
\arguments{
\item{x}{A \code{GRaster}. If more than one layer is in the \code{GRaster}, only the first will be used (with a warning).}

\item{round}{Logical: If \code{TRUE} (default), values in the raster will be rounded first before conversion to a vector. This causes cells that are adjacent that have the same (rounded) values to be combined into a single polygon. For more control, see \code{\link[=clump]{clump()}}.}

\item{smooth}{Logical: If \code{TRUE}, round the corners of square features. Default is \code{FALSE}.}
}
\value{
A \code{GVector}.
}
\description{
\code{\link[=as.polygons]{as.polygons()}} converts a \code{GRaster} to a "polygons" \code{GVector}. After running this function, \code{\link[=cleanGeom]{cleanGeom()}} may be useful to use to "tidy up" the vector.
}
\examples{
\dontrun{
# NB This example is in a "dontrun{}" block because it requires users to have
# GRASS GIS Version 8+ installed on their system.

# IMPORTANT #1: If you already have a GRASS session started, you will need to
# run the line below and the last line in this example to work with it again.
# If you have not started a GRASS session, you can skip this step and go to
# step #2.
opts. <- getFastOptions()

# IMPORTANT #2: Select the appropriate line below and change as necessary to
# where GRASS is installed on your system.
grassDir <- "/Applications/GRASS-8.3.app/Contents/Resources" # Mac
grassDir <- "C:/Program Files/GRASS GIS 8.3" # Windows
grassDir <- "/usr/local/grass" # Linux

# Setup
library(terra)

# Elevation
madElev <- fastData("madElev")

# Start GRASS session for examples only:
faster(x = madElev, grassDir = grassDir,
workDir = tempdir(), location = "examples") # line only needed for examples

# Convert SpatRaster to GRaster:
elev <- fast(madElev)

# To speed things up, first group cells of similar value:
elevClumps <- clump(elev, minDiff = 0.0115)

# Convert to polygons:
rastToPolys <- as.polygons(elevClumps)
plot(rastToPolys)

# IMPORTANT #3: Revert back to original GRASS session if needed.
fastRestore(opts.)
removeSession("examples")

}
}
\seealso{
\code{\link[=as.points]{as.points()}}, \code{\link[=as.lines]{as.lines()}}, \code{\link[terra:as.polygons]{terra::as.polygons()}}, \code{\link[=cleanGeom]{cleanGeom()}}, and \strong{GRASS} module \code{r.to.vect}
}
