% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.r
\name{c,GRaster-method}
\alias{c,GRaster-method}
\alias{c}
\alias{c,GVector-method}
\title{"Stack" GRasters and combine GVectors}
\usage{
\S4method{c}{GRaster}(x, ...)

\S4method{c}{GVector}(x, ..., df = FALSE)
}
\arguments{
\item{x}{A \code{GRaster} or a \code{GVector}.}

\item{...}{One or more \code{GRaster}s, one or more \code{GVector}s, a list of \code{GRaster}s, or a list of \code{GVector}s. You can use a mix of lists and individual rasters or vectors.}

\item{df}{If \code{TRUE}, copy the data frame of the \code{x} \code{GVector} to the output. If \code{FALSE} (default), the output will not have a data frame.}
}
\value{
A \code{GRaster} or \code{GVector}.
}
\description{
\code{GRaster}s can be "stacked" using this function, effectively creating a multi-layered raster. Note that this is different from creating a 3-dimensional raster, though such an effect can be emulated using stacking. \code{GVector}s can be combined into a single vector.  Stacks can only be created when:
\itemize{
\item All objects are in the same \strong{GRASS} \link[=tutorial_locations]{location and mapset}.
\item All objects are the same class (either all \code{GRaster}s or all \code{GVector}s).
\item For \code{GRaster}s:
\itemize{
\item Horizontal extents are the same.
\item Horizontal dimensions are the same.
\item The topology (2- or 3-dimensional) must be the same. If 3D, then all rasters must have the same number of depths and vertical extents.
}
\item For \code{GVector}s:
\itemize{
\item The geometry (points, lines, or polygons) must be the same.
If features (boundaries, lines, etc.) of \code{GVector}s that are combined are identical or nearly identical, they can be cleaned using \code{\link[=removeDupNodes]{removeDupNodes()}} and \code{\link[=snap]{snap()}}.
}
}
}
\examples{
\dontrun{
# NB This example is in a "dontrun{}" block because it requires users to have
# GRASS GIS Version 8+ installed on their system.

# IMPORTANT #1: If you already have a GRASS session started, you will need to
# run the line below and the last line in this example to work with it again.
# If you have not started a GRASS session, you can skip this step and go to
# step #2.

# IMPORTANT #2: Select the appropriate line below and change as necessary to
# where GRASS is installed on your system.
grassDir <- "/Applications/GRASS-8.2.app/Contents/Resources" # Mac
grassDir <- 'C:/Program Files/GRASS GIS 8.2' # Windows
grassDir <- '/usr/local/grass' # Linux

# setup
library(sf)
library(terra)

# example data
madElev <- fastData('madElev')
madCoast0 <- fastData('madCoast0')
madRivers <- fastData('madRivers')
madDypsis <- fastData('madDypsis')

# start GRASS session for examples only
wd <- forwardSlash(tempdir()) # only for examples

faster(crs = madElev, grassDir = grassDir,
workDir = wd, location = 'examples') # line only needed for examples

### GRaster properties
######################

# convert SpatRasters to GRasters
elev <- fast(madElev)

dim(elev) # rows, columns, depths, layers
nrow(elev) # rows
ncol(elev) # columns
ndepth(elev) # depths
nlyr(elev) # layers

res(elev) # resolution

ncell(elev) # cells
ncell3d(elev) # cells (3D rasters only)

topology(elev) # number of dimensions
is.2d(elev) # is it 2D?
is.3d(elev) # is it 3D?

minmax(elev) # min/max values

# information on the GRASS session in which the GRaster is located
location(elev) # location
mapset(elev) # mapset

# "gnames" of the object (its name in GRASS)
gnames(elev)

# coordinate reference system
crs(elev)
st_crs(elev)

# extent (bounding box)
ext(elev)
st_bbox(elev)

# data type
datatype(elev)

# names
names(elev)

### GVector properties
######################

# convert sf vectors to GVectors
coast <- fast(madCoast4)
rivers <- fast(madRivers)
dypsis <- fast(madDypsis)

# extent
ext(rivers)
st_bbox(rivers) # extent

# coordinate reference system
crs(rivers)
st_crs(rivers)

# column names and data types
names(coast)
datatype(coast)

# session information
location(rivers) # GRASS location
mapset(rivers) # GRASS mapset

# points, lines, or polygons?
geomtype(dypsis)
geomtype(rivers)
geomtype(coast)

is.points(dypsis)
is.points(coast)

is.lines(rivers)
is.lines(dypsis)

is.polygons(coast)
is.polygons(dypsis)

# dimensions
nrow(rivers) # how many spatial features
ncol(rivers) # hay many columns in the data frame

# 2- or 3D
topology(rivers) # dimensionality
is.2d(elev) # is it 2D?
is.3d(elev) # is it 3D?

### operations on GVectors
##########################

# convert to data frame
as.data.frame(rivers)

# IMPORTANT #3: Revert back to original GRASS session if needed.
fastRestore(opts.)
removeSession('examples')

}
}
