% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colbind.r
\name{colbind,GVector-method}
\alias{colbind,GVector-method}
\alias{colbind}
\alias{colbind,data.frame-method}
\alias{colbind,matrix-method}
\alias{colbind,data.table-method}
\title{Add columns to the data table of a GVector}
\usage{
\S4method{colbind}{GVector}(x, ...)

\S4method{colbind}{data.frame}(x, ...)

\S4method{colbind}{matrix}(x, ...)

\S4method{colbind}{data.table}(x, ...)
}
\arguments{
\item{x}{A \code{GVector}, a \code{data.frame}, or a \code{data.table}. If \code{x} is a \code{GVector}, then subsequent items must be \code{data.frame}s, \code{data.table}s, or \code{matrices}.}

\item{...}{A \code{GVector}, or \code{data.frame}s, \code{data.table}s, and/or \code{matrices}. If \code{x} is a \code{data.frame}, \code{data.table}, or \code{matrix}, then at least one argument in \code{...} must be a \code{GVector}.}
}
\value{
A \code{GVector}.
}
\description{
\code{colbind()} is like \code{\link[=cbind]{cbind()}}, except that it adds columns to the data table of a \code{GVector}. To combine rows, see \code{\link[=c]{c()}}. You combine multiple \code{data.frame}s, \code{data.table}s, or \code{matrices}, but you can only combine them with one \code{GVector} at a time.
}
\examples{
if (grassStarted()) {

# Setup
library(sf)

# Rivers vector
madRivers <- fastData("madRivers")

# Convert sf to a GVector
rivers <- fast(madRivers)

# Subset rivers vector
rivers1 <- rivers[1:2]
rivers2 <- rivers[10:11]

# Concatenate rivers
riversCombo <- c(rivers1, rivers2)
riversCombo

# Add columns
newCol <- data.frame(x = 1:11)
riversCols <- colbind(rivers, newCol)
riversCol

as.data.frame(riversCol)

}
}
\seealso{
\code{\link[=c]{c()}}
}
