% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparable.r
\name{comparable,GRaster,GRaster-method}
\alias{comparable,GRaster,GRaster-method}
\alias{comparable}
\alias{comparable,GRaster,GVector-method}
\alias{comparable,GVector,GRaster-method}
\alias{comparable,GVector,GVector-method}
\title{Determine if two GRasters and/or GVectors are geographically comparable}
\usage{
\S4method{comparable}{GRaster,GRaster}(
  x,
  y,
  fail = TRUE,
  warn = TRUE,
  extent = TRUE,
  zextent = TRUE,
  dim = TRUE,
  dim3d = TRUE,
  res = TRUE,
  res3d = TRUE,
  topo = TRUE,
  nlayers = FALSE
)

\S4method{comparable}{GRaster,GVector}(
  x,
  y,
  fail = TRUE,
  warn = TRUE,
  extent = FALSE,
  zextent = FALSE,
  topo = FALSE
)

\S4method{comparable}{GVector,GRaster}(
  x,
  y,
  fail = TRUE,
  warn = TRUE,
  extent = FALSE,
  zextent = FALSE,
  topo = FALSE
)

\S4method{comparable}{GVector,GVector}(
  x,
  y,
  fail = TRUE,
  warn = TRUE,
  extent = FALSE,
  zextent = FALSE,
  topo = TRUE,
  geometry = FALSE
)
}
\arguments{
\item{x, y}{\code{GRaster}s or \code{GVector}s.}

\item{fail}{Logical: If \code{TRUE} (default), throw an error with an explanation if the objects are not comparable. If \code{FALSE} (default), return \code{TRUE} or \code{FALSE}.}

\item{warn}{Logical: If \verb{TRUE (default), display a warning if a condition is not met. This only comes into effect if }fail\code{is}TRUE`.}

\item{extent}{Logical: If \code{TRUE}, test for same extent. By default, is \code{TRUE} for raster-raster comparison and \code{FALSE} for all others.}

\item{zextent}{Logical: Test for same vertical extents (3D only). By default, this is \code{TRUE} for raster-raster comparisons, and \code{FALSE} for all others.}

\item{res, res3d}{Logical (raster-raster only): If \code{TRUE} (default), test for same resolution in 2D and 3D.}

\item{topo}{Logical: Test for same topology (2D or 3D). By default, this is \code{TRUE} for raster-raster comparisons, and \code{FALSE} for all others.}

\item{nlayers}{Logical (raster-raster comparison only): If \code{FALSE} (default), do not compare number of layers. If \code{TRUE}, then return \code{FALSE} or fail if the number of layers in each raster is different.}

\item{geometry}{Logical (vector-vector comparison only): Compare geometry. Default is \code{FALSE}.}

\item{dim.dim3d}{Logical (raster-raster only): If \code{TRUE} (default), test for same dimensions in 2D and 3D.}
}
\value{
Logical (invisibly), or side effect of throwing an error.
}
\description{
\code{comparable()} compares geographic metadata between two \code{GRaster}s, two \code{GVector}s, or a \code{GRaster} and a \code{GVector}. In many cases, spatial objects must be comparable for them to "interact" (e.g., conducting arithmetic operations, masking, etc.).
}
