% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convHull.r
\name{convHull,GVector-method}
\alias{convHull,GVector-method}
\alias{convHull}
\title{Minimum convex hull around a spatial vector}
\usage{
\S4method{convHull}{GVector}(x)
}
\arguments{
\item{x}{A \code{GVector}.}
}
\value{
A \code{GVector}.
}
\description{
Create a minimum convex hull around a spatial vector.
}
\seealso{
[terra::convHull(), \code{sf::st_convex_hull()}, \strong{GRASS} module \code{v.hull}
}
