% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyGSpatial.r
\name{copyGSpatial,GRaster-method}
\alias{copyGSpatial,GRaster-method}
\alias{copyGSpatial}
\alias{copyGSpatial,GVector-method}
\alias{copyGSpatial,character-method}
\title{Make a copy of an object in GRASS}
\usage{
\S4method{copyGSpatial}{GRaster}(x)

\S4method{copyGSpatial}{GVector}(x)

\S4method{copyGSpatial}{character}(x, rastOrVect)
}
\arguments{
\item{x}{\code{GRaster}, \code{GVector}, or character: The object or the \code{gnames} of the object to be copied. Can take multi-layered objects or multiple \code{gnames}.}

\item{rastOrVect}{Character: Type of object to copy. Either \code{raster} or \code{vector}. Partial matching is used, and case does not matter.}
}
\value{
Character vector representing the new \code{gnames} of each object, plus makes a copy of the given object(s) in \strong{GRASS}.
}
\description{
Create a copy of a \code{GRaster} or \code{GVector} in \strong{GRASS}.  This function is used internally and is of little use to most users.  This only creates a copy of the object in the \strong{GRASS} session--to make a \code{GRaster} or \code{GVector}, \code{\link[=makeGRaster]{makeGRaster()}} or \code{\link[=makeGVector]{makeGVector()}} need to be called after making the copy. Note that if the object is multi-layered, then a copy is made of each layer.
}
