% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop.r
\name{crop,GRaster-method}
\alias{crop,GRaster-method}
\alias{crop}
\alias{crop,GVector-method}
\title{Remove parts of a GRaster or GVector}
\usage{
\S4method{crop}{GRaster}(x, y)

\S4method{crop}{GVector}(x, y)
}
\arguments{
\item{x}{A \code{GRaster} or \code{GVector} to be cropped.}

\item{y}{A \code{GRaster} or \code{GVector} to serve as a template for cropping.}
}
\value{
A \code{GRaster} or \code{GVector}.
}
\description{
[\verb{crop()] removes parts of a }GRaster\code{or}GVector` that fall "outside" another raster or vector.
}
\examples{
man/examples/ex_crop.r

}
\seealso{
\code{\link[terra:crop]{terra::crop()}}; \code{\link[sf:st_crop]{sf::st_crop()}}
}
