% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.r
\name{datatype}
\alias{datatype}
\title{Get the datatype of a 'GRaster'}
\usage{
datatype(x, type)
}
\arguments{
\item{x}{A \code{GRaster}.}

\item{format}{Either \code{'GRASS'} (default), \code{'terra'} (\strong{terra} package data types; see \code{\link[terra:writeRaster]{terra::writeRaster()}}), or \code{'GDAL'} (see \link[=https://gdal.org/user/raster_data_model.html#raster-band]{GDAL: Raster Band}).}
}
\value{
Character.
}
\description{
Returns the data type of a \code{GRaster}. Note that this may not necessarily be the optimal data type for ensuring the smallest file size given the values to be saved. Precedence is given to retaining data (vs. truncating values) over size on disk.
}
