% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_nrow_ncol_ncell.r
\name{nrow}
\alias{nrow}
\alias{ncol}
\alias{ndepth}
\alias{ncell}
\alias{ncell3d}
\title{Number of rows in a \code{GRaster}}
\usage{
ndepth(x)

ncell(x)

ncell3d(x)
}
\description{
Number of rows in a \code{GRaster}

Number of columns in a \code{GRaster}

Number of depths in a 3D \code{GRaster}

Number of cells in a \code{GRaster}

Number of cells in a 3D \code{GRaster}
}
