% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backdoor.r
\name{.backdoor}
\alias{.backdoor}
\title{Setup fasterRaster for ABS}
\usage{
.backdoor(ver = "84")
}
\arguments{
\item{ver}{Character: \strong{GRASS}: e.g., "83" or "84".}
}
\value{
\code{TRUE} (invisibly).
}
\description{
This is a secret function to be used for faster development of \strong{fasterRaster}. It calls \code{\link[=faster]{faster()}} to set the install directory for \strong{GRASS}, increases default memory, and number of cores. The function assumes development is on a Windows machine.
}
\keyword{internal}
