% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.proj.r
\name{.g.proj}
\alias{.g.proj}
\title{Call GRASS \code{g.proj} module}
\usage{
.g.proj()
}
\value{
Displays current projection information for the active "project/location" in \strong{GRASS}.
}
\description{
This function calls the \strong{GRASS} module \code{g.region} to display information on the projection of the current \strong{GRASS} "project".
}
\keyword{internal}
