% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rastInfo_vectInfo.r
\name{.rastInfo}
\alias{.rastInfo}
\title{Information on rasters and vectors in 'GRASS'}
\usage{
.rastInfo(gname)
}
\arguments{
\item{gname}{\code{GRASS} name for a raster, 3D raster, or vector.}
}
\value{
Reports extent, dimensions, resolution, bottom/top, etc.
}
\description{
Information on rasters and vectors in 'GRASS'
}
\keyword{internal}
