% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_rastInfo_vectInfo.r
\name{.rastInfo}
\alias{.rastInfo}
\title{Information on rasters and vectors in 'GRASS'}
\usage{
.rastInfo(gname)
}
\description{
Private. Reports extent, dimensions, resolution, bottom/top, etc.
}
\keyword{internal}
