% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rastInfo_vectInfo.r
\name{.rastInfo}
\alias{.rastInfo}
\title{Metadata on rasters and vectors in GRASS}
\usage{
.rastInfo(x)
}
\arguments{
\item{x}{A \code{GRaster}, \code{GVector}, or \code{sources}.}
}
\value{
Metadata on the extent, dimensions, resolution, bottom/top, etc. of rasters and vectors in \strong{GRASS}.
}
\description{
\code{.rastInfo()} and \code{.vectInfo()} fetch metadata on rasters and vectors in \strong{GRASS}. The \code{print()}, \code{show()}, and \code{summary()} functions can be used to display this metadata.
}
\keyword{internal}
