% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_sharedArgs.r
\name{.sharedArgs_grassToR}
\alias{.sharedArgs_grassToR}
\title{Shared argument(s)}
\arguments{
\item{grassToR}{Logical or \code{NULL}. If \code{TRUE} or \code{NULL} (default) then the output will be returned to \code{R}. If \code{FALSE}, then the output is left in the \code{GRASS} session (with the name given by \code{outGrassName}). You can set this argument for all functions at once using \code{\link{fasterOptions}}.}
}
\description{
Shared argument(s)
}
\keyword{internal}
