% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_sharedArgs.r
\name{.sharedArgs_inVectName}
\alias{.sharedArgs_inVectName}
\title{Shared argument(s)}
\arguments{
\item{inVectName}{The name of the input vector in \code{GRASS}. This is useful only if you are "chaining" \code{fasterRaster} functions together and wish to refer to the input vector later. The default value is \code{'vect'}.}
}
\description{
Shared argument(s)
}
