% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expanse.r
\name{expanse,GRaster-method}
\alias{expanse,GRaster-method}
\alias{expanse}
\title{Calculate area of non-NA cells}
\usage{
\S4method{expanse}{GRaster}(x, unit = "m")
}
\arguments{
\item{x}{A \code{GRaster}.}

\item{unit}{Character: Units of the output. Valid values include \code{"m"} (meters-squared), \code{"km"} (kilometers-squared), or \code{"ha"} (hectares). Partial matching is used.}
}
\value{
A \code{data.frame} with two columns, one for each layer's name and one for the summed area of all non-\code{NA} cells in the layer.
}
\description{
\code{expanse()} calculates the area of all non-\code{NA} cells in meters-squared.
}
