% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expanse.r
\name{expanse,GRaster-method}
\alias{expanse,GRaster-method}
\alias{expanse}
\title{Calculate area of non-NA cells}
\usage{
\S4method{expanse}{GRaster}(x, unit = "m")
}
\arguments{
\item{x}{A \code{GRaster}.}

\item{unit}{Character: Units of the output. Valid values include \code{"m"} or \verb{"meters"`` (meters-squared), }"km"\code{or}"kilometers"\verb{(kilometers-squared),}"ha"\code{or}"hectares"\verb{(hectares),}"mi"\code{or}"miles" (miles-squared), or \code{"ya"} or \code{"yards"} (yards-squared). Partial matching is used.}
}
\value{
A \code{data.frame} with two columns, one for each layer's name and one for the summed area of all non-\code{NA} cells in the layer.
}
\description{
\code{expanse()} calculates the area of all non-\code{NA} cells in meters-squared.
}
\examples{
\dontrun{
# NB This example is in a "dontrun{}" block because it requires users to have
# GRASS GIS Version 8+ installed on their system.

# IMPORTANT #1: If you already have a GRASS session started, you will need to
# run the line below and the last line in this example to work with it again.
# If you have not started a GRASS session, you can skip this step and go to
# step #2.
opts. <- getFastOptions()

# IMPORTANT #2: Select the appropriate line below and change as necessary to
# where GRASS is installed on your system.
grassDir <- "/Applications/GRASS-8.3.app/Contents/Resources" # Mac
grassDir <- "C:/Program Files/GRASS GIS 8.3" # Windows
grassDir <- "/usr/local/grass" # Linux

# setup
library(terra)

# elevation raster
madElev <- fastData("madElev")

# Start GRASS session for examples only.
faster(x = madElev, grassDir = grassDir,
workDir = tempdir(), location = "examples") # line only needed for examples

# Convert to fasterRaster format.
elev <- fast(madElev)

# Calculate area of non-NA cells:
expanse(elev)

# IMPORTANT #3: Revert back to original GRASS session if needed.
fastRestore(opts.)
removeSession("examples")

}
}
