% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportRastToGrass.r
\name{exportRastToGrass}
\alias{exportRastToGrass}
\title{Write raster to an existing GRASS session}
\usage{
exportRastToGrass(rast, vname = "rast", tempDir = raster::tmpDir())
}
\arguments{
\item{rast}{Raster layer.}

\item{vname}{Name of raster in GRASS.}

\item{tempDir}{Temporary directory.}
}
\value{
Nothing (exports a raster to a GRASS session so it can be used by other functions).
}
\description{
Write a raster to an existing GRASS session with handling for very big rasters.
}
