% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportRastToGrass.r
\name{exportRastToGrass}
\alias{exportRastToGrass}
\title{Export raster to an existing GRASS session}
\usage{
exportRastToGrass(rast, grassName = "rast")
}
\arguments{
\item{rast}{A \code{SpatRaster} raster from the \pkg{terra} package.}

\item{grassName}{What to name the raster in \pkg{GRASS}.}
}
\value{
Nothing (exports a raster to a \code{GRASS} session so it can be used by other functions).
}
\description{
Export a raster to an existing \code{GRASS} session.
}
\examples{
man/examples/ex_exportXToGrass.r

}
