% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportRastToGrass.r
\name{exportVectToGrass}
\alias{exportVectToGrass}
\title{Export a spatial vector to an existing GRASS session}
\usage{
exportVectToGrass(vect, inVectName = "vect")
}
\arguments{
\item{vect}{A \code{SpatVector} (\pkg{terra} package) or \code{sf} (\pkg{sf} package) spatial vector object.}

\item{inVectName}{Name of vector in \code{GRASS}. By default, this will be \code{'vect'}.}
}
\value{
\code{TRUE} (invisibly, if the vector was successfully exported). The function also exports a vector to a \code{GRASS} session so it can be used by other functions.
}
\description{
Export a spatial vector to an existing \code{GRASS} session.
}
\examples{
man/examples/ex_exportXToGrass.R

}
\seealso{
\code{\link[rgrass]{write_VECT}} in \pkg{rgrass}
}
