% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportVectToGrass.r
\name{exportVectToGrass}
\alias{exportVectToGrass}
\title{Write vector to an existing GRASS session}
\usage{
exportVectToGrass(vect, grassName = "vect", tempDir = raster::tmpDir())
}
\arguments{
\item{vect}{Raster layer.}

\item{grassName}{Name of vector in GRASS.}

\item{tempDir}{Temporary directory.}
}
\value{
Nothing (exports a vector to a GRASS session so it can be used by other functions).
}
\description{
Write a raster to an existing GRASS session with handling for very big rasters.
}
