% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext.r
\name{ext}
\alias{ext}
\title{Spatial extent of a 'GRaster' or 'GVector'}
\usage{
ext(x, vector)
}
\arguments{
\item{x}{A \code{GRaster} or \code{GVector}.}

\item{vector}{Logical: if \code{FALSE} (default), return a \code{SpatExtent} object. If \code{TRUE}, return a vector.}
}
\value{
A numeric vector.
}
\description{
Spatial extent of a 'GRaster' or 'GVector'.
}
\examples{
\dontshow{if (man/examples/example_slots.r) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontshow{\}) # examplesIf}
}
