% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.r
\name{extract,GRaster,GVector-method}
\alias{extract,GRaster,GVector-method}
\alias{extract}
\alias{extract,GRaster,data.frame-method}
\alias{extract,GRaster,data.table-method}
\alias{extract,GRaster,matrix-method}
\alias{extract,GRaster,numeric-method}
\alias{extract,GVector,GVector-method}
\alias{extract,GVector,data.frame-method}
\alias{extract,GVector,data.table-method}
\alias{extract,GVector,matrix-method}
\alias{extract,GVector,numeric-method}
\title{Extract values from a GRaster at locations in a points GVector}
\usage{
\S4method{extract}{GRaster,GVector}(x, y, xy = FALSE, cats = FALSE)

\S4method{extract}{GRaster,data.frame}(x, y, xy = FALSE, cats = FALSE)

\S4method{extract}{GRaster,data.table}(x, y, xy = FALSE, cats = FALSE)

\S4method{extract}{GRaster,matrix}(x, y, xy = FALSE, cats = FALSE)

\S4method{extract}{GRaster,numeric}(x, y, xy = FALSE, cats = FALSE)

\S4method{extract}{GVector,GVector}(x, y, xy = FALSE)

\S4method{extract}{GVector,data.frame}(x, y, xy = FALSE)

\S4method{extract}{GVector,data.table}(x, y, xy = FALSE)

\S4method{extract}{GVector,matrix}(x, y, xy = FALSE)

\S4method{extract}{GVector,numeric}(x, y, xy = FALSE)
}
\arguments{
\item{x}{A \code{GRaster} or \code{GVector}.}

\item{y}{A \code{points} \code{GVector}, \emph{or} a \code{data.frame} or \code{matrix} where the first two columns represent longitude and latitude (in that order), \emph{or} a two-element numeric vector where the first column represents longitude and the second latitude. Values of \code{x} will be extracted from the points in \code{y}.}

\item{xy}{Logical: If \code{TRUE}, also return the coordinates of each point. Default is \code{FALSE.}}

\item{cats}{Logical (extracting from a raster): If \code{TRUE} and \code{x} is a \link[=tutorial_raster_data_types]{categorical raster}, then return the category labels instead of the values. Default is `FALSE.}
}
\value{
A \code{data.frame} or \code{data.table}.
}
\description{
\code{extract()} obtains the values of a \code{GRaster} or \code{GVector} associated with the locations of a set of points.
}
\examples{
\dontrun{
# NB This example is in a "dontrun{}" block because it requires users to have
# GRASS GIS Version 8+ installed on their system.

# IMPORTANT #1: If you already have a GRASS session started, you will need to
# run the line below and the last line in this example to work with it again.
# If you have not started a GRASS session, you can skip this step and go to
# step #2.
opts. <- getFastOptions()

# IMPORTANT #2: Select the appropriate line below and change as necessary to
# where GRASS is installed on your system.
grassDir <- "/Applications/GRASS-8.3.app/Contents/Resources" # Mac
grassDir <- "C:/Program Files/GRASS GIS 8.3" # Windows
grassDir <- "/usr/local/grass" # Linux

# setup
library(sf)
library(terra)

# Example data: elevation raster and points vector
madElev <- fastData("madElev")
madCover <- fastData("madCover")
madCoast4 <- fastData("madCoast4")
madDypsis <- fastData("madDypsis")

# start GRASS session for examples only
faster(x = madElev, grassDir = grassDir,
workDir = tempdir(), location = "examples") # line only needed for examples

# Convert to fasterRaster formats:
elev <- fast(madElev)
cover <- fast(madCover)
coast <- fast(madCoast4)
dypsis <- fast(madDypsis)

# Get values of elevation at points where Dypsis species are located:
extract(elev, dypsis, xy = TRUE)

# Extract from categorical raster:
vals <- extract(cover, dypsis)
vals

cats <- extract(cover, dypsis, cats = TRUE)
cats

# Extract from a vector:
table <- extract(coast, dypsis, xy = TRUE)
head(table) # first 3 are outside vector, next 3 are inside

# IMPORTANT #3: Revert back to original GRASS session if needed.
restoreSession(opts.)
removeSession("examples")

}
}
\seealso{
\code{\link[terra:extract]{terra::extract()}}, and modules \code{r.what} and \code{v.what} in \strong{GRASS}
}
