% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.r
\name{extract,GRaster,GVector-method}
\alias{extract,GRaster,GVector-method}
\alias{extract}
\alias{extract,GRaster,data.frame-method}
\alias{extract,GRaster,matrix-method}
\alias{extract,GRaster,numeric-method}
\title{Extract values from a raster at locations in a points vector}
\usage{
\S4method{extract}{GRaster,GVector}(x, y, xy = FALSE, cat = FALSE)

\S4method{extract}{GRaster,data.frame}(x, y, xy = FALSE, cat = FALSE)

\S4method{extract}{GRaster,matrix}(x, y, xy = FALSE, cat = FALSE)

\S4method{extract}{GRaster,numeric}(x, y, xy = FALSE, cat = FALSE)
}
\arguments{
\item{x}{A \code{GRaster}.}

\item{y}{A \code{points} \code{GVector}, a \code{data.frame} or \code{matrix} where the first two columns represent longitude and latitude (in that order), or a two-element numeric vector where the first column represents longitude and the second latitude}

\item{xy}{Logical: If \code{TRUE}, also return the coordinates of each point. Default is \code{FALSE.}}

\item{cat}{Logical: If \code{TRUE}, report the category label rather than the cell value. Default is `FALSE.}
}
\value{
A \code{data.frame}.
}
\description{
\code{extract()} obtains the values of a raster associated with the locations of a points vector. Note that only points vectors are accommodated--lines and areas vectors cannot me used.
}
\examples{
\dontrun{
# NB This example is in a "dontrun{}" block because it requires users to have
# GRASS GIS Version 8+ installed on their system.

# IMPORTANT #1: If you already have a GRASS session started, you will need to
# run the line below and the last line in this example to work with it again.
# If you have not started a GRASS session, you can skip this step and go to
# step #2.
opts. <- getFastOptions()

# IMPORTANT #2: Select the appropriate line below and change as necessary to
# where GRASS is installed on your system.
grassDir <- "/Applications/GRASS-8.3.app/Contents/Resources" # Mac
grassDir <- "C:/Program Files/GRASS GIS 8.3" # Windows
grassDir <- "/usr/local/grass" # Linux

# setup
library(sf)
library(terra)

# Example data: elevation raster and points vector
madElev <- fastData("madElev")
madDypsis <- fastData("madDypsis")

# start GRASS session for examples only
faster(x = madElev, grassDir = grassDir,
workDir = tempdir(), location = "examples") # line only needed for examples

# Convert to fasterRaster format(s):
elev <- fast(madElev)
dypsis <- fast(madDypsis)

# Get values of elevation at points where Dypsis species are located:
vals <- extract(elev, dypsis, xy = TRUE)

# IMPORTANT #3: Revert back to original GRASS session if needed.
fastRestore(opts.)
removeSession("examples")

}
}
\seealso{
\code{\link[terra:extract]{terra::extract()}}, and modules \code{r.what} and \code{v.what} in \strong{GRASS}
}
