% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast.r
\name{fast}
\alias{fast}
\title{Export a raster or vector to 'GRASS'}
\usage{
fast(x, rastOrVect = NULL)
}
\arguments{
\item{x}{Any one of:
\itemize{
\item A \code{SpatRaster} or \code{stars} raster. Rasters can have one or more layers. They will retain their "layerdness" in most \strong{fasterRaster} functions.
\item A \code{SpatVector} or \code{sf} spatial vector.
\item A character string with the path and filename of a raster or vector to be loaded directly into \strong{GRASS}. The function will attempt to ascertain the type of object from the file extension (raster or vector), but it can help to indicate which it is using the \code{rastOrVect} argument if it is unclear.
}}

\item{rastOrVect}{Either \code{NULL} (default) or character (\code{raster} or \code{vector}). If \code{x} is a raster or vector already in \strong{R}, this does not need to be specified. However, if \code{x} is a filename, then the function will try to ascertain the data type from the filenames, but sometimes this will fail. In that case, it can help to specify if the file holds a raster or vector. Partial matching is used.}
}
\value{
A \code{GRaster} or \code{GVector}.
}
\description{
To use a raster or vector in with most \strong{fasterRaster} functions, you need to 1) initiate a \strong{GRASS} session with \code{\link[=fastStart]{fastStart()}}, 2) then use this function to either convert a raster or vector in \strong{R} to a \code{GRaster} or \code{GVector}, or load a raster or vector from disk directly into the \code{GRaster} or \code{GVector} format. \strong{GRASS} supports loading from disk a variety of \href{https://grass.osgeo.org/grass82/manuals/r.in.gdal.html}{raster} and \href{https://grass.osgeo.org/grass82/manuals/v.in.ogr.html}{vector} formats.
}
\seealso{
\code{\link[rgrass:readRAST]{rgrass::read_RAST()}} and \code{\link[rgrass:readVECT]{rgrass::read_VECT()}}; \strong{GRASS} modules \href{https://grass.osgeo.org/grass82/manuals/r.in.gdal}{r.in.gdal} and \href{https://grass.osgeo.org/grass82/manuals/r.in.gdal}{v.in.ogr}
}
