% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast.r
\name{fast,character-method}
\alias{fast,character-method}
\alias{fast}
\alias{fast,SpatRaster-method}
\alias{fast,SpatVector-method}
\alias{fast,sf-method}
\title{Create a GRaster or GVector}
\usage{
\S4method{fast}{character}(
  x,
  rastOrVect = NULL,
  method = NULL,
  fallback = TRUE,
  wrap = FALSE,
  warn = TRUE
)

\S4method{fast}{SpatRaster}(x, method = NULL, fallback = TRUE, wrap = FALSE, warn = TRUE)

\S4method{fast}{SpatVector}(x, warn = TRUE)

\S4method{fast}{sf}(x, warn = TRUE)
}
\arguments{
\item{x}{Any one of:
\itemize{
\item A \code{SpatRaster} raster. Rasters can have one or more layers. They will retain their "layerdness" in most \strong{fasterRaster} functions.
\item A \code{SpatVector} or \code{sf} spatial vector.
\item A character string with the path and filename of a raster or vector to be loaded directly into \strong{GRASS}. The function will attempt to ascertain the type of object from the file extension (raster or vector), but it can help to indicate which it is using the \code{rastOrVect} argument if it is unclear.
}}

\item{rastOrVect}{Either \code{NULL} (default) or character (\code{'raster'} or \code{'vector'}). If \code{x} is a raster or vector already in \strong{R}, this does not need to be specified. However, if \code{x} is a filename, then the function will try to ascertain whether it represents a raster or a vector, but sometimes this will fail. In that case, it can help to specify if the file holds a raster or vector. Partial matching is used.}

\item{method}{Character or \code{NULL} (rasters only): If \code{x} does not have the same coordinate reference system as the currently active \strong{GRASS} \link[=tutorial_sessions]{session}, then it will be projected when it is imported. You may need to specify which method is used to conduction the transformation. Partial matching is used.
\itemize{
\item \code{NULL} (default): Automatically choose based on raster properties (\code{near} for categorical data, \code{bilinear} for continuous data)
\item \code{'near'}: Nearest neighbor. Best for categorical data, and often a poor choice for continuous data.
\item \code{'bilinear'}: Bilinear interpolation (default for non-categorical data; uses weighted values from 4 cells).
\item \code{'bicubic'}: Bicubic interpolation (uses weighted values from 16 cells).
\item \code{'lanczos'}: Lanczos interpolation (uses weighted values from 25 cells).
}}

\item{fallback}{Logical (rasters only): If \code{TRUE} (default), then use "lower" resampling methods to fill in \code{NA} cells when a "higher" method is used. For example, if \code{method = 'bicubic'}, \code{NA} cells will be filled in using the \code{bilinear} method, except when that results in \code{NA}s, in which case the \code{near} method will be used. Fallback causes fewer cells to revert to \code{NA} values, so may be better at capturing complex "edges" (e.g., coastlines). Fallback does increase processing time because each "lower" method must be applied, then results merged.}

\item{wrap}{Logical (rasters only): When projecting rasters that "wrap around" (i.e., whole-world rasters or rasters that have edges that actually circle around to meet on the globe), \code{wrap} should be \code{TRUE} to avoid removing rows and columns from the "edge" of the map. The default is \code{FALSE}.}

\item{warn}{Logical: If \code{TRUE}, display a warning when projecting the vector or raster.}
}
\value{
A \code{GRaster} or \code{GVector}.
}
\description{
To use most \strong{fasterRaster} functions, you need to
\enumerate{
\item Initiate a \strong{GRASS} session with \code{\link[=faster]{faster()}};
\item Then use \code{fast()} to either:
\itemize{
\item Convert a \code{SpatRaster}, \code{SpatVector} or \code{sf} object to a \code{GRaster} or \code{GVector},
\item Load a raster or vector from a file directly into the \code{GRaster} or \code{GVector} format using \code{fast()}. \strong{GRASS} supports loading from disk a variety of raster formats (see the \strong{GRASS} manual page for \code{r.in.gdal}) and vector formats (see the \strong{GRASS} manual page for \code{v.in.ogr}), though not all of them will work with this function.
}
}
}
\details{
When projecting a raster, the "fallback" methods in \code{r.import} are actually used, even though the \code{method} argument takes the strings for non-fallback methods. See the manual page for the \code{r.import} \strong{GRASS} module.
}
\examples{
\dontrun{
# NB This example is in a "dontrun{}" block because it requires users to have
# GRASS GIS Version 8+ installed on their system.

# IMPORTANT #1: If you already have a GRASS session started, you will need to
# run the line below and the last line in this example to work with it again.
# If you have not started a GRASS session, you can skip this step and go to
# step #2.
opts. <- getFastOptions()

# IMPORTANT #2: Select the appropriate line below and change as necessary to
# where GRASS is installed on your system.

grassDir <- "/Applications/GRASS-8.2.app/Contents/Resources" # Mac
grassDir <- 'C:/Program Files/GRASS GIS 8.2' # Windows
grassDir <- '/usr/local/grass' # Linux

# setup
library(sf)
library(terra)

# example data
madElev <- fastData('madElev')
madRivers <- fastData('madRivers')

# start GRASS session for examples only
faster(x = madElev, grassDir = grassDir,
workDir = tempdir(), location = 'examples') # line only needed for examples

# convert a SpatRaster to a GRaster
elev <- fast(madElev)
elev

# convert an sf vector to a GVector
rivers <- fast(madRivers)
rivers

# IMPORTANT #3: Revert back to original GRASS session if needed.
fastRestore(opts.)
removeSession('examples')

}
}
\seealso{
\code{\link[rgrass:readRAST]{rgrass::read_RAST()}} and \code{\link[rgrass:readVECT]{rgrass::read_VECT()}}, plus \strong{GRASS} modules \code{r.in.gdal}, \code{r.import}, and \code{v.in.ogr}.
}
