% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastData.r
\name{fastData}
\alias{fastData}
\title{Get one of the example rasters or spatial vectors}
\usage{
fastData(x)
}
\arguments{
\item{x}{The name of the raster or spatial vector to get. All of these represent a portion of the eastern coast of Madagascar.

Spatial vectors (objects of class \code{sf} from the \strong{sf} `package):
\itemize{
\item `madCoast0: Outline of the region (polygon)
\item \code{madCoast4}: Outlines of the Fokontanies (Communes) of the region (polygons)
\item \code{madDypsis}: Records of plants of the genus \emph{Dypsis} (points)
\item \code{madRivers}: Major rivers (lines)
}

Rasters (objects of class \code{SpatRaster} from the \strong{terra} package):
\itemize{
\item \code{madChelsa}: Bioclimatic variables
\item \code{madElev}: Elevation
\item \code{madElevAnt}: Elevation of the Antanambe Commune
\item \code{madElevMan}: Elevation of the Manompana Commune
\item \code{madForest2000}: Forest cover in year 2000
\item \code{madForest2014}: Forest cover in year 2014
}}
}
\value{
A \code{SpatRaster} or \code{sf} spatial vector.
}
\description{
This function is a simple way to get example rasters or spatial vector datasets that come with \strong{fasterRaster}.
}
\seealso{
\link{madCoast0}, \link{madCoast4}, \link{madDypsis}, \link{madElev}, \link{madElevAnt}, \link{madElevMan}, \link{madForest2000}, \link{madForest2014}, \link{madRivers}
}
