% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastData.r
\name{fastData}
\alias{fastData}
\title{Get one of the example rasters or spatial vectors}
\usage{
fastData(x)
}
\arguments{
\item{x}{The name of the raster or spatial vector to get. All of these represent a portion of the eastern coast of Madagascar.

Spatial vectors (objects of class \code{sf} from the \strong{sf} `package):
\itemize{
\item `madCoast0: Outline of the region (polygon)
\item \code{madCoast4}: Outlines of the Fokontanies (Communes) of the region (polygons)
\item \code{madDypsis}: Records of plants of the genus \emph{Dypsis} (points)
\item \code{madRivers}: Major rivers (lines)
}

Rasters (objects of class \code{SpatRaster} from the \strong{terra} package):
\itemize{
\item \code{madChelsa}: Bioclimatic variables
\item \code{madCover}: Land cover (also see `madCoverCats)
\item \code{madElev}: Elevation
\item \code{madForest2000}: Forest cover in year 2000
\item \code{madForest2014}: Forest cover in year 2014
}

Data frames
\itemize{
\item \code{appFunsTable}: Table of functions usable by \code{\link[=app]{app()}}.
\item \code{madCoverCats}: Land cover values and categories.
}}
}
\value{
A \code{SpatRaster} or \code{sf} spatial vector.
}
\description{
This function is a simple way to get example rasters or spatial vector datasets that come with \strong{fasterRaster}.
}
\examples{

### vector data
library(sf)

# for vector data, we can use data(*) or fastData(*):
data(madCoast0) # same as next line
madCoast0 <- fastData("madCoast0") # same as previous
madCoast0

madCoast4 <- fastData("madCoast4")
madCoast4

madRivers <- fastData("madRivers")
madRivers

madDypsis <- fastData("madDypsis")
madDypsis

### raster data
library(terra)

# for raster data, we can get the file directly or using fastData(*):
rastFile <- system.file("extdata/madElev.tif", package="fasterRaster")
madElev <- terra::rast(rastFile)

madElev <- fastData("madElev") # same as previous two lines
madElev

madForest2000 <- fastData("madForest2000")
madForest2000

madForest2014 <- fastData("madForest2014")
madForest2014

madChelsa <- fastData("madChelsa")
madChelsa

madCover <- fastData("madCover")
madCover
madCover <- droplevels(madCover)
levels(madCover) # levels in the raster
ncat(madCover) # number of categories
}
\seealso{
\link{madCoast0}, \link{madCoast4}, \link{madCover}, \link{madCoverCats}, \link{madDypsis}, \link{madElev}, \link{madForest2000}, \link{madForest2014}, \link{madRivers}
}
