% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_options.r
\name{faster}
\alias{faster}
\title{Set or get options shared across "fasterRaster" functions}
\usage{
faster(..., default = FALSE, restore = FALSE)
}
\arguments{
\item{...}{Either a character (the name of an option), or an option and the value of the option using an \code{option = value} pattern. These include:
\itemize{
\item \code{grassDir} (character): The folder in which \strong{GRASS} is installed on your computer. Typically, this option is set when you run \code{\link[=faster]{faster()}}. Depending on your operating system, your install directory will look something like this:
\itemize{
\item Windows: \code{"C:/Program Files/GRASS GIS 8.3"}
\item Mac OS: \code{"/Applications/GRASS-8.3.app/Contents/Resources"}
\item Linux: \code{"/usr/local/grass"}
}
\item \code{addonsDir} (character): Folder in which \strong{GRASS} addons are stored. If \code{NA} and \code{grassDir} is not \code{NA}, this will be assumed to be \code{file.path(grassDir, "addons")}. The default values is \code{NA}.
\item \code{cores} (integer/numeric integer): Number of processor cores to use on a task. The default is 2. Some \strong{GRASS} modules are parallelized.
\item \code{memory} (integer/numeric): The amount of memory to allocate to a task, in GB, for \strong{GRASS}. The default is 1024 MB (i.e., 1 GB). Some \strong{GRASS} modules can take advantage of more memory.
\item \code{useDataTable} (logical): If \code{FALSE} (default), use \code{data.frame}s when going back and forth between data tables of \code{GVector}s and \strong{R}. This can be slow for very large data tables. If \code{TRUE}, use \code{data.table}s from the \strong{data.table} package. This can be much faster, but it might require you to know how to use \code{data.table}s if you want to manipulate them in \strong{R}. You can always convert them to \code{data.frame}s using \code{\link[base:as.data.frame]{base::as.data.frame()}}.
\item \code{verbose} (logical): If \code{TRUE}, show \strong{GRASS} messages and otherwise hidden slots in classes. This is mainly used for debugging, so most users will want to keep this at its default, \code{FALSE}.
\item \code{workDir} (character): The folder in which \strong{GRASS} rasters, vectors, and other objects are created and manipulated. By default, this is given by \code{\link[=tempdir]{tempdir()}}.
}}

\item{default}{Logical: Return the default value(s) of the option(s). The default value of \code{default} is \code{FALSE}.}

\item{restore}{Logical: If \code{TRUE}, the all options will be reset to their default values. The default is \code{FALSE}.}
}
\value{
If values of options changed, then a named list of option values \emph{before} they were changed is returned invisibly.

If option values are requested, a named list with option values is returned (not invisibly).
}
\description{
\code{faster()} either sets or gets options used across \strong{fasterRaster} functions. If the \code{default} argument is \code{TRUE}, then the default value(s) of options is returned (but value(s) are kept unchanged). If the \code{restore} argument is \code{TRUE}, then all options will be set to their default values. You cannot simultaneously set options and restore their default values.
}
\examples{
if (grassStarted()) {

# Setup
library(sf)
library(terra)

# Example data
madElev <- fastData("madElev")
madRivers <- fastData("madRivers")

# Convert a SpatRaster to a GRaster
elev <- fast(madElev)
elev

# Convert an sf vector to a GVector
rivers <- fast(madRivers)
rivers

}
}
