% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterContour.r
\name{fasterContour}
\alias{fasterContour}
\title{Contour vectors from a raster}
\usage{
fasterContour(
  rast,
  step = NULL,
  minlevel = NULL,
  maxlevel = NULL,
  levels = seq(minValue(rast), maxValue(rast), length.out = 5),
  cut = 2,
  grassDir = options()$grassDir,
  grassToR = TRUE,
  outGrassName = "contours",
  ...
)
}
\arguments{
\item{rast}{A \code{SpatRaster} or the name of a raster already in an existing \code{GRASS} session, with cell values representing elevation (typically in meters).}

\item{step}{Numeric. Increment between contour levels.}

\item{minlevel, maxlevel}{Numeric or \code{NULL} (default). Minimum and maximum contour levels. If \code{NULL} and \code{step} is not \code{NULL}, then the minimum and maximum values in the raster will be used.}

\item{levels}{Numeric vector. Levels of values in \code{rast} at which contours should be drawn. You can specify contour levels using this argument or by providing values for \code{step}, \code{minlevel}, and \code{maxlevel}. \code{levels} will override use of \code{step}, if both of them are specified.}

\item{cut}{Integer >= 0. Minimum number of points necessary to generate a contour line. A value of 0 implies no limit. Default is 2.}

\item{grassDir}{Name of the folder path in which \code{GRASS} is installed. For a Windows system, this might be something like \code{'C:/Program Files/GRASS GIS 8.2'}. For a Mac system, this might be like \code{"/Applications/GRASS-8.2.app/Contents/Resources"}. If you are using a lot of \pkg{fasterRaster} functions, it can be easier to set this once using \code{\link{fasterOptions}}. By default, this is \code{NULL}, in which case the install path for \code{GRASS} will be searched for, but if not found, will result in an error.}

\item{grassToR}{If \code{TRUE} (default) then the output will be returned to \pkg{R}. If \code{FALSE}, then the output is left in the \code{GRASS} session and named the value in \code{outGrassName}. You can then refer to the output in subsequent \pkg{fasterRaster} functions and avoid the need to re-import it. This is called "chaining" functions together, and it can save a lot of computational time.}

\item{outGrassName}{Character. Name of output in \code{GRASS}. This is useful if you want to refer to the output object in \code{GRASS} later in a session by "chaining" functions.}

\item{...}{Arguments to pass to \code{\link[rgrass]{execGRASS}} when calculating horizon height (i.e., function \code{r.horizon} in \code{GRASS}).}
}
\value{
If \code{grassToR} if \code{TRUE}, then a \code{SpatialLines} or \code{SpatialLinesDataFrame} object with the same coordinate reference system as \code{rast}. Regardless, a vector is written into the \code{GRASS} session. The name of this vector is given by \code{outGrassName}.
}
\description{
This function creates a vector (\code{SpatialLines} or \code{SpatialLinesDataFrame}) object from a raster representing contour lines in the raster. It utilizes the \code{GRASS} function \code{r.contour}.
}
\details{
See the documentation for the \code{GRASS} module \code{r.contour}{https://grass.osgeo.org/grass82/manuals/r.contour.html}.
}
\examples{
man/examples/ex_fasterContour.r

}
\seealso{
\code{\link[terra]{contour}}
}
