% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterCRS.r
\name{fasterCRS}
\alias{fasterCRS}
\title{Coordinate reference system of the current GRASS sesion}
\usage{
fasterCRS(nice = FALSE)
}
\arguments{
\item{nice}{If \code{TRUE}, then print the CRS in a formatted manner and return it invisibly. Default is \code{FALSE}.}
}
\value{
Nothing (displays information on the coordinate reference system of the current \code{GRASS} session).
}
\description{
Display information on the coordinate reference system of the current \code{GRASS} sesion.
}
\examples{
\dontrun{

library(sf)
library(terra)

# start a GRASS session with a raster
madElev <- fasterData('madElev')
initGrass(madElev, inRastName = 'madElev')

# import a vector
madCoast0 <- fasterData('madCoast0')
exportVectToGrass(madCoast0, inVectName = 'madCoast0')

# get information on the raster and vector
fasterInfoRast('madElev')
fasterInfoVect('madCoast0')

# information on the coordinate reference system
fasterCRS() # "plain" format
fasterCRS(TRUE) # nice for the eyes

# spatial extent of the GRASS session
fasterExt() # returns a SpatExtent object
fasterExt(FALSE) # returns a vector

# spatial resolution
fasterRes()

# names of all rasters and vectors in the GRASS session
fasterLs()

}

}
\seealso{
\href{https://grass.osgeo.org/grass82/manuals/g.proj.html}{\code{g.list}} in \code{GRASS}
}
