% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterCRS.r
\name{fasterCRS}
\alias{fasterCRS}
\title{Coordinate reference system of the current GRASS sesion}
\usage{
fasterCRS(nice = FALSE)
}
\arguments{
\item{nice}{If \code{TRUE}, then print the CRS in a formatted manner and return it invisibly. Default is \code{FALSE}.}
}
\value{
Nothing (displays information on the coordinate reference system of the current \code{GRASS} session).
}
\description{
Display information on the coordinate reference system of the current \code{GRASS} sesion.
}
\examples{
\dontrun{

library(sf)
library(terra)

# start a GRASS session with a raster
madElev <- fasterData('madElev')
initGrass(madElev, inRastName = 'madElev')

# import a vector
madRivers <- fasterData('madRivers')
fasterVect(madRivers, inVectName = 'madRivers')

# what's in the GRASS session?
fasterLs('rasters') # what rasters?
fasterLs('vectors') # what vectors?
fasterLs() # what rasters and vectors?

# information on the raster and vector
fasterInfo('madElev')
fasterInfo('madCoast0')
fasterInfo() # everything

# information on the coordinate reference system of the GRASS session
fasterCRS() # "plain" format
fasterCRS(TRUE) # nice for the eyes

# spatial extent of the GRASS session
fasterExt('madElev') # extent of madElev raster
fasterExt('madRivers') # extent of madRivers vector
fasterExt() # maximum extent of all spatial objects

}
}
\seealso{
\href{https://grass.osgeo.org/grass82/manuals/g.proj.html}{\code{g.list}} in \code{GRASS}
}
