% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterData.r
\name{fasterData}
\alias{fasterData}
\title{Get one of the example rasters or spatial vectors}
\usage{
fasterData(x)
}
\arguments{
\item{x}{The name of the raster or spatial vector to get. All of these represent a portion of the eastern coast of Madagascar. \cr
Spatial vectors (objects of class \code{sf} from the \pkg{sf} package):
\itemize{
    \item    \code{'madCoast0'}: Outline of the region (polygon).
    \item    \code{'madCoast4'}: Outlines of the Fokontanies (Communes) of the region (polygons).
    \item    \code{'madDypsis'}: Records of plants of the genus \emph{Dypsis} (points).
    \item    \code{'madRivers'}: Major rivers (lines).
}
Rasters (objects of class \code{SpatRaster} from the \pkg{terra} package):
\itemize{
    \item    \code{'madChelsa'}: Bioclimatic variables (\code{SpatRaster}).
    \item    \code{'madElev'}: Elevation (\code{SpatRaster}).
    \item    \code{'madElevAnt'}: Elevation of the Antanambe Commune(\code{SpatRaster}).
    \item    \code{'madElevMan'}: Elevation of the Manompana Commune (\code{SpatRaster}).
    \item    \code{'madForest2000'}: Forest cover in year 2000 (\code{SpatRaster}).
    \item    \code{'madForest2014'}: Forest cover in year 2014 (\code{SpatRaster}).
}}
}
\value{
A \code{SpatRaster} or \code{sf} spatial vector.
}
\description{
This function is a simple way to get one the several example rasters or spatial vector datasets that come with \pkg{fasterRaster}. These include.
}
\examples{

### vector data
library(sf)

# for vector data, we can use data(*) or fasterData(*):
data(madCoast0) # same as next line
madCoast <- fasterData('madCoast0') # same as previous
plot(madCoast0)

madCoast <- fasterData('madCoast4')
plot(madCoast4, add = TRUE)

madRivers <- fasterData('madRivers')
plot(madRivers, add = TRUE)

### raster data
library(terra)

# for raster data, we can get the file directly or using fasterData(*):
rastFile <- system.file('extdata/madElev.tif'), package='fasterRaster')
madElev <- terra::rast(rastFile)

madElev <- fasterData('madElev') # same as previus two lines
plot(madElev)

madForest2000 <- fasterData('madForest2000')
plot(madForest2000, col = 'forestgreen', add = TRUE)

madForest2014 <- fasterData('madForest2014')
plot(madForest2014, col = 'green', add = TRUE)
}
\seealso{
\code{\link{madCoast0}}, \code{\link{madCoast4}}, \code{\link{madDypsis}}, \code{\link{madElev}}, \code{\link{madElevAnt}}, \code{\link{madElevMan}}, \code{\link{madForest2000}}, \code{\link{madForest2014}}, \code{\link{madRivers}}
}
