% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterExt.r
\name{fasterExt}
\alias{fasterExt}
\title{Spatial extent of a GRASS session}
\usage{
fasterExt(terra = TRUE)
}
\arguments{
\item{terra}{If \code{TRUE} (default), the object that is returned is an \code{\link[terra]{SpatExtent}} object. If \code{FALSE}, then the bounding coordinates are returned as a vector.}
}
\value{
An object of class \code{\link[terra]{SpatExtent}} or a numeric vector.
}
\description{
This function returns the spatial extent of a \code{GRASS} session. \code{GRASS} sessions must have an extent, and this is usually set by the first raster or vector that is imported into them. Any subsequent rasters or vectors are cropped to this extent (which may or may not be desirable). By default, the function returns an object of class \code{\link[terra]{SpatExtent}} from the \pkg{terra} package. \cr
Note that retrieving this information is somewhat of a hack, so it only returns the extent of the "default" location with the "PERMANENT" mapset. These will work for nearly all users of \pkg{fasterRaster}, but if you create non-typical locations or mapsets using \code{\link{initGrass}}, then this function may return erroneous results.
}
\examples{
\dontrun{

library(sf)
library(terra)

# start a GRASS session with a raster
madElev <- fasterData('madElev')
initGrass(madElev, inRastName = 'madElev')

# import a vector
madCoast0 <- fasterData('madCoast0')
exportVectToGrass(madCoast0, inVectName = 'madCoast0')

# get information on the raster and vector
fasterInfoRast('madElev')
fasterInfoVect('madCoast0')

# information on the coordinate reference system
fasterCRS() # "plain" format
fasterCRS(TRUE) # nice for the eyes

# spatial extent of the GRASS session
fasterExt() # returns a SpatExtent object
fasterExt(FALSE) # returns a vector

# spatial resolution
fasterRes()

# names of all rasters and vectors in the GRASS session
fasterLs()

}

}
\seealso{
\code{\link{fasterRes}}, \code{\link{fasterInfoRast}}, and \code{\link{fasterInfoVect}} in \pkg{fasterRaster};  \code{\link[terra]{SpatExtent}} and \code{\link[terra]{ext}} in the \code{\link[terra]{terra}} package
}
