% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterFunctions.r
\docType{data}
\name{fasterFunctions}
\alias{fasterFunctions}
\title{Table of functions in fasterRaster and equivalents in \pkg{GRASS} and terra}
\format{
An object of class \code{'data.table'}.
}
\usage{
data(fasterFunctions)
}
\description{
Table of functions in \pkg{fasterRaster} and equivalents in \pkg{GRASS} and \pkg{terra}, plus a link to the \pkg{GRASS} function.
}
\examples{
fasterHelp('fasterRasterize') # lookup by fasterRaster name
fasterHelp('rasterize', terra = TRUE) # lookup by terra name
fasterHelp('v.to.rast', grass = TRUE) # lookup by GRASS name
fasterHelp('st_buffer', sf = TRUE) # lookup by sf name
fasterHelp('st_buffer', terra = TRUE, sf = TRUE) # warning
fasterHelp('v.to', grass = TRUE, approx = TRUE) # approximate lookup

# The next two lines return the same thing, a table with all GRASS functions
# represented in fasterRaster, and their terra equivalents.
fasterHelp()
data(fasterFunctions)

}
\keyword{functions}
