% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterHelp.r
\name{fasterHelp}
\alias{fasterHelp}
\title{Lookup equivalent functions in fasterRaster, \code{GRASS}, and terra}
\usage{
fasterHelp(x = NULL, what = "fasterRaster", approx = FALSE)
}
\arguments{
\item{x}{Name of the function to lookup. This can be a function in \pkg{fasterRaster}, \pkg{terra}, or \pkg{GRASS}. By default, it will be assumed to be the \pkg{fasterRaster} function. If this is \code{NULL} (default), it will return the entire crosswalk table.}

\item{what}{Which sofware/package does the function appear in?  Partial matches are OK, and case does not matter. This can be any of:
\itemize{
\item \code{'fasterRaster'}: Default
\item \code{'terra'}
\item \code{'GRASS'}
\item \code{'sf'}
}}
}
\value{
Nothing (displays a equivalent functions on the screen).
}
\description{
Lookup equivalent functions in fasterRaster, \pkg{GRASS}, and terra, plus get the URL of the help file for the \code{GRASS} function, if it exists.
}
\examples{
fasterHelp('fasterRasterize') # lookup by fasterRaster name
fasterHelp('rasterize', terra = TRUE) # lookup by terra name
fasterHelp('v.to.rast', grass = TRUE) # lookup by GRASS name
fasterHelp('st_buffer', sf = TRUE) # lookup by sf name
fasterHelp('st_buffer', terra = TRUE, sf = TRUE) # warning
fasterHelp('v.to', grass = TRUE, approx = TRUE) # approximate lookup

# The next two lines return the same thing, a table with all GRASS functions
# represented in fasterRaster, and their terra equivalents.
fasterHelp()
data(fasterFunctions)

}
