% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterInfoRast.r
\name{fasterInfoRast}
\alias{fasterInfoRast}
\alias{fasterInfoVect}
\title{Information on a raster or vector in a \code{GRASS} session}
\usage{
fasterInfoRast(x, flags = NULL)

fasterInfoVect(x, flags = NULL)
}
\arguments{
\item{x}{The name of the raster or vector in the R session. This must be in quotes.}

\item{flags}{Either \code{NULL} (default) or flags to affect the information shown. Flags should be in quotes and without the preceeding "-" sign (example: \code{flags = 'e'}).}
}
\value{
Nothing (displays information on the raster or vector).
}
\description{
Displays information on a raster or vector already in an R session. For further
}
\examples{
\dontrun{

library(sf)
library(terra)

# start a GRASS session with a raster
madElev <- fasterData('madElev')
initGrass(madElev, inRastName = 'madElev')

# import a vector
madCoast0 <- fasterData('madCoast0')
exportVectToGrass(madCoast0, inVectName = 'madCoast0')

# get information on the raster and vector
fasterInfoRast('madElev')
fasterInfoVect('madCoast0')

# information on the coordinate reference system
fasterCRS() # "plain" format
fasterCRS(TRUE) # nice for the eyes

# spatial extent of the GRASS session
fasterExt() # returns a SpatExtent object
fasterExt(FALSE) # returns a vector

# spatial resolution
fasterRes()

# names of all rasters and vectors in the GRASS session
fasterLs()

}

}
\seealso{
\code{\link{fasterLs}}, \href{https://grass.osgeo.org/grass82/manuals/r.info.html}{\code{r.info}} and \href{https://grass.osgeo.org/grass82/manuals/v.info.html}{\code{v.info}} in \code{GRASS}
}
