% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterInfoRast.r
\name{fasterInfoRast}
\alias{fasterInfoRast}
\alias{fasterInfoVect}
\title{Information on a raster or vector in a GRASS session}
\usage{
fasterInfoRast(rastName, flags = NULL)

fasterInfoVect(rastName, flags = NULL)
}
\arguments{
\item{flags}{Either \code{NULL} (default) or flags to affect the information shown. Flags should be in quotes and without the preceeding "-" sign (example: \code{flags = 'e'}).}

\item{name}{The name of the raster or vector in the R session. This must be in quotes.}
}
\value{
Nothing (displays information on the raster or vector).
}
\description{
Displays information on a raster or vector already in an R session. For further
}
\examples{

\donttest{

rastFile <- system.file('extdata', 'madElev.tif', package='fasterRaster')
madElev <- rast(rastFile)

# start a GRASS session
initGrass(madElev, rastName = 'elev')
fasterInfoRast('madElev')

# start a GRASS session
exportVectToGrass(madCoast0, vectName = 'madCoast0')
fasterInfoVect('madCoast0')

}

}
