% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterInfoRast.r
\name{fasterInfoRast}
\alias{fasterInfoRast}
\alias{fasterInfoVect}
\title{Information on a raster or vector in a \code{GRASS} session}
\usage{
fasterInfoRast(rastName, flags = NULL)

fasterInfoVect(rastName, flags = NULL)
}
\arguments{
\item{flags}{Either \code{NULL} (default) or flags to affect the information shown. Flags should be in quotes and without the preceeding "-" sign (example: \code{flags = 'e'}).}

\item{name}{The name of the raster or vector in the R session. This must be in quotes.}
}
\value{
Nothing (displays information on the raster or vector).
}
\description{
Displays information on a raster or vector already in an R session. For further
}
\examples{
\dontrun{

library(sf)
library(terra)

# start a GRASS session with a raster and get information on it
madElev <- fasterData('madElev')
initGrass(madElev, rastName = 'madElev')
fasterInfoRast('madElev')

# start a GRASS session with a vector and get information on it
madCoast0 <- fasterData('madCoast0')
exportVectToGrass(madCoast0, vectName = 'madCoast0')
fasterInfoVect('madCoast0')

}

}
\seealso{
\href{https://grass.osgeo.org/grass82/manuals/r.info.html}{\code{r.info}} and \href{https://grass.osgeo.org/grass82/manuals/v.info.html}{\code{v.info}} in \code{GRASS}
}
