% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterLs.r
\name{fasterLs}
\alias{fasterLs}
\title{Names of all rasters and/or spatial vectors in a GRASS session}
\usage{
fasterLs(type = c("raster", "vector"), incTemp = FALSE)
}
\arguments{
\item{type}{The type of spatial dataset for which to display names. This can include \code{'rasters'} (all rasters) and/or \code{'vectors'} (all spatial vectors). Partial matching is supported.}

\item{incTemp}{If \code{TRUE}, then remove any temporary files from the results. By default, these are not returned as part of the results. Some \pkg{fasterRaster} functions create temporary rasters or vectors which begin with the string "\code{__TEMPTEMP}".}
}
\value{
Nothing (displays names of rasters and/or vectors).
}
\description{
Displays the names of all rasters and/or vectors that have been exported to or created in a GRASS session.
}
\examples{
\dontrun{

library(sf)
library(terra)

# start a GRASS session with a raster
madElev <- fasterData('madElev')
initGrass(madElev, inRastName = 'madElev')

# import a vector
madCoast0 <- fasterData('madCoast0')
exportVectToGrass(madCoast0, inVectName = 'madCoast0')

# get information on the raster and vector
fasterInfoRast('madElev')
fasterInfoVect('madCoast0')

# information on the coordinate reference system
fasterCRS() # "plain" format
fasterCRS(TRUE) # nice for the eyes

# spatial extent of the GRASS session
fasterExt() # returns a SpatExtent object
fasterExt(FALSE) # returns a vector

# spatial resolution
fasterRes()

# names of all rasters and vectors in the GRASS session
fasterLs()

}

}
\seealso{
\code{\link{fasterInfoRast}}, \code{\link{fasterInfoVect}}, \href{https://grass.osgeo.org/grass82/manuals/g.list.html}{\code{g.list}} in \code{GRASS}
}
