% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterOptions.r
\name{fasterOptions}
\alias{fasterOptions}
\title{Set GRASS directory for fasterRaster functions}
\description{
Most functions in \pkg{fasterRaster} use the \code{grassDir} option for specifying where \pkg{GRASS} is installed. However, defining this argument every time can be cumbersome if you are using a lot of \code{fasterRaster} functions. Instead, you can define \code{grassDir} just once in \code{fasterOptions}, and subsequent calls to \pkg{fasterRaster} functions will automatically use that directory.
}
\section{Usage}{

To set the \code{grassDir} argument for all \pkg{fasterRaster} functions at once, simply do this: \cr\cr

\code{grassDir <- 'C:/Program Files/GRASS GIS 8.2'} \cr
\code{settings(grassDir = grassDir)} \cr\cr

To reset the \pkg{GRASS} path, simply do this: \cr \cr

\code{settings(grassDir = NULL)} \cr
}

\keyword{options}
