% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterRastDistance.r
\name{fasterRastDistance}
\alias{fasterRastDistance}
\title{Distance distance between cells with \code{NA}s and closest non-\code{NA} cells}
\usage{
fasterRastDistance(
  rast,
  metric = "euclidean",
  meters = TRUE,
  fillNAs = TRUE,
  grassDir = NULL,
  alreadyInGrass = FALSE,
  grassToR = TRUE,
  outGrassName = "distance",
  ...
)
}
\arguments{
\item{rast}{Either a raster or the name of a raster in an existing GRASS session.}

\item{metric}{Character, indicates type of distance to calculate:
\itemize{
\item \code{euclidean} (default): Euclidean distance
\item \code{geodesic}: geographic distance (suggested to use with \code{meters = TRUE}).
\item \code{squared}: Squared Euclidean distance
\item \code{maximum}: Maximum Euclidean distance
\item \code{manhattan}: Manhattan distance (i.e., "taxicab" distance, distance along cells going only north-south and east-west and never along a diagonal).
}}

\item{meters}{Logical, if \code{TRUE} then distance is in meters. If \code{FALSE} then distance is in map units.}

\item{fillNAs}{Logical, if \code{TRUE} (default) then fill code{NA} cells with distance between them and closest non-\code{NA}. If \code{TRUE} then replace value in non-{NA} cells with distance between them and nearest \code{NA} cell.}

\item{grassDir}{Character or \code{NULL} (default). Name of the directory in which GRASS is installed. Example: \code{'C:/Program Files/GRASS GIS 7.8'}. If this is \code{NULL}, R will search for the directory in which GRASS is installed. This usually fails, or if it succeeds, takes several minutes.}

\item{alreadyInGrass}{Logical, if \code{FALSE} (default) then start a new GRASS session and import the raster named in \code{rast}. If \code{FALSE}, use a raster already in GRASS with the name given by \code{rast}. The latter is useful if you are chaining \pkg{fasterRaster} functions together and the first function initializes the session. The first function should use \code{alreadyInGrass = FALSE} and subsequent functions should use \code{alreadyInGrass = TRUE} then use their \code{rast} (or \code{vect}) arguments to name the raster (or vector) that was made by the previous function.}

\item{grassToR}{Logical, if \code{TRUE} (default) then the product of the calculations will be returned to R. If \code{FALSE}, then the product is left in the GRASS session and named \code{distance}. The latter case is useful (and faster) when chaining several \pkg{fasterRaster} functions together.}

\item{outGrassName}{Character. Name of output in GRASS. This is useful if you want to refer to the output object in GRASS later in a session.}

\item{...}{Arguments to pass to \code{\link[rgrass7]{execGRASS}} when used for rasterizing (i.e., function \code{r.grow.distance} in GRASS).}
}
\value{
If \code{grassToR} if \code{TRUE}, then a raster with the same extent, resolution, and coordinate reference system as \code{vect}. Regardless, a raster with the name given by \code{outGrassName} is written into the GRASS session.
}
\description{
This function is a potentially faster version of the \code{\link[raster]{distance}} function in the \pkg{raster} package which it replaces values in all \code{NA} cells with the distance between them and the closest non-\code{NA} cell. Alternatively, it fills in values of non-\code{NA} cells with the distance between them and the closest \code{NA} cell. Note that the \code{distance} function also calculates distances between a raster and a spatial vector object, but this functionality is reproduced in \code{\link[fasterRaster]{fasterVectToRastDistance}}.
}
\details{
See the documentation for the GRASS module \code{r.grow.distance} at \url{https://grass.osgeo.org/grass78/manuals/r.grow.distance.html}.
}
\examples{
\donttest{
# change this according to where GRASS 7 is installed on your system
grassDir <- 'C:/Program Files/GRASS GIS 7.8' # example for a PC
grassDir <- "/Applications/GRASS-7.8.app/Contents/Resources" # for a Mac

data(madForest2000)

# could also use distance() function from the raster package which is
# slower in this example
distToForest <- fasterRastDistance(rast=madForest2000,
fillNAs=TRUE, grassDir=grassDir)
# distToForest <- distance(madForest2000)
par(mfrow=c(1, 2))
plot(madForest2000, 'Forest', col='forestgreen')
plot(distToForest, main='Distance to Forest (m)')
}
}
\seealso{
\code{\link[raster]{distance}}, \code{\link[fasterRaster]{fasterVectToRastDistance}}
}
