% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterRaster.r
\docType{package}
\name{fasterRaster}
\alias{fasterRaster-package}
\alias{fasterRaster}
\title{"fasterRaster": Faster raster and spatial vector processing using "GRASS GIS"}
\description{

}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.earthSkySea.org}
  \item \url{https://adamlilith.github.io/fasterRaster/}
  \item Report bugs at \url{https://github.com/adamlilith/fasterRaster/issues}
}

}
\author{
Adam B. Smith
}
\keyword{internal}
