% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterRaster.r
\docType{package}
\name{fasterRaster}
\alias{fasterRaster}
\title{'fasterRaster': Faster raster and spatial vector processing using 'GRASS GIS'}
\description{
Processing of large-in-memory/-on disk rasters and spatial vectors using \strong{GRASS GIS}. Most functions in the \strong{terra} and \strong{sf} packages are recreated, and can stand in for them when rasters or vectors are extremely large in memory and/or on disk. Processing of medium-sized and smaller spatial objects will nearly always be faster using \strong{terra} or \strong{sf}. To use most of the functions you must have the stand-alone version of \strong{GRASS} 8.0 or higher (not the **OSGeoW4*installer version).
\subsection{Most useful functions/links:}{

A \link[=tutorial]{quick-start tutorial}\cr
\code{\link[=fastStart]{fastStart()}}: Initiate a **GRASS*session\cr
\code{\link[=fast]{fast()}}: Create \code{GRaster}s or \code{GVector}s from \code{SpatRaster}s, \code{SpatVector}s, or \code{sf} or \code{stars} objects, or from files\cr
\code{\link[=rast]{rast()}}, \code{\link[=vect]{vect()}}, \code{\link[=st_as_sf]{st_as_sf()}}, and \code{\link[=stars]{stars()}}: Convert \code{GRaster}s and \code{GVector}s to \code{SpatRaster}s, \code{SpatVector}s, \code{sf} vectors, or \code{stars} rasters\cr
\code{\link[=writeRaster]{writeRaster()}} and \code{\link[=writeVector]{writeVector()}}: Save \code{GRaster}s or \code{GVector}s to disk\cr
\code{\link[=setFastOptions]{setFastOptions()}} and \code{\link[=getFastOptions]{getFastOptions()}}: Set options for working with \strong{fasterRaster}\cr
}

\subsection{Properties of **GRASS*rasters}{

\code{\link[=crs]{crs()}}: Coordinate reference system\cr
\code{\link[=dim]{dim()}}: Number of rows and columns\cr
\code{\link[=ext]{ext()}}: Spatial extent\cr
\code{\link[=datatype]{datatype()}}: Data type\cr
\code{\link[=minmax]{minmax()}}: Minimum and maximum values across all non-\code{NA} cells\cr
\code{\link[=names]{names()}}: \code{GRaster} names\cr
\code{\link[=ncol]{ncol()}}: Number of columns\cr
\code{\link[=ncell]{ncell()}}: Number of cells\cr
\code{\link[=ncell3d]{ncell3d()}}: Number of cells of a 3D raster\cr
\code{\link[=ndepth]{ndepth()}}: Number of depths of a 3D raster\cr
\code{\link[=nlyr]{nlyr()}}: Number of layers\cr
\code{\link[=nrow]{nrow()}}: Number of rows\cr
\code{\link[=res]{res()}}: Spatial resolution\cr
\code{\link[=st_crs]{st_crs()}}: Coordinate reference system\cr
\code{\link[=topology]{topology()}}: Dimensionality of a raster (2D or 3D)\cr
\code{\link[=zExt]{zExt()}}: Vertical extent
}

\subsection{Functions that operate on \code{GRasters}}{

\code{\link[=as.contour]{as.contour()}}: Contour lines from a \code{GRaster}\cr
\code{\link[=c]{c()}}: "Stack" two or more \code{GRaster}s\cr
}

\subsection{Properties of `GVectors'}{

\code{\link[=crs]{crs()}}: Coordinate reference system\cr
\code{\link[=ext]{ext()}}: Spatial extent\cr
\code{\link[=geometry]{geometry()}}: Type of vector (points, lines, polygons)\cr
\code{\link[=st_crs]{st_crs()}}: Coordinate reference system\cr
\code{\link[=topology]{topology()}}: Dimensions of vector coordinates (2D or 3D)\cr
\code{\link[=zExt]{zExt()}}: Vertical extent\cr
}

\subsection{Functions that operate on vectors}{

Something here
}

\subsection{Converting between data types}{

\code{\link[=as.points]{as.points()}}, \code{\link[=as.lines]{as.lines()}}, and \code{\link[=as.polygons]{as.polygons()}}: Convert a \code{GRaster} to a \code{GVector}\cr
\code{\link[=rasterize]{rasterize()}}: Convert a \code{GVector} to a \code{GRaster}\cr
\code{\link[=fast]{fast()}}: Create \code{GRaster}s or \code{GVector}s from \code{SpatRaster}s, \code{SpatVector}s, or \code{sf} or \code{stars} objects, or from files\cr
\code{\link[=rast]{rast()}}: Convert a \code{GRaster} to a \code{SpatRaster}\cr
\code{\link[=vect]{vect()}}: Convert a \code{GVector} to a \code{SpatVector}\cr
\code{\link[=st_as_sf]{st_as_sf()}}: Convert a \code{GVector} to a \code{sf} vectors\cr
\code{\link[=stars]{stars()}}: Convert a \code{GVector} to a \code{stars} raster\cr
}

\subsection{General purpose functions}{

\code{\link[=appendLists]{appendLists()}}: Append values to elements of a list from another list\cr
\code{\link[=compareFloat]{compareFloat()}}: Compare values accounting for differences due to floating point precision\cr
\code{\link[=forwardSlash]{forwardSlash()}}: Replace backslash with forward slash\cr
\code{\link[=removeSession]{removeSession()}}: Delete a GRASS session (location, mapset(s), and all associated files)\cr
}

\subsection{Functions that operate on *\emph{GRASS}"sessions" (seldom used by most users):}{

\code{\link[=fastRestore]{fastRestore()}}: Restore a previous \strong{GRASS} session or switch \strong{GRASS} locations/mapsets\cr
\code{\link[=fastLocation]{fastLocation()}}: Get current \strong{GRASS}  "location"\cr
\code{\link[=fastMaspet]{fastMaspet()}}: Get current \strong{GRASS} "mapset"\cr
}

\subsection{Functions that operate on *\emph{GRASS}"regions" (seldom used by most users):}{

\code{\link[=regionDim]{regionDim()}}: Change or reprot the active region's resolution\cr
\code{\link[=regionExt]{regionExt()}}: Change or report the active region's extent\cr
\code{\link[=regionRes]{regionRes()}}: Change or reprot the active region's dimensions\cr
\code{\link[=regionReshape]{regionReshape()}}: Change or report the active region's extent and resolution\cr
}

\subsection{Esoteric tutorials}{

\link[=location]{working directories and **GRASS*locations and mapsets}\cr
}
}
\author{
Adam B. Smith
}
