% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterRaster.r
\docType{package}
\name{fasterRaster}
\alias{fasterRaster}
\alias{fasterRaster-package}
\title{"fasterRaster": Faster raster and spatial vector processing using "GRASS GIS"}
\description{
Processing of large-in-memory/-on disk rasters and spatial vectors using \strong{GRASS GIS}. Most functions in the \strong{terra} and \strong{sf} packages are recreated. Processing of medium-sized and smaller spatial objects will nearly always be faster using \strong{terra} or \strong{sf}. To use most of the functions you must have the stand-alone version of \strong{GRASS} version 8.0 or higher (not the \strong{OSGeoW4} installer version). Note that due to developer choices, results will not always be strictly comparable between \strong{terra}, \strong{sf}, and \strong{fasterRaster}.
\subsection{Most useful tutorials and functions:}{

A \link[=tutorial_getting_started]{quick-start} tutorial\cr
A tutorial on \link[=tutorial_raster_data_types]{raster data types}\cr
\code{\link[=faster]{faster()}}: Initiate a \strong{GRASS} session\cr
\code{\link[=fast]{fast()}}: Convert a \code{SpatRaster}, \code{SpatVector}, or \code{sf} vector to \strong{fasterRaster}'s raster format (\code{GRaster}s) or vector format (\code{GVector}s), or load one from a file\cr
\code{\link[=rast]{rast()}}, \code{\link[=vect]{vect()}}, and \code{\link[=st_as_sf]{st_as_sf()}}: Convert \code{GRaster}s and \code{GVector}s to \code{SpatRaster}s, \code{SpatVector}s, or \code{sf} vectors\cr
\code{\link[=writeRaster]{writeRaster()}} and \code{\link[=writeVector]{writeVector()}}: Save \code{GRaster}s or \code{GVector}s to disk\cr
\code{\link[=restoreSession]{restoreSession()}}: Revert to another \strong{GRASS} \link[=tutorial_sessions]{"location" or "mapset"}, or restart a \strong{GRASS} session saved to disk\cr
\code{\link[=setFastOptions]{setFastOptions()}} and \code{\link[=getFastOptions]{getFastOptions()}}: Set options for working with \strong{fasterRaster}\cr
}

\subsection{Properties of \code{GRasters}}{

\code{\link[=crs]{crs()}}: Coordinate reference system\cr
\code{\link[=datatype]{datatype()}}: Data type\cr
\code{\link[=dim]{dim()}}: Number of rows and columns\cr
\code{\link[=ext]{ext()}}, \code{\link[=N]{N()}}, \code{\link[=S]{S()}}, \code{\link[=E]{E()}}, \code{\link[=W]{W()}}, \code{\link[=top]{top()}}, and \code{\link[=bottom]{bottom()}}: Spatial extent\cr
\code{\link[=freq]{freq()}}: Frequencies of cell values in a raster\cr
\code{\link[=global]{global()}}: Summary statistics\cr
\code{\link[=is.2d]{is.2d()}} and \code{\link[=is.3d]{is.3d()}}: Is an object 2- or 3-dimensional?\cr
\code{\link[=is.int]{is.int()}}, \code{\link[=is.float]{is.float()}}, \code{\link[=is.doub]{is.doub()}}: \code{GRaster} data type (integer/float/double)\cr
\code{\link[=is.factor]{is.factor()}}: Does a raster represent categorical data?\cr
\code{\link[=is.lonlat]{is.lonlat()}}: Is an object projected (e.g., in WGS84)?\cr
\code{\link[=levels]{levels()}}: Names of levels in a categorical \code{GRaster}\cr
\code{\link[=location]{location()}}: \strong{GRASS} "location" of an object or the active session\cr
\code{\link[=mapset]{mapset()}}: \strong{GRASS} "mapset" of an object or the active session\cr
\code{\link[=minmax]{minmax()}}: Minimum and maximum values across all non-\code{NA} cells\cr
\code{\link[=names]{names()}}: \code{GRaster} names\cr
\code{\link[=ncol]{ncol()}}: Number of columns\cr
\code{\link[=nacell]{nacell()}}: Number of \code{NA} cells\cr
\code{\link[=ncell]{ncell()}}: Number of cells\cr
\code{\link[=ncell3d]{ncell3d()}}: Number of cells of a 3D \code{GRaster}\cr
\code{\link[=ndepth]{ndepth()}}: Number of depths of a 3D \code{GRaster}\cr
\code{\link[=nlyr]{nlyr()}}: Number of layers\cr
\code{\link[=nonnacell]{nonnacell()}}: Number of non-\code{NA} cells\cr
\code{\link[=nrow]{nrow()}}: Number of rows\cr
\code{\link[=nlevels]{nlevels()}}: Number of categories\cr
\code{\link[=res]{res()}}, \code{\link[=xres]{xres()}}, \code{\link[=yres]{yres()}}, and \code{\link[=zres]{zres()}}: Spatial resolution\cr
\code{\link[=sources]{sources()}}: Name of the \code{GRaster} in \strong{GRASS}\cr
\code{\link[=st_bbox]{st_bbox()}}: Spatial extent\cr
\code{\link[=st_crs]{st_crs()}}: Coordinate reference system\cr
\code{\link[=topology]{topology()}}: Dimensionality (2D or 3D)\cr
\code{\link[=zext]{zext()}}: Vertical extent\cr
\code{\link[=zres]{zres()}}: Vertical resolution\cr
}

\subsection{Functions that operate on or create \code{GRasters}}{

\link{Arithmetic}: Mathematical operations on \code{GRaster}s: \code{+}, \code{-}, \code{*}, \code{/}, \code{^}, \code{\%\%} (modulus), \code{\%/\%} (integer division)\cr
\link[=Comparison]{Equality}: \code{<}, \code{<=}, \code{==}, \code{!=}, \code{>=}, and \code{>}, plus \code{\%in\%} (for categorical rasters only)\cr

Single-layer functions (applied to each layer of a \code{GRaster}):
\itemize{
\item \code{NA}s: \code{\link[=is.na]{is.na()}} and \code{\link[=not.na]{not.na()}} \cr
\item Trigonometry: \code{\link[=sin]{sin()}}, \code{\link[=cos]{cos()}}, \code{\link[=tan]{tan()}}, \code{\link[=asin]{asin()}}, \code{\link[=acos]{acos()}}, \code{\link[=atan]{atan()}}, \code{\link[=atan2]{atan2()}} \cr
\item Logarithms and powers: \code{\link[=exp]{exp()}}, \code{\link[=log]{log()}}, \code{\link[=ln]{ln()}}, \code{\link[=log1p]{log1p()}}, \code{\link[=log2]{log2()}}, \code{\link[=log10]{log10()}}, \code{\link[=sqrt]{sqrt()}} \cr
\item Rounding: \code{\link[=round]{round()}}, \code{\link[=floor]{floor()}}, \code{\link[=ceiling]{ceiling()}}, \code{\link[=trunc]{trunc()}} \cr
\item Signs: \code{\link[=abs]{abs()}} \cr
}

Multi-layer functions (applied across layers of a "stack" of \code{GRaster}s):
\itemize{
\item Numeration: \code{\link[=sum]{sum()}}, \code{\link[=count]{count()}} \cr
\item Central tendency: \code{\link[=mean]{mean()}}, \code{\link[=mmode]{mmode()}}, \code{\link[=median]{median()}} \cr
\item Dispersion: \code{\link[=sd]{sd()}}, \code{\link[=var]{var()}}, \code{\link[=sdpop]{sdpop()}}, \code{\link[=varpop]{varpop()}}, \code{\link[=nunique]{nunique()}}, \code{\link[=range]{range()}}, \code{\link[=quantile]{quantile()}}, \code{\link[=skewness]{skewness()}}, \code{\link[=kurtosis]{kurtosis()}}
\item Extremes: \code{\link[=min]{min()}}, \code{\link[=max]{max()}}, \code{\link[=which.min]{which.min()}}, \code{\link[=which.max]{which.max()}} \cr
}
}

\subsection{Functions that operate on or create \code{GRaster}s}{

The operators \code{$} (\link[=subset_dollar]{dollar}) and \code{[[} (\link[=subset_double_square_brackets]{double-square brackets}) can be used to subset or remove specific layers of a \code{GRaster}.\cr
The \verb{[<-} (\link[=replace_single_square_bracket]{single-bracket assign}) operator can be used to replace values of cells of a \code{GRaster}.\cr
The assign operators, \verb{$<-} (\link[=replace_dollar]{dollar assign}), \verb{[[<-} (\link[=replace_double_square_brackets]{double-square brackets assign}), and \verb{[add<-][add]} can be used to replace specific layers of a \code{GRaster}.\cr
\code{\link[=as.int]{as.int()}}, \code{\link[=as.float]{as.float()}}, \code{\link[=as.doub]{as.doub()}}: Change data type (integer/float/double)\cr
\code{\link[=as.contour]{as.contour()}}: Contour lines from a \code{GRaster}\cr
\code{\link[=as.lines]{as.lines()}}: Convert a \code{GRaster} to a "lines" vector\cr
\code{\link[=as.points]{as.points()}}: Convert a \code{GRaster} to a "points" vector\cr
\code{\link[=as.polygons]{as.polygons()}}: Convert a \code{GRaster} to a "polygons" vector\cr
\code{\link[=aggregate]{aggregate()}}: Aggregate values of \code{GRaster} cells into larger cells\cr
\code{\link[=buffer]{buffer()}}: Create a buffer around non-\code{NA} cells\cr
\code{\link[=app]{app()}}: Apply a user-defined function to multiple layers of a \code{GRaster} (with helper functions \code{\link[=appFuns]{appFuns()}} and \code{\link[=appCheck]{appCheck()}})\cr
\code{\link[=c]{c()}}: "Stack" two or more rasters\cr
\code{\link[=clump]{clump()}}: Group adjacent cells with similar values\cr
\code{\link[=cor]{cor()}}: Correlation matrix between layers of a \code{GRaster}\cr
\code{\link[=cov]{cov()}}: Covariance matrix between layers of a \code{GRaster}\cr
\code{\link[=crop]{crop()}}: Remove parts of a \code{GRaster}\cr
\code{\link[=denoise]{denoise()}}: Remove "noise" from a \code{GRaster} using a principal components analysis (PCA)\cr
\code{\link[=distance]{distance()}}: Distance to non-\code{NA} cells, or vice versa\cr
\code{\link[=extend]{extend()}}: Add rows and columns to a \code{GRaster}\cr
\code{\link[=extract]{extract()}}: Extract values from a \code{GRaster} at specific points\cr
\code{\link[=focal]{focal()}}: Calculate cell values based on values of nearby cells\cr
\code{\link[=fractalRast]{fractalRast()}}: Create a fractal \code{GRaster}\cr
\code{\link[=global]{global()}}: Summary statistics across cells of each \code{GRaster} layer\cr
\code{\link[=shade]{hillshade()}}: Create a hillshade \code{GRaster}\cr
\code{\link[=hist]{hist()}}: Histogram of \code{GRaster} values\cr
\code{\link[=horizonHeight]{horizonHeight()}}: Horizon height\cr
\code{\link[=kernel]{kernel()}}: Kernel density estimator of points.\cr
\code{\link[=longlat]{longlat()}}: Create longitude/latitude rasters.\cr
\code{\link[=mask]{mask()}}: Remove values in a \code{GRaster} based on values in another \code{GRaster} or vector\cr
\code{\link[=merge]{merge()}}: Combine two or more rasters with different extents and fill in \code{NA}s\cr
\code{\link[=noise]{noise()}}: Remove coarse-scale trends from a \code{GRaster}, leaving just fine-scale "noise"\cr
\code{\link[=pairs]{pairs()}}: Plot correlations between \code{GRaster} layers\cr
\code{\link[=pca]{pca()}}: Apply a principal components analysis (PCA) to a \code{GRaster}\cr
\code{\link[=pcs]{pcs()}}: Retrieve a principal components model from a PCA \code{GRaster} generated using \code{pca()}\cr
\code{\link[=plot]{plot()}}: Display a \code{GRaster}\cr
\code{\link[=plotRGB]{plotRGB()}}: Display a multispectral \code{GRaster} using red, blue, green, and alpha channels\cr
\code{\link[=project]{project()}}: Change coordinate reference system and cell size\cr
\code{\link[=resample]{resample()}}: Change cell size\cr
\code{\link[=rnormRast]{rnormRast()}}: A random \code{GRaster} with values drawn from a normal distribution\cr
\code{\link[=runifRast]{runifRast()}}: A random \code{GRaster} with values drawn from a uniform distribution\cr
\code{\link[=selectRange]{selectRange()}}: Select values from rasters in a stack based on values in another \code{GRaster}\cr
\code{\link[=spatSample]{spatSample()}}: Randomly points from a \code{GRaster}\cr
\code{\link[=spDepRast]{spDepRast()}}: Create a random \code{GRaster} with or without spatial dependence\cr
\code{\link[=sun]{sun()}}: Solar radiance and irradiance\cr
\code{\link[=terrain]{terrain()}}: Slope, aspect, curvature, and partial slopes\cr
\code{\link[=thinLines]{thinLines()}}: Reduce linear features on a \code{GRaster} so linear features are 1 cell wide\cr
\code{\link[=trim]{trim()}}: Remove rows and columns from a \code{GRaster} that are all \code{NA}\cr
\code{\link[=viewshed]{viewshed()}}: Areas visible from points on a raster\cr
}

\subsection{Functions operating on categorical rasters}{

\code{\link[=activeCat]{activeCat()}}: Column that defines category labels\cr
\verb{[activeCat<-]}: Set column that defines category labels\cr
\code{\link[=addCats]{addCats()}}: Add new columns to a "levels" table\cr
\verb{[addCats<-]}: Add new rows (levels) to a "levels" table\cr
\code{\link[=categories]{categories()}}: Set "levels" table for specific layers of a categorical raster\cr
\code{\link[=catNames]{catNames()}}: Names of each "levels" table\cr
\code{\link[=cats]{cats()}}: "Levels" table of a categorical raster\cr
\code{\link[=droplevels]{droplevels()}}: Remove one or more levels\cr
\code{\link[=freq]{freq()}}: Frequency of each category across cells of a raster\cr
\code{\link[=is.factor]{is.factor()}}: Is a raster categorical?\cr
\code{\link[=levels]{levels()}}: "Levels" table of a categorical raster\cr
\link{levels<-}: Set "levels" table of a categorical raster\cr
\code{\link[=minmax]{minmax()}}: "Lowest" and "highest" category values of categorical rasters (when argument \code{levels = TRUE})\cr
\code{\link[=missingCats]{missingCats()}}: Values that have no category assigned to them\cr
\code{\link[=nlevels]{nlevels()}}: Number of levels\cr
}

\subsection{Functions for analysis of remote sensing rasters}{

\code{\link[=toarAster]{toarAster()}}: Convert digital number ASTER rasters to radiance, and radiance to reflectance\cr
\code{\link[=toarLandsat]{toarLandsat()}}: Convert digital number LANDSAT rasters to radiance, and radiance to reflectance\cr
\code{\link[=vegIndex]{vegIndex()}}: Vegetation indices from surface reflectance\cr
}

\subsection{Properties of \strong{fasterRaster} vectors (\code{GVector}s)}{

\code{\link[=crs]{crs()}}: Coordinate reference system\cr
\code{\link[=datatype]{datatype()}}: Data type of fields\cr
\code{\link[=dim]{dim()}}: Number of geometries and columns\cr
\code{\link[=expanse]{expanse()}}: Area of polygons or length of lines\cr
\code{\link[=ext]{ext()}}, \code{\link[=N]{N()}}, \code{\link[=S]{S()}}, \code{\link[=E]{E()}}, \code{\link[=W]{W()}}, \code{\link[=top]{top()}}, and \code{\link[=bottom]{bottom()}}: Spatial extent\cr
\code{\link[=geomtype]{geomtype()}}: Type of vector (points, lines, polygons)\cr
\code{\link[=is.2d]{is.2d()}} and \code{\link[=is.3d]{is.3d()}}: Is an object 2- or 3-dimensional?\cr
\code{\link[=is.lonlat]{is.lonlat()}}: Is an object projected (e.g., in WGS84)?\cr
\code{\link[=is.points]{is.points()}}, \code{\link[=is.lines]{is.lines()}}, \code{\link[=is.polygons]{is.polygons()}}: Does a \code{GVector} represent points, lines, or polygons?\cr
\code{\link[=location]{location()}}: \strong{GRASS} "location" of an object or the active session\cr
\code{\link[=mapset]{mapset()}}: \strong{GRASS} "mapset" of an object or the active session\cr
\code{\link[=names]{names()}}: Names of \code{GVector} fields\cr
\code{\link[=ncol]{ncol()}}: Number of fields\cr
\code{\link[=ngeom]{ngeom()}}: Number of geometries (points, lines, polygons)\cr
\code{\link[=nrow]{nrow()}}: Number of rows in a vector data table\cr
\code{\link[=nsubgeom]{nsubgeom()}}: Number of subgeometries (points, lines, polygons that make up single- and multipart geometries)\cr
\code{\link[=sources]{sources()}}: Name of the vector in \strong{GRASS}\cr
\code{\link[=st_bbox]{st_bbox()}}: Spatial extent\cr
\code{\link[=st_crs]{st_crs()}}: Coordinate reference system\cr
\code{\link[=topology]{topology()}}: Dimensionality (2D or 3D)\cr
\code{\link[=zext]{zext()}}: Vertical extent\cr
}

\subsection{Functions that operate on or create \code{GVector}s}{

The \code{[} (\link[=subset_single_bracket]{single bracket}) operator, can be used to subset geometries of a \code{GRaster}.\cr
The \code{$} \link[=subset_dollar]{dollar} and \code{[[} (\link[=subset_double_square_brackets]{double-square brackets}) operators can be used to get columns of a \code{GVector}'s data table.\cr
The \verb{$<-} (\link[=replace_dollar]{dollar replace}) operator can be used to replace specific columns of a \code{GVector}'s data table.\cr
\code{\link[=aggregate]{aggregate()}}: Combine \code{GVector} geometries\cr
\code{\link[=as.data.frame]{as.data.frame()}}: Convert a \code{GVector}'s attribute table to a \code{data.frame}\cr
\code{\link[=as.data.table]{as.data.table()}}: Convert a \code{GVector}'s attribute table to a \code{data.table}\cr
\code{\link[=as.points]{as.points()}}: Extract vertex coordinates from a "lines" or "polygons" \code{GVector}\cr
\code{\link[=buffer]{buffer()}}: Create a polygon around/inside a \code{GVector}\cr
\code{\link[=cleanGeom]{cleanGeom()}}: Fix undesirable geometries of a \code{GVector}\cr
\code{\link[=connectors]{connectors()}}: Create lines connecting nearest features of two \code{GVector}s\cr
\code{\link[=convHull]{convHull()}}: Minimum convex hull\cr
\code{\link[=crds]{crds()}}: Extract coordinates of a \code{GVector}\cr
\code{\link[=crop]{crop()}}: Remove parts of a \code{GVector}\cr
\code{\link[=delaunay]{delaunay()}}: Delaunay triangulation\cr
\code{\link[=disagg]{disagg()}}: Separate multipart geometries into singlepart geometries\cr
\code{\link[=distance]{distance()}}: Distance between geometries in two \code{GVector}, or from a \code{GVector} to cells of a \code{GRaster}\cr
\code{\link[=erase]{erase()}} or \code{-}: Remove part of a \code{GVector} that overlaps with another\cr
\code{\link[=expanse]{expanse()}}: Area of polygons or length of lines\cr
\code{\link[=extract]{extract()}}: Extract values from a \code{GVector} at specific points\cr
\code{\link[=fillHoles]{fillHoles()}}: Fill holes in a \code{GVector} (requires 8.4+)\cr
\code{\link[=grid]{grid()}}: Create a grid \code{GVector}\cr
\code{\link[=head]{head()}}: First rows of a \code{GVector}'s data table\cr
\code{\link[=hexagons]{hexagons()}}: Create a hexagonal grid\cr
\code{\link[=intersect]{intersect()}} or \code{*}: Intersection of two \code{GVectors}.\cr
\code{\link[=kernel]{kernel()}}: Kernel density estimator of points.\cr
\code{\link[=project]{project()}}: Change coordinate reference system\cr
\code{\link[=rasterize]{rasterize()}}: Convert a \code{GVector} to a \code{GRaster}\cr
\code{\link[=simplifyGeom]{simplifyGeom()}}: Remove vertices\cr
\code{\link[=smoothGeom]{smoothGeom()}}: Remove "angular" aspects of features\cr
\code{\link[=st_as_sf]{st_as_sf()}}: Convert a \code{GVector} to a \code{sf} vector\cr
\code{\link[=st_buffer]{st_buffer()}}: Create a polygon around/inside a \code{GVector}\cr
\code{\link[=st_distance]{st_distance()}}: Distance between geometries in two \code{GVector}\cr
\code{\link[=tail]{tail()}}: Last rows of a \code{GVector}'s data table\cr
\code{\link[=thinPoints]{thinPoints()}}: Reduce number of points in same raster cell\cr
\code{\link[=union]{union()}} or \code{+}: Combine two \code{GVector}s\cr
\code{\link[=xor]{xor()}} or \code{/}: Select parts of polygons not shared by two \code{GVector}s\cr
}

\subsection{Converting between data types}{

\code{\link[=as.contour]{as.contour()}}: Convert a \code{GRaster} to a \code{GVector} representing contour lines\cr
\code{\link[=as.doub]{as.doub()}}: Convert a \code{GRaster} to a double-floating point raster (*\strong{GRASS} data type \code{DCELL})\cr
\code{\link[=as.data.frame]{as.data.frame()}}: Convert \code{GVector} to a \code{data.frame}\cr
\code{\link[=as.data.table]{as.data.table()}}: Convert \code{GVector} to a \code{data.table}\cr
\code{\link[=as.float]{as.float()}}: Convert a \code{GRaster} to a floating-point raster (*\strong{GRASS} data type \code{FCELL})\cr
\code{\link[=as.int]{as.int()}}: Convert a \code{GRaster} to an integer raster (*\strong{GRASS} data type \code{CELL})\cr
\code{\link[=as.points]{as.points()}}, \code{\link[=as.lines]{as.lines()}}, and \code{\link[=as.polygons]{as.polygons()}}: Convert a \code{GRaster} to a \code{GVector}\cr
\code{\link[=fast]{fast()}}: Convert a \code{SpatRaster}, \code{SpatVector}, or \code{sf} vector to a \code{GRaster} or \code{GVector}, or load one from a file\cr
\code{\link[=categories]{categories()}} and \link{levels<-}: Convert an integer raster to a categorical ("factor") raster.
\code{\link[=rast]{rast()}}: Convert a \code{GRaster} to a \code{SpatRaster}\cr
\code{\link[=rasterize]{rasterize()}}: Convert a \code{GVector} to a \code{GRaster}\cr
\code{\link[=st_as_sf]{st_as_sf()}}: Convert a \code{GVector} to a \code{sf} vector\cr
\code{\link[=vect]{vect()}}: Convert a \code{GVector} to a \code{SpatVector}\cr
}

\subsection{Functions that operate on \strong{GRASS} "sessions":}{

\code{\link[=crs]{crs()}}: Coordinate reference system of the current location\cr
\code{\link[=location]{location()}}: \strong{GRASS} "location" of an object or the active session\cr
\code{\link[=mapset]{mapset()}}: \strong{GRASS} "mapset" of an object or the active session\cr
\code{\link[=restoreSession]{restoreSession()}}: Restore a previous \strong{GRASS} session or switch \strong{GRASS} locations/mapsets\cr
\code{\link[=removeSession]{removeSession()}}: Delete a \strong{GRASS} session (location, mapset(s), and all associated files)\cr
}

\subsection{General purpose functions}{

\code{\link[=appendLists]{appendLists()}}: Append values to elements of a list from another list\cr
\code{\link[=compareFloat]{compareFloat()}}: Compare values accounting for differences due to floating point precision\cr
\code{\link[=compareGeom]{compareGeom()}}: Determine if geographic metadata is same between \code{GRaster}s and/or \code{GVector}s\cr
\code{\link[=dropRows]{dropRows()}}: Remove rows from a \code{data.frame} or \code{data.table}\cr
\code{\link[=forwardSlash]{forwardSlash()}}: Replace backslash with forward slash\cr
\code{\link[=grassInfo]{grassInfo()}}: \strong{GRASS} version and citation\cr
\code{\link[=pmatchSafe]{pmatchSafe()}}: Partial matching of strings with error checking\cr
\code{\link[=replaceNAs]{replaceNAs()}}: Replace \code{NA}s in columns of a \code{data.table} or \code{data.frame}, or in a vector\cr
\code{\link[=rstring]{rstring()}}: Create a string statistically likely to be unique\cr
\code{\link[=seqToSQL]{seqToSQL()}}: Format a numeric series into an SQL value call.\cr
\code{\link[=update]{update()}}: Refresh metadata in a \code{GRaster} or \code{GVector} object.\cr
}

\subsection{Functions that operate on \strong{GRASS} "regions" (useful mostly to developers):}{

\code{\link[=region]{region()}}: Change or report the active region's extent and resolution\cr
\code{\link[=regionDim]{regionDim()}}: Change or report the active region's resolution (also \code{\link[=dim]{dim()}} and related functions, with no arguments)
\code{\link[=regionExt]{regionExt()}}: Change or report the active region's extent (also \code{\link[=ext]{ext()}} and related functions, with no arguments)\cr
\code{\link[=regionRes]{regionRes()}}: Change or report the active region's dimensions (also \code{\link[=res]{res()}} and related functions, with no arguments)\cr
}

\subsection{Example data}{

\link{appFunsTable} (see also \code{\link[=appFuns]{appFuns()}}): Functions usable by the \code{\link[=app]{app()}} function\cr
\link{madChelsa}: Climate rasters for of a portion of eastern Madagascar\cr
\link{madCoast0}, \link{madCoast4}, and \link{madCoast}: Borders of an eastern portion of Madagascar\cr
\link{madCover}: Land cover raster\cr
\link{madCoverCats}: Table of land cover classes\cr
\link{madDypsis}: Specimens records of species in the genus \emph{Dypsis}\cr
\link{madElev}: Elevation raster\cr
\link{madForest2000} and \link{madForest2014}: Forest cover in 2000 and 2014\cr
\link{madLANDSAT}: Surface reflectance in 2023\cr
\link{madRivers}: Rivers vector\cr
\link{vegIndices}: Vegetation indices that can be calculated using \code{\link[=vegIndex]{vegIndex()}}.
}

\subsection{Esoteric tutorials}{

\link[=tutorial_sessions]{Sessions, locations, and mapsets}\cr
\link[=tutorial_raster_data_types]{Raster data types}\cr
\link[=tutorial_regions]{Regions}\cr
\link[=tutorial_undocumented_functions]{Undocumented functions}\cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.earthSkySea.org}
  \item Report bugs at \url{https://github.com/adamlilith/fasterRaster/issues}
}

}
\author{
Adam B. Smith
}
\keyword{internal}
