% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterRasterize.r
\name{fasterRasterize}
\alias{fasterRasterize}
\title{Rasterize using a call to GRASS GIS.}
\usage{
fasterRasterize(x, y, use = "value", value = 1, field = NULL,
  grassLoc = NULL, ...)
}
\arguments{
\item{x}{SpatialPoints, SpatialPointsDataFrame, SpatialLines, SpatialLinesDataFrame, SpatialPolygons, or SpatialPolygonsDataFrame or the full path and name of such an object.}

\item{y}{Raster* object, serves as a template for the new raster or the full path and name of such an object.}

\item{use}{Character, indicates the types of values to be "burned" to the raster. Options include
\itemize{
\item \code{value} (default): a user-define value given by \code{value}
\item \code{field}: values directly from a field in \code{x} named by \code{field}
\item \code{category}: values according to which polygon is covered by a cell named in \code{field}
\item \code{z}: z-coordinate (points or contours only)
\item \code{direction}: flow direction (lines only)
}}

\item{value}{Numeric, value to burn to each cell overlapped by the spatial object in \code{x}.}

\item{field}{Name of column in \code{x} with values or category labbels to which to burn to the raster.}

\item{grassLoc}{Either \code{NULL} or a 3-element character vector. If the latter, the first element is the base path to the installation of GRASS, the second the version number, and the third the install type for GRASS.  For example, \code{c('C:/OSGeo4W64/', 'grass-7.4.1', 'osgeo4W')}. See \code{\link[link2GI]{linkGRASS7}} for further help. If \code{NULL} (default) the an installation of GRASS is searched for; this may take several minutes.}

\item{...}{Arguments to pass to \code{\link[rgrass7]{execGRASS}} when used for rasterizing (i.e., function \code{v.to.rast} in GRASS).}
}
\value{
A raster with the same extent, resolution, and coordinate reference system as \code{y}.
}
\description{
This function is a potentially faster version of the \code{\link[raster]{rasterize}} function in the \pkg{ratser} package which converts a spatial points, lines, or polygon into a raster based on a "template" raster. All cells covered by the spatial object can either have values taken from the spatial object or a user-defined.
}
\seealso{
\code{\link[raster]{rasterize}}
}
