% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterRasterize.r
\name{fasterRasterize}
\alias{fasterRasterize}
\title{Rasterize using a call to GRASS GIS.}
\usage{
fasterRasterize(
  vect,
  rast,
  use = "value",
  value = 1,
  field = NULL,
  burn = NULL,
  grassDir = NULL,
  alreadyInGrass = FALSE,
  grassToR = TRUE,
  ...
)
}
\arguments{
\item{vect}{Either a SpatialPoints, SpatialPointsDataFrame, SpatialLines, SpatialLinesDataFrame, SpatialPolygons, or SpatialPolygonsDataFrame or the name of such a vector dataset already in a GRASS session.}

\item{rast}{Either a raster or the name of a raster in an existing GRASS session. This serves as a template for the new raster.}

\item{use}{Character, indicates the types of values to be "burned" to the raster. Options include
\itemize{
\item \code{value} (default): a user-define value given by \code{value}
\item \code{field}: values directly from a field in \code{vect} named by \code{field}
\item \code{category}: values according to which polygon is covered by a cell named in \code{field}
\item \code{z}: z-coordinate (points or contours only)
\item \code{direction}: flow direction (lines only)
}}

\item{value}{Numeric, value to burn to each cell overlapped by the spatial object in \code{vect}.}

\item{field}{Name of column in \code{vect} with values or category labbels to which to burn to the raster.}

\item{grassDir}{Either \code{NULL} or a 3-element character vector. If the latter, the first element is the base path to the installation of GRASS, the second the version number, and the third the install type for GRASS.  For example, \code{c('C:/OSGeo4W64/', 'grass-7.4.1', 'osgeo4W')}. See \code{\link[link2GI]{linkGRASS7}} for further help. If \code{NULL} (default) the an installation of GRASS is searched for; this may take several minutes.}

\item{alreadyInGrass}{Logical, if \code{FALSE} (default) then start a new GRASS session and import the raster named in \code{rast}. If \code{FALSE}, use a raster already in GRASS with the name given by \code{rast}. The latter is useful if you are chaining \pkg{fasterRaster} functions together and the first function initializes the session. The first function should use \code{alreadyInGrass = FALSE} and subsequent functions should use \code{alreadyInGrass = TRUE} then use their \code{rast} (or \code{vect}) arguments to name the raster (or vector) that was made by the previous function.}

\item{grassToR}{Logical, if \code{TRUE} (default) then the product of the calculations will be returned to R. If \code{FALSE}, then the product is left in the GRASS session and named \code{vectToRast}. The latter case is useful (and faster) when chaining several \pkg{fasterRaster} functions together.}

\item{...}{Arguments to pass to \code{\link[rgrass7]{execGRASS}} when used for rasterizing (i.e., function \code{v.to.rast} in GRASS).}
}
\value{
If \code{grassToR} if \code{TRUE}, then a raster with the same extent, resolution, and coordinate reference system as \code{rast}. Otherwise, a raster with the name of \code{vectToRast} is written into the GRASS session.
}
\description{
This function is a potentially faster version of the \code{\link[raster]{rasterize}} function in the \pkg{raster} package which converts a spatial points, lines, or polygon into a raster based on a "template" raster. All cells covered by the spatial object can either have values taken from the spatial object or a user-defined.
}
\details{

}
\examples{
\donttest{
# change this according to where GRASS 7 is installed on your system
grassDir <- 'C:/Program Files/GRASS GIS 7.8'

data(mad0)
data(madForest2000)

# could also use rasterize() or mask() from the raster package which may
# be faster in this example
madMask <- fasterRasterize(mad0, madForest2000, grassDir=grassDir)
# madMask <- rasterize(mad0, madForest2000)
# madMask <- mask(madForest2000, mad0)
plot(madMask, main='Portion of Eastern Madagascar')
plot(mad0, add=TRUE)
}
}
\seealso{
\code{\link[raster]{rasterize}}
}
