% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterRasterize.r
\name{fasterRasterize}
\alias{fasterRasterize}
\title{Rasterize a spatial vector}
\usage{
fasterRasterize(
  vect,
  rast,
  use = "value",
  value = 1,
  field = NULL,
  burn = NULL,
  grassDir = options()$grassDir,
  grassToR = TRUE,
  outGrassName = "vectToRast",
  ...
)
}
\arguments{
\item{vect}{A \code{Spatvector}, an \code{sf} object, or the name of a spatial vector already in an existing \code{GRASS} session.}

\item{rast}{A \code{SpatRaster} or the name of a raster already in an existing \code{GRASS} session.}

\item{use}{Character, indicates the types of values to be "burned" to the raster. Options include
\itemize{
\item \code{value} (default): a user-define value given by \code{value}
\item \code{field}: values directly from a field in \code{vect} named by \code{field}
\item \code{category}: values according to which polygon is covered by a cell named in \code{field}
\item \code{z}: z-coordinate (points or contours only)
\item \code{direction}: flow direction (lines only)
}}

\item{value}{Numeric, value to burn to each cell overlapped by the spatial object in \code{vect}.}

\item{field}{Name of column in \code{vect} with values or category labbels to which to burn to the raster.}

\item{burn}{\code{NULL} or any of \code{'point'}, \code{'line'}, \code{'area'}, \code{'boundary'}, or \code{'centroid'}. This determines the manner in which the vector data is "burned" to the raster. If \code{NULL} (default), then SpatialPoints and SpatialPointsDataFrame objects will rasterized as points, SpatialLines and SpatialLinesDataFrame objects as lines, and SpatialPolygons and SpatialPolygonsDataFrame objects as areas. See \url{https://grass.osgeo.org/grass82/manuals/v.to.rast.html} for more details.}

\item{grassDir}{Name of the folder path in which \code{GRASS} is installed. For a Windows system, this might be something like \code{'C:/Program Files/GRASS GIS 8.2'}. For a Mac system, this might be like \code{"/Applications/GRASS-8.2.app/Contents/Resources"}. If you are using a lot of \pkg{fasterRaster} functions, it can be easier to set this once using \code{\link{fasterOptions}}. By default, this is \code{NULL}, in which case the install path for \code{GRASS} will be searched for, but if not found, will result in an error.}

\item{grassToR}{If \code{TRUE} (default) then the output will be returned to \pkg{R}. If \code{FALSE}, then the output is left in the \code{GRASS} session and named the value in \code{outGrassName}. You can then refer to the output in subsequent \pkg{fasterRaster} functions and avoid the need to re-import it. This is called "chaining" functions together, and it can save a lot of computational time.}

\item{outGrassName}{Character. Name of output in \code{GRASS}. This is useful if you want to refer to the output object in \code{GRASS} later in a session by "chaining" functions.}

\item{...}{Arguments to pass to \code{\link[rgrass]{execGRASS}} when used for rasterizing (i.e., function \code{v.to.rast} in \code{GRASS}).}
}
\value{
If \code{grassToR} if \code{TRUE}, then a raster with the same extent, resolution, and coordinate reference system as \code{rast}. Otherwise, a raster with the name given by \code{outRastName} is written into the \code{GRASS} session.
}
\description{
This function is a potentially faster version of the \code{\link[terra]{rasterize}} function in the \pkg{terra} package which converts a spatial points, lines, or polygon into a raster based on a "template" raster. All cells covered by the spatial object can either have values taken from the spatial object or a user-defined.
}
\details{
See the documentation for the \code{GRASS} module \code{v.to.rast}{https://grass.osgeo.org/grass82/manuals/v.to.rast.html}.
}
\examples{
man/examples/ex_fasterRasterize.r

}
\seealso{
\code{\link[terra]{rasterize}}
}
