% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterRename.r
\name{fasterRename}
\alias{fasterRename}
\title{Rename a raster or vector in an existing GRASS session}
\usage{
fasterRename(from, to, type = NULL)
}
\arguments{
\item{from, to}{Names of the raster or vector to rename.}

\item{type}{Either \code{'raster'} and/or \code{'vector'}, or \code{NULL} (default). This specifies the type of file to be renamed. Partial matching is allowed. If left as \code{NULL} (default), the function will rename any raster or vector with the given name. Note that unlike in \pkg{R}, \code{GRASS} can have rasters with the same name as vectors.}
}
\value{
The function invisibly returns \code{TRUE} if the desired rasters and/or vectors were named, and \code{FALSE} if raster and/or vector to be renamed did not exist in the \code{GRASS} session. Notably, a raster or vector or both are renamed in an existing \code{GRASS} session.
}
\description{
Rename a raster or vector in an existing GRASS session.
}
\examples{
\dontrun{

library(sf)
library(terra)

# start a GRASS session with a raster
madElev <- fasterData('madElev')
initGrass(madElev, inRastName = 'madElev')

# import a vector
madCoast0 <- fasterData('madCoast0')
exportVectToGrass(madCoast0, inVectName = 'madCoast0')

# names of all rasters and vectors in the GRASS session
fasterLs()

# rename raster
fasterRename('madElev', 'elevation')

# remove raster
fasterRm('elevation')
fasterLs()

# remove vector
fasterRm('madCoast0')
fasterLs()

}

}
\seealso{
\code{\link{fasterLs}}, \code{\link{fasterRm}}, \href{https://grass.osgeo.org/grass82/manuals/g.rename.html}{\code{g.list}} in \code{GRASS}
}
