% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterSurfFractal.r
\name{fasterSurfFractal}
\alias{fasterSurfFractal}
\title{Raster with a fractal pattern}
\usage{
fasterSurfFractal(
  rast,
  dimension = 2.05,
  grassDir = options()$grassDir,
  grassToR = TRUE,
  outGrassName = "surfFractal",
  ...
)
}
\arguments{
\item{rast}{A \code{SpatRaster} or the name of a raster already in an existing \code{GRASS} session.}

\item{dimension}{Numeric. Fractal dimension. Must be >2 and <3. Default is 2.05.}

\item{grassDir}{Name of the folder path in which \code{GRASS} is installed. For a Windows system, this might be something like \code{'C:/Program Files/GRASS GIS 8.2'}. For a Mac system, this might be like \code{"/Applications/GRASS-8.2.app/Contents/Resources"}. If you are using a lot of \pkg{fasterRaster} functions, it can be easier to set this once using \code{\link{fasterOptions}}. By default, this is \code{NULL}, in which case the install path for \code{GRASS} will be searched for, but if not found, will result in an error.}

\item{grassToR}{If \code{TRUE} (default) then the output will be returned to \pkg{R}. If \code{FALSE}, then the output is left in the \code{GRASS} session and named the value in \code{outGrassName}. You can then refer to the output in subsequent \pkg{fasterRaster} functions and avoid the need to re-import it. This is called "chaining" functions together, and it can save a lot of computational time.}

\item{outGrassName}{Character. Name of output in \code{GRASS}. This is useful if you want to refer to the output object in \code{GRASS} later in a session by "chaining" functions.}

\item{...}{Arguments to pass to \code{\link[rgrass]{execGRASS}}.}
}
\value{
If \code{grassToR} if \code{TRUE}, then a raster object with the same coordinate reference system, resolution, and extent as \code{rast}. Regardless, a raster is written into the \code{GRASS} session with the name given by \code{outGrassName}.
}
\description{
This function creates a raster with a fractal pattern. It utilizes the \code{GRASS} function \code{r.surf.fractal}.
}
\details{
See the documentation for the \code{GRASS} module \code{r.surf.fractal}{https://grass.osgeo.org/grass82/manuals/r.surf.fractal.html}.
}
\examples{
man/examples/ex_fasterSurfFractal.r

}
