% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterVectorize.r
\name{fasterVectorize}
\alias{fasterVectorize}
\title{Convert a raster to a vector (points, lines, or polygons)}
\usage{
fasterVectorize(rast, vectType, agg = TRUE, smooth = FALSE,
  grassLoc = NULL, grassInit = TRUE, grassToR = TRUE, ...)
}
\arguments{
\item{rast}{Either a raster or the full path and name of a raster object. The raster values must be 1 or 0 (or \code{NA}). The fragmentation index applies to the state of the entity represented by the 1's.}

\item{vectType}{Character, Indicates type of output: \code{point}, \code{line} (not supported yet), or \code{area}.}

\item{agg}{Logical, if \code{TRUE} (default) then union all points/lines/polygons with the same value into the same "multipart" polygon. This may or may not be desirable. For example, if the raster is vectorized into a polygons object each cell will become a separate polygon. Using this option will merge cells with the same value (even if they are not spatially adjacent one another).}

\item{smooth}{Logical, if \code{TRUE} then "round" cell corners by connecting the midpoints of corner cells (which leaves out the corner-most triangle of that cell). This option only applies if \code{vectType} is \code{area}. Default is \code{FALSE}.}

\item{grassLoc}{Either \code{NULL} or a 3-element character vector. If the latter, the first element is the base path to the installation of GRASS, the second the version number, and the third the install type for GRASS.  For example, \code{c('C:/OSGeo4W64/', 'grass-7.4.1', 'osgeo4W')}. See \code{\link[link2GI]{linkGRASS7}} for further help. If \code{NULL} (default) the an installation of GRASS is searched for; this may take several minutes.}

\item{grassInit}{Logical, if \code{TRUE} (default) then a new GRASS session is initialized. If \code{FALSE} then it is assumed a GRASS session has been initialized using the raster in \code{rast}. The latter is useful if you are chaining \pkg{fasterRaster} functions together and the first function initializes the session.}

\item{grassToR}{Logical, if \code{TRUE} (default) then the product of the calculations will be returned to R. If \code{FALSE}, then the product is left in the GRASS session and named \code{rastToVect}. The latter case is useful (and faster) when chaining several \pkg{fasterRaster} functions together.}

\item{...}{Arguments to pass to \code{\link[rgrass7]{execGRASS}} when used for converting a raster to a vector (i.e., function \code{r.to.vect} in GRASS).}

\item{calcDensity}{Logical, if \code{TRUE} then calculate density in the moving window. This will create a raster named \code{density} in the GRASS environment if \code{grassToR} is \code{FALSE} or return a raster named \code{density} if \code{grassToR} is \code{TRUE}. Default is \code{FALSE}.}

\item{calcConnect}{Logical, if \code{TRUE} then calculate a connectivity raster (conditional probability a cell with a value of 1 has a value that is also 1) in the moving window. This will create a raster named \code{connect} in the GRASS environment if \code{grassToR} is \code{FALSE} or return a raster named \code{connect} if \code{grassToR} is \code{TRUE}. Default is \code{FALSE}.}
}
\value{
If \code{grassToR} if \code{TRUE}, then a SpatialPointsDataFrame, SpatialLinesDataFrame, or a SpatialPolygonsDataFrame with the same coordinate reference system as \code{rast}. The field named \code{value} will have the raster values. Otherwise, vector object named \code{vectToRast} a  will be written into the GRASS session.
}
\description{
This function is a potentially faster version of the function \code{\link[raster]{rasterToPolygons}} in the \pkg{raster} package. It can convert a raster to points or polygons (conversion to lines is not yet supported, although it is possible using the \code{r.to.vect} function in GRASS).
}
\details{
See (r.to.vect)[https://grass.osgeo.org/grass74/manuals/r.to.vect.html] for more details.  Note that if you get an error saying "", then you should add the EPSG code to the beginning of the raster and vector coordinate reference system string (their "proj4string"). For example, \code{proj4string(x) <- CRS('+init=epsg:32738')}. EPSG codes for various projections, datums, and locales can be found at (Spatial Reference)[http://spatialreference.org/].
}
\examples{
\dontrun{
# change this according to where GRASS 7 is installed on your system
grassLoc <- c('C:/OSGeo4W64/', 'grass-7.4.1', 'osgeo4W')

data(madForest2000)

# could also use rasterToPolygons() which is
# probably faster in this example
forestPoly <- fasterVectorize(madForest2000,
	vectType='area', grassLoc=grassLoc)
# forestPoly <- rasterToPolygons(madForest2000, dissolve=TRUE)
plot(forestPoly, main='Forest as Polygon')
}
}
\seealso{
\code{\link[raster]{rasterToPolygons}}, \code{\link[fasterRaster]{fasterRasterize}}
}
