% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterVectToRastDistance.r
\name{fasterVectToRastDistance}
\alias{fasterVectToRastDistance}
\title{Distance between raster cells and nearest spatial vector feature}
\usage{
fasterVectToRastDistance(
  rast,
  vect,
  metric = "euclidean",
  meters = TRUE,
  invert = FALSE,
  grassDir = options()$grassDir,
  grassToR = TRUE,
  inRastName = "rast",
  inVectName = "vect",
  outGrassName = "distToVectRast",
  ...
)
}
\arguments{
\item{rast}{A \code{SpatRaster} or the name of a raster already in an existing \code{GRASS} session.}

\item{vect}{A \code{Spatvector}, an \code{sf} object, or the name of a spatial vector already in an existing \code{GRASS} session.}

\item{metric}{Character, indicates type of distance to calculate:
\itemize{
\item \code{euclidean}: Euclidean distance
\item \code{geodesic} (default): geographic distance (suggested to use with \code{meters = TRUE}).
\item \code{squared}: Squared Euclidean distance
\item \code{maximum}: Maximum Euclidean distance
\item \code{manhattan}: Manhattan distance (i.e., "taxicab" distance, distance along cells going only north-south and east-west and never along a diagonal).
}}

\item{meters}{Logical, if \code{TRUE} then distance is in meters. If \code{FALSE} then distance is in map units.}

\item{invert}{Logical, if \code{TRUE} then for cells covered by a vector object calculate distance to nearest cell \emph{not} covered by a vector object.}

\item{grassDir}{Name of the folder path in which \code{GRASS} is installed. For a Windows system, this might be something like \code{'C:/Program Files/GRASS GIS 8.2'}. For a Mac system, this might be like \code{"/Applications/GRASS-8.2.app/Contents/Resources"}. If you are using a lot of \pkg{fasterRaster} functions, it can be easier to set this once using \code{\link{fasterOptions}}. By default, this is \code{NULL}, in which case the install path for \code{GRASS} will be searched for, but if not found, will result in an error.}

\item{grassToR}{If \code{TRUE} (default) then the output will be returned to \pkg{R}. If \code{FALSE}, then the output is left in the \code{GRASS} session and named the value in \code{outGrassName}. You can then refer to the output in subsequent \pkg{fasterRaster} functions and avoid the need to re-import it. This is called "chaining" functions together, and it can save a lot of computational time.}

\item{inRastName}{The name of the input raster in \code{GRASS}. This argument is useful if you are "chaining" \pkg{fasterRaster} functions together and wish to refer to the input raster later. The default value is the name of the raster (i.e., \code{names(rast)}).}

\item{inVectName}{The name of the input vector in \code{GRASS}. This is useful only if you are "chaining" \pkg{fasterRaster} functions together and wish to refer to the input vector later. The default value is \code{'vect'}.}

\item{outGrassName}{Character. Name of output in \code{GRASS}. This is useful if you want to refer to the output object in \code{GRASS} later in a session by "chaining" functions.}

\item{...}{Arguments to pass to \code{\link[rgrass]{execGRASS}} when used for rasterizing (i.e., function \code{r.grow.distance} in \code{GRASS}).}
}
\value{
If \code{grassToR} if \code{TRUE}, then a raster with the same extent, resolution, and coordinate reference system as \code{vect}. Regardless, a raster with the name of \code{distToVect} is written into the \code{GRASS} session.
}
\description{
Calculate the distance between each cell in a raster and the nearest feature in a spatial points, lines, or polygon object. Alternatively, calculate the distance from any cell covered by a vector object and the nearest cell \emph{not} covered by a vector object.
}
\examples{
\dontrun{

# change this according to where GRASS is installed on your system
grassDir <- 'C:/Program Files/GRASS GIS 8.2' # example for a PC
grassDir <- "/Applications/GRASS-8.2.app/Contents/Resources" # for a Mac

library(sf)
library(terra)

madForest2000 <- fasterData('madForest2000')
madRivers <- fasterData('madRivers')

distToRiver <- fasterVectToRastDistance(rast=madForest2000,
vect=madRivers, grassDir=grassDir)

plot(distToRiver, main='Distance to Rivers (m)')
plot(madRivers, col='blue', add=TRUE)

}
}
\seealso{
\code{\link[terra]{distance}} in \pkg{terra}; \code{\link{fasterRastDistance}} in \pkg{fasterRaster}; \href{https://grass.osgeo.org/grass82/manuals/r.grow.distance.html}{\code{r.grow.distance}} in \code{GRASS}
}
