% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.r
\name{global,GRaster-method}
\alias{global,GRaster-method}
\alias{global}
\alias{global,missing-method}
\title{Summary statistics for GRasters}
\usage{
\S4method{global}{GRaster}(x, fun = "mean", prob = 0.5, ...)

\S4method{global}{missing}(x, ...)
}
\arguments{
\item{x}{A \code{GRaster} or missing.  If missing, then a vector of all of the accepted function names is returned.}

\item{fun}{Character vector: The name of the function(s):
\itemize{
\item \code{"*"}: All of the statistics below.
\item \code{"countNonNA"}: Total number of non-\code{NA} cells.
\item \code{"countNA"}: Total number of \code{NA} cells.
\item \code{"cv"}: Sample coefficient of variation (expressed as a proportion of the mean).
\item \code{"cvpop"}: Population coefficient of variation (expressed as a proportion of the mean).
\item \code{"max"} and \code{"min"}: Highest and lowest values across non-\code{NA} cells.
\item \code{"mean"} (default): Average.
\item \code{"meanAbs"}: Mean of absolute values.
\item \code{"median"}: Median.
\item \code{"quantile"}: Quantile (see also argument \code{prob}).
\item \code{"range"}: Range--note that as per \code{\link[terra:global]{terra::global()}}, the minimum and maximum are reported, not the actual range.
\item \code{"sd"}: Sample standard deviation.
\item \code{"sdpop"}: Population standard deviation.
\item \code{"sum"}: Sum.
\item \code{"var"}: Sample variance.
\item \code{"varpop"}: Population variance.
}}

\item{prob}{Numeric: Quantile at which to calculate \code{quantile}. Only a single value between 0 and 1 is allowed.}

\item{...}{Other arguments (unused).}
}
\value{
If \code{x} is missing, the function returns a character vector of all accepted function names. If \code{x} is a \code{GRaster}, a data frame with the specified statistics is returned.
}
\description{
\code{global()} calculates a summary statistic across all the cells of a \code{GRaster}. It returns a single value for each layer of the raster.
}
\examples{
if (grassStarted()) {

# Setup
library(terra)

# Example data
madElev <- fastData("madElev")

# Convert a SpatRaster to a GRaster:
elev <- fast(madElev)

# Calculate global statistics:
global(elev, fun = c("mean", "countNonNA", "countNA", "var", "varpop"))

}
}
\seealso{
\code{\link[terra:global]{terra::global()}} and \strong{GRASS} module \code{r.univar}
}
