% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.r
\name{global,GRaster-method}
\alias{global,GRaster-method}
\alias{global}
\alias{global,missing-method}
\title{Summary statistics for GRasters}
\usage{
\S4method{global}{GRaster}(x, fun = "mean", prob = NULL, ...)

\S4method{global}{missing}(x, ...)
}
\arguments{
\item{x}{A \code{GRaster} or missing.  If missing, then a vector of all of the accepted function names is returned.}

\item{fun}{Character vector: The name of the function(s):
\itemize{
\item \code{'countNA'}: Total number of \code{NA} cells.
\item \code{'countNonNA'}: Total number of non-\code{NA} cells.
\item \code{'cv'}: Sample coefficient of variation (expressed as a proportion of the mean).
\item \code{'cvpop'}: Population coefficient of variation (expressed as a proportion of the mean).
\item \code{'max'} and \code{'min'}: Highest and lowest values across non-\code{NA} cells.
\item \code{'mean'} (default): Average.
\item \code{'meanAbs'}: Mean of absolute values.
\item \code{'median'}: Median.
\item \code{'quantile'}: Quantile (see also argument \code{prob}).
\item \code{'range'}: Range.
\item \code{'sd'}: Sample standard deviation.
\item \code{'sdpop'}: Population standard deviation.
\item \code{'sum'}: Sum.
\item \code{'var'}: Sample variance.
\item \code{'varpop'}: Population variance.
}}

\item{prob}{Numeric: Quantile at which to calculate \code{quantile}. Only a single value in the range \link{0, 1} is allowed.}

\item{...}{Other arguments (unused).}
}
\value{
If \code{x} is missing, the function returns a character vector of all accepted function names. If \code{x} is a \code{GRaster}, a data frame with the specified statistics is returned.
}
\description{
\code{global()} calculates a summary statistic across all the cells of a \code{GRaster}. It returns a single value for each layer of the raster.
}
\examples{
\dontrun{
# NB This example is in a "dontrun{}" block because it requires users to have
# GRASS GIS Version 8+ installed on their system.

# IMPORTANT #1: If you already have a GRASS session started, you will need to
# run the line below and the last line in this example to work with it again.
# If you have not started a GRASS session, you can skip this step and go to
# step #2.
opts. <- getFastOptions()

# IMPORTANT #2: Select the appropriate line below and change as necessary to
# where GRASS is installed on your system.

grassDir <- "/Applications/GRASS-8.2.app/Contents/Resources" # Mac
grassDir <- 'C:/Program Files/GRASS GIS 8.2' # Windows
grassDir <- '/usr/local/grass' # Linux

# setup
library(sf)
library(terra)

# example data
madElev <- fastData('madElev')

# start GRASS session for examples only
faster(grassDir = grassDir, crs = madElev,
workDir = tempdir(), location = 'examples') # line only needed for examples

# convert a SpatRaster to a GRaster
elev <- fast(madElev)
elevs <- c(elev, elev, log10(elev) - 1, sqrt(elev))
names(elevs) <- c('elev1', 'elev2', 'log_elev', 'sqrt_elev')

global(elevs, 'mean')
global(elevs, 'sum')
global(elevs, 'var')
global(elevs, 'meanAbs')

global(elev, 'median')
global(elev, 'quantile', prob=0.95)

# get a vector of all accepted functions and calculate them all
# ... can take a while!
funs <- global()
global(elevs, funs, prob=0.95)

# IMPORTANT #3: Revert back to original GRASS session if needed.
fastRestore(opts.)
removeSession('examples')

}
}
\seealso{
\code{\link[terra:global]{terra::global()}} and module \code{r.univar} in **GRASS
}
