% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.r
\name{global,GRaster-method}
\alias{global,GRaster-method}
\alias{global}
\alias{global,missing-method}
\title{Summary statistics for GRasters}
\usage{
\S4method{global}{GRaster}(x, fun = "mean", prob = NULL, ...)

\S4method{global}{missing}(x, ...)
}
\arguments{
\item{x}{A \code{GRaster} or missing.  If missing, then a vector of all of the accepted function names is returned.}

\item{fun}{Character vector: The name of the function(s):
\itemize{
\item \code{"countNonNA"}: Total number of non-\code{NA} cells.
\item \code{"countNA"}: Total number of \code{NA} cells.
\item \code{"cv"}: Sample coefficient of variation (expressed as a proportion of the mean).
\item \code{"cvpop"}: Population coefficient of variation (expressed as a proportion of the mean).
\item \code{"max"} and \code{"min"}: Highest and lowest values across non-\code{NA} cells.
\item \code{"mean"} (default): Average.
\item \code{"meanAbs"}: Mean of absolute values.
\item \code{"median"}: Median.
\item \code{"quantile"}: Quantile (see also argument \code{prob}).
\item \code{"range"}: Range.
\item \code{"sd"}: Sample standard deviation.
\item \code{"sdpop"}: Population standard deviation.
\item \code{"sum"}: Sum.
\item \code{"var"}: Sample variance.
\item \code{"varpop"}: Population variance.
}}

\item{prob}{Numeric: Quantile at which to calculate \code{quantile}. Only a single value between 0 and 1 is allowed.}

\item{...}{Other arguments (unused).}
}
\value{
If \code{x} is missing, the function returns a character vector of all accepted function names. If \code{x} is a \code{GRaster}, a data frame with the specified statistics is returned.
}
\description{
\code{global()} calculates a summary statistic across all the cells of a \code{GRaster}. It returns a single value for each layer of the raster.
}
\examples{
\dontrun{
# NB This example is in a "dontrun{}" block because it requires users to have
# GRASS GIS Version 8+ installed on their system.

# IMPORTANT #1: If you already have a GRASS session started, you will need to
# run the line below and the last line in this example to work with it again.
# If you have not started a GRASS session, you can skip this step and go to
# step #2.
opts. <- getFastOptions()

# IMPORTANT #2: Select the appropriate line below and change as necessary to
# where GRASS is installed on your system.

grassDir <- "/Applications/GRASS-8.3.app/Contents/Resources" # Mac
grassDir <- "C:/Program Files/GRASS GIS 8.3" # Windows
grassDir <- "/usr/local/grass" # Linux

# setup
library(terra)

# example data
madElev <- fastData("madElev")

# start GRASS session for examples only
faster(x = madElev, grassDir = grassDir,
workDir = tempdir(), location = "examples") # line only needed for examples

# Convert a SpatRaster to a GRaster:
elev <- fast(madElev)

# Create a random raster without spatial dependence:
rand <- randRast(elev)
plot(rand)

# IMPORTANT #3: Revert back to original GRASS session if needed.
restoreSession(opts.)
removeSession("examples")

}
}
\seealso{
\code{\link[terra:global]{terra::global()}} and \strong{GRASS} module \code{r.univar}
}
