% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_classes_and_slot_methods.r
\docType{class}
\name{GSession-class}
\alias{GSession-class}
\title{'fasterRaster' classes for locations, rasters, and vectors}
\value{
An object of class \code{GSession}, \code{GSpatial}, \code{GRaster}, or \code{GVector}.
}
\description{
S4 classes that contains \strong{GRASS} object representations in \strong{R}. Most users will manipulate objects using these classes, but do not need to know the details.
}
\details{
\code{GRaster} and \code{GVector} are the main "working" classes for \strong{fasterRaster} functions--most user will not need to engage directly with the others. \code{GSpatial} is contained by \code{GRaster} and \code{GVector} and is the general spatial class. It may be extended to include \strong{GRASS} regions later. \code{GSession} is the location class and contained by all the rest.
}
\section{Slots}{

\describe{
\item{\code{location}}{Character: The \strong{GRASS} \link{location} of the object. Default value is \code{default}. Typically hidden to users.}

\item{\code{mapset}}{Character: The \strong{GRASS} mapset. Defaiult value is \code{PERMANENT}. Typically hidden to users.}

\item{\code{gname}}{Character: Name of the object in \strong{GRASS}. These names are made on-the-fly and provide the pointer to the object from \strong{R} to \strong{GRASS}.  Changing them will break the connection.}

\item{\code{name}}{Character  (\code{GRaster}s only): Name of a raster or each raster layer in \strong{R} (obtained typically using \code{\link[=names]{names()}}).}

\item{\code{crs}}{Character: Coordinate reference systems string (preferably in WKT2 format).}

\item{\code{dimensions}}{Vector of four integers (\code{GRaster}s only): Number of rows, columns, depths (for 3D rasters), and layers of a "stack" of rasters.}

\item{\code{extent}}{Numeric vector with four values: Extent of the object listed in order from westernmost longitude, easternmost longitude, southernmost latitude, northernmost latitude.}

\item{\code{ztop,zbottom}}{Numeric: Top- and bottom-most elevations of 3D \code{GRaster}s and \code{GVector}s.}

\item{\code{topology}}{Character: Valid values are \verb{2D} (2-dimensional--most rasters and vectors) or \verb{3D} (3-dimensional--e.g., as in LIDAR data).}

\item{\code{geometry}}{Character (\code{GVectors}s only): Either \code{points}, \code{lines}, or \code{polygons}.}

\item{\code{datatypeGRASS}}{Character (\code{GRaster}s only): Type of data stored in a raster, as interpreted by \code{GRASS}. This is either \code{CELL} (integers), \code{FCELL} (floating-point values), or \code{DCELL} (double-values).}

\item{\code{resolution}}{Vector of two numeric values (\code{GRaster}s only): Size of a raster cell in the east-west direction and in the north-south direction.}

\item{\code{numCategories}}{Integer (\code{GRaster}s only): Number of categories. Must be >0.}

\item{\code{minVal,maxVal}}{Numeric (\code{GRaster}s only): Minimum and maximum value across all cells.}
}}

