% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist.r
\name{hist,GRaster-method}
\alias{hist,GRaster-method}
\alias{hist}
\title{Plot a histogram of raster values}
\usage{
\S4method{hist}{GRaster}(x, layer, maxnl = 16, bins = 30, freq = TRUE, ...)
}
\arguments{
\item{x}{A \code{GRaster}.}

\item{layer}{Character, numeric, or integer: Indicates which layer of a multi-layer \code{GRaster} for which to plot a histogram. The layer can be identified using its \code{\link[=name]{name()}} (character) or index (numeric or integer). If this is missing, then up to \code{maxnl} layers are plotted.}

\item{maxnl}{Maximum number of layers for which to create histograms. This is 16 by default, but ignored if \code{layer} is defined.}

\item{bins}{Positive numeric integer: Number of bins in which to divide values of a raster with continuous values. For \code{integer} and categorical rasters, each value is tallied.}

\item{freq}{Logical: If \code{TRUE} (default), plot the frequency of values. If \code{FALSE}, plot the density of values (i.e., the number in each bin divided by the total number of cells with non-\code{NA} values).}

\item{...}{Arguments to pass to \code{\link[graphics:barplot]{graphics::barplot()}}.}
}
\value{
A named list of \code{data.frame}s (invisibly), one per layer plotted, and creates a graph.
}
\description{
This function creates a histogram of values in \code{GRaster}. The function is modeled after \code{\link[graphics:hist]{graphics::hist()}}, but actually uses \code{\link[graphics:barplot]{graphics::barplot()}}.
}
\examples{
\dontrun{
# NB This example is in a "dontrun{}" block because it requires users to have
# GRASS GIS Version 8+ installed on their system.

# IMPORTANT #1: If you already have a GRASS session started, you will need to
# run the line below and the last line in this example to work with it again.
# If you have not started a GRASS session, you can skip this step and go to
# step #2.
opts. <- getFastOptions()

# IMPORTANT #2: Select the appropriate line below and change as necessary to
# where GRASS is installed on your system.
grassDir <- "/Applications/GRASS-8.3.app/Contents/Resources" # Mac
grassDir <- "C:/Program Files/GRASS GIS 8.3" # Windows
grassDir <- "/usr/local/grass" # Linux

# Setup
library(terra)

# Example data
madLand <- fastData("madLand")
madRivers <- fastData("madRivers")

# Start GRASS session for examples only:
faster(x = madLand, grassDir = grassDir,
workDir = tempdir(), location = "examples") # line only needed for examples

# Convert SpatRaster to GRaster and SpatVector to GVector
land <- fast(madLand)
rivers <- fast(madRivers)

# Plot:
plot(land[[1L]])
plot(rivers, add = TRUE)

# Histogram:
hist(land)

# Plot surface reflectance in RGB:
plotRGB(land, 3, 2, 1) # "natural" color
plotRGB(land, 4, 1, 2, stretch = "lin") # emphasize near-infrared (vegetation)

# IMPORTANT #3: Revert back to original GRASS session if needed.
restoreSession(opts.)
removeSession("examples")

}
}
