% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels.r
\name{levels<-,GRaster,data.frame-method}
\alias{levels<-,GRaster,data.frame-method}
\alias{levels<-}
\alias{levels<-,GRaster,data.table-method}
\alias{levels<-,GRaster,SpatRaster-method}
\alias{categories,GRaster-method}
\alias{categories}
\alias{levels,GRaster-method}
\alias{levels}
\title{Set and get categories for categorical rasters}
\usage{
\S4method{levels}{GRaster,data.frame}(x) <- value

\S4method{levels}{GRaster,data.table}(x) <- value

\S4method{levels}{GRaster,SpatRaster}(x) <- value

\S4method{categories}{GRaster}(x, layer = 1, value, active = 1)

\S4method{levels}{GRaster}(x)
}
\arguments{
\item{x}{A \code{GRaster}.}

\item{value}{A \code{data.frame}, \code{data.table}, a list of \code{data.frames} or \code{data.tables} with one per raster layer, or a categorical \code{SpatRaster}. The table's first column is the "value" column and must contain numeric values (of class \code{numeric} or \code{character}). If a \code{SpatRaster} is supplied, then its categories will be transferred to the \code{GRaster}.}
}
\value{
Values returned are:
\code{levels} list of \code{data.frame}s or \code{data.table}s, one per raster.
\verb{levels <-} A categorical \code{GRaster}.
}
\description{
\code{GRaster}s can represent \link[=tutorial_raster_data_types]{categorical data}. Cell values are actually integers, each corresponding to a category, such as "desert" or "wetland." A categorical raster is associated with a table that matches each value to a category name. The two \code{levels()} functions are related:
\itemize{
\item \code{levels()}: Reports category values and their codes (only those represented in the raster).
\item \verb{levels() <-}: Sets category values.
}
}
\examples{
\dontrun{
# NB This example is in a "dontrun{}" block because it requires users to have
# GRASS GIS Version 8+ installed on their system.

# IMPORTANT #1: If you already have a GRASS session started, you will need to
# run the line below and the last line in this example to work with it again.
# If you have not started a GRASS session, you can skip this step and go to
# step #2.
opts. <- getFastOptions()

# IMPORTANT #2: Select the appropriate line below and change as necessary to
# where GRASS is installed on your system.
grassDir <- "/Applications/GRASS-8.3.app/Contents/Resources" # Mac
grassDir <- "C:/Program Files/GRASS GIS 8.3" # Windows
grassDir <- "/usr/local/grass" # Linux

# Setup
library(terra)

# Example data: Land cover raster
madCover <- fastData("madCover")

# Start GRASS session for examples only
faster(x = madCover, grassDir = grassDir,
workDir = tempdir(), location = "examples") # line only needed for examples

# Convert categorical SpatRaster to categorical GRaster:
cover <- fast(madCover)

# set categories
cover # note... no categories
levs <- levels(madCover)
levs <- levs[[1]]
cover <- categories(cover, layer = 1, value = levs)

cover # Note display of "min"/"max" categories and number of categories
is.factor(cover) # Is the raster categorical?
ncat(cover) # Number of categories
levels(cover) # Categories

# Re-assign levels
value <- c(20, 30, 40, 50, 120, 130, 140, 170, 210)
label <- c("Cropland", "Cropland", "Forest", "Forest", "Forest", "Shrubland", "Herbaceous", "Flooded", "Water")

cats <- data.frame(value = value, label = label)
cover <- categories(cover, layer = 1, value = cats)
cover

# We can implement logical operations on categorical rasters
cover < "Forest" # 1 for cells with a value < 40, 0 otherwise
cover <= "Forest" # 1 for cells with a value < 120, 0 otherwise
cover == "Forest" # 1 for cells with value of 40-120, 0 otherwise
cover != "Forest" # 1 for cells with value that is not 40-120, 0 otherwise
cover > "Forest" # 1 for cells with a value > 120, 0 otherwise
cover >= "Forest" # 1 for cells with a value >= 120, 0 otherwise

cover \%in\% c("Cropland", "Forest") # 1 for cropland/forest cells, 0 otherwise

# IMPORTANT #3: Revert back to original GRASS session if needed.
fastRestore(opts.)
removeSession("examples")

}
}
\seealso{
\code{\link[=ncat]{ncat()}}, \code{\link[terra:factors]{terra::cats()}}, \code{\link[terra:factors]{terra::levels()}}, \code{\link[terra:factors]{terra::addCats()}}, \code{\link[terra:factors]{terra::droplevels()}}, \link[=tutorial_categorical_rasters]{categorical rasters}
}
