% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels.r
\name{levels<-,GRaster,data.frame-method}
\alias{levels<-,GRaster,data.frame-method}
\alias{levels<-}
\alias{levels<-,GRaster,data.table-method}
\alias{levels<-,GRaster,list-method}
\alias{levels,GRaster-method}
\alias{levels}
\title{Set and get categories for categorical rasters}
\usage{
\S4method{levels}{GRaster,data.frame}(x) <- value

\S4method{levels}{GRaster,data.table}(x) <- value

\S4method{levels}{GRaster,list}(x) <- value

\S4method{levels}{GRaster}(x)
}
\arguments{
\item{x}{A \code{GRaster}.}

\item{value}{A \code{data.frame}, \code{data.table}, or a list of \code{data.frames} or \code{data.tables}, one per raster layer. The table's first column is the "value" column and must contain numeric values (of class \code{numeric} or \code{character}). The second column defines category labels.  (Subsequent columns are ignored.) Note that in the "value" column you can specify ranges using formatting like \code{"1.3:10.2"} (to denote the given category should be assigned to all values in the range from 1.3 to 10.2).}
}
\value{
Values returned are:
\code{levels} list of \code{data.frame}s or \code{data.table}s, one per raster.
\verb{levels <-} A categorical \code{GRaster}.
}
\description{
\code{GRaster}s can represent \link[=tutorial_raster_data_types]{categorical data}. Cell values can be integers or non-integers, each corresponding to a category, such as "desert" or "wetland." A categorical raster is associated with a table that matches each value or range of values to a category name. The two \code{levels()} functions are related:
\itemize{
\item \code{levels()}: Reports category values and their codes (only those represented in the raster).
\item \verb{levels() <-}: Sets category values.
}
}
\examples{
\dontrun{
# NB This example is in a "dontrun{}" block because it requires users to have
# GRASS GIS Version 8+ installed on their system.

# IMPORTANT #1: If you already have a GRASS session started, you will need to
# run the line below and the last line in this example to work with it again.
# If you have not started a GRASS session, you can skip this step and go to
# step #2.
opts. <- getFastOptions()

# IMPORTANT #2: Select the appropriate line below and change as necessary to
# where GRASS is installed on your system.
grassDir <- "/Applications/GRASS-8.3.app/Contents/Resources" # Mac
grassDir <- "C:/Program Files/GRASS GIS 8.3" # Windows
grassDir <- "/usr/local/grass" # Linux

# Setup
library(terra)

# Example data: Land cover raster
madCover <- fastData("madCover")

# Start GRASS session for examples only
faster(x = madCover, grassDir = grassDir,
workDir = tempdir(), location = "examples") # line only needed for examples

# Convert categorical SpatRaster to categorical GRaster
cover <- fast(madCover)

# The GRaster inherited the SpatRaster's categories:
cover # Note display of "min"/"max" categories and number of categories
is.factor(cover) # Is the raster categorical?
ncat(cover) # Number of categories
levels(cover) # Categories

# Assign new levels. Note we can use ranges of values.
value <- c("11:30", "40:110", "120:150", "160:180", "190", "200:220", "230")
label <- c("Cropland", "Forest", "Grass/shrubland", "Flooded", "Artifical",
     "Bare/Water/Ice", "NA")

cats <- data.frame(value = value, label = label)
levels(cover) <- cats



# IMPORTANT #3: Revert back to original GRASS session if needed.
fastRestore(opts.)
removeSession("examples")

}
}
\seealso{
\code{\link[=ncat]{ncat()}}, \code{\link[terra:factors]{terra::cats()}}, \code{\link[terra:factors]{terra::levels()}}, \code{\link[terra:factors]{terra::addCats()}}, \code{\link[terra:factors]{terra::droplevels()}}, \link[=tutorial_categorical_rasters]{categorical rasters}
}
