% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.r
\name{.location,GLocation-method}
\alias{.location,GLocation-method}
\alias{.location}
\alias{.location,missing-method}
\title{GRASS "location" of an object or the active session}
\usage{
\S4method{.location}{GLocation}(x)

\S4method{.location}{missing}(x)
}
\arguments{
\item{x}{Either:
\itemize{
\item Missing: Reports location of currently active \link[=tutorial_locations_mapsets]{session}.
\item A \code{GLocation} object or an object that contains the \code{GLocation} class (i.e., a \code{GSpatial} object: a \code{GRaster} or \code{GVector}).
}}
}
\value{
A character string.
}
\description{
\strong{GRASS} \link[=tutorial_locations_mapsets]{"locations"} are sets of one or more rasters and/or vectors with the same coordinate reference systems, and may or may not represent the same actual location on Earth. \strong{GRASS} "mapsets" are like sub-folders of locations, and are collections of rasters and/or vectors typically related to the same general project. Typical users will not need to make changes to the default location (called "location") or mapset (called "PERMANENT").
}
\seealso{
\code{\link[=.mapset]{.mapset()}}
}
