% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations_and_mapsets.r
\name{location}
\alias{location}
\title{An explanation of GRASS "locations" and "mapsets"}
\description{
Most users of \strong{fasterRaster} will not need to understand or explicitly use \strong{GRASS} locations and mapsets. The default settings allow use of functions in \strong{GRASS} without users needing to explicitly define locations and mapsets. This help page is porvided to assist power users who may wish to use different locations/mapsets or develop their own applications based on \strong{fasterRaster}.

A \strong{GRASS} location is a set of vectors and/or rasters with the same coordinate reference system (CRS), that may or may not represent the same location on Earth. Rather, a location can be thought of as a folder in which spatial objects of the same CRS are stored. Mapsets are subfolders inside a location, and typically represent different projects or sub-projects. Locations are themselves inside a working directory. By default, the working directory is the temporary directory on the user's system (assigned using \code{\link[=tempdir]{tempdir()}}).  Hence, the "containments" of these components is as:

Multiple \code{mapset}s can be inside a single \code{location}, which is inside a \verb{working directory}

Rasters and vectors must typically be in the same location and mapset to be operated on by a single function, although they can be copied between locations/mapsets.  When the CRS is different between locations, rasters and vectors are automaticlaly reprojected "on the fly". This can be OK for non-essential work, but in practice, it is better to use \code{\link[=project]{project()}} to ensure the projecting is done in the manner you prefer.

User can exercise full control over working directory, location, and mapset when they initiate a \strong{GRASS} session using \code{\link[=fastStart]{fastStart()}}. The ellipse (\code{...}) argument allows users to assign a mapset or location, and the \code{workDir} argument allows for a user-defined working directory. The latter is especially helpful if the user wishes to keep the \strong{GRASS} session and restart it with the same spatial objects.

Users can switch between locations, mapsets, and working directories that have been created using \code{\link[=fastSrart]{fastSrart()}} using \code{\link[=sessionRestore]{sessionRestore()}}.
}
\keyword{tutorial}
