% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.r
\name{location,GSession-method}
\alias{location,GSession-method}
\alias{location}
\alias{location,missing-method}
\title{GRASS "location" of an object or the active session}
\usage{
\S4method{location}{GSession}(x)

\S4method{location}{missing}(x)
}
\arguments{
\item{x}{Either:
\itemize{
\item Missing: Reports location of currently active \link[=tutorial_sessions]{session}.
\item A \code{GSession} object or an object that contains the \code{GSession} class (i.e., a \code{GSpatial} object: a \code{GRaster} or \code{GVector}).
}}
}
\value{
A character string.
}
\description{
\strong{GRASS} \link[=tutorial_sessions]{"locations"} are sets of one or more rasters and/or vectors with the same coordinate reference systems, and may or may not represent the same actual location on Earth. \strong{GRASS} "mapsets" are like sub-folders of locations, and are collections of rasters and/or vectors typically related to the same general project. Typical users will not need to make changes to the default location (called "location") or mapset (called "PERMANENT").
}
\examples{
\dontrun{
# NB This example is in a "dontrun{}" block because it requires users to have
# GRASS GIS Version 8+ installed on their system.

# IMPORTANT #1: If you already have a GRASS session started, you will need to
# run the line below and the last line in this example to work with it again.
# If you have not started a GRASS session, you can skip this step and go to
# step #2.
opts. <- getFastOptions()

# IMPORTANT #2: Select the appropriate line below and change as necessary to
# where GRASS is installed on your system.
grassDir <- "/Applications/GRASS-8.3.app/Contents/Resources" # Mac
grassDir <- "C:/Program Files/GRASS GIS 8.3" # Windows
grassDir <- "/usr/local/grass" # Linux

# setup
library(terra)

# example data
madElev <- fastData("madElev")

# set up one location
faster(x = madElev, grassDir = grassDir,
workDir = tempdir(), location = "examples1") # line only needed for examples

# convert a SpatRaster to a GRaster using the first location
me1 <- fast(madElev)

# set up another location
faster(crs = madElev, grassDir = grassDir,
workDir = wd, location = "examples2") # line only needed for examples

# convert a SpatRaster to a GRaster using the second location
me2 <- fast(madElev)

location(me1)
location(me2)

# What GRASS location and mapset are we currently in?
location()
mapset()

# What coordinate reference system?
crs()

# We cannot do any work on me1 because we"re in the "examples2" location. We
# need to switch back to "examples1":
restoreSession(location = "examples1")

# IMPORTANT #3: Revert back to original GRASS session if needed.
restoreSession(opts.)
removeSession("examples1")
removeSession("examples2")

}
}
\seealso{
\code{\link[=mapset]{mapset()}}
}
