% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationRestore.r
\name{.locationRestore,character-method}
\alias{.locationRestore,character-method}
\alias{.locationRestore}
\alias{.locationRestore,integer-method}
\alias{.locationRestore,numeric-method}
\alias{.locationRestore,GSpatial-method}
\title{Revert to a previously-created "GRASS" "location"}
\usage{
\S4method{.locationRestore}{character}(x)

\S4method{.locationRestore}{integer}(x)

\S4method{.locationRestore}{numeric}(x)

\S4method{.locationRestore}{GSpatial}(x)
}
\arguments{
\item{x}{Either:
\itemize{
\item A character: Name of the "location" in \strong{GRASS}.
\item An integer: Index of the "location" in \code{.fasterRaster$locations}.
\item A \code{GSpatial} object (usually a \code{GRaster} or \code{GVector}).
}

Any of these can be found using \code{.locationFind()}.}
}
\value{
An object of class \code{GLocation} (invisibly) if successful. An error will likely result if not.
}
\description{
This function resets the connection to a previously-created \strong{GRASS} "location". The session must have been already created using \code{\link[=fast]{fast()}} in the current \strong{R} session. This function is typically only of use to developers.
}
\keyword{internal}
