% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.r
\name{.locations}
\alias{.locations}
\title{Meta-data on active GRASS locations}
\usage{
.locations(warn = TRUE)
}
\arguments{
\item{warn}{Logical: If \code{TRUE} (default), display a warning if no \strong{GRASS} "locations" have been created.}
}
\value{
A named \code{list}. The names are the "location's" names and the values are the coordinate reference strings.
}
\description{
\strong{GRASS} uses "locations"to store sets of rasters and vectors with the same coordinate reference system (CRS). These rasters and vectors may or may not be in the same actual location on Earth--they just have the same CRS. This function returns information on all of the \strong{GRASS} "locations" that have been initialized. It is mainly useful for developers.

NB: \strong{fasterRaster} always uses the "PERMANENT" mapset.
}
\keyword{internal}
