% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madChelsa.r
\docType{data}
\name{madChelsa}
\alias{madChelsa}
\title{Rasters of bioclimatic variables for an eastern portion of Madagascar}
\format{
An object of class \code{SpatRaster} in unprojected (WGS84) coordinates.
}
\source{
\doi{https://doi.org/10.1038/sdata.2017.122}
}
\description{
Rasters of bioclimatic variables for an eastern portion of Madagascar from CHELSA version 2.1 in unprojected (WGS84) coordinates. Values represent averages across 1980-2010. Only these BIOCLIM variables are included:
* \code{bio1}: Mean annual temperature (deg C)
* \code{bio7}: Temperature annual range (hottest - coldest month temperature; deg C)
* \code{bio12}: Total annual precipitation (mm)
* \code{bio15}: Precipitation seasonality (unitless)
}
\references{
Karger, D.N., Conrad, O., Böhner, J., Kawohl, T., Kreft, H., Soria-Auza, R.W., Zimmermann, N.E., Linder, H.P., and Kessler, M.  2017.  Climatologies at high resolution for the earth’s land surface areas. \emph{Scientific Data} 4:170122. \doi{https://doi.org/10.1038/sdata.2017.122}
}
\keyword{Madagascar}
\keyword{climate}
