% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madChelsa.r, R/madElev.r
\docType{data}
\name{madChelsa}
\alias{madChelsa}
\title{Rasters of bioclimatic variables for an eastern portion of Madagascar}
\format{
An object of class \code{'SpatRaster'} in unprojected (WGS84) coordinates.

An object of class \code{'SpatRaster'}. Values are mean meters above sea level.
}
\source{
\doi{https://doi.org/10.1038/sdata.2017.122}

\href{https://lta.cr.usgs.gov/GMTED2010}{GMTED2010}
}
\description{
Rasters of bioclimatic variables for an eastern portion of Madagascar from CHELSA version 2.1. Values represent averages across 1980-2010.

Rasters of bioclimatic variables for an eastern portion of Madagascar from CHELSA.
}
\examples{

### vector data
library(sf)

# for vector data, we can use data(*) or fasterData(*):
data(madCoast0) # same as next line
madCoast <- fasterData('madCoast0') # same as previous
plot(madCoast0)

madCoast <- fasterData('madCoast4')
plot(madCoast4, add = TRUE)

madRivers <- fasterData('madRivers')
plot(madRivers, add = TRUE)

### raster data
library(terra)

# for raster data, we can get the file directly or using fasterData(*):
rastFile <- system.file('extdata/madElev.tif'), package='fasterRaster')
madElev <- terra::rast(rastFile)

madElev <- fasterData('madElev') # same as previus two lines
plot(madElev)

madForest2000 <- fasterData('madForest2000')
plot(madForest2000, col = 'forestgreen', add = TRUE)

madForest2014 <- fasterData('madForest2014')
plot(madForest2014, col = 'green', add = TRUE)

### vector data
library(sf)

# for vector data, we can use data(*) or fasterData(*):
data(madCoast0) # same as next line
madCoast <- fasterData('madCoast0') # same as previous
plot(madCoast0)

madCoast <- fasterData('madCoast4')
plot(madCoast4, add = TRUE)

madRivers <- fasterData('madRivers')
plot(madRivers, add = TRUE)

### raster data
library(terra)

# for raster data, we can get the file directly or using fasterData(*):
rastFile <- system.file('extdata/madElev.tif'), package='fasterRaster')
madElev <- terra::rast(rastFile)

madElev <- fasterData('madElev') # same as previus two lines
plot(madElev)

madForest2000 <- fasterData('madForest2000')
plot(madForest2000, col = 'forestgreen', add = TRUE)

madForest2014 <- fasterData('madForest2014')
plot(madForest2014, col = 'green', add = TRUE)
}
\references{
Karger, D.N., Conrad, O., Böhner, J., Kawohl, T., Kreft, H., Soria-Auza, R.W., Zimmermann, N.E., Linder, H.P., and Kessler, M.  2017.  Climatologies at high resolution for the earth’s land surface areas. \emph{Scientific Data} 4:170122. \doi{https://doi.org/10.1038/sdata.2017.122}

Danielson, J.J. and Gesch, D.B. 2010. Global Multi-resolution Terrain Elevation Data 2010 (GMTED2010). US Geological Survey. Open-File Report 2011-1073 (\href{https://pubs.usgs.gov/of/2011/1073/pdf/of2011-1073.pdf}{GMTED2010})
}
\keyword{Madagascar}
