% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madChelsa.r
\docType{data}
\name{madChelsa}
\alias{madChelsa}
\title{Rasters of bioclimatic variables for an eastern portion of Madagascar}
\format{
An object of class \code{SpatRaster} in unprojected (WGS84) coordinates.
}
\source{
\doi{https://doi.org/10.1038/sdata.2017.122}
}
\description{
Rasters of bioclimatic variables for an eastern portion of Madagascar from CHELSA version 2.1 in unprojected (WGS84) coordinates. Values represent averages across 1980-2010. Only these BIOCLIM variables are included:
* \code{bio1}: Mean annual temperature (deg C)
* \code{bio7}: Temperature annual range (hottest - coldest month temperature; deg C)
* \code{bio12}: Total annual precipitation (mm)
* \code{bio15}: Precipitation seasonality (unitless)
}
\examples{

### vector data
library(sf)

# for vector data, we can use data(*) or fastData(*):
data(madCoast0) # same as next line
madCoast0 <- fastData("madCoast0") # same as previous
madCoast0

madCoast4 <- fastData("madCoast4")
madCoast4

madRivers <- fastData("madRivers")
madRivers

madDypsis <- fastData("madDypsis")
madDypsis

### raster data
library(terra)

# for raster data, we can get the file directly or using fastData(*):
rastFile <- system.file("extdata/madElev.tif", package="fasterRaster")
madElev <- terra::rast(rastFile)

madElev <- fastData("madElev") # same as previous two lines
madElev

madForest2000 <- fastData("madForest2000")
madForest2000

madForest2014 <- fastData("madForest2014")
madForest2014

madChelsa <- fastData("madChelsa")
madChelsa

madCover <- fastData("madCover")
madCover
madCover <- dropLevels(madCover)
levels(madCover)
}
\references{
Karger, D.N., Conrad, O., Böhner, J., Kawohl, T., Kreft, H., Soria-Auza, R.W., Zimmermann, N.E., Linder, H.P., and Kessler, M.  2017.  Climatologies at high resolution for the earth’s land surface areas. \emph{Scientific Data} 4:170122. \doi{https://doi.org/10.1038/sdata.2017.122}
}
\keyword{Madagascar}
\keyword{climate}
