% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madCoast0.r
\docType{data}
\name{madCoast0}
\alias{madCoast0}
\title{Spatial vector of a portion of the coastline of Madagascar}
\format{
An object of class \code{'sf'}.
}
\source{
Database of Global Administrative Areas Version 2.8 (\href{https://gadm.org}{GADM})
}
\usage{
data(madCoast4)
}
\description{
Borders of a selected portion of Madagascar
}
\examples{
man/examples/madData.r

}
\keyword{Madagascar}
