% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madDypsis.r
\docType{data}
\name{madDypsis}
\alias{madDypsis}
\title{Spatial points vector of records of Dypsis in eastern Madagascar}
\format{
An object of class \code{sf}.
}
\source{
\link[=https://gbif.org]{Global Biodiversity Information Facility (GBIF)}
}
\usage{
data(madDypsis)
}
\description{
Spatial points vector of herbarium specimens and observations of plants in the genus \emph{Dypsis}  (slender, evergreen palms) from a portion of eastern Madagascar.
}
\keyword{Dypsis}
\keyword{Madagascar}
