% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madElev.r
\docType{data}
\name{madElev}
\alias{madElev}
\title{Digital elevation model (DEM) of a portion of the coastline of Madagascar}
\format{
An object of class \code{'SpatRaster'}. Values are mean meters above sea level.
}
\source{
\href{https://lta.cr.usgs.gov/GMTED2010}{GMTED2010}
}
\usage{
data(madElev)
}
\description{
Rasterized digital elevation model (DEM) of a portion of Madagascar at 7.5-arcmin spatial resolution.
}
\examples{
man/examples/madData.r

}
\references{
Danielson, J.J. and Gesch, D.B. 2010. Global Multi-resolution Terrain Elevation Data 2010 (GMTED2010). US Geological Survey. Open-File Report 2011-1073 (\href{https://pubs.usgs.gov/of/2011/1073/pdf/of2011-1073.pdf}{GMTED2010})
}
\keyword{elevation}
