% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madElev.r
\docType{data}
\name{madElev}
\alias{madElev}
\title{Elevation raster for an eastern portion of Madagascar}
\format{
An object of class \code{'SpatRaster'}. Values are mean meters above sea level.
}
\source{
\href{https://worldclim.org}{WorldClim}
}
\description{
Elevation raster for an eastern portion of Madagascar.
}
\examples{

### vector data
library(sf)

# for vector data, we can use data(*) or fasterData(*):
data(madCoast0) # same as next line
madCoast <- fasterData('madCoast0') # same as previous
plot(madCoast0)

madCoast <- fasterData('madCoast4')
plot(madCoast4, add = TRUE)

madRivers <- fasterData('madRivers')
plot(madRivers, add = TRUE)

### raster data
library(terra)

# for raster data, we can get the file directly or using fasterData(*):
rastFile <- system.file('extdata/madElev.tif'), package='fasterRaster')
madElev <- terra::rast(rastFile)

madElev <- fasterData('madElev') # same as previus two lines
plot(madElev)

madForest2000 <- fasterData('madForest2000')
plot(madForest2000, col = 'forestgreen', add = TRUE)

madForest2014 <- fasterData('madForest2014')
plot(madForest2014, col = 'green', add = TRUE)
}
\keyword{Madagascar}
\keyword{elevation}
