% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madForest2000.r
\docType{data}
\name{madForest2000}
\alias{madForest2000}
\title{Spatial data: forest cover in 2000 for a portion of Madagascar}
\format{An object of class \code{'raster'}. See \code{\link[raster]{raster}}. Values are forest (1) or not forest (\code{NA}).}
\source{
\href{https://doi.org/10.1016/j.biocon.2018.04.008}{Viellendent et al. 2018}
}
\usage{
data(madForest2000)
}
\description{
Raster of occurrence/non-occurrence of forest cover in a portion of Madagascar. Cells are 30-m in resolution. Values represent forest (1) or non-forest (\code{NA}).
}
\examples{
data(madForest2000)
plot(madForest2000)
}
\references{
Vielledent, G., Grinand, C., Rakotomala, F.A., Ranaivosoa, R., Rakotoarijaona, J-R., Allnutt, T.F., and Achard, F.  2018.  Combining global tree cover loss data with historical national forest cover maps to look at six decades of deforestation and forest fragmentation in Madagascar.  Biological Conservation 222:189-197. (\href{https://www.ncbi.nlm.nih.gov/pubmed/23979570}{PubMed})
}
\keyword{datasets}
