% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madRivers.r
\docType{data}
\name{madRivers}
\alias{madRivers}
\title{Major rivers in a selected portion of Madagascar}
\format{
An object of class \code{sf}.
}
\source{
\link[=http://www.diva-gis.org/gdata]{DIVA-GIS}
}
\usage{
data(madRivers)
}
\description{
Spatial lines object of major rivers in a portion of Madagascar.
}
\keyword{Madagascar}
