% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madRivers.r
\docType{data}
\name{madRivers}
\alias{madRivers}
\title{Major rivers in a selected portion of Madagascar}
\format{
An object of class \code{sf}.
}
\source{
\href{http://www.diva-gis.org/gdata}{DIVA-GIS}
}
\usage{
data(madRivers)
}
\description{
Spatial lines object of major rivers in a portion of Madagascar.
}
\examples{

### vector data
library(sf)

# for vector data, we can use data(*) or fastData(*):
data(madCoast0) # same as next line
madCoast0 <- fastData("madCoast0") # same as previous
madCoast0

madCoast4 <- fastData("madCoast4")
madCoast4

madRivers <- fastData("madRivers")
madRivers

madDypsis <- fastData("madDypsis")
madDypsis

### raster data
library(terra)

# for raster data, we can get the file directly or using fastData(*):
rastFile <- system.file("extdata/madElev.tif", package="fasterRaster")
madElev <- terra::rast(rastFile)

madElev <- fastData("madElev") # same as previous two lines
madElev

madForest2000 <- fastData("madForest2000")
madForest2000

madForest2014 <- fastData("madForest2014")
madForest2014

madChelsa <- fastData("madChelsa")
madChelsa
}
\keyword{Madagascar}
