% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_GRaster_functions_byLayer.r
\name{is.na,GRaster-method}
\alias{is.na,GRaster-method}
\alias{is.na}
\alias{not.na,GRaster-method}
\alias{not.na}
\alias{abs,GRaster-method}
\alias{abs}
\alias{sin,GRaster-method}
\alias{sin}
\alias{cos,GRaster-method}
\alias{cos}
\alias{tan,GRaster-method}
\alias{tan}
\alias{asin,GRaster-method}
\alias{asin}
\alias{acos,GRaster-method}
\alias{acos}
\alias{atan,GRaster-method}
\alias{atan}
\alias{atan2,GRaster,GRaster-method}
\alias{atan2}
\alias{exp,GRaster-method}
\alias{exp}
\alias{log1p,GRaster-method}
\alias{log1p}
\alias{log,GRaster-method}
\alias{log}
\alias{ln,GRaster-method}
\alias{ln}
\alias{log2,GRaster-method}
\alias{log2}
\alias{log10,GRaster-method}
\alias{log10}
\alias{sqrt,GRaster-method}
\alias{sqrt}
\alias{round,GRaster-method}
\alias{round}
\alias{floor,GRaster-method}
\alias{floor}
\alias{ceiling,GRaster-method}
\alias{ceiling}
\alias{trunc,GRaster-method}
\alias{trunc}
\title{Mathematical operations on single GRasters}
\usage{
\S4method{is.na}{GRaster}(x)

\S4method{not.na}{GRaster}(x, falseNA = FALSE)

\S4method{abs}{GRaster}(x)

\S4method{sin}{GRaster}(x)

\S4method{cos}{GRaster}(x)

\S4method{tan}{GRaster}(x)

\S4method{asin}{GRaster}(x)

\S4method{acos}{GRaster}(x)

\S4method{atan}{GRaster}(x)

\S4method{atan2}{GRaster,GRaster}(y, x)

\S4method{exp}{GRaster}(x)

\S4method{log1p}{GRaster}(x)

\S4method{log}{GRaster}(x, base = exp(1))

\S4method{ln}{GRaster}(x)

\S4method{log2}{GRaster}(x)

\S4method{log10}{GRaster}(x)

\S4method{sqrt}{GRaster}(x)

\S4method{round}{GRaster}(x, digits = 0)

\S4method{floor}{GRaster}(x)

\S4method{ceiling}{GRaster}(x)

\S4method{trunc}{GRaster}(x)
}
\arguments{
\item{x, y}{\code{GRaster}s.}

\item{falseNA}{Function \code{not.na()}, logical: If \code{FALSE} (default), non-\code{NA} cells will be converted to 1, and \code{NA} cells to 0. If \code{TRUE}, non-\code{NA} cells will be converted to  and \code{NA} cells will stay as \code{NA}.}

\item{base}{Numeric: Base of the logarithm.}

\item{digits}{Numeric: Number of digits to round to. If negative, then rounding is to the nearest power of 10. For example, if \code{digits = -2}, then the \code{GRaster} values are rounded to the nearest 100.}
}
\value{
A \code{GRaster}.
}
\description{
You can apply mathematical operators to \code{GRaster}s. These include:\cr
\itemize{
\item \code{NA}s: \code{is.na()}, \code{!is.na()}, and \code{not.na()}
\item Absolute value: \code{abs()}
\item Trigonometric functions (assumes values are in radians): \code{cos()}, \code{sin()}, \code{tan()}, \code{acos()}, \code{asin()}, \code{atan()}, \code{atan2()}
\item Exponential and logarithmic functions: \code{exp()}, \code{log()} (natural log), \code{ln()} (also natural log), \code{log1p()} (same as \code{log(x + 1)}), \code{log2()} (log, base 2), \code{log10()} (log, base 10)
\item Power functions: \code{sqrt()}, \code{x^y}
\item Rounding: \code{round()}, \code{floor()}, \code{ceiling()}, \code{trunc()}
}
}
\examples{
\dontrun{
# NB This example is in a "dontrun{}" block because it requires users to have
# GRASS GIS Version 8+ installed on their system.

# IMPORTANT #1: If you already have a GRASS session started, you will need to
# run the line below and the last line in this example to work with it again.
# If you have not started a GRASS session, you can skip this step and go to
# step #2.
opts. <- getFastOptions()

# IMPORTANT #2: Select the appropriate line below and change as necessary to
# where GRASS is installed on your system.

grassDir <- "/Applications/GRASS-8.3.app/Contents/Resources" # Mac
grassDir <- "C:/Program Files/GRASS GIS 8.3" # Windows
grassDir <- "/usr/local/grass" # Linux

# setup
library(sf)
library(terra)

# example data
madElev <- fastData("madElev")

# start GRASS session for examples only
faster(x = madElev, grassDir = grassDir,
workDir = tempdir(), location = "examples") # line only needed for examples

# convert a SpatRaster to a GRaster
elev <- fast(madElev)
elevs <- c(elev, elev, log10(elev) - 1, sqrt(elev))
names(elevs) <- c("elev1", "elev2", "log_elev", "sqrt_elev")

elev
elevs

# do some math
elev + 100
elev - 100
elev * 100
elev / 100
elev ^ 2
elev \%/\% 100 # divide then round down
elev \%\% 100 # modulus

100 + elev
100 \%/\% elev
100 \%\% elev

elevs + 100
100 + elevs

# math with logicals
elev + TRUE
elev - TRUE
elev * TRUE
elev / TRUE
elev ^ TRUE
elev \%/\% TRUE # divide then round down
elev \%\% TRUE # modulus

elevs + TRUE
TRUE + elevs

# raster interacting with raster
elev + elev
elev - elev
elev * elev
elev / elev
elev ^ log(elev)
elev \%/\% sqrt(elev) # divide then round down
elev \%\% sqrt(elev) # modulus

elevs + elev
elev * elevs

# sign
abs(-1 * elev)
abs(elevs)
sqrt(elevs)

# trigonometry
sin(elev)
cos(elev)
tan(elev)

asin(elev)
acos(elev)
atan(elev)

atan(elevs)
atan2(elev, elev^1.2)
atan2(elevs, elev^1.2)
atan2(elev, elevs^1.2)
atan2(elevs, elevs^1.2)

# logarithms
exp(elev)
log(elev)
ln(elev)
log2(elev)
log1p(elev)
log10(elev)
log(elev, 3)

log(elevs)

# rounding
round(elev + 0.5)
floor(elev + 0.5)
ceiling(elev + 0.5)
trunc(elev + 0.5)

# comparison
elev < 100
elev <= 100
elev == 100
elev != 100
elev > 100
elev >= 100

elev + 100 < 2 * elev

elevs > 10
10 > elevs

# mathematical functions on two or more rasters
mean(elevs)
mmode(elevs)
median(elevs)
nunique(elevs)

sum(elevs)
count(elevs)
min(elevs)
max(elevs)
range(elevs)
skewness(elevs)
kurtosis(elevs)

which.min(elevs)
which.max(elevs)

slope(elevs)
intercept(elevs)
r2(elevs)
tvalue(elevs)

sd(elevs)
sdpop(elevs)
var(elevs)
varpop(elevs)
quantile(elevs, 0.1)

# IMPORTANT #3: Revert back to original GRASS session if needed.
fastRestore(opts.)
removeSession("examples")

}
}
