% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mow.r
\name{mow}
\alias{mow}
\title{Remove unused rasters and vectors from the GRASS cache}
\usage{
mow(x, type = NULL, keep = NULL, verbose = TRUE, ask = TRUE)
}
\arguments{
\item{x}{Either missing (default) or an environment.}

\item{type}{Either \code{NULL} or a character vector. If \code{NULL}, all rasters and vectors in the \strong{GRASS} cache are candidates for deletion. Otherwise, this can be either \code{"raster"}, \code{"vector"}, or both.}

\item{keep}{Either \code{NULL} (default) or a \code{list()} of \code{GRaster}s and/or \code{GVector}s that you want to retain. The rasters and vectors in \strong{GRASS} pointed to by these objects will not be deleted.}

\item{verbose}{Logical: If \code{TRUE} (default), report progress.}

\item{ask}{Logical: If \code{TRUE} (default), prompt for reassurance.}
}
\value{
Invisibly returns a list with the number of rasters and vectors deleted.
}
\description{
\strong{fasterRaster} attempts to delete rasters and vectors in the \strong{GRASS} cache. This function will a) search the current \strong{GRASS} "project/location" cache for all rasters/vectors there; and b) remove any of them that are not pointed to by an object in the active \strong{R} environment. Only objects in the currently active \strong{GRASS} project/location will be removed (see \code{vignette("project_mapset", package = "fasterRaster")}).

Note that calling this function inside another function's environment without providing an argument for \code{x} can be very \strong{dangerous}, as it will detect objects outside that environment, and thus delete any rasters/vectors outside that environment.
}
\examples{
if (grassStarted()) {

if (FALSE) mow()

}
}
\seealso{
Option \code{clean} in \code{\link[=faster]{faster()}}
}
