% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mow.r
\name{mow}
\alias{mow}
\title{Remove unused rasters and vectors from the GRASS cache}
\usage{
mow(x, type = NULL, verbose = TRUE, ask = TRUE)
}
\arguments{
\item{x}{Either missing (default) or an environment.}

\item{type}{Either \code{NULL} or a character vector. If \code{NULL}, all rasters and vectors in the \strong{GRASS} cache are candidates for deletion. Otherwise, this can be either \code{"raster"}, \code{"vector"}, or both.}

\item{verbose}{Logical: If \code{TRUE} (default), report progress.}

\item{ask}{Logical: If \code{TRUE} (default), prompt for reassurance.}
}
\value{
Invisibly returns a list with the number of rasters and vectors deleted.
}
\description{
\strong{fasterRaster} attempts to remove intermediate rasters and vectors from the \strong{GRASS} cache as they are not needed, but files can still accumulate, especially if you remove \code{GRaster}s or \code{GVector}s from the \strong{R} working environment (e.g., by using \code{\link[=rm]{rm()}} or simply using the same variable name for different rasters/vectors). This function will a) search the \strong{GRASS} cache for all rasters/vectors there; and b) remove any of them that are not pointed to by an object in the active \strong{R} environment.

Note that calling this function inside another function's environment can be very dangerous, as it will only be able to see objects in that environment, and thus delete any rasters/vectors outside that environment.

Note also that this function will only clean the current \strong{GRASS} "project"/"location" (see \code{vignette("10_projects_locations_mapsets", package = "fasterRaster")}).
}
\examples{
if (grassStarted()) {

if (FALSE) mow()

}
}
\seealso{
Option \code{clean} in \code{\link[=faster]{faster()}}
}
