% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.r
\name{names,GRaster-method}
\alias{names,GRaster-method}
\alias{names}
\alias{names<-,GRaster-method}
\alias{names<-}
\alias{names,GVector-method}
\title{Name(s) of a GRaster or fields of a GVector}
\usage{
\S4method{names}{GRaster}(x)

\S4method{names}{GRaster}(x) <- value

\S4method{names}{GVector}(x)
}
\arguments{
\item{x}{A \code{GRaster} or \code{GVector}.}

\item{value}{Character: Name(s) to assign to the raster(s).}
}
\value{
Character vector.
}
\description{
Names(s) of a \code{GRaster} or of fields of a \code{GVector}.
}
\details{
\code{GRaster}s have two types of names, a "gnames", which is used internally to point to the \strong{GRASS} representation of the raster (this is not of use to most users and is hidden by default), and an "rname", which is taken from the "\strong{R}" name of the raster or the filename. This function returns the "rname".
}
\examples{
\dontrun{
# NB This example is in a "dontrun{}" block because it requires users to have
# GRASS GIS Version 8+ installed on their system.

# IMPORTANT #1: If you already have a GRASS session started, you will need to
# run the line below and the last line in this example to work with it again.
# If you have not started a GRASS session, you can skip this step and go to
# step #2.
opts. <- getFastOptions()

# IMPORTANT #2: Select the appropriate line below and change as necessary to
# where GRASS is installed on your system.
grassDir <- "/Applications/GRASS-8.3.app/Contents/Resources" # Mac
grassDir <- "C:/Program Files/GRASS GIS 8.3" # Windows
grassDir <- "/usr/local/grass" # Linux

# setup
library(terra)

# example data
madElev <- fastData("madElev")
madForest2000 <- fastData("madForest2000")

# start GRASS session for examples only
faster(x = madElev, grassDir = grassDir,
workDir = tempdir(), location = "examples") # line only needed for examples

# convert SpatRasters to GRasters
elev <- fast(madElev)
forest <- fast(madForest2000)

### GRaster properties

# dimensions
dim(elev) # rows, columns, depths, layers
nrow(elev) # rows
ncol(elev) # columns
ndepth(elev) # depths
nlyr(elev) # layers

res(elev) # resolution (2D)
res3d(elev) # resolution (3D)
zres(elev) # vertical resolution

# cell counts
ncell(elev) # cells
ncell3d(elev) # cells (3D rasters only)

# number of NA and non-NA cells
nacell(elev)
nonnacell(elev)

# topology
topology(elev) # number of dimensions
is.2d(elev) # is it 2-dimensional?
is.3d(elev) # is it 3-dimensional?

minmax(elev) # min/max values

# information on the GRASS session in which the GRaster is located
location(elev) # location
mapset(elev) # mapset

# "gnames" of the object (its name in GRASS)
gnames(elev)

# coordinate reference system
crs(elev)
st_crs(elev)

# extent (bounding box)
ext(elev)
st_bbox(elev)

# vertical extent (not defined for this raster)
zext(elev)

# data type
datatype(elev)

# convert data type
as.cell(elev) # integer; note that "elev" is already of type "CELL"
as.fcell(elev) # floating-precision
as.dcell(elev) # double-precision

# concatenating multiple GRasters
rasts <- c(elev, forest)
rasts

# number of layers
nlyr(rasts)

# names
names(rasts)
names(rasts) <- c("elev_meters", "forest")
rasts

# cell frequencies
freq(elev)
freq(2 * elev)
freq(rasts)
freq(rasts, value = 1)

# plotting
plot(elev)

# IMPORTANT #3: Revert back to original GRASS session if needed.
fastRestore(opts.)
removeSession("examples")

}
}
\seealso{
\code{\link[terra:names]{terra::names()}}
}
