% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs.r
\name{pairs,GRaster-method}
\alias{pairs,GRaster-method}
\alias{pairs}
\title{Scatterplot of values in each GRaster layer against the others}
\usage{
\S4method{pairs}{GRaster}(x, n = NULL, ...)
}
\arguments{
\item{x}{A \code{GRaster} with two or more layers.}

\item{n}{A numeric integer, integer, or \code{NULL} (default): Number of cells to sample. If \code{NULL}, 50\% of the total number of cells will be used.}

\item{...}{Arguments to send to \code{\link[graphics:pairs]{graphics::pairs()}} (which typically sends them to \code{\link[graphics:plot.default]{graphics::plot()}}). Special arguments for affecting certain panels include:
\itemize{
\item Correlation panels -- \code{cor.cex}: Size of the values of the correlation coefficients.
\item Histogram panels -- \code{hist.col}: Histogram color.
\item Dot-plot panels -- \code{colramp}: Function taking an integer \code{n} as input and returning \code{n} names of colors. The default is: \code{colorRampPalette(c("white", blues9))}.
}}
}
\value{
Nothing (creates a plot).
}
\description{
\code{pairs()} generates a scatterplot between values of cells in each layer of a \code{GRaster} against all the other layers.
}
\examples{
if (grassStarted()) {

# Setup
library(terra)

# Example data
madElev <- fastData("madElev")
madForest2000 <- fastData("madForest2000")

# Convert SpatRasters to GRasters
elev <- fast(madElev)
forest <- fast(madForest2000)

### GRaster properties

# plotting
plot(elev)

# dimensions
dim(elev) # rows, columns, depths, layers
nrow(elev) # rows
ncol(elev) # columns
ndepth(elev) # depths
nlyr(elev) # layers

res(elev) # resolution (2D)
res3d(elev) # resolution (3D)
zres(elev) # vertical resolution
xres(elev) # vertical resolution
yres(elev) # vertical resolution
zres(elev) # vertical resolution (NA because this is a 2D GRaster)

# cell counts
ncell(elev) # cells
ncell3d(elev) # cells (3D rasters only)

# number of NA and non-NA cells
nacell(elev)
nonnacell(elev)

# topology
topology(elev) # number of dimensions
is.2d(elev) # is it 2-dimensional?
is.3d(elev) # is it 3-dimensional?

minmax(elev) # min/max values

# "names" of the object
names(elev)

# coordinate reference system
crs(elev)
st_crs(elev)
coordRef(elev)

# extent (bounding box)
ext(elev)

# vertical extent (not defined for this raster)
zext(elev)

# data type
datatype(elev) # fasterRaster type
datatype(elev, "GRASS") # GRASS type
datatype(elev, "terra") # terra type
datatype(elev, "GDAL") # GDAL type

is.integer(elev)
is.float(elev)
is.double(elev)
is.factor(elev)

# convert data type
as.int(elev) # integer; note that "elev" is already of type "integer"
as.float(elev) # floating-precision
as.doub(elev) # double-precision

# assigning
pie <- elev
pie[] <- pi # assign all cells to the value of pi
pie

# concatenating multiple GRasters
rasts <- c(elev, forest)
rasts

# subsetting
rasts[[1]]
rasts[["madForest2000"]]

# replacing
rasts[[2]] <- 2 * forest
rasts

# adding layers
rasts[[3]] <- elev > 500 # add a layer
rasts <- c(rasts, sqrt(elev)) # add another
add(rasts) <- ln(elev)
rasts

# names
names(rasts)
names(rasts) <- c("elev_meters", "2_x_forest", "high_elevation", "sqrt_elev", "ln_elev")
rasts

# remove a layer
rasts[["2_x_forest"]] <- NULL
rasts

# number of layers
nlyr(rasts)

# correlation and covariance matrices
madLANDSAT <- fastData("madLANDSAT")
landsat <- fast(madLANDSAT) # projects matrix
layerCor(landsat) # correlation
layerCor(landsat, fun = 'cov') # covariance

}
}
