% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.r
\name{pca,GRaster-method}
\alias{pca,GRaster-method}
\alias{pca}
\title{Apply a principal component analysis (PCA) to a GRaster}
\usage{
\S4method{pca}{GRaster}(x, scale = TRUE, scores = FALSE)
}
\arguments{
\item{x}{A \code{GRaster} with two or more layers.}

\item{scale}{Logical: If \code{TRUE}, input layers will be rescaled by dividing each layer by its overall population standard deviation. Note that rasters will always be centered (have their mean subtracted from values). Centering and scaling is recommended when rasters values are in different units. The default is \code{FALSE} (do not scale).}

\item{scores}{Logical: If \code{TRUE}, the \code{prcomp} object will have the scores attached to it. This can vastly increase the size of the object and take a lot of memory if the input raster has many cells. It will also take more time. If \code{FALSE} (default), then skip returning scores.}
}
\value{
A multi-layer \code{GRaster} with one layer per principal component axis. The \code{\link[=pcs]{pcs()}} function can be used on the output raster to retrieve a \code{prcomp} object from the raster, which includes rotations (loadings) and proportions of variance explained.
}
\description{
\code{pca()} applies a principal component analysis to layers of a \code{GRaster}.
}
\examples{
\dontrun{
# NB This example is in a "dontrun{}" block because it requires users to have
# GRASS GIS Version 8+ installed on their system.

# IMPORTANT #1: If you already have a GRASS session started, you will need to
# run the line below and the last line in this example to work with it again.
# If you have not started a GRASS session, you can skip this step and go to
# step #2.
opts. <- getFastOptions()

# IMPORTANT #2: Select the appropriate line below and change as necessary to
# where GRASS is installed on your system.
grassDir <- "/Applications/GRASS-8.3.app/Contents/Resources" # Mac
grassDir <- "C:/Program Files/GRASS GIS 8.3" # Windows
grassDir <- "/usr/local/grass" # Linux

# Setup
library(terra)

# Climate raster:
madChelsa <- fastData("madChelsa")

# Start GRASS session for examples only:
faster(x = madChelsa, grassDir = grassDir,
workDir = tempdir(), location = "examples") # line only needed for examples

# Convert a SpatRaster to a GRaster:
chelsa <- fast(madChelsa)

# Generate raster with layers representing principal component predictions:
pcRast <- pca(chelsa, scale = TRUE)
plot(pcRast)

# Get information on the PCA:
prinComp <- pcs(pcRast)

prinComp
summary(prinComp)
plot(prinComp)

# IMPORTANT #3: Revert back to original GRASS session if needed.
restoreSession(opts.)
removeSession("examples")

}
}
\seealso{
\code{\link[stats:prcomp]{stats::prcomp()}}; module \code{i.pca} in \strong{GRASS}
}
