% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot,GRaster,missing-method}
\alias{plot,GRaster,missing-method}
\alias{plot}
\alias{plot,GVector,missing-method}
\title{Display a raster or vector}
\usage{
\S4method{plot}{GRaster,missing}(x, y, maxcell = 5e+05, ...)

\S4method{plot}{GVector,missing}(x, y, simplify = FALSE, maxGeoms = 10000, ...)
}
\arguments{
\item{x}{A \code{GRaster} or \code{GVector}.}

\item{y}{Missing--leave as empty.}

\item{maxcell}{Positive integer (rasters only): Maximum number of cells to display. When simplifying, \code{\link[=aggregate]{aggregate()}} will be applied so that it has approximately this number of cells. The default is 500000.}

\item{...}{Other arguments to send to \code{\link[terra:plot]{terra::plot()}}.}

\item{simplify}{Logical (vectors only): If \code{TRUE} (default), then simplify the \code{GVector} before plotting. This can save time for very large vectors. However, details in the vector may appear inaccurate.}

\item{maxGeoms}{Positive integer (vectors only): Maximum number of features before simplification is used (\code{simplify} must also be \code{TRUE}).}
}
\value{
Nothing (displays a raster or vector).
}
\description{
\code{plot()} displays a \code{GRaster} or \code{GVector}.

This function is essentially a hack, as it it not possible to dependably call the appropriate \strong{GRASS} modules and display a raster or vector without potential confusion on the user side. Instead, this function 1) simplifies the focal \code{GRaster} or \code{GVector} (if needed); 2) writes it to disk as a \code{SpatRaster} or \code{SpatVector}; and 3) plots the object.
}
\examples{
\dontrun{
# NB This example is in a "dontrun{}" block because it requires users to have
# GRASS GIS Version 8+ installed on their system.

# IMPORTANT #1: If you already have a GRASS session started, you will need to
# run the line below and the last line in this example to work with it again.
# If you have not started a GRASS session, you can skip this step and go to
# step #2.
opts. <- getFastOptions()

# IMPORTANT #2: Select the appropriate line below and change as necessary to
# where GRASS is installed on your system.
grassDir <- "/Applications/GRASS-8.3.app/Contents/Resources" # Mac
grassDir <- "C:/Program Files/GRASS GIS 8.3" # Windows
grassDir <- "/usr/local/grass" # Linux

# Setup
library(terra)

# Example data
madLand <- fastData("madLand")
madRivers <- fastData("madRivers")

# Start GRASS session for examples only:
faster(x = madLand, grassDir = grassDir,
workDir = tempdir(), location = "examples") # line only needed for examples

# Convert SpatRaster to GRaster and SpatVector to GVector
land <- fast(madLand)
rivers <- fast(madRivers)

# Plot:
plot(land[[1L]])
plot(rivers, add = TRUE)

# Histogram:
hist(land)

# Plot surface reflectance in RGB:
plotRGB(land, 3, 2, 1) # "natural" color
plotRGB(land, 4, 1, 2, stretch = "lin") # emphasize near-infrared (vegetation)

# Make composite map from RGB and plot in grayscale:
comp <- compositeRGB(r = land[[3]], g = land[[2]], b = land[[1]])
grays <- paste0("gray", 0:100)
plot(comp, col = grays)

# IMPORTANT #3: Revert back to original GRASS session if needed.
restoreSession(opts.)
removeSession("examples")

}
}
\seealso{
\code{\link[terra:plot]{terra::plot()}}
}
