% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast.r
\name{rast,GRaster-method}
\alias{rast,GRaster-method}
\alias{rast}
\title{Convert a GRaster to a SpatRaster}
\usage{
\S4method{rast}{GRaster}(x, ...)
}
\arguments{
\item{x}{A \code{GRaster}.}

\item{...}{Additional arguments to send to \code{\link[=writeRaster]{writeRaster()}} (typically unneeded, though \code{bigTiff} may be of use if the raster is large).}
}
\value{
A \code{SpatRaster} (\strong{terra} package).
}
\description{
The \strong{fasterRaster} version of the \code{rast()} function converts a \code{GRaster} to a \code{SpatRaster} (from the \strong{terra} package).
}
\seealso{
\code{\link[terra:rast]{terra::rast()}}
}
