% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterize.r
\name{rasterize,GVector,GRaster-method}
\alias{rasterize,GVector,GRaster-method}
\alias{rasterize}
\title{Convert a GVector to a GRaster}
\usage{
\S4method{rasterize}{GVector,GRaster}(x, y, background = NA, byGeom = FALSE, collapse = TRUE)
}
\arguments{
\item{x}{A \code{GVector}.}

\item{y}{A \code{GRaster}: The new raster will have the same extent and resolution as this raster.}

\item{background}{Numeric or \code{NA} (default): Value to put in cells that are not \code{NA}.}

\item{byGeom, collapse}{Logical:
\itemize{
\item If \code{byGeom} is \code{TRUE} and \code{collapse} is \code{TRUE}, the output will have just one raster layer, but with values that correspond to each geometry (i.e., each geometry will get a different value). If geometries overlap, then category value of the one on top will be assigned to the relevant cells.
\item If \code{byGeom} is \code{TRUE}, and \code{collapse} is \code{FALSE}, then create a separate raster layer for each geometry in \code{x}.
}}
}
\value{
A \code{GRaster}.
}
\description{
The \code{rasterize()} function converts a \code{GVector} into a \code{GRaster}.
}
\examples{
if (grassStarted()) {

# Setup
library(sf)
library(terra)

# Elevation raster, outline of a part of Madagascar, and rivers vector:
madElev <- fastData("madElev") # raster
madDypsis <- fastData("madDypsis") # points vector
madRivers <- fastData("madRivers") # lines vector
madCoast4 <- fastData("madCoast4") # polygons vector

# Convert to GRaster and GVectors:
elev <- fast(madElev)
dypsis <- fast(madDypsis)
coast <- fast(madCoast4)
rivers <- fast(madRivers)

# Convert points, line, and polygons vectors to rasters:
points <- rasterize(dypsis, elev)
lines <- rasterize(rivers, elev)
polys <- rasterize(coast4, elev)

plot(points)
plot(lines)
plot(polys)

# Change background value:
polys <- rasterize(coast4, elev, background = -1)
plot(polys)

# By geometry onto one rasters:
polys <- rasterize(coast4, elev, byGeom = TRUE, collapse = TRUE)
plot(polys)

# By geometry, separate rasters:
polys <- rasterize(coast4, elev, byGeom = TRUE, collapse = FALSE)
plot(polys)

}
}
\seealso{
\code{\link[terra:rasterize]{terra::rasterize()}}, module \href{https://grass.osgeo.org/grass84/manuals/v.to.rast.html}{\code{v.to.rast}} in \strong{GRASS}
}
