% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind.r
\name{rbind.GVector}
\alias{rbind.GVector}
\alias{rbind}
\title{Combine one or more GVectors}
\usage{
rbind.GVector(...)
}
\arguments{
\item{...}{One or more \code{GVector}s.}
}
\value{
A \code{GVector}.
}
\description{
\code{rbind()} combines two or more \code{GVector}s of the same type (points, lines, or polygons) and same coordinate reference system. You can speed operations by putting the ector that is largest in memory first in \code{rbind(...)}. If the \code{GVector}s have data tables, these will also be combined using \code{rbind()} if their column names and data types match.
}
\examples{
if (grassStarted()) {

# Setup
library(sf)

# Rivers vector
madRivers <- fastData("madRivers")

# Convert sf to a GVector
rivers <- fast(madRivers)

# Subset rivers vector
rivers1 <- rivers[1:2]
rivers2 <- rivers[10:11]

# Concatenate rivers
riversCombo <- rbind(rivers1, rivers2)
riversCombo

# Add columns
newCol <- data.frame(x = 1:11)
riversCols <- cbind(rivers, newCol)
riversCol

as.data.frame(riversCol)

}
}
\seealso{
\code{\link[=rbind]{rbind()}}, \code{\link[=cbind]{cbind()}}, \code{\link[=c]{c()}}
}
