% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeSession.r
\name{removeSession}
\alias{removeSession}
\title{Delete a GRASS session (location, mapset(s), and all associated files)}
\usage{
removeSession(location, mapset = NULL, workDir = NULL)
}
\arguments{
\item{location}{Name of the \strong{GRASS} location.}

\item{mapset}{Either \code{NULL} (default) or the name of the mapset to delete. If \code{NULL}, then all mapsets in the given location will be deleted.}

\item{workDir}{Either \code{NULL} (default) or a character string of the directory in which the location to be removed resides. If \code{NULL}, then the working directory will be obtained from \code{getFastOptions('workDir')}.}
}
\value{
If successful, \code{TRUE} (invisibly). If not, then \code{FALSE} (also invisibly), plus a warning.
}
\description{
\code{removeSession()} deletes a \strong{GRASS} "session" \link[=location]{location and mapsets} and all rasters and vectors therein. This function should be used cautiously, and is mainly provided as a way to clean up after running examples.
}
