% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res.r
\name{res}
\alias{res}
\title{Spatial resolution of a 'GRaster'}
\usage{
res(x)
}
\arguments{
\item{x}{A \code{GRaster}.}
}
\value{
A numeric vector. The first value is the length of cells in the x-direction and the second the length of cells in the y-direction.
}
\description{
Spatial resolution of a \code{GRaster}.
}
