% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res.r
\name{res,GRegion-method}
\alias{res,GRegion-method}
\alias{res}
\alias{xres,GRegion-method}
\alias{xres}
\alias{yres,GRegion-method}
\alias{yres}
\alias{zres,GRegion-method}
\alias{zres}
\alias{res3d,GRegion-method}
\alias{res3d}
\title{Spatial resolution}
\usage{
\S4method{res}{GRegion}(x)

\S4method{xres}{GRegion}(x)

\S4method{yres}{GRegion}(x)

\S4method{zres}{GRegion}(x)

\S4method{res3d}{GRegion}(x)
}
\arguments{
\item{x}{A \code{GRaster}.}
}
\value{
A numeric vector. For both \code{res()} and \code{res3d()}, the first value is the length of cells in the x-direction and the second the length of cells in the y-direction. For \code{res3d()} the third value is height of a voxel (the z-direction). \code{xres()}, \code{yres()}, and \code{tbres()} each return a single value.
}
\description{
Spatial resolution of a \code{GRaster}:

\code{res()}: 2-dimensional resolution (x and y).\cr\cr
\code{res3d()}: 3-dimensinal resolution (z, y, and z).\cr\cr
\code{xres()}, \code{yres()}, and \code{zres()}: East-west resolution, north-south resolution, and top-bottom resolution.
}
\examples{
\dontrun{
# NB This example is in a "dontrun{}" block because it requires users to have
# GRASS GIS Version 8+ installed on their system.

# IMPORTANT #1: If you already have a GRASS session started, you will need to
# run the line below and the last line in this example to work with it again.
# If you have not started a GRASS session, you can skip this step and go to
# step #2.
opts. <- getFastOptions()

# IMPORTANT #2: Select the appropriate line below and change as necessary to
# where GRASS is installed on your system.
grassDir <- "/Applications/GRASS-8.3.app/Contents/Resources" # Mac
grassDir <- 'C:/Program Files/GRASS GIS 8.3' # Windows
grassDir <- '/usr/local/grass' # Linux

# setup
library(terra)

# example data
madElev <- fastData('madElev')
madForest2000 <- fastData('madForest2000')

# start GRASS session for examples only
faster(x = madElev, grassDir = grassDir,
workDir = tempdir(), location = 'examples') # line only needed for examples

# convert SpatRasters to GRasters
elev <- fast(madElev)
forest <- fast(madForest2000)

### GRaster properties

# dimensions
dim(elev) # rows, columns, depths, layers
nrow(elev) # rows
ncol(elev) # columns
ndepth(elev) # depths
nlyr(elev) # layers

res(elev) # resolution (2D)
res3d(elev) # resolution (3D)
zres(elev) # vertical resolution

# cell counts
ncell(elev) # cells
ncell3d(elev) # cells (3D rasters only)

# topology
topology(elev) # number of dimensions
is.2d(elev) # is it 2-dimensional?
is.3d(elev) # is it 3-dimensional?

minmax(elev) # min/max values

# information on the GRASS session in which the GRaster is located
location(elev) # location
mapset(elev) # mapset

# "gnames" of the object (its name in GRASS)
gnames(elev)

# coordinate reference system
crs(elev)
st_crs(elev)

# extent (bounding box)
ext(elev)
st_bbox(elev)

# vertical extent (not defined for this raster)
zext(elev)

# data type
datatype(elev)

# convert data type
as.cell(elev) # integer; note that "elev" is already of type "CELL"
as.fcell(elev) # floating-precision
as.dcell(elev) # double-precision

# concatenating multiple GRasters
rasts <- c(elev, forest)
rasts

# number of layers
nlyr(rasts)

# names
names(rasts)
names(rasts) <- c('elev_meters', 'forest')
rasts

# cell frequencies
freq(elev)
freq(2 * elev)
freq(rasts)
freq(rasts, value = 1)

# IMPORTANT #3: Revert back to original GRASS session if needed.
fastRestore(opts.)
removeSession('examples')

}
}
\seealso{
\code{\link[terra:dimensions]{terra::res()}}
}
