% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.r
\name{resample,GRaster,GRaster-method}
\alias{resample,GRaster,GRaster-method}
\alias{resample}
\alias{resample,GRaster,numeric-method}
\title{Change the cell size of a GRaster}
\usage{
\S4method{resample}{GRaster,GRaster}(x, y, method = NULL)

\S4method{resample}{GRaster,numeric}(x, y, method = NULL)
}
\arguments{
\item{x}{The \code{GRaster} to resample.}

\item{y}{Either a \code{GRaster} to serve as a template, or a numeric vector with two or three values. If a numeric vector, the values represent east-west and north-south resolution for 2D rasters, or east-west, north-south, and top-bottom resolution for 3D rasters.}

\item{method}{Character or \code{NULL}: Method to use to assign values to cells. Partial matching is used.
\itemize{
\item \code{NULL} (default): Automatically choose based on raster properties (\code{near} for categorical data, \code{bilinear} for continuous data)
\item \code{'near'}: Nearest neighbor. Best for categorical data, and often a poor choice for continuous data.  If \code{\link[=ncat]{ncat()}} is >0 for all rasters, this method will be used by default.
\item \code{'bilinear'}: Bilinear interpolation (default for non-categorical data; uses weighted values from 4 cells).
\item \code{'bicubic'}: Bicubic interpolation (uses weighted values from 16 cells).
\item \code{'lanczos'}: Lanczos interpolation (uses weighted values from 25 cells).
}}
}
\value{
A \code{GRaster}.
}
\description{
\code{resample()} changes the cell size (resolution) of a \code{GRaster} using either another raster as a template or a user-defined resolution. Note that the extent of the output raster may be expanded to accommodate an integer number of cells.
}
\examples{
\dontrun{
# NB This example is in a "dontrun{}" block because it requires users to have
# GRASS GIS Version 8+ installed on their system.

# IMPORTANT #1: If you already have a GRASS session started, you will need to
# run the line below and the last line in this example to work with it again.
# If you have not started a GRASS session, you can skip this step and go to
# step #2.

# IMPORTANT #2: Select the appropriate line below and change as necessary to
# where GRASS is installed on your system.
grassDir <- "/Applications/GRASS-8.2.app/Contents/Resources" # Mac
grassDir <- 'C:/Program Files/GRASS GIS 8.2' # Windows
grassDir <- '/usr/local/grass' # Linux

# setup
library(terra)

# elevation raster, climate raster, rivers vector
madElev <- fastData('madElev')

# start GRASS session for examples only
wd <- forwardSlash(tempdir())

# initiate GRASS session
faster(crs = madElev, grassDir = grassDir,
workDir = wd, location = 'examples') # line only needed for examples

elev <- fast(madElev)

# resample raster to 120 x 120 m
elevResamp <- resample(elev, c(120, 120), method='bilinear')
elev
elevResamp

# IMPORTANT #3: Revert back to original GRASS session if needed.
sessionRestore(opts.)
removeSession('exampleFrom')
removeSession('exampleTo')

}
}
