% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnormRast.r
\name{rnormRast,GRaster-method}
\alias{rnormRast,GRaster-method}
\alias{rnormRast}
\title{A random raster with or without spatial dependence}
\usage{
\S4method{rnormRast}{GRaster}(
  x,
  mu = 0,
  sigma = 1,
  dist = 0,
  exponent = 1,
  delay = 0,
  seed = round(1e+06 * runif(1))
)
}
\arguments{
\item{x}{A \code{GRaster}. The output will have the same extent and dimensions as this raster.}

\item{mu, sigma}{Numeric: Mean and sample standard deviation of output.}

\item{dist}{Numeric: Maximum distance of spatial autocorrelation (in map units--typically meters). Default is 0 (no spatial autocorrelation).}

\item{exponent}{Numeric > 0: Distance decay exponent.}

\item{delay}{Numeric >= 0: Values >0 force the distance decay of similarity to remain constant up to this distance. Beyond this distance, the decay exponent takes effect. Default is 0.}

\item{seed}{Numeric integer: Random seed.}
}
\value{
A \code{GRaster}.
}
\description{
'rnormRast()' creates a raster with random values in cells. Across the raster, values are approxmately normally distributed. However, spatial dependence can be introduced, though all together the values will still be normally distributed.
}
\examples{
\dontrun{
# NB This example is in a "dontrun{}" block because it requires users to have
# GRASS GIS Version 8+ installed on their system.

# IMPORTANT #1: If you already have a GRASS session started, you will need to
# run the line below and the last line in this example to work with it again.
# If you have not started a GRASS session, you can skip this step and go to
# step #2.
opts. <- getFastOptions()

# IMPORTANT #2: Select the appropriate line below and change as necessary to
# where GRASS is installed on your system.
grassDir <- "/Applications/GRASS-8.3.app/Contents/Resources" # Mac
grassDir <- "C:/Program Files/GRASS GIS 8.3" # Windows
grassDir <- "/usr/local/grass" # Linux

# setup
library(sf)
library(terra)

# elevation raster
madElev <- fastData("madElev")

# start GRASS session for examples only
faster(x = madElev, grassDir = grassDir,
workDir = tempdir(), location = "examples") # line only needed for examples

# Convert a SpatRaster to a GRaster:
elev <- fast(madElev)

# Create a raster with random values:
rand <- randRast(elev)
plot(rand)

# Create a raster with spatial dependence in values:
spatDep <- randRast(elev, dist = 3000)
plot(spatDep)


# IMPORTANT #3: Revert back to original GRASS session if needed.
fastRestore(opts.)
removeSession("examples")

}
}
\seealso{
\code{\link[=fractalRast]{fractalRast()}} and module \code{r.random.surface} in \strong{GRASS}
}
