% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstring.r
\name{rstring}
\alias{rstring}
\title{Make a nearly-guaranteed unique string}
\usage{
rstring(n, x = 10)
}
\arguments{
\item{n}{Numeric integer: How many strings to make (default is 1).}

\item{x}{Numeric integer: Number of letters and digits to use to make the string. Default is 10, leading to a probability of two matching random strings of <1E-38.}
}
\value{
Character.
}
\description{
\code{rstring()} makes a string that is statically extremely likely to be unique.
}
\examples{

rstring(1)
rstring(5)
rstring(c(5, 2))

}
