% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleRast.r
\name{sampleRast,GRaster-method}
\alias{sampleRast,GRaster-method}
\alias{sampleRast}
\title{Randomly sample cells from a GRaster}
\usage{
\S4method{sampleRast}{GRaster}(
  x,
  size,
  prop = FALSE,
  maskvalues = NA,
  updatevalue = NULL,
  fail = TRUE,
  seed = NULL
)
}
\arguments{
\item{x}{A \code{GRaster}.}

\item{size}{Numeric: Number of cells or proportion of cells to select.}

\item{prop}{Logical: If \code{TRUE}, the value of \code{size} will be interpreted as a proportion of cells. The default is \code{FALSE} (\code{size} is interpreted as the number of cells to select).}

\item{maskvalues}{Numeric vector, including \code{NA}, or \code{NULL} (default): Values in the raster to select from. All others will be ignored. If this is \code{NULL}, then only non-\code{NA} cells will be selected for retention.}

\item{updatevalue}{Numeric or \code{NULL} (default): Value to assign to masked cells. If \code{NULL}, then the values in the input raster are retained.}

\item{fail}{Logical: If \code{TRUE} (default), and \code{size} is greater than the number of non-\code{NA} cells in \code{x}, then fail.}

\item{seed}{\code{NULL} (default) or numeric: If \code{NULL}, then a random seed will be generated for the random number generator. Otherwise a seed can be provided.}
}
\value{
A \code{GRaster}.
}
\description{
\code{sampleRast()} randomly samples cells from non-\code{NA} cells of a raster. The output will be a raster with selected non-\code{NA} cells, and all other cells set to \code{NA}. To generate random points, see \code{\link[=spatSample]{spatSample()}}.
}
\examples{
if (grassStarted()) {

# Setup
library(sf)
library(terra)

# Example data
madElev <- fastData("madElev") # raster
madForest <- fastData("madForest2000") # raster
madCoast <- fastData("madCoast4") # vector

# Convert to GRasters and GVectors
elev <- fast(madElev)
forest <- fast(madForest)
coast <- fast(madCoast)

ant <- coast[coast$NAME_4 == "Antanambe"] # subset

# Random cells in non-NA cells:
rand <- sampleRast(elev, 10000)
plot(rand)
nonnacell(rand)

# Use custom values for the mask:
randCustomMask <- sampleRast(elev, 10000, maskvalues = 1:20)
plot(randCustomMask)

# Force selected values to a custom value:
randCustomUpdate <- sampleRast(elev, 10000, updatevalue = 7)
plot(randCustomUpdate)

# Custom values for mask and set selected cells to custom value:
randAll <- sampleRast(elev, 10000, maskvalues = 1:20, updatevalue = 7)
plot(randAll)

}
}
\seealso{
\code{\link[=spatSample]{spatSample()}}; \code{\link[terra:sample]{terra::spatSample()}}, module \code{r.random} in \strong{GRASS}
}
