% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothGeom.r
\name{smoothGeom,GVector-method}
\alias{smoothGeom,GVector-method}
\alias{smoothGeom}
\title{Smooth the geometry of a vector}
\usage{
\S4method{smoothGeom}{GVector}(x, method = "Hermite", dist = NULL, angle = 3)
}
\arguments{
\item{x}{A \code{GVector}.}

\item{method}{Character: Method used to smooth line segments. Partial matching is used, and case does not matter:
\itemize{
\item \code{"Hermite"}: Hermite interpolation (default): Guarantees that the output vector always passes through the original points. This method adds points (possibly many) by constructing cubic splines with points approximately \code{dist} apart. The number of points can be reduced by specifying a smaller value of \code{angle}, which specifies the minimum angle between two successive line segments.
\item \code{"Chaiken"}: Chaiken's algorithm: Guarantees that the new vector always touches the midpoint of each original line segment. The points on the new line are at least \code{dist} apart.
}}

\item{dist}{Numeric > 0 of \code{NULL} (default): Minimum distance (see \code{method}). Units are in map units. If \code{NULL}, then 2\% of the minimum of the x-, y-, and z-extent will be used.}

\item{angle}{Numeric > 0: Maximum angle for the Hermite algorithm. Default is 3.}
}
\value{
A \code{GVector}.
}
\description{
\code{smoothGeom()} makes line segments of a vector appear less angular.
}
\examples{
if (grassStarted()) {

# Setup
library(sf)
library(terra)

# Example data
madCoast0 <- fastData("madCoast0")
madRivers <- fastData("madRivers")

### Simplify geometry (remove nodes)
####################################

rivers <- fast(madRivers)

vr <- simplifyGeom(rivers, tolerance = 2000)
dp <- simplifyGeom(rivers, tolerance = 2000, method = "dp")
dpr <- simplifyGeom(rivers, tolerance = 2000, method = "dpr", prop=0.5)
rw <- simplifyGeom(rivers, tolerance = 2000, method = "rw")

plot(st_geometry(madRivers), col="gray", lwd=3)
plot(vr, col="blue", add=TRUE)
plot(dp, col="red", add=TRUE)
plot(dpr, col="chartreuse", add=TRUE)
plot(rw, col="orange", add=TRUE)

legend("topright",
   legend = c(
	  "Original",
      "Vertex reduction",
      "Douglas-Peucker",
      "Douglas-Peucker reduction",
      "Reumann-Witkam"
	),
	col = c("gray", "blue", "red", "chartreuse", "orange"),
	lwd = c(3, 1, 1, 1, 1)
)

### Smooth geometry
###################

hermite <- smoothGeom(rivers, dist = 2000, angle = 3)
chaiken <- smoothGeom(rivers, method = "Chaiken", dist = 2000)

plot(st_geometry(madRivers), col="gray", lwd=2)
plot(hermite, col="blue", add=TRUE)
plot(chaiken, col="red", add=TRUE)

legend("topright",
   legend = c(
	  "Original",
      "Hermite",
      "Chaiken"
	),
	col = c("gray", "blue", "red"),
	lwd = c(2, 1, 1, 1, 1)
)

### Clean geometry
##################

noDangs <- removeDangles(rivers, tolerance = 2000)

plot(st_geometry(madRivers), col="blue")
plot(noDangs, col="red", add=TRUE)

legend("topright",
   legend = c(
	  "Original",
      "No dangles"
	),
	col = c("blue", "red")
)

}
}
\seealso{
\code{\link[=simplifyGeom]{simplifyGeom()}}, \code{\link[terra:simplify]{terra::simplifyGeom()}}, \link[=breakPolys]{geometry cleaning}
}
