% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatSample.r
\name{spatSample,GRaster-method}
\alias{spatSample,GRaster-method}
\alias{spatSample}
\title{Random points from a raster}
\usage{
\S4method{spatSample}{GRaster}(
  x,
  size,
  prop = FALSE,
  maskvalues = NA,
  updatevalue = NULL,
  fail = TRUE,
  seed = NULL
)
}
\arguments{
\item{x}{A \code{GRaster}.}

\item{size}{Integer > 0: Number of cells or proportion of cells to select.}

\item{prop}{Logical: If \code{TRUE}, the value of \code{size} will be interpreted as a proportion of cells. The default is \code{FALSE} (\code{size} is interpreted as the number of cells to select).}

\item{maskvalues}{Numeric vector, including \code{NA}, or \code{NULL} (default): Values in the raster to select from. All others will be ignored. If this is \code{NULL}, then only non-\code{NA} cells will be selected for retention.}

\item{updatevalue}{Numeric or \code{NULL} (default): Value to assign to masked cells. If \code{NULL}, then the values in the input raster are retained.}

\item{fail}{Logical: If \code{TRUE} (default), and \code{size} is greater than the number of non-\code{NA} cells in \code{x}, then fail.}

\item{seed}{\code{NULL} (default) or numeric: If \code{NULL}, then a random seed will be generated for the random number generator. Otherwise a seed can be provided.}
}
\value{
A \code{GRaster}.
}
\description{
\code{spatSample()} randomly samples points from non-\code{NA} cells of a raster. Points will be located at cell centers.
}
\examples{
\dontrun{
# NB This example is in a "dontrun{}" block because it requires users to have
# GRASS GIS Version 8+ installed on their system.

# IMPORTANT #1: If you already have a GRASS session started, you will need to
# run the line below and the last line in this example to work with it again.
# If you have not started a GRASS session, you can skip this step and go to
# step #2.
opts. <- getFastOptions()

# IMPORTANT #2: Select the appropriate line below and change as necessary to
# where GRASS is installed on your system.

grassDir <- "/Applications/GRASS-8.3.app/Contents/Resources" # Mac
grassDir <- "C:/Program Files/GRASS GIS 8.3" # Windows
grassDir <- "/usr/local/grass" # Linux

# setup
library(sf)
library(terra)

# example data
madElev <- fastData("madElev") # raster
madForest <- fastData("madForest2000") # raster
madCoast <- fastData("madCoast4") # vector
madAnt <- madCoast[madCoast$NAME_4 == "Antanambe", ]

# start GRASS session for examples only
faster(x = madElev, grassDir = grassDir,
workDir = tempdir(), location = "examples") # line only needed for examples

# convert to GRasters and GVectors
elev <- fast(madElev)
forest <- fast(madForest)
ant <- fast(madAnt)

# Random cells in non-NA cells:
rand <- spatSample(elev, 10000)
plot(rand)
nonnacell(rand)

# Use custom values for the mask:
randCustomMask <- spatSample(elev, 10000, maskvalues = 1:20)
plot(randCustomMask)

# Force selected values to a custom value:
randCustomUpdate <- spatSample(elev, 10000, updatevalue = 7)
plot(randCustomUpdate)

# Custom values for mask and set selected cells to custom value:
randAll <- spatSample(elev, 10000, maskvalues = 1:20, updatevalue = 7)
plot(randAll)

# IMPORTANT #3: Revert back to original GRASS session if needed.
fastRestore(opts.)
removeSession("examples")

}
}
\seealso{
\code{\link[=spatSample]{spatSample()}}, \strong{GRASS} module \code{r.random}
}
