% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startFast.r
\name{startFast}
\alias{startFast}
\title{Initialize a 'GRASS' session}
\usage{
startFast(grassDir, workDir = tempdir(), ...)
}
\arguments{
\item{grassDir}{Folder in which \strong{GRASS} is installed on your computer. This will look different depending on the operating system and verison of \strong{GRASS} you have installed. Here are some examples:
\itemize{
\item Windows: \code{'C:/Program Files/GRASS GIS 8.3'}
\item Mac OS: \code{"/Applications/GRASS-8.3.app/Contents/Resources"}
\item Linux: \code{'/usr/local/grass'}
}}

\item{workDir}{Folder in which \strong{GRASS} rasters and vectors are to be saved. The default is to use \code{\link[=tempdir]{tempdir()}}, which works for most users who will export rasters/vectors they want from the \strong{GRASS} session before quitting \strong{R}.}

\item{...}{Options to send to \code{\link[=setFastOptions]{setFastOptions()}}. These should be in \code{option = value} format.}
}
\value{
\code{TRUE} if successful (invisibly). If not successful, the function will usually either fail or return \code{FALSE} with a warning.
}
\description{
This function initializes a \strong{GRASS} session in a particulatr folder. You need to run this function before you use most functions in \strong{fasterRaster} (just once). You can also use the function to switch your \strong{GRASS} session to a different folder, even if one is already started.
}
\seealso{
Guide to getting \href{tutorial_starting}{started} with \strong{fasterRaster}.
}
