% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stretch.r
\name{stretch,GRaster-method}
\alias{stretch,GRaster-method}
\alias{stretch}
\title{Rescale values in a raster}
\usage{
\S4method{stretch}{GRaster}(x, minv = 0, maxv = 255, minq = 0, maxq = 1, smin = NA, smax = NA)
}
\arguments{
\item{x}{A \code{GRaster}.}

\item{minv, maxv}{Numeric: Minumum and maximum values to which to rescale values.}

\item{minq, maxq}{Numeric: Specifies range of values to rescale, given by their quantiles. The default is to stretch all values (the 0th and 100th quantiles). One or both are  ignored if \code{smin} and/or \code{smax} are provided.}

\item{smin, smax}{Numeric: Specifies range of values to rescale.}
}
\value{
A \code{GRaster}.
}
\description{
\code{stretch()} rescales the values in a \code{GRaster}. All values can be rescaled, or just values in a user-defined range. This range can be given by specifying either the lower and upper bounds of the range using \code{smin} and \code{smax}, and/or by the quantiles (across all cells of the raster) using \code{minq} and \code{maxq}.
}
\seealso{
\code{\link[terra:stretch]{terra::stretch()}} and module \code{r.rescale} in \strong{GRASS} (not used on this function)
}
