% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thin.r
\name{thin,GRaster-method}
\alias{thin,GRaster-method}
\alias{thin}
\title{Reduce linear features on a raster so linear features are 1 cell wide}
\usage{
\S4method{thin}{GRaster}(x, iter = 200)
}
\arguments{
\item{x}{A \code{GRaster}.}

\item{iter}{Numeric integer: Number of iterations (default is 200).}
}
\value{
A \code{GRaster}.
}
\description{
The \code{thin()} function attempts to reduce linear features on a raster to just 1 cell wide. You may need to run \code{thin()} multiple times on the same raster (or experiment with the \code{iter} argument) to get acceptable output. \code{thin()} can be helpful run on a raster before using \code{as.lines()}.
}
