% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thinPoints.r
\name{thinPoints,GVector,GRaster-method}
\alias{thinPoints,GVector,GRaster-method}
\alias{thinPoints}
\title{Reduce number of points in same raster cell}
\usage{
\S4method{thinPoints}{GVector,GRaster}(x, y, n = 1)
}
\arguments{
\item{x}{A "points" \code{GVector}.}

\item{y}{A \code{GRaster}.}

\item{n}{Integer or numeric integer: Maximum number of points to remain in a cell. The default is 1.}
}
\value{
A "points" \code{GVector}.
}
\description{
This function thins a "points" \code{GVector} so that it has no more than \code{n} points per grid cell in a raster.
}
\examples{
man/examples/ex_thinPoints.r

}
