% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorial_undocumented_functions.r
\name{tutorial_undocumented_functions}
\alias{tutorial_undocumented_functions}
\title{Undocumented functions}
\description{
\strong{fasterRaster} interfaces with \strong{GRASS GIS} to process rasters and spatial vector data. It is intended as an add-on to the \strong{terra} and \strong{sf} packages, and relies heavily on them. For most rasters and vectors that are small or medium-sized in memory/disk, those packages will almost always be faster. They may also be faster for very large objects.  But when they aren't, \strong{fasterRaster} can step in.
\subsection{Installing \strong{fasterRaster} has a number of undocumented functions. These are used internally, but may be of utility to power-users, so are documented here.}{
\subsection{\code{.copyGSpatial()}}{

Creates in \strong{GRASS} a copy of a \code{GSpatial} object (a \code{GRaster} or a \code{GVector}).
}

\subsection{\code{dbName()} and \code{dbLayer()}}{

\strong{GRASS} names of the data table and data layer associated with a vector.
}

\subsection{\code{.gnames()}}{

Gets the name(s) of the file(s) in \strong{GRASS} that represent(s) the object(s).
}

\subsection{\code{.ls()}}{

Lists the \code{gnames} of all objects in the current \strong{GRASS} session.
}

\subsection{\code{.makeGNames()}}{

Makes one or more statistically unique strings that can be used as file names to represent rasters or vectors in \strong{GRASS}.
}

\subsection{\code{.refresh()}}{

Refresh metadata held within a \code{GRaster} or `GVector`` by querying the object in \strong{GRASS}.
}

\subsection{\code{.rename()}}{

Rename a \strong{GRASS} raster or vector.
}

\subsection{\code{.rm()}}{

Delete all \strong{GRASS} files representing rasters or vectors.
}

\subsection{\code{.rastInfo()}}{

Metadata for a \strong{GRASS} raster.
}

\subsection{\code{.vectInfo()}}{

Metadata for a \strong{GRASS} vector.
}

}
}
\keyword{tutorial}
