% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vAsDataTable.r
\name{.vAsDataTable}
\alias{.vAsDataTable}
\title{Convert a GRASS vector's attribute table to a data.table}
\usage{
.vAsDataTable(x)
}
\arguments{
\item{x}{A \code{GVector} or the name of a vector in \strong{GRASS}.}
}
\value{
A \code{data.table} or \code{NULL} if the vector is not attached to a database.
}
\description{
\strong{GRASS} vectors can be linked to an attribute table, which can be exported from \strong{GRASS} to \strong{R} using this function. \strong{This function is mostly of use to developers.}

Values in the \code{cat} column are not necessarily unique--if a value appears more than once, the set of features they index are (in other software) called "multipart" features. The table can have more columns with metadata for each feature.

This function is typically used by developers.
}
\keyword{internal}
