% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vect.r
\name{vect,GVector-method}
\alias{vect,GVector-method}
\alias{vect}
\title{Convert a GVector to a SpatVector}
\usage{
\S4method{vect}{GVector}(x, ...)
}
\arguments{
\item{x}{A \code{GVector}.}

\item{...}{Additional arguments to send to \code{\link[=writeRaster]{writeRaster()}} (typically unneeded, though \code{bigTiff} may be of use if the raster is large).}
}
\value{
A \code{SpatVector} (\strong{terra} package).
}
\description{
The \strong{fasterRaster} version of the \code{vect()} function converts a \code{GVector} to a \code{SpatVector} (from the \strong{terra} package).
}
