% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vHasDatabase.r
\name{.vHasDatabase}
\alias{.vHasDatabase}
\title{Tests if a GRASS vector is linked to an attribute table}
\usage{
.vHasDatabase(x)
}
\arguments{
\item{x}{A \code{GVector} or the \code{\link[=sources]{sources()}} name of a vector in \strong{GRASS}.}
}
\value{
Logical.
}
\description{
\strong{GRASS} vectors can be lined to one or more attribute tables, or "databases.". This function tests to see if the vector does indeed have a database. This function is typically used by developers.
}
\keyword{internal}
