% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vNames.r
\name{.vNames}
\alias{.vNames}
\title{Names of columns of a GRASS vector's attribute table}
\usage{
.vNames(x)
}
\arguments{
\item{x}{A \code{GVector} or the name of a vector in \strong{GRASS}.}
}
\value{
Character vector.
}
\description{
This function returns the column names of a \strong{GRASS} vector's attribute table.
}
\keyword{internal}
