% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workDir.r
\name{.workDir,GLocation-method}
\alias{.workDir,GLocation-method}
\alias{.workDir}
\title{Get a GLocation's working directory}
\usage{
\S4method{.workDir}{GLocation}(x)
}
\arguments{
\item{x}{A \code{GLocation} object.}
}
\value{
Character.
}
\description{
This function returns the working directory of a \code{GLocation} object.
}
\keyword{internal}
