% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeRaster.r
\name{writeRaster,GRaster,character-method}
\alias{writeRaster,GRaster,character-method}
\alias{writeRaster}
\alias{writeRaster,missing,missing-method}
\title{Save a GRaster to disk}
\usage{
\S4method{writeRaster}{GRaster,character}(
  x,
  filename,
  overwrite = FALSE,
  datatype = NULL,
  mm = FALSE,
  warn = TRUE,
  ...
)

\S4method{writeRaster}{missing,missing}(x, filename)
}
\arguments{
\item{x}{A \code{GRaster} or missing: If missing, a table of supported file types is reported.}

\item{filename}{Character: Path and file name.}

\item{overwrite}{Logical: If \code{FALSE} (default), do not save over existing file(s).}

\item{datatype}{\code{NULL} (default) or character: The datatype of the values stored in non-ASCII rasters. If \code{NULL}, this will be ascertained from the raster. This can any of:\tabular{lllll}{
   \strong{\code{fasterRaster}} \tab \strong{\code{terra}} \tab \strong{\code{GRASS}} \tab \strong{\code{GDAL}} \tab \strong{Values} \cr
   \code{integer} \tab \code{INT1U} \tab \code{CELL} \tab \code{Byte} \tab Integer values from 0 to 255 \cr
   \code{integer} \tab \code{INT2U} \tab \code{CELL} \tab \code{UInt16} \tab Integer values from 0 to 65,534 \cr
   \code{integer} \tab \code{INT2S} \tab \code{CELL} \tab \code{Int16} \tab Integer values from -32,767 to -32,767 \cr
   \code{integer} \tab na \tab \code{CELL} \tab \code{Int32}* \tab Integer values from 0 to 4,294,967,295 \cr
   \code{integer} \tab \code{INT4S} \tab \code{CELL} \tab \code{Int32}* \tab Integer values from -2,147,483,647 to 2,147,483,647 \cr
   \code{float} \tab \code{FLT4S} \tab \code{FCELL} \tab \code{Float32} \tab Values from -3.4E+38 to 3.4E+38, including decimal values \cr
   \code{double} \tab \code{FLT8S} \tab \code{DCELL} \tab \code{Float64} \tab Values from -1.79E+308 to 1.79E+308, including decimal values \cr
   \code{factor} \tab \code{INT}* \tab \code{CELL} \tab \verb{INT*} \tab Integer values corresponding to categories \cr
}


\code{*} Note that when the raster is written to disk, the GDAL datatype is used, regardless of whether the \strong{fasterRaster}-, \strong{terra}-, or \strong{GRASS}-style datatype was provided.
\verb{**} Depends on the integers (signed/unsigned, range of values). Categorical rasters will have a CSV file with category values and labels saved with them. The file name will be the same as the raster's file name, but end in extension ".csv".}

\item{mm}{Logical: If \code{TRUE}, call \code{\link[terra:minmax]{terra::setMinMax()}} on the raster to ensure it has metadata on the minimum and maximum values. For large rasters, this can take a long time, so the default value of \code{mm} is \code{FALSE}. This is only useful if you assign a raster to the output of \code{writeRaster}, as in \code{x <- writeRaster(my_raster, filename = './raster_file.tif', mm = TRUE)}.}

\item{warn}{Logical: If \code{TRUE} (default), display a warning if the \code{datatype} argument does not match the value given by \code{datatype(x, "GDAL")}.}

\item{...}{Additional arguments. These can include:
\itemize{
\item \code{compressTiff}: Character or \code{NULL}: Type of compression for GeoTIFF files:
\itemize{
\item \code{"DEFLATE"} (default)
\item \code{"LZW"}
\item \code{"PACKBITS"}
\item \code{"LZMA"}
\item \code{NULL}: No compression is used, but the file can still be reduced in size by using zip, gzip, or other compressions.
}
\item \code{bigTiff}: Logical: If \code{TRUE}, and the file format is a GeoTIFF and would be larger than 4 GB (regardless of compression), then the file will be saved in BIGTIFF format.
\item \code{format}: Character, indicating file format. This is usually ascertained from the file extension, but in case this fails, it can be stated explicitly. When using other formats, you may have to specify the \code{createopts} argument, too (see help page for \strong{GRASS} module \code{r.out.gdal}). Two common formats include:
\itemize{
\item \code{"GTiff"} (default): GeoTIFF \code{filename} ends in \code{.tif}
\item \code{"ASC"}: ASCII \code{filename} ends in \code{.asc}
}
\item Additional arguments to send to \strong{GRASS} modules \code{r.out.gdal} and \code{r.out.ascii}.
\item \code{precision}: Numeric: For ASCII files, you may need to state the number of significant digits. 32-bit values have 7 digits and 64-bit values have 16. So in these cases the argument would be \code{precision=7} or \code{precision=16}.
}}
}
\value{
A \code{SpatRaster} (invisibly). A raster is also saved to disk.
}
\description{
This function saves a \code{GRaster} to disk directly from a \strong{GRASS} session. It is faster than using \code{\link[=rast]{rast()}}, then saving the output of that to disk (because \code{rast()} actually save the raster to disk, anyway).

The function will attempt to ascertain the file type to be ascertained from the file extension, but you can specify the format using the \code{format} argument (see entry for \code{...}). You can see a list of supported formats by simply using this function with no arguments, as in \code{writeRaster()}, or by consulting the online help page for the \strong{GRASS} module \href{https://grass.osgeo.org/grass84/manuals/r.out.gdal.html}{\code{r.out.gdal}}. Only the \code{GeoTIFF} file format is guaranteed to work for multi-layered rasters.

The function will attempt to optimize the \code{datatype} argument, but this can take a long time. You can speed this up by setting \code{datatype} manually. Note that if you are saving a "stack" of \code{GRaster}s with different \code{datatype}s, the one with the highest information density will be used (e.g., low-bit integer < high-bit integer < floating-point < double-floating point). This can make rasters with lower datatypes much larger on disk. In these cases, it make be best to save rasters with similar \code{datatype}s together.
}
\examples{
if (grassStarted()) {

# Setup
library(terra)

# Example data
madElev <- fastData("madElev")

# Convert SpatRaster to GRaster
elev <- fast(madElev)

### Save GRaster to disk (using temporary file)
filename <- tempfile()
filename <- paste0(filename, ".tif")
writeRaster(elev, filename)

### Load raster from disk
elev2 <- fast(filename)
elev2

}
}
\seealso{
\code{\link[terra:writeRaster]{terra::writeRaster()}}, module \href{https://grass.osgeo.org/grass84/manuals/r.out.gdal.html}{\code{r.out.gdal}} in \strong{GRASS}
}
