% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeRaster.r
\name{writeRaster,GRaster,character-method}
\alias{writeRaster,GRaster,character-method}
\alias{writeRaster}
\alias{writeRaster,missing,missing-method}
\title{Save a GRaster to disk}
\usage{
\S4method{writeRaster}{GRaster,character}(x, filename, datatype = NULL, overwrite = FALSE, ...)

\S4method{writeRaster}{missing,missing}(x, filename)
}
\arguments{
\item{x}{A \code{GRaster} or missing: If missing, a table of supported file types is reported.}

\item{filename}{Character: Path and file name.}

\item{datatype}{\code{NULL} (default) or character: The datatype of the values stored in non-ASCII rasters. If \code{NULL}, this will be ascertained from the raster. This can any of:\tabular{lllll}{
   \strong{\code{fasterRaster}} \tab \strong{\code{terra}} \tab \strong{\code{GRASS}} \tab \strong{\code{GDAL}} \tab \strong{Values} \cr
   \code{integer} \tab \code{INT1U} \tab \code{CELL} \tab \code{Byte} \tab Integer values from 0 to 255 \cr
   \code{integer} \tab \code{INT2U} \tab \code{CELL} \tab \code{UInt16} \tab Integer values from 0 to 65,534 \cr
   \code{integer} \tab \code{INT2S} \tab \code{CELL} \tab \code{Int16} \tab Integer values from -32,767 to -32,767 \cr
   \code{integer} \tab \code{INT4S} \tab \code{CELL} \tab \code{Int32} \tab Integer values from -2,147,483,647 to 2,147,483,647 \cr
   \code{float} \tab \code{FLT4S} \tab \code{FCELL} \tab \code{Float32} \tab Values from -3.4e+38 to 3.4e+38, including decimal values \cr
   \code{double} \tab \code{FLT8S} \tab \code{DCELL} \tab \code{Float64} \tab Values from -1.7e+308 to 1.7e+308, including decimal values \cr
   \code{factor} \tab \code{INT}* \tab \code{CELL} \tab * \tab Integer values corresponding to categories \cr
}


\code{*} Depends on the integers (signed/unsigned, range of values). Categorical rasters will have a CSV file with category values and labels saved with them. The file name will be the same as the raster's file name, but end in extension ".csv".}

\item{overwrite}{Logical: If \code{FALSE} (default), do not save over existing file(s).}

\item{...}{Additional arguments. These can include:
\itemize{
\item \code{compressTiff}: Character or \code{NULL}: Type of compression for GeoTIFF files:
\itemize{
\item \code{"DEFLATE"} (default)
\item \code{"LZW"}
\item \code{"PACKBITS"}
\item \code{"LZMA"}
\item \code{NULL}: No compression is used, but the file can still be reduced in size by using zip, gzip, or other compressions.
}
\item \code{bigTiff}: Logical: If \code{TRUE}, and the file format is a GeoTIFF and would be larger than 4 GB (regardless of compression), then the file will be saved in BIGTIFF format.
\item \code{format}: Character, indicating file format. This is usually ascertained from the file extension, but in case this fails, it can be stated explicitly. When using other formats, you may have to specify the \code{createopts} argument, too (see help page for \strong{GRASS} module \code{r.out.gdal}). Two common formats include:
\itemize{
\item \code{"GTiff"} (default): GeoTIFF \code{filename} ends in \code{.tif}
\item \code{"ASC"}: ASCII \code{filename} ends in \code{.asc}
}
\item Additional arguments to send to \strong{GRASS} modules \code{r.out.gdal} and \code{r.out.ascii}.
\item \code{precision}: Numeric: For ASCII files, you may need to state the number of significant digits. 32-bit values have 7 digits and 64-bit values have 16. So in these cases the argument would be \code{precision=7} or \code{precision=16}.
}}
}
\value{
A \code{GRaster} or a \code{stars} raster. A raster is also saved to disk.
}
\description{
This function saves a \code{GRaster} to disk directly from a \strong{GRASS} session. It is faster than using \code{\link[=rast]{rast()}}, then saving the output of that to disk (because \code{rast()} actually save the raster to disk, anyway).

The function will attempt to ascertain the file type to be ascertained from the file extension, but you can specify the format using the \code{format} argument (see entry for \code{...}). You can see a list of supported formats by simply using this function with no arguments, as in \code{writeRaster()}, or by consulting the online help page for the \strong{GRASS} module \code{r.out.gdal}. Only the \code{GeoTIFF} file format is guaranteed to work for multi-layered rasters.

\strong{Note}: Rasters will be trimmed to their smallest extent such that no rows or columns are all \code{NA}s. You can add \code{NA} rows and columns to a \code{SpatRaster} using \code{\link[terra:extend]{terra::extend()}}.
}
\examples{
\dontrun{
# NB This example is in a "dontrun{}" block because it requires users to have
# GRASS GIS Version 8+ installed on their system.

# IMPORTANT #1: If you already have a GRASS session started, you will need to
# run the line below and the last line in this example to work with it again.
# If you have not started a GRASS session, you can skip this step and go to
# step #2.
opts. <- getFastOptions()

# IMPORTANT #2: Select the appropriate line below and change as necessary to
# where GRASS is installed on your system.
grassDir <- "/Applications/GRASS-8.3.app/Contents/Resources" # Mac
grassDir <- "C:/Program Files/GRASS GIS 8.3" # Windows
grassDir <- "/usr/local/grass" # Linux

# setup
library(terra)

# example data
madElev <- fastData("madElev")

# start GRASS session for examples only
faster(x = madElev, grassDir = grassDir,
workDir = tempdir(), location = "examples") # line only needed for examples

# convert SpatRaster to GRaster
elev <- fast(madElev)

### save GRaster to disk (using temporary file)
filename <- tempfile()
filename <- paste0(filename, ".tif")
writeRaster(elev, filename)

### load raster from disk
elev2 <- fast(filename)
elev2

# IMPORTANT #3: Revert back to original GRASS session if needed.
fastRestore(opts.)
removeSession("examples")

}
}
\seealso{
\code{\link[terra:writeRaster]{terra::writeRaster()}}
}
